+++ src/presolve/presolve.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 2.4 - 07/02/2011 AT 11:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-03-30 AT 09:00 GMT.
 
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -285,6 +285,8 @@
 
       USE GALAHAD_SPECFILE_double ! specfile manipulation
 
+      USE GALAHAD_SPACE_double    ! allocating, extending and deleting storage
+
 !     Matrix storage schemes
 
       USE GALAHAD_SYMBOLS,                                                     &
@@ -391,7 +393,23 @@
 !     Make the PRESOLVE calls public
 
       PUBLIC :: PRESOLVE_initialize, PRESOLVE_read_specfile, PRESOLVE_apply,   &
-                PRESOLVE_restore,    PRESOLVE_terminate, SMT_put, SMT_get
+                PRESOLVE_restore, PRESOLVE_terminate,                          &
+                PRESOLVE_full_initialize, PRESOLVE_full_terminate,             &
+                PRESOLVE_import_problem, PRESOLVE_transform_problem,           &
+                PRESOLVE_restore_solution, PRESOLVE_information,               &
+                SMT_put, SMT_get
+
+!----------------------
+!   I n t e r f a c e s
+!----------------------
+
+     INTERFACE PRESOLVE_initialize
+       MODULE PROCEDURE PRESOLVE_initialize, PRESOLVE_full_initialize
+     END INTERFACE PRESOLVE_initialize
+
+     INTERFACE PRESOLVE_terminate
+       MODULE PROCEDURE PRESOLVE_terminate, PRESOLVE_full_terminate
+     END INTERFACE PRESOLVE_terminate
 
 !-------------------------------------------------------------------------------
 !   P r e c i s i o n
@@ -1084,7 +1102,7 @@
 !                  Default : 10.**8 in double precision,
 !                            10.**4 in single precision.
 
-      END TYPE
+      END TYPE PRESOLVE_control_type
 
 !   Note that the default double precision settings correpond to the following
 !   specfile, which is optionnaly read by PRESOLVE (see the description of
@@ -1476,7 +1494,7 @@
 !                   It is printed out on device errout at the end of execution
 !                   if control%print_level >= TRACE.
 
-      END TYPE
+      END TYPE PRESOLVE_inform_type
 
 !-------------------------------------------------------------------------------
 !         The structure that saves information between the various calls
@@ -1813,7 +1831,19 @@
                                         ! problem (only used if
                                         ! control%check_dual_feasibility>=BASIC)
 
-      END TYPE
+      END TYPE PRESOLVE_data_type
+
+!  - - - - - - - - - - - -
+!   full_data derived type
+!  - - - - - - - - - - - -
+
+      TYPE, PUBLIC :: PRESOLVE_full_data_type
+        LOGICAL :: f_indexing
+        TYPE ( PRESOLVE_data_type ) :: PRESOLVE_data
+        TYPE ( PRESOLVE_control_type ) :: PRESOLVE_control
+        TYPE ( PRESOLVE_inform_type ) :: PRESOLVE_inform
+        TYPE ( QPT_problem_type ) :: prob
+      END TYPE PRESOLVE_full_data_type
 
[…]
