+++ src/presolve/C/presolvetf.c
@@ -1,0 +1,184 @@
+/* presolvetf.c */
+/* Full test for the PRESOLVE C interface using Fortran sparse matrix indexing */
+
+#include <stdio.h>
+#include <math.h>
+#include "galahad_presolve.h"
+
+int main(void) {
+
+    // Derived types
+    void *data;
+    struct presolve_control_type control;
+    struct presolve_inform_type inform;
+
+    // Set problem data
+    int n = 6; // dimension
+    int m = 5; // number of general constraints
+    int H_ne = 1; // Hesssian elements
+    int H_row[] = {1};   // row indices, NB lower triangle
+    int H_col[] = {1};    // column indices, NB lower triangle
+    int H_ptr[] = {1, 2, 2, 2, 2, 2, 2}; // row pointers
+    double H_val[] = {1.0};   // values
+    double g[] = {1.0, 1.0, 1.0, 1.0, 1.0, 1.0}; // linear term in the objective
+    double f = 1.0;  // constant term in the objective
+    int A_ne = 8; // Jacobian elements
+    int A_row[] = {3, 3, 3, 4, 4, 5, 5, 5}; // row indices
+    int A_col[] = {3, 4, 5, 3, 6, 4, 5, 6}; // column indices
+    int A_ptr[] = {1, 1, 1, 4, 6, 9}; // row pointers
+    double A_val[] = {1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0}; // values
+    double c_l[] = { 0.0, 0.0, 2.0, 1.0, 3.0};   // constraint lower bound
+    double c_u[] = {1.0, 1.0, 3.0, 3.0, 3.0};   // constraint upper bound
+    double x_l[] = {-3.0, 0.0, 0.0, 0.0, 0.0, 0.0}; // variable lower bound
+    double x_u[] = {3.0, 1.0, 1.0, 1.0, 1.0, 1.0}; // variable upper bound
+
+    // Set output storage
+    char st;
+    int status;
+
+    printf(" Fortran sparse matrix indexing\n\n");
+
+    printf(" basic tests of qp storage formats\n\n");
+
+    for( int d=1; d <= 7; d++){
+
+        // Initialize PRESOLVE
+        presolve_initialize( &data, &control, &status );
+
+        // Set user-defined control options
+        control.f_indexing = true; // Fortran sparse matrix indexing
+
+        switch(d){
+            case 1: // sparse co-ordinate storage
+                st = 'C';
+                presolve_import_problem( &control, &data, &status, n, m,
+                           "coordinate", H_ne, H_row, H_col, NULL, H_val, g, f,
+                           "coordinate", A_ne, A_row, A_col, NULL, A_val,
+                           c_l, c_u, x_l, x_u,
+                           n_trans, m_trans, H_ne_trans, A_ne_trans );
+                break;
+            printf(" case %1i break\n",d);
+            case 2: // sparse by rows
+                st = 'R';
+                presolve_import_problem( &control, &data, &status, n, m,
+                       "sparse_by_rows", H_ne, NULL, H_col, H_ptr, H_val, g, f,
+                       "sparse_by_rows", A_ne, NULL, A_col, A_ptr, A_val,
+                       c_l, c_u, x_l, x_u,
+                       n_trans, m_trans, H_ne_trans, A_ne_trans );
+                break;
+            case 3: // dense
+                st = 'D';
+                int H_dense_ne = n*(n+1)/2; // number of elements of H
+                int A_dense_ne = m*n; // number of elements of A
+                double H_dense[] = {1.0,
+                                    0.0, 0.0,
+                                    0.0, 0.0, 0.0,
[…]
