+++ include/galahad_presolve.h
@@ -224,12 +224,12 @@
   \a restoration phase.
 
   If the number of problem transformations exceeds
-  \p control.transf_buffer_size},  the transformation buffer size,
+  \p control.transf_buffer_size,  the transformation buffer size,
   then they are saved in a ``history'' file, whose
   name may be chosen by specifying the control.transf_file_name control
   parameter,  When this is the case, this file
-  is subsequently reread by \p presolve_restore. It must not be altered
-  by the user.
+  is subsequently reread by \p presolve_restore_solution. It must not be 
+  altered by the user.
 
   Overall, the presolving process follows one of the two sequences:
 
@@ -288,8 +288,7 @@
   if it is an upper bound, or unconstrained in sign if the variables is fixed;
   and this convention is reversed in \f$z\_{sign}\f$ = -1}. The values of
   \f$z_{sign}\f$ and \f$y_{sign}\f$ may be chosen by setting the corresponding
-  components of the \p control structure to {\sym \galsympositive} or {\sym
-  \galsymnegative}
+  components of the \p control structure to \p 1 or \p -1.
 
   \subsection presolve_references Reference
 
@@ -311,7 +310,7 @@
       the dimensions of the transformed problem
   - \link presolve_transform_problem \endlink - apply the presolve algorithm
       to transform the data
-  - \link presolve_recover_solution \endlink - restore the solution from
+  - \link presolve_restore_solution \endlink - restore the solution from
          that of the transformed problem
   - \link presolve_information \endlink (optional) - recover information about
     the solution and solution process
@@ -480,12 +479,12 @@ struct presolve_control_type {
 
     /// \brief
     /// The maximum number of problem transformations, cumulated
-    /// over all calls to PRESOLVE.
+    /// over all calls to \p presolve.
     int max_nbr_transforms;
 
     /// \brief
     /// The maximum number of analysis passes for problem analysis
-    /// during a single call of PRESOLVE_apply.
+    /// during a single call of \p presolve_transform_problem.
     int max_nbr_passes;
 
     /// \brief
@@ -526,7 +525,7 @@ struct presolve_control_type {
     /// \li 2 reports the identity of each problem
     ///              transformation
     /// \li 3 reports more details
-    /// \li 4 reports LOTS of information.
+    /// \li 4 reports lots of information.
     /// \li 5 reports a completely silly amount of information
     int print_level;
 
@@ -631,8 +630,8 @@ struct presolve_control_type {
     /// The name of the file (to be) used for storing
     /// problem transformation on disk.
     /// NOTE: this parameter must be identical for all calls to
-    /// PRESOLVE following PRESOLVE_read_specfile. It can
-    /// then only be changed after calling PRESOLVE_terminate.
+    /// \p presolve following \p presolve_read_specfile. It can
+    /// then only be changed after calling presolve_terminate.
     char transf_file_name[31];
 
     /// \brief
@@ -650,18 +649,14 @@ struct presolve_control_type {
 
     /// \brief
     /// Determines whether or not the multipliers corresponding
-    /// to constraints that are inactive at the unreduced point
-    /// corresponding to the reduced point on input of RESTORE
+    /// to constraints that are inactive at the unreduced point corresponding
+    /// to the reduced point on input to \p presolve_restore_solution
     /// must be set to zero. Possible values are:
     /// associated with the general linear constraints.
     /// \li 0 All multipliers corresponding to inactive
     ///     inequality constraints are forced to zero,
[…]
