+++ src/presolve/C/presolve_ciface.f90
@@ -74,19 +74,19 @@
       INTEGER ( KIND = C_INT ) :: final_y_bounds
       INTEGER ( KIND = C_INT ) :: check_primal_feasibility
       INTEGER ( KIND = C_INT ) :: check_dual_feasibility
-      LOGICAL ( KIND = C_BOOL ) :: get_q
-      LOGICAL ( KIND = C_BOOL ) :: get_f
-      LOGICAL ( KIND = C_BOOL ) :: get_g
-      LOGICAL ( KIND = C_BOOL ) :: get_H
-      LOGICAL ( KIND = C_BOOL ) :: get_A
-      LOGICAL ( KIND = C_BOOL ) :: get_x
-      LOGICAL ( KIND = C_BOOL ) :: get_x_bounds
-      LOGICAL ( KIND = C_BOOL ) :: get_z
-      LOGICAL ( KIND = C_BOOL ) :: get_z_bounds
-      LOGICAL ( KIND = C_BOOL ) :: get_c
-      LOGICAL ( KIND = C_BOOL ) :: get_c_bounds
-      LOGICAL ( KIND = C_BOOL ) :: get_y
-      LOGICAL ( KIND = C_BOOL ) :: get_y_bounds
+!     LOGICAL ( KIND = C_BOOL ) :: get_q
+!     LOGICAL ( KIND = C_BOOL ) :: get_f
+!     LOGICAL ( KIND = C_BOOL ) :: get_g
+!     LOGICAL ( KIND = C_BOOL ) :: get_H
+!     LOGICAL ( KIND = C_BOOL ) :: get_A
+!     LOGICAL ( KIND = C_BOOL ) :: get_x
+!     LOGICAL ( KIND = C_BOOL ) :: get_x_bounds
+!     LOGICAL ( KIND = C_BOOL ) :: get_z
+!     LOGICAL ( KIND = C_BOOL ) :: get_z_bounds
+!     LOGICAL ( KIND = C_BOOL ) :: get_c
+!     LOGICAL ( KIND = C_BOOL ) :: get_c_bounds
+!     LOGICAL ( KIND = C_BOOL ) :: get_y
+!     LOGICAL ( KIND = C_BOOL ) :: get_y_bounds
       REAL ( KIND = wp ) :: pivot_tol
       REAL ( KIND = wp ) :: min_rel_improve
       REAL ( KIND = wp ) :: max_growth_factor
@@ -94,6 +94,8 @@
 
     TYPE, BIND( C ) :: presolve_inform_type
       INTEGER ( KIND = C_INT ) :: status
+      INTEGER ( KIND = C_INT ) :: status_continue   ! ignore - doxygen fix
+      INTEGER ( KIND = C_INT ) :: status_continued  ! ignore - doxygen fix
       INTEGER ( KIND = C_INT ) :: nbr_transforms
       CHARACTER ( KIND = C_CHAR ), DIMENSION( 3, 81 ) :: message
     END TYPE presolve_inform_type
@@ -155,19 +157,19 @@
     ! Logicals
     fcontrol%dual_transformations = ccontrol%dual_transformations
     fcontrol%redundant_xc = ccontrol%redundant_xc
-    fcontrol%get_q = ccontrol%get_q
-    fcontrol%get_f = ccontrol%get_f
-    fcontrol%get_g = ccontrol%get_g
-    fcontrol%get_H = ccontrol%get_H
-    fcontrol%get_A = ccontrol%get_A
-    fcontrol%get_x = ccontrol%get_x
-    fcontrol%get_x_bounds = ccontrol%get_x_bounds
-    fcontrol%get_z = ccontrol%get_z
-    fcontrol%get_z_bounds = ccontrol%get_z_bounds
-    fcontrol%get_c = ccontrol%get_c
[…]

[…]
