### 2022-04-24 #4291 @melissawm
Hello, folks - sorry for the delay!

@nimgould would you recommend any reviewers from the [list of potential reviewers](https://bit.ly/joss-reviewers)?

Also - because different people have been involved in the past, I just want to confirm that the authorship for GALAHAD 4.0 is correct.

Finally, I have had experience with GALAHAD and know the documentation is pretty extensive, but while we wait for the reviewers would it be possible to have an entry point for the documentation? Could be a page on the wiki, or an html or PDF document that guides the user through the different documents.

Thank you!
