### 2022-05-05 1fc6a7d
bug fixes in output in arc and tru (thanks Tangi Migot!)
