+++ src/cro/cros2.f90
@@ -1,0 +1,128 @@
+! THIS VERSION: GALAHAD 4.1 - 11/05/2022 AT 07:30 GMT.
+   PROGRAM GALAHAD_CRO_EXAMPLE2 ! AFIRO example from Netlib LP set
+   USE GALAHAD_CRO_double       ! double precision version
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   TYPE ( CRO_data_type ) :: data
+   TYPE ( CRO_control_type ) :: control        
+   TYPE ( CRO_inform_type ) :: inform
+   INTEGER :: i
+   INTEGER, PARAMETER :: n = 32, m = 27, m_equal = 8, a_ne = 83, h_ne = 0
+   INTEGER, DIMENSION( h_ne ) :: H_col
+   INTEGER, DIMENSION( n + 1 ) :: H_ptr
+   REAL ( KIND = wp ), DIMENSION( h_ne ) :: H_val
+   INTEGER, DIMENSION( a_ne ) :: A_col
+   INTEGER, DIMENSION( m + 1 ) :: A_ptr
+   REAL ( KIND = wp ), DIMENSION( a_ne ) :: A_val
+   REAL ( KIND = wp ), DIMENSION( n ) :: G, X_l, X_u, X, Z
+   REAL ( KIND = wp ), DIMENSION( m ) :: C_l, C_u, C, Y
+   INTEGER, DIMENSION( m ) :: C_stat
+   INTEGER, DIMENSION( n ) :: X_stat
+! start problem data
+   H_ptr = 0 ! pointers to H columns (this is AFIRO, a linear program)
+   g = 0.0_wp ! objective gradient
+   G( 2 ) = - 0.4_wp ; g( 13 ) = - 0.32_wp ; G( 17 ) = - 0.6_wp
+   G( 29 ) = - 0.48_wp ; g( 32 ) = 10.0_wp
+   A_val  = (/  1.0_wp, -1.0_wp,  1.0_wp, -1.06_wp,  1.0_wp,                   &
+                1.0_wp, -1.0_wp,  1.4_wp,  1.0_wp, -1.0_wp,                    &
+                1.0_wp, -1.0_wp, -1.0_wp, -1.0_wp,  1.0_wp,                    &
+               -0.86_wp, -0.96_wp, -1.06_wp, -1.06_wp,  1.0_wp,                &
+               -1.0_wp, -1.0_wp,  1.0_wp,  1.0_wp, -1.0_wp,                    &
+               -1.0_wp,  1.0_wp,  1.0_wp, -1.0_wp,  1.0_wp,                    &
+                1.0_wp, -0.43_wp,  1.0_wp,  1.0_wp,  1.40_wp,                  &
+               -1.0_wp, -0.43_wp,  1.0_wp, -0.37_wp, -0.43_wp,                 &
+               -0.39_wp,  1.0_wp,  1.0_wp,  1.0_wp,  1.0_wp,                   &
+                1.0_wp,  1.0_wp, -1.0_wp, -1.0_wp,  1.0_wp,                    &
+               -1.0_wp,  1.0_wp,  1.0_wp, -1.0_wp,  1.0_wp,                    &
+               -1.0_wp,  2.364_wp,  2.249_wp,  2.408_wp,  2.219_wp,            &
+                2.429_wp,  2.386_wp,  2.279_wp, -1.0_wp,  2.191_wp,            &
+                0.109_wp, -1.0_wp,  0.1_wp,  0.10_wp, -1.0_wp,                 &
[…]
