### 2022-05-17 f9db503
rename ccqp as cdqp in anticipation of new ccqp improvements to cro
