+++ src/qp/qp.f90
@@ -51,7 +51,7 @@
       USE GALAHAD_QPC_double
       USE GALAHAD_CQP_double
       USE GALAHAD_DQP_double
-      USE GALAHAD_CCQP_double
+      USE GALAHAD_CDQP_double
       USE GALAHAD_LMS_double, ONLY: LMS_apply_lbfgs
 
       IMPLICIT NONE
@@ -148,7 +148,7 @@
         LOGICAL :: generate_sif_file = .FALSE.
 
 !  quadratic programming solver. Possible values are
-!   qpa, qpb, qpc, cqp, dqp, ccqp
+!   qpa, qpb, qpc, cqp, dqp, cdqp
 
         CHARACTER ( LEN = 30 ) :: quadratic_programming_solver =               &
            "qpc" // REPEAT( ' ', 27 )
@@ -192,9 +192,9 @@
 
         TYPE ( DQP_control_type ) :: DQP_control
 
-!  control parameters for CCQP
+!  control parameters for CDQP
 
-        TYPE ( CCQP_control_type ) :: CCQP_control
+        TYPE ( CDQP_control_type ) :: CDQP_control
 
       END TYPE
 
@@ -305,9 +305,9 @@
 
         TYPE ( DQP_inform_type ) :: DQP_inform
 
-!  inform parameters for CCQP
+!  inform parameters for CDQP
 
-        TYPE ( CCQP_inform_type ) :: CCQP_inform
+        TYPE ( CDQP_inform_type ) :: CDQP_inform
       END TYPE
 
    CONTAINS
@@ -352,8 +352,8 @@
       control%CQP_control%prefix = '" - CQP:"                     '
       CALL DQP_initialize( data, control%DQP_control, inform%DQP_inform  )
       control%DQP_control%prefix = '" - DQP:"                     '
-      CALL CCQP_initialize( data, control%CCQP_control, inform%CCQP_inform  )
-      control%CQP_control%prefix = '" - CCQP:"                    '
+      CALL CDQP_initialize( data, control%CDQP_control, inform%CDQP_inform  )
+      control%CQP_control%prefix = '" - CDQP:"                    '
 
       inform%status = GALAHAD_ok
 
@@ -515,7 +515,7 @@
       control%QPC_control%infinity = control%infinity
       control%CQP_control%infinity = control%infinity
       control%DQP_control%infinity = control%infinity
-      control%CCQP_control%infinity = control%infinity
+      control%CDQP_control%infinity = control%infinity
 
 !  Read the specfiles for PRESOLVE, QPA, QPB, QPC, CQP and DQP
 
@@ -535,8 +535,8 @@
                                 alt_specname = TRIM( alt_specname ) // '-CQP' )
         CALL DQP_read_specfile( control%DQP_control, device,                   &
                                 alt_specname = TRIM( alt_specname ) // '-DQP' )
-        CALL CCQP_read_specfile( control%CCQP_control, device,                 &
-                                alt_specname = TRIM( alt_specname ) // '-CCQP' )
+        CALL CDQP_read_specfile( control%CDQP_control, device,                 &
+                                alt_specname = TRIM( alt_specname ) // '-CDQP' )
       ELSE
         CALL SCALE_read_specfile( control%SCALE_control, device )
         CALL PRESOLVE_read_specfile( device, control%PRESOLVE_control,         &
@@ -546,7 +546,7 @@
         CALL QPC_read_specfile( control%QPC_control, device )
         CALL CQP_read_specfile( control%CQP_control, device )
         CALL DQP_read_specfile( control%DQP_control, device )
-        CALL CCQP_read_specfile( control%CCQP_control, device )
[…]
