+++ src/ccqp/useccqp.f90
@@ -1,0 +1,1436 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-05-17 AT 15:00 GMT.
+
+!-*-*-*-*-*-*-*-  G A L A H A D   U S E C Q P   M O D U L E  -*-*-*-*-*-*-*-*-
+
+!  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
+!  Principal authors: Nick Gould and Dominique Orban
+
+!  History -
+!   originally released as usecqp with GALAHAD Version 2.4, January 1st 2010
+!   modified to form useccqp with GALAHAD 4.1, May 17th 2022
+
+!  For full documentation, see
+!   http://galahad.rl.ac.uk/galahad-www/specs.html
+
+    MODULE GALAHAD_USECCQP_double
+
+!    ---------------------------------------------------------------
+!    | CUTEst/AMPL interface to CCQP, an interior-point crossover  |
+!    | algorithm for convex quadratic & least-distance programming |
+!    ---------------------------------------------------------------
+
+!$    USE omp_lib
+      USE CUTEst_interface_double
+!NOT95USE GALAHAD_CPU_time
+      USE GALAHAD_CLOCK
+      USE GALAHAD_RAND_double
+      USE GALAHAD_QPT_double
+      USE GALAHAD_SORT_double, ONLY: SORT_reorder_by_rows
+      USE GALAHAD_NORMS_double, ONLY: TWO_NORM
+      USE GALAHAD_CCQP_double
+      USE GALAHAD_SLS_double
+      USE GALAHAD_PRESOLVE_double
+      USE GALAHAD_SPECFILE_double
+      USE GALAHAD_STRING, ONLY: STRING_upper_word
+      USE GALAHAD_COPYRIGHT
+      USE GALAHAD_SYMBOLS,                                                     &
+          ACTIVE                => GALAHAD_ACTIVE,                             &
+          TRACE                 => GALAHAD_TRACE,                              &
+          DEBUG                 => GALAHAD_DEBUG,                              &
+          GENERAL               => GALAHAD_GENERAL,                            &
+          ALL_ZEROS             => GALAHAD_ALL_ZEROS
+      USE GALAHAD_SCALE_double
+
+      IMPLICIT NONE
+
+      PRIVATE
+      PUBLIC :: USE_CCQP
+
+    CONTAINS
+
+!-*-*-*-*-*-*-*-*-*-   U S E _ C Q P  S U B R O U T I N E   -*-*-*-*-*-*-*-*-
+
+     SUBROUTINE USE_CCQP( input, close_input )
+
+!  --------------------------------------------------------------------
+!
+!  Solve the quadratic program from CUTEst
+!
+!     minimize     1/2 x(T) H x + g(T) x
+!
+!     subject to     c_l <= A x <= c_u
+!                    x_l <=  x <= x_u
+!
+!  using the GALAHAD package GALAHAD_CCQP
+!
+!  --------------------------------------------------------------------
+
+!  Dummy argument
+
+      INTEGER, INTENT( IN ) :: input
+      LOGICAL, OPTIONAL, INTENT( IN ) :: close_input
+
+!  Parameters
+
+      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
+      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
+      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
+      REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 19
+
[…]
