+++ src/ccqp/C/ccqp_ciface.f90
@@ -1,0 +1,958 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-05-18 AT 08:30 GMT.
+
+!-*-*-*-*-*-*-*-  G A L A H A D _  C C Q P    C   I N T E R F A C E  -*-*-*-*-*-
+
+!  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
+!  Principal authors: Jaroslav Fowkes & Nick Gould
+
+!  History -
+!    originally released GALAHAD Version 4.1. May 18th 2022
+
+!  For full documentation, see
+!   http://galahad.rl.ac.uk/galahad-www/specs.html
+
+  MODULE GALAHAD_CCQP_double_ciface
+    USE iso_c_binding
+    USE GALAHAD_common_ciface
+    USE GALAHAD_CCQP_double, ONLY:                                             &
+        f_ccqp_control_type   => CCQP_control_type,                            &
+        f_ccqp_time_type      => CCQP_time_type,                               &
+        f_ccqp_inform_type    => CCQP_inform_type,                             &
+        f_ccqp_full_data_type => CCQP_full_data_type,                          &
+        f_ccqp_initialize     => CCQP_initialize,                              &
+        f_ccqp_read_specfile  => CCQP_read_specfile,                           &
+        f_ccqp_import         => CCQP_import,                                  &
+        f_ccqp_reset_control  => CCQP_reset_control,                           &
+        f_ccqp_solve_qp       => CCQP_solve_qp,                                &
+        f_ccqp_solve_sldqp    => CCQP_solve_sldqp,                             &
+        f_ccqp_information    => CCQP_information,                             &
+        f_ccqp_terminate      => CCQP_terminate
+
+    USE GALAHAD_SBLS_double_ciface, ONLY:                                      &
+        sbls_inform_type,                                                      &
+        sbls_control_type,                                                     &
+        copy_sbls_inform_in   => copy_inform_in,                               &
+        copy_sbls_inform_out  => copy_inform_out,                              &
+        copy_sbls_control_in  => copy_control_in,                              &
+        copy_sbls_control_out => copy_control_out
+
+    USE GALAHAD_FDC_double_ciface, ONLY:                                       &
+        fdc_inform_type,                                                       &
+        fdc_control_type,                                                      &
+        copy_fdc_inform_in   => copy_inform_in,                                &
+        copy_fdc_inform_out  => copy_inform_out,                               &
+        copy_fdc_control_in  => copy_control_in,                               &
+        copy_fdc_control_out => copy_control_out
+
+    USE GALAHAD_FIT_double_ciface, ONLY:                                       &
+        fit_inform_type,                                                       &
+        fit_control_type,                                                      &
[…]
