+++ src/ccqp/ccqps3.f90
@@ -1,0 +1,48 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-05-18 AT 08:15 GMT.
+   PROGRAM GALAHAD_CCQP_EXAMPLE3
+   USE GALAHAD_CCQP_double         ! double precision version
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   TYPE ( QPT_problem_type ) :: p
+   TYPE ( CCQP_data_type ) :: data
+   TYPE ( CCQP_control_type ) :: control
+   TYPE ( CCQP_inform_type ) :: inform
+   INTEGER :: s
+   INTEGER, PARAMETER :: n = 3, m = 2
+   INTEGER, ALLOCATABLE, DIMENSION( : ) :: C_stat, B_stat
+! start problem data
+   ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
+   ALLOCATE( p%C( m ), p%C_l( m ), p%C_u( m ) )
+   ALLOCATE( p%X( n ), p%Y( m ), p%Z( n ) )
+   ALLOCATE( B_stat( n ), C_stat( m ) )
+   p%new_problem_structure = .TRUE.           ! new structure
+   p%n = n ; p%m = m ; p%f = 1.0_wp           ! dimensions & objective constant
+   p%G = (/ 0.0_wp, 2.0_wp, 0.0_wp /)         ! objective gradient
+   p%C_l = (/ 1.0_wp, 2.0_wp /)               ! constraint lower bound
+   p%C_u = (/ 2.0_wp, 2.0_wp /)               ! constraint upper bound
+   p%X_l = (/ - 1.0_wp, - infinity, - infinity /) ! variable lower bound
+   p%X_u = (/ 1.0_wp, infinity, 2.0_wp /)     ! variable upper bound
+   p%X = 0.0_wp ; p%Y = 0.0_wp ; p%Z = 0.0_wp ! start from zero
+!  dense storage format
+   CALL SMT_put( p%H%type, 'DENSE', s )     ! Specify dense
+   CALL SMT_put( p%A%type, 'DENSE', s )     ! storage for H and A
+   ALLOCATE( p%H%val( n * ( n + 1 ) / 2 ) )
+   ALLOCATE( p%A%val( n * m ) )
+   p%H%val = (/ 1.0_wp, 0.0_wp, 2.0_wp, 0.0_wp, 1.0_wp, 3.0_wp /) ! Hessian H
+   p%A%val = (/ 2.0_wp, 1.0_wp, 0.0_wp, 0.0_wp, 1.0_wp, 1.0_wp /) ! Jacobian A
+! problem data complete
+   CALL CCQP_initialize( data, control, inform ) ! Initialize control parameters
+!  control%print_level = 1
+   control%infinity = infinity                  ! Set infinity
+   CALL CCQP_solve( p, data, control, inform, C_stat, B_stat ) ! Solve
+   IF ( inform%status == 0 ) THEN               !  Successful return
+     WRITE( 6, "( ' CCQP: ', I0, ' iterations  ', /,                           &
+    &     ' Optimal objective value =',                                        &
+    &       ES12.4, /, ' Optimal solution = ', ( 5ES12.4 ) )" )                &
+     inform%iter, inform%obj, p%X
+   ELSE                                         !  Error returns
+     WRITE( 6, "( ' CCQP_solve exit status = ', I6 ) " ) inform%status
+   END IF
+   CALL CCQP_terminate( data, control, inform )  !  delete internal workspace
+   END PROGRAM GALAHAD_CCQP_EXAMPLE3

[…]
