### 2022-05-22 caed5b1
improvements to ccqp
