+++ src/slls/C/sllstf.c
@@ -1,0 +1,321 @@
+/* sllstf.c */
+/* Full test for the SLLS C interface using fortran sparse matrix indexing */
+
+#include <stdio.h>
+#include <math.h>
+#include <string.h>
+#include "galahad_slls.h"
+
+// define max
+
+#define max(a,b)             \
+({                           \
+    __typeof__ (a) _a = (a); \
+    __typeof__ (b) _b = (b); \
+    _a > _b ? _a : _b;       \
+})
+
+// Custom userdata struct
+struct userdata_type {
+   double scale;
+};
+
+// Function prototypes
+int prec( int n, const double v[], double p[], const void * );
+
+int main(void) {
+
+    // Derived types
+    void *data;
+    struct slls_control_type control;
+    struct slls_inform_type inform;
+
+    // Set user data
+    struct userdata_type userdata;
+    userdata.scale = 1.0;
+
+    // Set problem data
+    int n = 10; // dimension
+    int m = n + 1; // number of residuals
+    int A_ne = 2 * n; // sparse Jacobian elements
+    int A_dense_ne = m * n; // dense Jacobian elements
+    // row-wise storage
+    int A_row[A_ne]; // row indices,
+    int A_col[A_ne]; // column indices
+    int A_ptr[m+1];  // row pointers
+    double A_val[A_ne]; // values
+    double A_dense[A_dense_ne]; // dense values
+    // column-wise storage
+    int A_by_col_row[A_ne]; // row indices,
+    int A_by_col_ptr[n+1];  // column pointers
+    double A_by_col_val[A_ne]; // values
+    double A_by_col_dense[A_dense_ne]; // dense values
+    double b[m];  // linear term in the objective
+    double x[n]; // variables
+    double z[n]; // dual variables
+    double c[m]; // residual
+    double g[n]; // gradient
+
+    // Set output storage
+    int x_stat[n]; // variable status
+    char st[3];
+    int i, l, status;
+
+    //   A = (  I  )  and b = ( i * e )
+    //       ( e^T )          ( n + 1 )
+
+    for( int i = 0; i < n; i++) b[i] = i + 1;
+    b[n] = n+1;
+
+    // A by rows
+
+    for( int i = 0; i < n; i++)
+    {
+      A_ptr[i] = i + 1;
+      A_row[i] = i + 1; A_col[i] = i + 1; A_val[i] = 1.0;
+    }
+    A_ptr[n] = n + 1;
+    for( int i = 0; i < n; i++)
+    {
+      A_row[n+i] = m; A_col[n+i] = i + 1; A_val[n+i] = 1.0;
+    }
+    A_ptr[m] = A_ne + 1;
+    l = - 1;
+    for( int i = 0; i < n; i++)
+    {
+      for( int j = 0; j < n; j++)
+      {
+        l = l + 1;
+        if ( i == j ) {
+          A_dense[l] = 1.0;
+        }
+        else {
+          A_dense[l] = 0.0;
+        }
+      }
+    }
+    for( int j = 0; j < n; j++)
+    {
+      l = l + 1;
+      A_dense[l] = 1.0;
+    }
+
+    // A by columns
+
+    l = - 1;
+    for( int j = 0; j < n; j++)
+    {
+      l = l + 1;  A_by_col_ptr[j] = l + 1;
+      A_by_col_row[l] = j + 1; A_by_col_val[l] = 1.0;
+      l = l + 1;
+      A_by_col_row[l] = m; A_by_col_val[l] = 1.0;
+    }
+    A_by_col_ptr[n] = A_ne;
+    l = - 1;
+    for( int j = 0; j < n; j++)
+    {
+      for( int i = 0; i < n; i++)
+      {
[…]
