+++ src/slls/sllss3.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-05-26 AT 07:15 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-06-03 AT 11:40 GMT
 ! Used to test components of package
    PROGRAM GALAHAD_SLLS_EXAMPLE3
    USE GALAHAD_SLLS_double         ! double precision version
@@ -6,23 +6,181 @@
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
    REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
    TYPE ( QPT_problem_type ) :: p
-   TYPE ( SLLS_data_type ) :: data
+!  TYPE ( SLLS_data_type ) :: data
    TYPE ( SLLS_control_type ) :: control
    TYPE ( SLLS_inform_type ) :: inform
    TYPE ( GALAHAD_userdata_type ) :: userdata
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_stat
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: D
+   INTEGER :: status, segment, max_steps
+   LOGICAL :: advance
+   REAL ( KIND = wp ) :: f_opt, t_opt, t_0, t_max, beta, eta, v_j
+   INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_stat, FREE
+   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: D, S, AD, AE, R, B, R_t
+   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: A_val
+   INTEGER, ALLOCATABLE, DIMENSION( : ) :: A_row, A_ptr
+   TYPE ( SLLS_search_data_type ) :: data
+   TYPE ( SLLS_reverse_type ) :: reverse
+   REAL ( KIND = wp ) :: scale, scale2, flipflop, x_j, d_j, a
+   INTEGER :: i, j, l, n_free
+!  INTEGER, PARAMETER :: n = 100
    INTEGER, PARAMETER :: n = 10
-   INTEGER :: j, status
-   REAL ( KIND = wp ) :: scale, scale2, flipflop
-   ALLOCATE( p%X( n ), D( n ) )
+!!    INTEGER, PARAMETER :: n = 4
+   INTEGER, PARAMETER :: m = n + 1
+   INTEGER, PARAMETER :: a_ne = 2 * n
+   ALLOCATE( p%X( n ), D( n ), FREE( n ), AD( m ), AE( m ), R( m ) )
+   ALLOCATE( B( m ), A_val( a_ne ), A_row( a_ne ), A_ptr( n + 1 ) )
+!  p%X = (/ 0.01_wp, 0.01_wp, 0.05_wp, 0.93_wp /)
+!  D = (/ -10.0_wp, - 1.0_wp, -10.0_wp, 21.0_wp /)
    scale = REAL( n, KIND = wp )
    scale2 = REAL( n * ( n + 1 ) / 2, KIND = wp )
+   flipflop = 1.0_wp
    DO j = 1, n
      p%X( j ) = 1.0_wp / scale
-     D( j ) = REAL( j, KIND = wp ) / scale2
+     D( j ) = flipflop * REAL( j, KIND = wp ) / scale2
+     flipflop = - 1.0_wp * flipflop
    END DO
-   CALL SLLS_simplex_projection_path( n, p%X, D, status )
+!  p%X = 0.0_wp
+!  p%X( n-1 ) = 1.0_wp
[…]
