+++ src/slls/slls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-05-26 AT 07:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-06-03 AT 11:40 GMT
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -29,7 +29,8 @@
      USE GALAHAD_STRING
      USE GALAHAD_SPACE_double
      USE GALAHAD_SORT_double, ONLY: SORT_heapsort_build,                       &
-                                    SORT_heapsort_smallest, SORT_partition
+                                    SORT_heapsort_smallest, SORT_partition,    &
+                                    SORT_quicksort
      USE GALAHAD_SBLS_double
      USE GALAHAD_NORMS_double
      USE GALAHAD_QPT_double
@@ -46,12 +47,14 @@
      PUBLIC :: SLLS_initialize, SLLS_read_specfile, SLLS_solve,                &
                SLLS_terminate,  SLLS_reverse_type, SLLS_data_type,             &
                SLLS_full_initialize, SLLS_full_terminate,                      &
+               SLLS_search_data_type,                                          &
                SLLS_subproblem_data_type, SLLS_exact_arc_search,               &
                SLLS_inexact_arc_search, SLLS_import, SLLS_import_without_a,    &
                SLLS_solve_given_a, SLLS_solve_reverse_a_prod,                  &
                SLLS_reset_control, SLLS_information, GALAHAD_userdata_type,    &
                QPT_problem_type, SMT_type, SMT_put, SMT_get, &
                SLLS_project_onto_simplex, SLLS_simplex_projection_path
+
 
 !----------------------
 !   I n t e r f a c e s
@@ -335,6 +338,16 @@
        TYPE ( CONVERT_inform_type ) :: CONVERT_inform
      END TYPE SLLS_inform_type
 
+!  - - - - - - - - - - - - - - -
+!   arc_search data derived type
+!  - - - - - - - - - - - - - - -
+
+     TYPE :: SLLS_search_data_type
+       INTEGER :: step
+       REAL ( KIND = wp ) :: ete, f_0, f_1_stop, gamma, rtr
+       REAL ( KIND = wp ) :: s_fixed, t, t_break, t_total
+       LOGICAL :: reverse, backwards
+     END TYPE SLLS_search_data_type
 !  - - - - - - - - - - - - - - -
 !   arc_search data derived type
 !  - - - - - - - - - - - - - - -
@@ -2124,45 +2137,45 @@
 !  ... using the available Jacobian ...
 
            IF (  data%explicit_a ) THEN
-             CALL SLLS_exact_arc_search( prob%m, prob%n, data%weight,          &
-                                         prob%X_l, prob%X_u,                   &
-                                         control%infinity, prob%X, prob%C,     &
-                                         data%S, data%X_status,                &
-                                         epsmch,                               &
-                                         control%alpha_max,                    &
-                                         control%arcsearch_max_steps,          &
-                                         control%out, control%print_level,     &
-                                         prefix, data%X_new, data%f_new,       &
-                                         data%phi_new,                         &
-                                         data%alpha_new, data%segments,        &
-                                         data%subproblem_data, userdata,       &
-                                         data%arc_search_status,               &
-                                         inform%alloc_status,                  &
-                                         inform%bad_alloc,                     &
-                                         A_ptr = data%A%ptr,                   &
-                                         A_row = data%A%row,                   &
-                                         A_val = data%A%val )
+!             CALL SLLS_exact_arc_search( prob%m, prob%n, data%weight,         &
+!                                         prob%X_l, prob%X_u,                  &
[…]
