+++ src/slls/slls.f90
@@ -353,25 +353,15 @@
 !  - - - - - - - - - - - - - - -
 
      TYPE :: SLLS_subproblem_data_type
-       INTEGER :: branch, n_break, n_free, n_fixed, n_a0, preconditioner
-       INTEGER :: nz_d_start, nz_d_end, nz_out_end, base_free, step
-       REAL ( KIND = wp ) :: f_alpha_dash, f_alpha_dashdash, stop_cg
-       REAL ( KIND = wp ) :: rho_alpha, rho_alpha_dash, rho_alpha_dashdash
-       REAL ( KIND = wp ) :: phi_alpha_dash, phi_alpha_dashdash
-       REAL ( KIND = wp ) :: alpha_i, alpha_next, delta_alpha, alpha, target
-       REAL ( KIND = wp ) :: f_s, f_c, f_i, f_l, f_q, gamma, gamma_a, gamma_f
-       REAL ( KIND = wp ) :: rho_c, rho_i, rho_l, rho_q, phi_s, phi_i
-       REAL ( KIND = wp ) :: mu, mu_a, mu_f
+       INTEGER :: branch, n_free, n_preconditioner, step
+       REAL ( KIND = wp ) :: stop_cg, gamma, etpe
        LOGICAL :: printp, printw, printd, printdd, debug
        LOGICAL :: present_a, present_asprod, reverse_asprod, present_afprod
        LOGICAL :: reverse_afprod, reverse_prec, present_prec, present_dprec
        LOGICAL :: recompute, regularization, preconditioned
        CHARACTER ( LEN = 1 ) :: direction
-       INTEGER, ALLOCATABLE, DIMENSION( : ) :: FREE, NZ_d, NZ_out, P_used
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: G, P, Q, R, S, U, W
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: R_a, R_f, X_a, D_f
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: BREAK_points
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X_debug, R_debug
+       INTEGER, ALLOCATABLE, DIMENSION( : ) :: FREE
+       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: G, P, Q, PG, PE
      END TYPE SLLS_subproblem_data_type
 
 !  - - - - - - - - - - -
@@ -394,7 +384,7 @@
        INTEGER :: out, error, print_level, start_print, stop_print, print_gap
        INTEGER :: arc_search_status, cgls_status, change_status
        INTEGER :: n_free, branch, cg_iter, preconditioner
-       INTEGER :: nz_in_start, nz_in_end, nz_out_end, maxit, cg_maxit
+       INTEGER :: maxit, cg_maxit
        INTEGER :: segments, max_segments, steps, max_steps, eval_status
        REAL :: time_start
        REAL ( KIND = wp ) :: norm_step, step, stop_cg, old_gnrmsq, pnrmsq
@@ -406,11 +396,11 @@
        LOGICAL :: direct_subproblem_solve, steepest_descent
        CHARACTER ( LEN = 6 ) :: string_cg_iter
        INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_status, OLD_status
-       INTEGER, ALLOCATABLE, DIMENSION( : ) :: NZ_in, NZ_out
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X_new, G, R, V, P
[…]
