+++ src/slls/sllss3.f90
@@ -6,28 +6,29 @@
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
    REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
    TYPE ( QPT_problem_type ) :: p
-!  TYPE ( SLLS_data_type ) :: data
+   TYPE ( SLLS_data_type ) :: data
    TYPE ( SLLS_control_type ) :: control
    TYPE ( SLLS_inform_type ) :: inform
    TYPE ( GALAHAD_userdata_type ) :: userdata
    INTEGER :: status, segment, max_steps
    LOGICAL :: advance
    REAL ( KIND = wp ) :: f_opt, t_opt, t_0, t_max, beta, eta, v_j
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_status, FREE
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: D, S, AD, AE, R, B, R_t
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: A_val
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: A_row, A_ptr
-   TYPE ( SLLS_search_data_type ) :: data
+   INTEGER, ALLOCATABLE, DIMENSION( : ) :: FREE, X_stat
+   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: D, S, AD, AE, R, R_t, DIAG
+   TYPE ( SLLS_search_data_type ) :: search_data
+   TYPE ( SLLS_subproblem_data_type ) :: subproblem_data
    TYPE ( SLLS_reverse_type ) :: reverse
-   REAL ( KIND = wp ) :: scale, scale2, flipflop, x_j, d_j, a
-   INTEGER :: i, j, l, n_free
-!  INTEGER, PARAMETER :: n = 100
-   INTEGER, PARAMETER :: n = 10
+   REAL ( KIND = wp ) :: scale, scale2, flipflop, x_j, d_j, a, val, f
+   INTEGER :: i, j, l, n_free, alloc_status
+   CHARACTER ( LEN = 80 ) :: bad_alloc
+   INTEGER, PARAMETER :: n = 100
+!  INTEGER, PARAMETER :: n = 10
 !!    INTEGER, PARAMETER :: n = 4
    INTEGER, PARAMETER :: m = n + 1
    INTEGER, PARAMETER :: a_ne = 2 * n
-   ALLOCATE( p%X( n ), D( n ), FREE( n ), X_status( n ), AD( m ), AE( m )
-   ALLOCATE( R( m ) ), B( m ), A_val( a_ne ), A_row( a_ne ), A_ptr( n + 1 ) )
+   ALLOCATE( p%X( n ), p%Z( n ), D( n ), DIAG( n ), FREE( n ), X_stat( n ) )
+   ALLOCATE( AD( m ), AE( m ), R( m ), p%B( m ), p%C( m ), p%G( n ) )
+   ALLOCATE( p%A%val( a_ne ), p%A%row( a_ne ), p%A%ptr( n + 1 ) )
 !  p%X = (/ 0.01_wp, 0.01_wp, 0.05_wp, 0.93_wp /)
 !  D = (/ -10.0_wp, - 1.0_wp, -10.0_wp, 21.0_wp /)
    scale = REAL( n, KIND = wp )
@@ -44,55 +45,66 @@
 !  D = 2.0_wp * D - 1.0_wp
    D = 1.0_wp ;  D(1) = - 1.0_wp ;  D(2) = 0.0_wp
    l = 1
+
[…]
