+++ src/slls/sllss4.f90
@@ -1,0 +1,184 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-06-12 AT 10:15 GMT
+   PROGRAM GALAHAD_SLLS_EXAMPLE4
+
+! use slls to solve a bilinear fitting problem 
+!  min 1/2 sum_i ( y^T A_i x - b_i)^2,
+! where x and y lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
+! using alternating solves with x free and y fixed, and vice versa
+
+! For this example, n = 3m/2,
+!
+!        <-   m   -> <-n-m->
+! A_i = ( 1 . 1 . 1 | i     )
+!       (   .       |   .   )
+!       (     1     |     i ) <-n-m
+!       (       .   .     . )
+!       (         1 | 1 . 1 )
+!
+! and b_i is y_*^T A_i x_* for given x_* and y_*
+
+   USE GALAHAD_SLLS_double         ! double precision version
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   TYPE ( QPT_problem_type ) :: p_x, p_y
+   TYPE ( SLLS_data_type ) :: data_x, data_y
+   TYPE ( SLLS_control_type ) :: control_x, control_y
+   TYPE ( SLLS_inform_type ) :: inform_x, inform_y
+   TYPE ( GALAHAD_userdata_type ) :: userdata
+   INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_stat, Y_stat
+   INTEGER :: i, j, ne, s
+   INTEGER, PARAMETER :: m = 20, n = 3 * m / 2
+   INTEGER, PARAMETER :: obs = 100
+   REAL ( KIND = wp ) :: X( n ), Y( m ), B( obs ), Aix( m ), ATiy( n )
+   REAL ( KIND = wp ) :: AX( obs, n ), AY( obs, m )
+! start problem data
+!  solution is at x = 0 for i=1:n/2 and 2/n for i=n/2+1:n
+   X( : n / 2 ) = 0.0_wp
+   X( n / 2 + 1 : ) = 1.0_wp / REAL( n / 2, KIND = wp )
+!  solution is at y = 2/m for i=1:m/2 and 0 for i=m/2+1:m 
+   Y( : m / 2 ) = 1.0_wp / REAL( m / 2, KIND = wp )
+   Y( m / 2 + 1 : ) = 0.0_wp
+
+!  set up storage for the x problem
+
+   CALL SMT_put( p_x%A%type, 'DENSE_BY_ROWS', s )
+   ALLOCATE( p_x%A%val(obs * n ), p_x%B( obs ), p_x%X( n ), X_stat( n ) )
+   p_x%m = obs ; p_x%n = n ; p_x%A%m = p_x%m ; p_x%A%n = p_x%n
+
+!  set up storage for the y problem
+
+   CALL SMT_put( p_y%A%type, 'DENSE_BY_ROWS', s )
+   ALLOCATE( p_y%A%val( obs * m ), p_y%B( obs ), p_y%X( m ), Y_stat( m ) )
+   p_y%m = obs ; p_y%n = m ; p_y%A%m = p_y%m ; p_y%A%n = p_y%n
+
+   DO i = 1, obs
+     CALL FORM_Aix( m, n, i, X, Aix )
+!    CALL FORM_ATiy( m, n, i, Y, ATiy )
[…]
