+++ src/slls/sllss4.f90
@@ -1,23 +1,15 @@
 ! THIS VERSION: GALAHAD 4.1 - 2022-06-12 AT 10:15 GMT
    PROGRAM GALAHAD_SLLS_EXAMPLE4
 
-! use slls to solve a bilinear fitting problem 
-!  min 1/2 sum_i ( y^T A_i x - b_i)^2,
+! use slls to solve a bilinear fitting problem
+!  min 1/2 sum_i ( xi y^T A_i x + b - mu_i)^2,
 ! where x and y lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
 ! using alternating solves with x free and y fixed, and vice versa
 
-! For this example, n = 3m/2,
-!
-!        <-   m   -> <-n-m->
-! A_i = ( 1 . 1 . 1 | i     )
-!       (   .       |   .   )
-!       (     1     |     i ) <-n-m
-!       (       .   .     . )
-!       (         1 | 1 . 1 )
-!
-! and b_i is y_*^T A_i x_* for given x_* and y_*
-
+! For this example, n = 3m/2, A_i is given by subroutine calls (below)
+! and mu_i is xi_* y_*^T A_i x_* + b_* for given x_*, y_*, xi_* and b_i*
    USE GALAHAD_SLLS_double         ! double precision version
+   USE GALAHAD_RAND_double
    IMPLICIT NONE
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
    REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
@@ -27,49 +19,78 @@
    TYPE ( SLLS_inform_type ) :: inform_x, inform_y
    TYPE ( GALAHAD_userdata_type ) :: userdata
    INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_stat, Y_stat
-   INTEGER :: i, j, ne, s
+   INTEGER :: i, j, ne, pass
+   TYPE ( RAND_seed ) :: seed
+   INTEGER, PARAMETER :: eg = 2
+   REAL ( KIND = wp ) :: val, f, xi, b, a_11, a_12, a_22, r_1, r_2
    INTEGER, PARAMETER :: m = 20, n = 3 * m / 2
    INTEGER, PARAMETER :: obs = 100
-   REAL ( KIND = wp ) :: X( n ), Y( m ), B( obs ), Aix( m ), ATiy( n )
+   INTEGER, PARAMETER :: pass_max = 3
+   INTEGER, PARAMETER :: sol = 2
+!  LOGICAL, PARAMETER :: pertub_observations = .TRUE.
+   LOGICAL, PARAMETER :: pertub_observations = .FALSE.
+   REAL ( KIND = wp ) :: X( n ), Y( m ), Aix( m ), ATiy( n ), MU( obs )
    REAL ( KIND = wp ) :: AX( obs, n ), AY( obs, m )
+
 ! start problem data
+
+! specify a solution
+
+   SELECT CASE( sol )
+! sparse solution
+   CASE ( 1 )
 !  solution is at x = 0 for i=1:n/2 and 2/n for i=n/2+1:n
-   X( : n / 2 ) = 0.0_wp
-   X( n / 2 + 1 : ) = 1.0_wp / REAL( n / 2, KIND = wp )
-!  solution is at y = 2/m for i=1:m/2 and 0 for i=m/2+1:m 
[…]
