+++ src/slls/slls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-06-03 AT 11:40 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-06-13 AT 09:00 GMT
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -6,7 +6,7 @@
 !  Principal author: Nick Gould
 
 !  History -
-!   originally released with GALAHAD Version 4.1. March 26th 2022
+!   originally released with GALAHAD Version 4.1. June 3rd 2022
 
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
@@ -1606,7 +1606,11 @@
            DO k = data%A%ptr( j ), data%A%ptr( j + 1 ) - 1
              val = val + data%A%val( k ) ** 2
            END DO
-           data%DIAG( j ) = val
+           IF ( val > zero ) THEN
+             data%DIAG( j ) = val
+           ELSE
+             data%DIAG( j ) = one
+           END IF
          END DO
        ELSE
          data%DIAG( : prob%n ) = one
@@ -1820,7 +1824,7 @@
 
        CALL SLLS_project_onto_simplex( prob%n, prob%X - prob%G( : prob%n ),    &
                                        data%X_new, i )
-       inform%norm_pg = MAXVAL( ABS( data%X_new -prob%X ) ) 
+       inform%norm_pg = MAXVAL( ABS( data%X_new -prob%X ) )
 
 !  print details of the current iteration
 
@@ -1929,7 +1933,7 @@
 
          nap = 0
          DO i = 1, data%n_free
-           j = data%FREE( i )  
+           j = data%FREE( i )
            data%AT_sbls%ptr( i ) = nap + 1
            DO k = data%A%ptr( j ), data%A%ptr( j + 1 ) - 1
              nap = nap + 1
@@ -2004,7 +2008,7 @@
 
          data%D( : prob%n ) = zero
          DO i = 1, data%n_free
-           j = data%FREE( i )  
+           j = data%FREE( i )
            data%D( j ) =  data%S( i ) - lambda * data%SBLS_sol( prob%m + i )
          END DO
 !write(6,*) ' ----------- nfree ', data%n_free
@@ -2044,7 +2048,7 @@
 write(6,"( ' nfree = ', I0, ' FREE = ', /, ( 10I8 ) )" ) &
  data%n_free, data%FREE( : data%n_free )
 !write(6,"( ' x = ', /, ( 5ES12.4 ) )" ) data%X_new( : prob%n )
-       
+
  300   CONTINUE ! mock CGLS loop
 
 !  find an improved point, X_new, by conjugate-gradient least-squares ...
@@ -2192,7 +2196,7 @@ write(6,"( ' nfree = ', I0, ' FREE = ', /, ( 10I8 ) )" ) &
 
 !  set A d and A e
 
-       data%AD( : prob%m ) = zero 
+       data%AD( : prob%m ) = zero
        IF (  data%explicit_a ) THEN
          DO j = 1, prob%n
            d_j = data%D( j )
@@ -2222,7 +2226,7 @@ write(6,"( ' nfree = ', I0, ' FREE = ', /, ( 10I8 ) )" ) &
 
  410   CONTINUE ! mock arc search loop
 
-!  find an improved point, X_new, by exact arc_search using the available 
+!  find an improved point, X_new, by exact arc_search using the available
 !  Jacobian ...
 
          IF (  data%explicit_a ) THEN
[…]
