### 2022-06-15 9d14691
added scaling factors to blls and slls, plus other minor updates
