+++ src/blls/blls.f90
@@ -13,16 +13,18 @@
 
    MODULE GALAHAD_BLLS_double
 
-!        ----------------------------------------------------------------
-!        |                                                              |
-!        | Solve the bound-constrained linear-least-squares problem     |
-!        |                                                              |
-!        |    minimize   1/2 || A x - b ||_2^2 + weight / 2 || x ||^2   |
-!        |    subject to     x_l <= x <= x_u                            |
-!        |                                                              |
-!        | using a preconditioned projected conjugate-gradient approach |
-!        |                                                              |
-!        ----------------------------------------------------------------
+!        -------------------------------------------------------------------
+!        |                                                                  |
+!        | Solve the bound-constrained linear-least-squares problem         |
+!        |                                                                  |
+!        |    minimize   1/2 || A x - b ||_W^2 + weight / 2 || x ||^2       |
+!        |    subject to     x_l <= x <= x_u,                               |
+!        |                                                                  |
+!        | where ||v|| and ||r||_W^2 are the Euclidean & weighted Euclidean |
+!        | norms defined by ||v||^2 = v^T v and ||r||_W^2 = r^T W r, using  |
+!        | a preconditioned projected conjugate-gradient approach           |
+!        |                                                                  |
+!        --------------------------------------------------------------------
 
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
@@ -394,7 +396,7 @@
        LOGICAL :: set_printt, set_printi, set_printw, set_printd, set_printe
        LOGICAL :: set_printm, printt, printi, printm, printw, printd, printe
        LOGICAL :: reverse, reverse_prod, explicit_a, use_aprod, header
-       LOGICAL :: direct_subproblem_solve, steepest_descent
+       LOGICAL :: direct_subproblem_solve, steepest_descent, w_eq_identity
        CHARACTER ( LEN = 6 ) :: string_cg_iter
        INTEGER, ALLOCATABLE, DIMENSION( : ) :: X_status, OLD_status
        INTEGER, ALLOCATABLE, DIMENSION( : ) :: NZ_in, NZ_out
@@ -784,14 +786,14 @@
 !-*-*-*-*-*-*-*-   B L L S _ S O L V E  S U B R O U T I N E   -*-*-*-*-*-*-*-*-
 
      SUBROUTINE BLLS_solve( prob, X_stat, data, control, inform, userdata,     &
-                            reverse, eval_APROD, eval_ASPROD, eval_AFPROD,     &
+                            W, reverse, eval_APROD, eval_ASPROD, eval_AFPROD,  &
                             eval_PREC )
 
 ! =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 !
 !  Solve the linear least-squares problem
 !
-!     minimize     q(x) = 1/2 || A x - b ||_2^2 + weight / 2 || x ||^2
+!     minimize     q(x) = 1/2 || A x - b ||_W^2 + weight / 2 || x ||^2
 !
 !     subject to   (x_l)_i <=   x_i  <= (x_u)_i , i = 1, .... , n,
 !
@@ -1063,6 +1065,10 @@
 !    character_pointer is a rank-one pointer array of type default character.
[…]
