+++ src/slls/slls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-06-13 AT 09:00 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-06-13 AT 16:00 GMT
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -13,16 +13,18 @@
 
    MODULE GALAHAD_SLLS_double
 
-!        ----------------------------------------------------------------
-!        |                                                              |
-!        | Solve the simplex-constrained linear-least-squares problem   |
-!        |                                                              |
-!        |    minimize   1/2 || A x - b ||_2^2 + weight / 2 || x ||^2   |
-!        |    subject to      e^T x = 1, x >= 0                         |
-!        |                                                              |
-!        | using a preconditioned projected conjugate-gradient approach |
-!        |                                                              |
-!        ----------------------------------------------------------------
+!        --------------------------------------------------------------------
+!        |                                                                  |
+!        | Solve the simplex-constrained linear-least-squares problem       |
+!        |                                                                  |
+!        |    minimize   1/2 || A x - b ||_W^2 + weight / 2 || x ||^2       |
+!        |    subject to      e^T x = 1, x >= 0                             |
+!        |                                                                  |
+!        | where ||v|| and ||r||_W^2 are the Euclidean & weighted Euclidean |
+!        | norms defined by ||v||^2 = v^T v and ||r||_W^2 = r^T W r, using  |
+!        | a preconditioned projected conjugate-gradient approach           |
+!        |                                                                  |
+!        --------------------------------------------------------------------
 
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
@@ -393,7 +395,7 @@
        LOGICAL :: set_printt, set_printi, set_printw, set_printd, set_printe
        LOGICAL :: set_printm, printt, printi, printm, printw, printd, printe
        LOGICAL :: reverse, reverse_prod, explicit_a, use_aprod, header
-       LOGICAL :: direct_subproblem_solve, steepest_descent
+       LOGICAL :: direct_subproblem_solve, steepest_descent, w_eq_identity
        CHARACTER ( LEN = 6 ) :: string_cg_iter
        INTEGER, ALLOCATABLE, DIMENSION( : ) :: FREE
        LOGICAL, ALLOCATABLE, DIMENSION( : ) :: FIXED, FIXED_old
@@ -778,7 +780,7 @@
 !-*-*-*-*-*-*-*-   S L L S _ S O L V E  S U B R O U T I N E   -*-*-*-*-*-*-*-*-
 
      SUBROUTINE SLLS_solve( prob, X_stat, data, control, inform, userdata,     &
-                            reverse, eval_APROD, eval_ASPROD, eval_AFPROD,     &
+                            W, reverse, eval_APROD, eval_ASPROD, eval_AFPROD,  &
                             eval_PREC )
 
 ! =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
@@ -1054,6 +1056,10 @@
 !    character_pointer is a rank-one pointer array of type default character.
 !    logical_pointer is a rank-one pointer array of type default logical.
 !
[…]
