### 2022-06-15 f855424
added two more tests for slls
