+++ src/slls/sllss5.f90
@@ -1,0 +1,178 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-06-14 AT 09:00 GMT
+   PROGRAM GALAHAD_SLLS_EXAMPLE5
+
+! use slls to solve a bilinear fitting problem
+!  min 1/2 sum_i ( xi a^T_i x + b - mu_i)^2,
+! where x lies in n regular simplex {x : e^T x = 1, x >= 0},
+! using alternating solves with x free and (xi,b) fixed, and vice versa
+
+   USE GALAHAD_SLLS_double         ! double precision version
+   USE GALAHAD_RAND_double
+   USE GALAHAD_NORMS_double
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   TYPE ( QPT_problem_type ) :: p_r
+   TYPE ( SLLS_data_type ) :: data_r
+   TYPE ( SLLS_control_type ) :: control_r
+   TYPE ( SLLS_inform_type ) :: inform_r
+   TYPE ( GALAHAD_userdata_type ) :: userdata
+   INTEGER :: i, j, ne, pass
+   TYPE ( RAND_seed ) :: seed
+   REAL ( KIND = wp ) :: val, f, xi, b, a_max
+   REAL ( KIND = wp ) :: a_11, a_12, a_22, r_1, r_2, alpha, beta
+   REAL :: time, time_start, time_total
+   INTEGER, PARAMETER :: n = 500
+   INTEGER, PARAMETER :: obs = 200
+   INTEGER, PARAMETER :: it_max = 100
+   INTEGER, PARAMETER :: pass_max = 30
+   INTEGER, DIMENSION( n ) :: R_stat
+   REAL ( KIND = wp ) :: R( n ), AT( n, obs ), MU( obs ), SIGMA( obs )
+
+!  set up storage for the r problem
+
+   CALL SMT_put( p_r%A%type, 'DENSE_BY_ROWS', i )
+   ALLOCATE( p_r%A%val(obs * n ), p_r%B( obs ), p_r%X( n ) )
+   p_r%m = obs ; p_r%n = n ; p_r%A%m = p_r%m ; p_r%A%n = p_r%n
+
+! read problem data. Store rows of A as columns of AT
+
+   OPEN( unit = 21, file = "G.txt" )
+   DO i = 1, obs
+     READ( 21, * ) AT( 1 : n, i )
+   END DO
+   CLOSE( unit = 21 )
+
+   OPEN( unit = 22, file = "observation.txt" )
+   a_max = 0.0_wp
+   DO i = 1, obs
+     READ( 22, * ) val, MU( i ), SIGMA( i )
+     MU( i ) = MU( i ) / SIGMA( i )
+     AT( 1 : n, i ) = AT( 1 : n, i ) / SIGMA( i )
+     a_max = MAX ( a_max, TWO_NORM( AT( 1 : n, i ) ) )
+!write(6,"( ' ||a||_2, b =', 2ES12.4 )" ) a_max, MU( i )
+   END DO
+   CLOSE( unit = 22 )
+
+!  scale A and consequently xi
+
+   AT = AT / a_max
+!stop
+!  starting point
+
+   R = 1.0_wp / REAL( n, KIND = wp )
+   xi = 1.0_wp ; b = 0.0_wp
+
+!  report initial residuals
+
+   f = 0.0_wp
+   DO i = 1, obs
+     alpha = DOT_PRODUCT( R( 1 : n ), AT( 1 : n, i ) )
+     beta = 1.0_wp / SIGMA( i )
[…]
