+++ src/slls/sllss6.f90
@@ -1,0 +1,297 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-06-12 AT 10:15 GMT
+   PROGRAM GALAHAD_SLLS_EXAMPLE6
+
+! use slls to solve a multilinear fitting problem
+!  min 1/2 sum_i ( xi sum_iu sum_iv sum_ix sum_iy a(i,iu,iv,ix,iy)
+!                  u_iu v_iv x_ix y_iy + b - mu_i)^2,
+! where u, v, w and x lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
+! using alternating solves with u free and v,x,y fixed, and vice versa
+
+   USE GALAHAD_SLLS_double         ! double precision version
+   USE GALAHAD_RAND_double
+   USE GALAHAD_NORMS_double
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   TYPE ( QPT_problem_type ) :: p_u, p_v, p_x, p_y
+   TYPE ( SLLS_data_type ) :: data_u, data_v, data_x, data_y
+   TYPE ( SLLS_control_type ) :: control_u, control_v, control_x, control_y
+   TYPE ( SLLS_inform_type ) :: inform_u, inform_v, inform_x, inform_y
+   TYPE ( GALAHAD_userdata_type ) :: userdata
+   INTEGER :: i, iu, iv, ix, iy, j, ne, pass
+   TYPE ( RAND_seed ) :: seed
+   REAL ( KIND = wp ) :: val, f, xi, b, a_11, a_12, a_22, r_1, r_2
+   REAL ( KIND = wp ) :: alpha, beta, prod, prod1, prod2, prod3, prod4
+   REAL :: time, time_start, time_total
+   INTEGER, PARAMETER :: nu = 4
+   INTEGER, PARAMETER :: nv = 8
+   INTEGER, PARAMETER :: nx = 12
+   INTEGER, PARAMETER :: ny = 16
+   INTEGER, PARAMETER :: obs = 80
+   INTEGER, PARAMETER :: pass_max = 3
+!  LOGICAL, PARAMETER :: pertub_observations = .TRUE.
+   LOGICAL, PARAMETER :: pertub_observations = .FALSE.
+   INTEGER :: U_stat( nu ), V_stat( nv ), X_stat( nx ), Y_stat( ny )
+   REAL ( KIND = wp ) :: U( nu ), V( nv ), X( nx ), Y( nx ), MU( obs )
+   REAL ( KIND = wp ) :: A( nu, nv, nx, ny, obs )
+
+! start problem data
+
+! generate sparse solution
+
+   CALL RAND_initialize( seed )
+   CALL RAND_random_real( seed, .FALSE., U( : nu ) )
+   U = MAX( U, 0.0_wp )
+   val = SUM( U( : nu ) ) ; U( : nu ) = U( : nu ) / val
+   CALL RAND_random_real( seed, .FALSE., V( : nv ) )
+   V = MAX( V, 0.0_wp )
+   val = SUM( V( : nv ) ) ; V( : nv ) = V( : nv ) / val
+   CALL RAND_random_real( seed, .FALSE., X( : nx ) )
+   X = MAX( X, 0.0_wp )
+   val = SUM( X( : nx ) ) ; X( : nx ) = X( : nx ) / val
+   CALL RAND_random_real( seed, .FALSE., Y( : ny ) )
+   Y = MAX( Y, 0.0_wp )
[…]
