### 2022-06-17 325389a
further additions to slls
