+++ src/slls/sllss6.f90
@@ -20,20 +20,27 @@
    TYPE ( GALAHAD_userdata_type ) :: userdata
    INTEGER :: i, iu, iv, ix, iy, j, ne, pass
    TYPE ( RAND_seed ) :: seed
-   REAL ( KIND = wp ) :: val, f, xi, b, a_11, a_12, a_22, r_1, r_2
-   REAL ( KIND = wp ) :: alpha, beta, prod, prod1, prod2, prod3, prod4
+   REAL ( KIND = wp ) :: val, f, xi, b, a_11, a_12, a_22, r_1, r_2, f_best
+   REAL ( KIND = wp ) :: beta, prod, prod1, prod2, prod3, prod4
    REAL :: time, time_start, time_total
-   INTEGER, PARAMETER :: nu = 4
-   INTEGER, PARAMETER :: nv = 8
-   INTEGER, PARAMETER :: nx = 12
-   INTEGER, PARAMETER :: ny = 16
-   INTEGER, PARAMETER :: obs = 80
-   INTEGER, PARAMETER :: pass_max = 3
-!  LOGICAL, PARAMETER :: pertub_observations = .TRUE.
-   LOGICAL, PARAMETER :: pertub_observations = .FALSE.
+!  INTEGER, PARAMETER :: nu = 4
+   INTEGER, PARAMETER :: nu = 10
+!  INTEGER, PARAMETER :: nv = 8
+   INTEGER, PARAMETER :: nv = 10
+!  INTEGER, PARAMETER :: nx = 12
+   INTEGER, PARAMETER :: nx = 10
+!  INTEGER, PARAMETER :: ny = 16
+   INTEGER, PARAMETER :: ny = 10
+!  INTEGER, PARAMETER :: obs = 80
+   INTEGER, PARAMETER :: obs = 10000
+   INTEGER, PARAMETER :: pass_max = 100
+   LOGICAL, PARAMETER :: prints = .FALSE.
+   REAL ( KIND = wp ), PARAMETER :: f_stop = ( 10.0_wp ) ** ( - 16 )
+   LOGICAL, PARAMETER :: pertub_observations = .TRUE.
+!  LOGICAL, PARAMETER :: pertub_observations = .FALSE.
    INTEGER :: U_stat( nu ), V_stat( nv ), X_stat( nx ), Y_stat( ny )
-   REAL ( KIND = wp ) :: U( nu ), V( nv ), X( nx ), Y( nx ), MU( obs )
-   REAL ( KIND = wp ) :: A( nu, nv, nx, ny, obs )
+   REAL ( KIND = wp ) :: U( nu ), V( nv ), X( nx ), Y( ny ), MU( obs )
+   REAL ( KIND = wp ) :: A( nu, nv, nx, ny, obs ), ALPHA( obs )
 
 ! start problem data
 
@@ -56,7 +63,7 @@
    xi = 2.0_wp ; b = 0.5_wp
 
 !  random A
-   CALL CPU_TIME( time_start )
+
    DO i = 1, obs
      DO iy = 1, ny
        DO ix = 1, nx
@@ -68,9 +75,6 @@
        END DO
      END DO
    END DO
-   CALL CPU_TIME( time ) ; time_total = time - time_start
-   WRITE( 6, "( ' total CPU time = ', F0.2 )" ) time_total
-   stop
 
 !  generate a set of observations, mu_i
 
@@ -124,88 +128,229 @@
 !  starting point
    U = 1.0_wp / REAL( nu, KIND = wp ) ; V = 1.0_wp / REAL( nv, KIND = wp )
    X = 1.0_wp / REAL( nx, KIND = wp ) ; Y = 1.0_wp / REAL( ny, KIND = wp )
-   xi = 2.0_wp ; b = 1.0_wp
[…]
