+++ src/slls/sllss7.f90
@@ -1,0 +1,526 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-06-16 AT 09:45 GMT
+   PROGRAM GALAHAD_SLLS_EXAMPLE7
+
+! use slls to solve a multilinear fitting problem
+!  min 1/2 sum_i ( xi sum_il sum_ir sum_it sum_ip a(il,ir,it,ip,i)
+!                  l_il r_ir theta_it y_phi + b - mu_i)^2,
+! where u, v, w and x lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
+! using alternating solves with u free and v,x,y fixed, and vice versa
+
+   USE GALAHAD_SLLS_double         ! double precision version
+   USE GALAHAD_NORMS_double
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
+   REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 20
+   TYPE ( QPT_problem_type ) :: p_l, p_r, p_t, p_p
+   TYPE ( SLLS_data_type ) :: data_l, data_r, data_t, data_p
+   TYPE ( SLLS_control_type ) :: control_l, control_r, control_t, control_p
+   TYPE ( SLLS_inform_type ) :: inform_l, inform_r, inform_t, inform_p
+   TYPE ( GALAHAD_userdata_type ) :: userdata
+   INTEGER :: i, il, ir, it, ip, ix, iy, j, ne, pass
+   REAL ( KIND = wp ) :: val, f, xi, b, a_11, a_12, a_22, r_1, r_2, f_best
+   REAL ( KIND = wp ) :: alpha, beta, prod1, prod2, prod3, prod4
+   REAL :: time, time_start, time_total
+   INTEGER, PARAMETER :: nl = 18
+   INTEGER, PARAMETER :: nr = 17
+   INTEGER, PARAMETER :: nt = 16
+   INTEGER, PARAMETER :: np = 15
+   INTEGER, PARAMETER :: qx = 60
+   INTEGER, PARAMETER :: qy = 60
+   INTEGER, PARAMETER :: obs = qx * qy
+   INTEGER, PARAMETER :: pass_max = 1000
+   LOGICAL, PARAMETER :: prints = .FALSE.
+!  LOGICAL, PARAMETER :: solution = .TRUE.
+   LOGICAL, PARAMETER :: solution = .FALSE.
+   REAL ( KIND = wp ), PARAMETER :: f_stop = ten ** ( - 16 )
+   REAL ( KIND = wp ), PARAMETER :: scale = ten ** 15
+   LOGICAL, PARAMETER :: pertub_observations = .TRUE.
+!  LOGICAL, PARAMETER :: pertub_observations = .FALSE.
+   INTEGER :: L_stat( nl ), R_stat( nr ), T_stat( nt ), P_stat( np )
+   REAL ( KIND = wp ) :: L( nl ), R( nr ), THETA( nt ), PHI( np )
+   REAL ( KIND = wp ) :: MU( obs ), SIGMA( obs ), RHS( qx, qy )
+   REAL ( KIND = wp ) :: A( nl, nr, nt, np, obs ), A_SAVE( obs )
+
+!  start problem data
+
+   CALL CPU_TIME( time_start )
+
+!  starting point
+   IF ( solution ) THEN
+     OPEN( unit = 23, file = "/numerical/matrices/sas/cylinder/l.txt" )
+     DO j = 1, nl
+       READ( 23, * ) L( j )
+     END DO
[…]
