### 2022-07-04 48a4859
added applications directory and sas example
