+++ src/apps/sas/sass.f90
@@ -1,1442 +1,0 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-07-04 AT 07:45 GMT
-   PROGRAM GALAHAD_SAS_EXAMPLE
-
-! use slls to solve a multilinear fitting problem
-!  min 1/2 sum_i ( xi sum_il sum_ir sum_it sum_ip a(il,ir,it,ip,i)
-!                  l_il r_ir theta_it y_phi + b - mu_i)^2,
-! where u, v, w and x lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
-! using alternating solves with u free and v,x,y fixed, and vice versa
-
-   USE GALAHAD_SLLS_double         ! double precision version
-   USE GALAHAD_SLS_double
-   USE GALAHAD_RAND_double
-   USE GALAHAD_SPACE_double
-   USE GALAHAD_NORMS_double
-   IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-   REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 20
-   TYPE ( QPT_problem_type ) :: p_l, p_r, p_t, p_p
-   TYPE ( SLLS_data_type ) :: data_l, data_r, data_t, data_p
-   TYPE ( SLLS_control_type ) :: control_l, control_r, control_t, control_p
-   TYPE ( SLLS_inform_type ) :: inform_l, inform_r, inform_t, inform_p
-   TYPE ( GALAHAD_userdata_type ) :: userdata
-   INTEGER :: i, il, ir, it, ip, ix, iy, j, k, ll, ne, status, alloc_status
-   INTEGER :: n, nl_free, nr_free, nt_free, np_free, iter, pass
-!  INTEGER :: repeat
-   REAL ( KIND = wp ) :: val, f, xi, b, dxi, db, a_11, a_12, a_22, r_1, r_2
-   REAL ( KIND = wp ) :: alpha, beta, f_best, prod1, prod2, prod3, prod4
-   REAL ( KIND = wp ) :: step, f_step, gts
-   REAL :: time, time_start, time_total
-!   INTEGER, PARAMETER :: nl = 5
-!   INTEGER, PARAMETER :: nr = 4
-!   INTEGER, PARAMETER :: nt = 3
-!   INTEGER, PARAMETER :: np = 2
-!   INTEGER, PARAMETER :: qx = 5
-!   INTEGER, PARAMETER :: qy = 5
-
-    INTEGER, PARAMETER :: nl = 18
-    INTEGER, PARAMETER :: nr = 17
-    INTEGER, PARAMETER :: nt = 16
-    INTEGER, PARAMETER :: np = 15
-    INTEGER, PARAMETER :: qx = 60
-    INTEGER, PARAMETER :: qy = 60
-
-   INTEGER, PARAMETER :: obs = qx * qy
-   INTEGER, PARAMETER :: pass_max = 1000
-   INTEGER, PARAMETER :: itref_max = 2
-   LOGICAL, PARAMETER :: printi = .TRUE.
-   LOGICAL, PARAMETER :: prints = .FALSE.
-!  LOGICAL, PARAMETER :: random_problem = .FALSE.
-   LOGICAL, PARAMETER :: random_problem = .TRUE.
-   LOGICAL, PARAMETER :: pertub_observations = .TRUE.
-!  LOGICAL, PARAMETER :: pertub_observations = .FALSE.
-   LOGICAL, PARAMETER :: solution = .TRUE.
-!  LOGICAL, PARAMETER :: solution = .FALSE.
-   LOGICAL, PARAMETER :: include_xi_and_b = .FALSE.
[…]
