+++ src/apps/sas/sase.f90
@@ -1,11 +1,11 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-07-04 AT 07:45 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-07-06 AT 07:45 GMT
    PROGRAM GALAHAD_SAS_EXAMPLE
 
 ! use slls to solve a multilinear fitting problem
 !  min 1/2 sum_i ( xi sum_il sum_ir sum_it sum_ip a(il,ir,it,ip,i)
 !                  l_il r_ir theta_it y_phi + b - mu_i)^2,
-! where u, v, w and x lie in n and m regular simplexes {z : e^T z = 1, z >= 0},
-! using alternating solves with u free and v,x,y fixed, and vice versa
+! where l, r, theta and phi lie in regular simplexes {z : e^T z = 1, z >= 0},
+! using alternating solves with l free and r, theta, phi fixed etc.
 
    USE GALAHAD_SLLS_double         ! double precision version
    USE GALAHAD_SLS_double
@@ -23,7 +23,7 @@
    TYPE ( GALAHAD_userdata_type ) :: userdata
    INTEGER :: i, il, ir, it, ip, ix, iy, j, k, ll, ne, status, alloc_status
    INTEGER :: n, nl_free, nr_free, nt_free, np_free, iter, pass
-!  INTEGER :: repeat
+   INTEGER :: repeat
    REAL ( KIND = wp ) :: val, f, xi, b, dxi, db, a_11, a_12, a_22, r_1, r_2
    REAL ( KIND = wp ) :: alpha, beta, f_best, prod1, prod2, prod3, prod4
    REAL ( KIND = wp ) :: step, f_step, gts
@@ -43,23 +43,34 @@
     INTEGER, PARAMETER :: qy = 60
 
    INTEGER, PARAMETER :: obs = qx * qy
-   INTEGER, PARAMETER :: pass_max = 1000
+   INTEGER, PARAMETER :: pass_max = 100
+!  INTEGER, PARAMETER :: pass_max = 1000
    INTEGER, PARAMETER :: itref_max = 2
+   INTEGER, PARAMETER :: repeats = 1
    LOGICAL, PARAMETER :: printi = .TRUE.
    LOGICAL, PARAMETER :: prints = .FALSE.
-!  LOGICAL, PARAMETER :: random_problem = .FALSE.
-   LOGICAL, PARAMETER :: random_problem = .TRUE.
+!  LOGICAL, PARAMETER :: random_problem = .FALSE.   ! true data
+   LOGICAL, PARAMETER :: random_problem = .TRUE.    ! invented data
    LOGICAL, PARAMETER :: pertub_observations = .TRUE.
 !  LOGICAL, PARAMETER :: pertub_observations = .FALSE.
    LOGICAL, PARAMETER :: solution = .TRUE.
 !  LOGICAL, PARAMETER :: solution = .FALSE.
-   LOGICAL, PARAMETER :: include_xi_and_b = .FALSE.
-   REAL ( KIND = wp ), PARAMETER :: f_stop = ten ** ( - 16 )
-   REAL ( KIND = wp ), PARAMETER :: scale = ten ** 15
-   REAL ( KIND = wp ), PARAMETER :: pert = 0.001_wp
-   REAL ( KIND = wp ), PARAMETER :: g_stop = ten ** ( - 10 )
-!  REAL ( KIND = wp ), PARAMETER :: small = ten ** ( - 6 )
+   LOGICAL, PARAMETER :: include_xi_and_b = .TRUE.
+!  LOGICAL, PARAMETER :: include_xi_and_b = .FALSE.
    LOGICAL, PARAMETER :: fresh_read = .FALSE.
+!  LOGICAL, PARAMETER :: fresh_read = .TRUE.
[…]
