### 2022-07-06 65abe32
improvements to slls
