+++ src/slls/slls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-06-13 AT 16:00 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-07-06 AT 08:00 GMT
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -31,12 +31,10 @@
      USE GALAHAD_STRING
      USE GALAHAD_SPACE_double
      USE GALAHAD_SORT_double, ONLY: SORT_heapsort_build,                       &
-                                    SORT_heapsort_smallest, SORT_partition,    &
-                                    SORT_quicksort
+                                    SORT_heapsort_smallest, SORT_quicksort
      USE GALAHAD_SBLS_double
      USE GALAHAD_NORMS_double
      USE GALAHAD_QPT_double
-     USE GALAHAD_QPP_double
      USE GALAHAD_QPD_double, ONLY: QPD_SIF
      USE GALAHAD_USERDATA_double
      USE GALAHAD_SPECFILE_double
@@ -172,10 +170,6 @@
 
        REAL ( KIND = wp ) :: weight = zero
 
-!  any bound larger than infinity in modulus will be regarded as infinite
-
-       REAL ( KIND = wp ) :: infinity = ten ** 19
-
 !  the required accuracy for the dual infeasibility
 
        REAL ( KIND = wp ) :: stop_d = ten ** ( - 6 )
@@ -223,7 +217,7 @@
        LOGICAL :: exact_arc_search = .TRUE.
 
 !  advance is true if an inexact exact arc_search can increase steps as well
-!   as decrease them
+!   as decrease them (currently not implemented)
 
        LOGICAL :: advance = .TRUE.
 
@@ -524,16 +518,15 @@
 !  maximum-number-of-arcsearch-steps                -1
 !  sif-file-device                                   52
 !  regularization-weight                             0.0D+0
-!  infinity-value                                    1.0D+19
 !  dual-accuracy-required                            1.0D-5
 !  complementary-slackness-accuracy-required         1.0D-5
 !  cg-relative-accuracy-required                     0.01
 !  cg-absolute-accuracy-required                     1.0D-8
-!  maximum-arcsearch-stepsize                       1.0D+20
-!  initial-arcsearch-stepsize                       1.0
-!  arcsearch-reduction-factor                       5.0D-1
-!  arcsearch-acceptance-tolerance                   1.0D-2
-!  stabilisation-weight                             0.0
+!  maximum-arcsearch-stepsize                        1.0D+20
+!  initial-arcsearch-stepsize                        1.0
+!  arcsearch-reduction-factor                        5.0D-1
+!  arcsearch-acceptance-tolerance                    1.0D-2
+!  stabilisation-weight                              0.0
 !  maximum-cpu-time-limit                            -1.0
 !  direct-subproblem-solve                           F
 !  exact-arc-search-used                             T
@@ -570,8 +563,7 @@
      INTEGER, PARAMETER :: cg_maxit = change_max + 1
      INTEGER, PARAMETER :: arcsearch_max_steps = cg_maxit + 1
      INTEGER, PARAMETER :: weight = arcsearch_max_steps + 1
-     INTEGER, PARAMETER :: infinity = weight + 1
-     INTEGER, PARAMETER :: stop_d = infinity + 1
+     INTEGER, PARAMETER :: stop_d = weight + 1
      INTEGER, PARAMETER :: stop_cg_relative = stop_d + 1
      INTEGER, PARAMETER :: stop_cg_absolute = stop_cg_relative + 1
[…]
