+++ doc/src/slls.tex
@@ -1,0 +1,1737 @@
+\documentclass{galahad}
+
+% set the package name
+
+\newcommand{\package}{slls}
+\newcommand{\packagename}{SLLS}
+\newcommand{\fullpackagename}{\libraryname\_\packagename}
+\newcommand{\solver}{{\tt \fullpackagename\_solve}}
+
+\begin{document}
+
+\galheader
+
+%%%%%%%%%%%%%%%%%%%%%% SUMMARY %%%%%%%%%%%%%%%%%%%%%%%%
+
+\galsummary
+This package uses a preconditioned, projected-gradient method
+to solve the {\bf simplex-constrained regularized linear least-squares problem}
+\disp{\mbox{minimize}\;\; q(\bmx) = \half \| \bmA \bmx - \bmb\|_2^2
++ \half \sigma \|\bmx\|^2,}
+where the variables $\bmx$ are required to lie within the
+{\bf regular simplex}
+\disp{\bme^T \bmx = 1 \tim{and} \bmx \geq 0,}
+where the $m$ by $n$ real matrix $\bmA$, the vector $\bmb$
+and the non-negative weights $\bmw$ and $\sigma$ are given, 
+where $\bme$ is the vector of ones, 
+and where the Euclidean and weighted-Euclidean norms
+are given by $\|\bmv\|_2^2 = \bmv^T \bmv$ and $\|\bmv\|_W^2 = \bmv^T \bmW \bmv$,
+respectively, with $\bmW = \mbox{diag}(\bmw)$.
+Full advantage is taken of any zero coefficients of the Jacobian matrix $\bmA$
+of the {\bf residuals} $\bmc(\bmx) = \bmA \bmx - \bmb$;
+the matrix need not be provided as there are options to obtain matrix-vector
+products involving $\bmA$ and its transpose either by reverse communication
+or from a user-provided subroutine.
+
+%%%%%%%%%%%%%%%%%%%%%% attributes %%%%%%%%%%%%%%%%%%%%%%%%
+
+\galattributes
+\galversions{\tt  \fullpackagename\_single, \fullpackagename\_double}.
+\galuses {\tt GALAHAD\_CPU\_time},
+{\tt GALAHAD\-\_SY\-M\-BOLS},
+{\tt GALAHAD\-\_\-SPACE},
+{\tt GALAHAD\_STRING},
+{\tt GALAHAD\_SORT},
+{\tt GALAHAD\_NORMS},
+{\tt GALAHAD\_CONVERT},
+{\tt GALAHAD\_SBLS},
+{\tt GALAHAD\_QPT},
+{\tt GALAHAD\_QPD},
+{\tt GALAHAD\_USERDATA}.
+{\tt GALAHAD\_SPECFILE}.
+\galdate July 2022.
+\galorigin N. I. M. Gould, STFC-Rutherford Appleton Laboratory.
+\gallanguage Fortran~95 + TR 15581 or Fortran~2003.
+
+%%%%%%%%%%%%%%%%%%%%%% HOW TO USE %%%%%%%%%%%%%%%%%%%%%%%%
+
+\galhowto
+
+%\subsection{Calling sequences}
+
+Access to the package requires a {\tt USE} statement such as
+
+\medskip\noindent{\em Single precision version}
+
+\hspace{8mm} {\tt USE \fullpackagename\_single}
+
+\medskip\noindent{\em Double precision version}
+
+\hspace{8mm} {\tt USE  \fullpackagename\_double}
+
+\medskip
+
+\noindent
+If it is required to use both modules at the same time, the derived types
+{\tt SMT\_type},
+{\tt QPT\_problem\_type},
+{\tt NLPT\_userdata\-\_type},
+{\tt \packagename\_time\_type},
[…]
