### 2022-07-12 #4291 @nimgould
I don't understand, I thought that this was being reviewed. What is going on? Nick

Get Outlook for Android<https://aka.ms/AAb9ysg>
