### 2022-07-13 #4291 @nimgould
Apologies ... I replied to Mellisa's email on the 6th of April, but my response to her request didn't show up here. Here it is again

1. I looked through the list, but no names jump out at me
(and the majority of people with an optimization preference,
seem to have python as their "go-to" language).

Of the more "mature" users that I am familiar with, I would
suggest any of (for example)

  Sven Leyffer ([leyffer@mcs.anl.gov](mailto:leyffer@mcs.anl.gov))
  Frank Curtis ([frank.e.curtis@gmail.com](mailto:frank.e.curtis@gmail.com))
  Julian Hall ([J.A.J.Hall@ed.ac.uk](mailto:J.A.J.Hall@ed.ac.uk))
  Ernesto Birgin ([egbirgin@ime.usp.br](mailto:egbirgin@ime.usp.br))

but maybe the journal doesn't approach people who aren't on the
list?

> Also - because different people have been involved in the past, I just
> want to confirm that the authorship for GALAHAD 4.0 is correct.
>

2. Both Dominique Orban and Philippe Toint were authors of GALAHAD 1.0
and contributed up to around 2.5. GALAHAD was fundamentally redesigned
for version 3.0, and more recently extended for 4.0, but neither
Dominique nor Philippe made any contributions (indeed, they have now
stated preferences for other languages, Matlab and julia, rather than
fortran and c). Of course there are legacy packages still in GALAHAD
that they contributed. So yes, the authorship reflects the work that
went into producing version 4.0.

> Finally, I have had experience with GALAHAD and know the documentation is pretty extensive, but while we wait for the reviewers would it be possible to have an entry point for the documentation? Could be a page on the wiki, or an html or PDF document that guides the user through the different documents.

3. The documents page

https://www.galahad.rl.ac.uk/doc.html

has this in mind. It provides a copy of the original design and scope
paper (that appeared in ACM TOMS back in 2004) as well as fortran and
c documentation for each of the main packages. The main "official" page

https://www.galahad.rl.ac.uk/

provides a broad outline of the scope of each package. The main README
tries to explain the installation process, and points to some of the
other READMEs for more technical details.

Is that enough? I'm not sure really what more to add.
