### 2022-07-13 8a8e374
addded complete test and documentation for slls
