+++ src/slls/slls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-07-06 AT 08:00 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-07-13 AT 14:20 GMT
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -52,8 +52,8 @@
                SLLS_inexact_arc_search, SLLS_import, SLLS_import_without_a,    &
                SLLS_solve_given_a, SLLS_solve_reverse_a_prod,                  &
                SLLS_reset_control, SLLS_information, GALAHAD_userdata_type,    &
-               QPT_problem_type, SMT_type, SMT_put, SMT_get, &
-               SLLS_project_onto_simplex, SLLS_simplex_projection_path, &
+               QPT_problem_type, SMT_type, SMT_put, SMT_get,                   &
+               SLLS_project_onto_simplex, SLLS_simplex_projection_path,        &
                SLLS_cgls
 
 
@@ -342,9 +342,11 @@
 
      TYPE :: SLLS_search_data_type
        INTEGER :: step
-       REAL ( KIND = wp ) :: ete, f_0, f_1_stop, gamma, rtr
-       REAL ( KIND = wp ) :: s_fixed, t, t_break, t_total
-       LOGICAL :: reverse, backwards
+       REAL ( KIND = wp ) :: ete, f_0, f_1_stop, gamma, rtr, xtx
+       REAL ( KIND = wp ) :: rho_0, rho_1, rho_2, s_fixed, t, t_break, t_total
+       LOGICAL :: present_a, reverse, backwards
+       INTEGER, ALLOCATABLE, DIMENSION( : ) :: NZ_out
+       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: V, P
      END TYPE SLLS_search_data_type
 !  - - - - - - - - - - - - - - -
 !   arc_search data derived type
@@ -352,7 +354,7 @@
 
      TYPE :: SLLS_subproblem_data_type
        INTEGER :: branch, n_preconditioner, step
-       REAL ( KIND = wp ) :: stop_cg, gamma, etpe
+       REAL ( KIND = wp ) :: stop_cg, gamma, gamma_old, etpe
        LOGICAL :: printp, printw, printd, printdd, debug
        LOGICAL :: present_a, present_asprod, reverse_asprod, present_afprod
        LOGICAL :: reverse_afprod, reverse_prec, present_prec, present_dprec
@@ -1547,7 +1549,7 @@
        IF ( inform%status /= GALAHAD_ok ) GO TO 910
 
        array_name = 'slls: reverse%NZ_out'
-       CALL SPACE_resize_array( prob%n, reverse%NZ_out, inform%status,         &
+       CALL SPACE_resize_array( prob%m, reverse%NZ_out, inform%status,         &
               inform%alloc_status, array_name = array_name,                    &
               deallocate_error_fatal = control%deallocate_error_fatal,         &
               exact_size = control%space_critical,                             &
@@ -1564,6 +1566,33 @@
 
        array_name = 'slls: reverse%P'
        CALL SPACE_resize_array( MAX( prob%m, prob%n ), reverse%P,              &
+              inform%status, inform%alloc_status, array_name = array_name,     &
+              deallocate_error_fatal = control%deallocate_error_fatal,         &
+              exact_size = control%space_critical,                             &
+              bad_alloc = inform%bad_alloc, out = control%error )
+       IF ( inform%status /= GALAHAD_ok ) GO TO 910
+     END IF
+
+     IF ( control%exact_arc_search .AND. PRESENT( eval_ASPROD ) .AND.          &
[…]
