+++ src/matlab/test_galahad_all.m
@@ -81,6 +81,19 @@ catch excpt
 end
 
 try
+    pkg = 'SLLS';
+    disp(sprintf('******* Testing %s *******', pkg));
+    ntests = ntests + 1;
+    test_galahad_slls
+    if inform.status ~= 0
+        pkg_failures = pkg_failures + 1;
+    end
+catch excpt
+    disp(sprintf('%d %d\n', err, pkg));
+    syntax_errors = syntax_errors + 1;
+end
+
+try
     pkg = 'DGO';
     disp(sprintf('******* Testing %s *******', pkg));
     ntests = ntests + 1;
