+++ src/matlab/galahad_slls.F90
@@ -1,54 +1,52 @@
 #include <fintrf.h>
 
-!  THIS VERSION: GALAHAD 3.3 - 12/12/2020 AT 17:05 GMT.
+!  THIS VERSION: GALAHAD 4.1 - 2022-07-13 AT 16:30 GMT.
 
 ! *-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
 !
-!                 MEX INTERFACE TO GALAHAD_BLLS
+!                 MEX INTERFACE TO GALAHAD_SLLS
 !
 ! *-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-
 !
 !  Given an m by n matrix A, an m-vector b, and a constant sigma >= 0, find
-!  a local mimimizer of the BOUND_CONSTRAINED LINER LEAST-SQUARES problem
+!  a local mimimizer of the SIMPLEX_CONSTRAINED LINER LEAST-SQUARES problem
 !    minimize 0.5 || A x - b ||^2 + 0.5 sigma ||x||^2
-!    subject to x_l <= x <= x_u
+!    subject to sum x_i = 1, x_i >= 0
 !  using a projection method.
 !  Advantage is taken of sparse A.
 !
 !  Simple usage -
 !
-!  to solve the bound-constrained quadratic program
+!  to solve the simplex_constrained liner least-squares problem
 !   [ x, inform, aux ]
-!     = galahad_blls( A, b,  x_l, x_u, control )
+!     = galahad_slls( A, b, control )
 !
 !  Sophisticated usage -
 !
 !  to initialize data and control structures prior to solution
 !   [ control ]
-!     = galahad_blls( 'initial' )
+!     = galahad_slls( 'initial' )
 !
-!  to solve the bound-constrained QP using existing data structures
+!  to solve the problem using existing data structures
 !   [ x, inform, aux ]
-!     = galahad_blls( 'existing', A, b, x_l, x_u, control )
+!     = galahad_slls( 'existing', A, b, control )
 !
 !  to remove data structures after solution
-!   galahad_blls( 'final' )
+!   galahad_slls( 'final' )
 !
 !  Usual Input -
 !    A: the m by n matrix A
 !    b: the m-vector b
-!    x_l: the n-vector x_l. The value -inf should be used for infinite bounds
-!    x_u: the n-vector x_u. The value inf should be used for infinite bounds
 !
 !  Optional Input -
 !    control, a structure containing control parameters.
 !      The components are of the form control.value, where
 !      value is the name of the corresponding component of
-!      the derived type BLLS_CONTROL as described in the
-!      manual for the fortran 90 package GALAHAD_BLLS.
+!      the derived type SLLS_CONTROL as described in the
+!      manual for the fortran 90 package GALAHAD_SLLS.
 !      In particular if the weight sigma is nonzero, it 
 !      should be passed via control.weight.
-!      See: http://galahad.rl.ac.uk/galahad-www/doc/blls.pdf
+!      See: http://galahad.rl.ac.uk/galahad-www/doc/slls.pdf
 !
 !  Usual Output -
 !   x: a global minimizer
@@ -58,16 +56,15 @@
 !   inform: a structure containing information parameters
 !      The components are of the form inform.value, where
 !      value is the name of the corresponding component of
-!      the derived type BLLS_INFORM as described in the manual for
-!      the fortran 90 package GALAHAD_BLLS.
[…]

[…]
