### 2022-07-16 ccfea1e
C interface to slls added
