+++ include/galahad_slls.h
@@ -1,9 +1,9 @@
-//* \file galahad_blls.h */
+//* \file galahad_slls.h */
 
 /*
- * THIS VERSION: GALAHAD 4.1 - 2022-05-26 AT 07:40 GMT.
+ * THIS VERSION: GALAHAD 4.1 - 2022-07-14 AT 09:40 GMT.
  *
- *-*-*-*-*-*-*-*-*-  GALAHAD_BLLS C INTERFACE  *-*-*-*-*-*-*-*-*-*-
+ *-*-*-*-*-*-*-*-*-  GALAHAD_SLLS C INTERFACE  *-*-*-*-*-*-*-*-*-*-
  *
  *  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
  *  Principal author: Jaroslav Fowkes & Nick Gould
@@ -15,29 +15,29 @@
  *   http://galahad.rl.ac.uk/galahad-www/specs.html
  */
 
-/*! \mainpage GALAHAD C package blls
+/*! \mainpage GALAHAD C package slls
 
-  \section blls_intro Introduction
+  \section slls_intro Introduction
 
-  \subsection blls_purpose Purpose
+  \subsection slls_purpose Purpose
 
   This package uses a preconditioned, projected-gradient method to solve the
    <b>simplex-constrained regularized linear least-squares problem</b>
-  \f[\mbox{minimize}\;\; q(x) = q(x) = \frac{1}{2} \| A x - b\|_2^2 + \frac{1}{2} \sigma \|x\|^2\f]
+  \f[\mbox{minimize}\;\; q(x) = \frac{1}{2} \| A x - b\|_2^2 + \frac{1}{2} \sigma \|x\|^2\f]
 \manonly
   \n
   minimize q(x) := 1/2 || A x - b ||^2 + sigma ||x||^2
   \n
 \endmanonly
-  subject to the simplex constraints
+  where \f$x\f$ is required to lie in the regular simplex
   \f[e^T x = 1 \;\;\mbox{and}\;\;  x_j \geq 0, \;\;\; j = 1, \ldots , n,\f]
 \manonly
   \n
    e^T x = 1 and x_j \[>=] 0, j = 1, ... , n,
   \n
 \endmanonly
-  where the \f$m\f$ by \f$n\f$ real matrix \f$A\f$, the vectors
-  \f$b\f$, \f$x^{l}\f$, \f$x^{u}\f$ and the non-negative weight
+  where the \f$m\f$ by \f$n\f$ real matrix \f$A\f$, the vector
+  \f$b\f$,and the non-negative weight
   \f$\sigma\f$ are given, and e is the vector of ones.
   Full advantage is taken of any zero
   coefficients of the Jacobian matrix \f$A\f$ of the <b>residuals</b>
@@ -45,23 +45,23 @@
   to obtain matrix-vector products involving \f$A\f$ and its transpose either
   by reverse communication or from a user-provided subroutine.
 
-  \subsection blls_authors Authors
+  \subsection slls_authors Authors
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
-  \subsection blls_date Originally released
+  \subsection slls_date Originally released
 
-  October 2019, C interface March 2022.
+  October 2019, C interface July 2022.
 
-  \subsection blls_terminology Terminology
+  \subsection slls_terminology Terminology
 
[…]
