+++ include/galahad_blls.h
@@ -671,7 +671,7 @@ void blls_solve_given_a( void **data,
                          const real_wp_ b[], 
                          const real_wp_ x_l[], 
                          const real_wp_ x_u[], 
-                         real_wp_ y[], 
+                         real_wp_ x[], 
                          real_wp_ z[], 
                          real_wp_ c[], 
                          real_wp_ g[], 
@@ -790,7 +790,7 @@ void blls_solve_reverse_a_prod( void **data,
                                 const real_wp_ b[], 
                                 const real_wp_ x_l[], 
                                 const real_wp_ x_u[], 
-                                real_wp_ y[], 
+                                real_wp_ x[], 
                                 real_wp_ z[], 
                                 real_wp_ c[], 
                                 real_wp_ g[], 
