+++ src/checkpoint/checkpoint.f90
@@ -1,9 +1,9 @@
-! THIS VERSION: GALAHAD 2.8 - 16/11/2015 AT 11:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-07-31 AT 10:30 GMT.
 
-!-*-*-*-*-*-  G A L A H A D    N O R M S   M O D U L E  -*-*-*-*-*-*-
+!-*-*-*-*-*-  G A L A H A D    C H E C K P O I N T   M O D U L E  -*-*-*-*-*-*-
 
 !  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
-!  Principal author: Nick Gould
+!  Principal authors: Daniel Robinson and Nick Gould
 !
 !  History -
 !   originally released GALAHAD Version 2.8. November 16th 2015
@@ -32,11 +32,10 @@
 
     CONTAINS
 
-!-*-*-*-  G A L A H A D   C H E C K P O I N T    S U B R O U T I N E   -*-*-*-
+!-*-*-*-  G A L A H A D   C H E C K P O I N T    S U B R O U T I N E   -*-*-*-*-
 
        SUBROUTINE CHECKPOINT( iter, time, measure, checkpointsIter,            &
                               checkpointsTime, low, up )
-       !SUBROUTINE CHECKPOINT( iter, measure, checkpoints, low, up )
 
 !  Find the iteration, iter, for which the criterion, measure, is
 !  first smaller than 10 ** -i, i = low, ..., up
@@ -46,7 +45,6 @@
        INTEGER, INTENT( IN ) :: iter, low, up
        REAL, INTENT( IN ) :: time
        REAL ( KIND = wp ), INTENT( IN ) :: measure
-       !INTEGER, INTENT( INOUT ), DIMENSION( low : up ) :: checkpoints
        INTEGER, INTENT( INOUT ), DIMENSION( low : up ) :: checkpointsIter
        REAL ( KIND = wp ), INTENT( INOUT ),                                    &
          DIMENSION( low : up ) :: checkpointsTime
@@ -56,12 +54,10 @@
        INTEGER :: i
 
        DO i = low, up
-         !IF ( checkpoints( i ) >= 0 ) CYCLE
          IF ( checkpointsIter( i ) >= 0 ) CYCLE
          IF ( measure <= ten ** ( - i ) ) THEN
-            !checkpoints( i ) = iter
-            checkpointsIter( i ) = iter
-            checkpointsTime( i ) = DBLE(time)
+           checkpointsIter( i ) = iter
+           checkpointsTime( i ) = REAL( time, KIND = wp )
          END IF
        END DO
 
