### 2022-08-01 5c95217
ununsed variables removed
