+++ src/lsp/lsp.f90
@@ -209,8 +209,8 @@
 !     subject to the bounds  x_l <=  x  <= x_u
 !     and constraints        c_l <= L x <= c_u ,
 !
-!  where x is a vector of n components ( x_1, .... , x_n ), A is an o by n 
-!  matrix, L is an m by n matrix, there are o obserservations b, and any of 
+!  where x is a vector of n components ( x_1, .... , x_n ), A is an o by n
+!  matrix, L is an m by n matrix, there are o obserservations b, and any of
 !  the bounds x_l, x_u, c_l, c_u may be infinite.
 !
 !  The reordered problem has the following properties:
@@ -316,14 +316,14 @@
 !                    with each the entries in each column in order of
 !                    increasing row indicies.
 !
-!    On exit, the components will most likely have been reordered. The 
+!    On exit, the components will most likely have been reordered. The
 !    output  matrix will be stored by columns, according to scheme (iii) above.
 !    However, if scheme (i) is used for input, the output A%col will contain
 !    the column numbers corresponding to the values in A%val, and thus in this
 !    case the output matrix will be available in both formats (i) and (iii).
 !
-!   B is a REAL array of length o, which must be set by the user to the value 
-!    of the observations, b. The i-th component of B, i = 1, ...., o should 
+!   B is a REAL array of length o, which must be set by the user to the value
+!    of the observations, b. The i-th component of B, i = 1, ...., o should
 !    contain the value of b_i. On exit, B will most likely have been reordered.
 !
 !   Hessian_kind is an INTEGER variable which defines the type of objective
@@ -603,8 +603,8 @@
 !          prob%n     >=  1
 !          prob%m     >=  0
 !          prob%o     >=  1
-!          prob%A%type and prob%L%type in 
-!            { 'DENSE', 'DENSE_BY_ROWS', 'DENSE_BY_COLUMNS', 
+!          prob%A%type and prob%L%type in
+!            { 'DENSE', 'DENSE_BY_ROWS', 'DENSE_BY_COLUMNS',
 !              'SPARSE_BY_ROWS', 'SPARSE_BY_COLUMNS', 'COORDINATE' }
 !       has been violated
 !
@@ -642,8 +642,7 @@
       INTEGER :: free, nonneg, lower, range, upper, nonpos, fixed, equality
       INTEGER :: l_free, l_lower, l_range, l_upper, l_equality
       INTEGER :: a_free, a_fixed
-      INTEGER :: d_free, o_free, d_nonneg, o_nonneg, d_lower, o_lower, d_range
-      INTEGER :: o_range, d_upper, o_upper, d_nonpos, o_nonpos, d_fixed, o_fixed
+      INTEGER :: d_free, d_nonneg, d_lower, d_range, d_upper, d_nonpos, d_fixed
       REAL ( KIND = wp ) :: xl, xu, cl, cu, val
       LOGICAL :: apy, apyl, apyu, apz, apzl, apzu
       CHARACTER ( LEN = 80 ) :: array_name
@@ -1239,7 +1238,7 @@
              exact_size = .TRUE. )
       IF ( inform%status /= GALAHAD_ok ) GO TO 900
 
-!  run through the rows of A to build the mapping array, and set the constant 
+!  run through the rows of A to build the mapping array, and set the constant
 !  term in the objective for rows with no free variables
 
       a_free = 0 ; a_fixed = prob%o ; prob%f = zero
@@ -1261,7 +1260,7 @@
         CALL SORT_inplace_permute( map%o, map%a_map, X = prob%DB )
 
[…]
