+++ src/cro/cro.f90
@@ -3655,12 +3655,6 @@
      TYPE ( CRO_control_type ), INTENT( IN ) :: control
      TYPE ( CRO_inform_type ), INTENT( INOUT ) :: inform
 
-!-----------------------------------------------
-!   L o c a l   V a r i a b l e s
-!-----------------------------------------------
-
-     CHARACTER ( LEN = 80 ) :: array_name
-
 !  deallocate workspace
 
      CALL CRO_terminate( data%cro_data, control, inform )
