+++ src/lsp/lspt.f90
@@ -1,32 +1,31 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-07-20 AT 15:25 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-08-02 AT 15:25 GMT.
    PROGRAM GALAHAD_LSP_EXAMPLE
-   USE GALAHAD_LSP_double                            ! double precision version
-   USE GALAHAD_LMS_double      
+   USE GALAHAD_LSP_double                    ! double precision version
    IMPLICIT NONE
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infty = 10.0_wp ** 20
+   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
    TYPE ( QPT_dimensions_type ) :: d
    TYPE ( LSP_map_type ) :: map
    TYPE ( LSP_control_type ) :: control        
    TYPE ( LSP_inform_type ) :: info
    TYPE ( QPT_problem_type ) :: p
-   TYPE ( LMS_control_type ) :: LMS_control
-   TYPE ( LMS_inform_type ) :: LMS_inform
    REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X_orig, Y_orig, Z_orig
    REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: S, Y
-   INTEGER :: n, m, h_ne, l_ne, smt_stat
+   INTEGER :: n, m, o, a_ne, l_ne, smt_stat
    INTEGER :: data_storage_type, i, j, status, iname = 0
    REAL ( KIND = wp ) :: delta
-   CHARACTER ( len = 1 ) :: st
+   CHARACTER ( len = 2 ) :: st
    CHARACTER ( len = 10 ) :: sname
+   INTEGER, PARAMETER :: coordinate = 1, sparse_by_rows = 2
+   INTEGER, PARAMETER :: sparse_by_columns = 3, dense = 4, dense_by_columns = 5
 
-!  GO TO 1
-   n = 3 ; m = 2 ; h_ne = 4 ; l_ne = 4 
-   ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
+   GO TO 1
+   n = 3 ; m = 2 ; o = 4 ; a_ne = 4 ; l_ne = 4 
+   ALLOCATE( p%B( o ), p%X_l( n ), p%X_u( n ) )
    ALLOCATE( p%C( m ), p%C_l( m ), p%C_u( m ) )
    ALLOCATE( p%X( n ), p%Y( m ), p%Z( n ), X_orig( n ) )
-   ALLOCATE( p%H%ptr( n + 1 ), p%L%ptr( m + 1 ) )
-   ALLOCATE( p%DG( n ), p%DX_l( n ), p%DX_u( n ), p%DC_l( m ), p%DC_u( m ) )
+   ALLOCATE( p%A%ptr( n + 1 ), p%L%ptr( m + 1 ) )
+   ALLOCATE( p%DB( o ), p%DX_l( n ), p%DX_u( n ), p%DC_l( m ), p%DC_u( m ) )
 
 !  ================
 !  error exit tests
@@ -35,21 +34,20 @@
    WRITE( 6, "( /, ' error exit tests ', / )" )
    st = ' '
 
-   p%Hessian_kind = - 1 ; p%gradient_kind = - 1
    p%n = n ; p%m = m ; p%f = 1.0_wp
[…]
