+++ src/qpt/qpt.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-04-03 AT 11:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-07-31 AT 09:30 GMT.
 !
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
