+++ src/lsp/lsp.f90
@@ -326,137 +326,10 @@
 !    of the observations, b. The i-th component of B, i = 1, ...., o should
 !    contain the value of b_i. On exit, B will most likely have been reordered.
 !
-!   Hessian_kind is an INTEGER variable which defines the type of objective
-!    function to be used. Possible values are
-!
-!     0  all the weights will be zero, and the analytic centre of the
-!        feasible region will be found. WEIGHT (see below) need not be set
-!
-!     1  all the weights will be one. WEIGHT (see below) need not be set
-!
-!     2  the weights will be those given by WEIGHT (see below)
-!
-!    <0  the Hessian H will be used
-!
-!   H is a structure of type SMT_type used to hold the LOWER TRIANGULAR part
-!    of H (except for the L-BFGS case). Eight storage formats are permitted:
-!
-!    i) sparse, co-ordinate
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 10 ) = TRANSFER( 'COORDINATE', H%type )
-!       H%val( : )   the values of the components of H
-!       H%row( : )   the row indices of the components of H
-!       H%col( : )   the column indices of the components of H
-!       H%ne         the number of nonzeros used to store
-!                    the LOWER TRIANGULAR part of H
-!
-!    ii) sparse, by rows
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 14 ) = TRANSFER( 'SPARSE_BY_ROWS', H%type )
-!       H%val( : )   the values of the components of H, stored row by row
-!       H%col( : )   the column indices of the components of H
-!       H%ptr( : )   pointers to the start of each row, and past the end of
-!                    the last row
-!
-!    iii) dense, by rows
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 5 ) = TRANSFER( 'DENSE', H%type )
-!       H%val( : )   the values of the components of H, stored row by row,
-!                    with each the entries in each row in order of
-!                    increasing column indicies.
-!
-!    iv) diagonal
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 8 ) = TRANSFER( 'DIAGONAL', H%type )
-!       H%val( : )   the values of the diagonals of H, stored in order
-!
-!   v) scaled identity
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 15) = 'SCALED-IDENTITY'
-!       H%val( 1 )  the value assigned to each diagonal of H
-!
-!   vi) identity
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 8 ) = 'IDENTITY'
-!
-!   vii) no Hessian
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 4 ) = 'ZERO' or 'NONE'
-!
-!   viii) L-BFGS Hessian
-!
-!       In this case, the following must be set:
-!
-!       H%type( 1 : 5 ) = 'LBFGS'
-!
-!       The Hessian in this case is available via the component H_lm below
-!
-!    On exit, the components will most likely have been reordered.
-!    The output  matrix will be stored by rows, according to scheme (ii) above,
-!    except for scheme (viii), for which a permutation will be set within H_lm.
-!    However, if scheme (i) is used for input, the output H%row will contain
-!    the row numbers corresponding to the values in H%val, and thus in this
[…]
