+++ doc/python/ugo.rst
@@ -1,0 +1,5 @@
+UGO
+===
+
+.. automodule:: galahad.ugo
+   :members:
