+++ doc/python/bgo.rst
@@ -1,0 +1,5 @@
+BGO
+===
+
+.. automodule:: galahad.bgo
+   :members:

[…]
