### 2022-08-14 514aae8
working towards automated python interfaces ... slowly
