### 2022-08-15 9519019
refinement of python install procedures
