+++ src/dgo/C/dgos3.c
@@ -13,7 +13,8 @@ int main(void) {
     struct dgo_inform_type inform;
 
     // Initialize DGO
-    dgo_initialize( &data, &control, &inform );
+    int status;
+    dgo_initialize( &data, &control, &status );
 
     // Set user-defined control options
     control.f_indexing = false; // C sparse matrix indexing (default)
@@ -24,31 +25,30 @@ int main(void) {
     int n = 3; // dimension
     int ne = 5; // Hesssian elements
     double x[] = {1.,1.,1.}; // start from one
-    double x_l[] = {-10.0,-10.0,-10.0}; 
+    double x_l[] = {-10.0,-10.0,-10.0};
     double x_u[] = {1.0,1.0,1.0};
     char H_type[] = "coordinate"; // specify co-ordinate storage
     int H_row[] = {0, 2, 1, 2, 2}; // Hessian H
     int H_col[] = {0, 0, 1, 1, 2}; // NB lower triangle
-    
+
     // Reverse-communication input/output
     int eval_status;
     double f;
     double g[n];
     double u[n], v[n];
-    double H_val[ne]; 
+    double H_val[ne];
     double freq = 10.0;
     double mag = 1000.0;
 
     // Set Hessian storage format, structure and problem bounds
-    int status;
-    dgo_import( &control, &data, &status, n, x_l, x_u, 
+    dgo_import( &control, &data, &status, n, x_l, x_u,
                 H_type, ne, H_row, H_col, NULL );
 
     // Solve the problem
     while(true){ // reverse-communication loop
 
         // Call DGO_solve
-        dgo_solve_reverse_with_mat( &data, &status, &eval_status, n, x, f, g, 
+        dgo_solve_reverse_with_mat( &data, &status, &eval_status, n, x, f, g,
                                     ne, H_val, u, v );
 
         // Evaluate f(x) and its derivatives as required
@@ -56,11 +56,11 @@ int main(void) {
             printf("DGO successful solve\n");
             break; // break while loop
         }else if(status == 2){ // obtain the objective function
-            f = pow(x[0] + x[2] + 4.0, 2) + pow(x[1] + x[2], 2) 
+            f = pow(x[0] + x[2] + 4.0, 2) + pow(x[1] + x[2], 2)
                 + mag*cos(freq*x[0]) + x[0] + x[1] + x[2];
             eval_status = 0; // record successful evaluation
         }else if(status == 3){ // obtain the gradient
-            g[0] = 2.0 * ( x[0] + x[2] + 4.0 ) 
+            g[0] = 2.0 * ( x[0] + x[2] + 4.0 )
                    - mag*freq*sin(freq*x[0]) + 1.0;;
             g[1] = 2.0 * ( x[1] + x[2] ) + 1.0;
             g[2] = 2.0 * ( x[0] + x[2] + 4.0 ) + 2.0 * ( x[1] + x[2] ) + 1.0;
@@ -73,7 +73,7 @@ int main(void) {
             H_val[4] = 4.0;
             eval_status = 0; // record successful evaluation
         }else if(status == 5){ // obtain Hessian-vector product
-            u[0] = u[0] + 2.0 * ( v[0] + v[2] ) 
+            u[0] = u[0] + 2.0 * ( v[0] + v[2] )
                     - mag*freq*freq*cos(freq*x[0]) * v[0];
[…]
