+++ src/bgo/C/bgos2.c
@@ -13,10 +13,10 @@ struct userdata_type {
 // Function prototypes
 int fun( int n, const double x[], double *f, const void * );
 int grad( int n, const double x[], double g[], const void * );
-int hessprod( int n, const double x[], double u[], const double v[], 
+int hessprod( int n, const double x[], double u[], const double v[],
               bool got_h, const void * );
-int shessprod( int n, const double x[], int nnz_v, const int index_nz_v[], 
-               const double v[], int *nnz_u, int index_nz_u[], double u[], 
+int shessprod( int n, const double x[], int nnz_v, const int index_nz_v[],
+               const double v[], int *nnz_u, int index_nz_u[], double u[],
                bool got_h, const void * );
 
 int main(void) {
@@ -27,7 +27,8 @@ int main(void) {
     struct bgo_inform_type inform;
 
     // Initialize BGO
-    bgo_initialize( &data, &control, &inform );
+    int status;
+    bgo_initialize( &data, &control, &status );
 
     // Set user-defined control options
     control.f_indexing = false; // C sparse matrix indexing (default)
@@ -45,20 +46,19 @@ int main(void) {
     int ne = 5; // Hesssian elements
     double x[] = {1.,1.,1.}; // start from one
     double infty = 1e20; // infinity
-    double x_l[] = {-infty,-infty, 0.}; 
+    double x_l[] = {-infty,-infty, 0.};
     double x_u[] = {1.1,1.1,1.1};
     char H_type[] = "absent"; // specify Hessian-vector products
 
     // Set storage
     double g[n]; // gradient
-    
+
     // Set Hessian storage format, structure and problem bounds
-    int status;
-    bgo_import( &control, &data, &status, n, x_l, x_u, 
+    bgo_import( &control, &data, &status, n, x_l, x_u,
                 H_type, ne, NULL, NULL, NULL );
 
     // Call BGO_solve
-    bgo_solve_without_mat( &data, &userdata, &status, n, x, g, 
+    bgo_solve_without_mat( &data, &userdata, &status, n, x, g,
                            fun, grad, hessprod, shessprod, NULL );
 
     // Record solution information
@@ -88,7 +88,7 @@ int main(void) {
     return 0;
 }
 
-// Objective function 
+// Objective function
 int fun(int n, const double x[], double *f, const void *userdata){
     struct userdata_type *myuserdata = (struct userdata_type *) userdata;
     double p = myuserdata->p;
@@ -101,7 +101,7 @@ int fun(int n, const double x[], double *f, const void *userdata){
 int grad(int n, const double x[], double g[], const void *userdata){
     struct userdata_type *myuserdata = (struct userdata_type *) userdata;
     double p = myuserdata->p;
-    
+
     g[0] = 2.0 * ( x[0] + x[2] + p ) - sin(x[0]);
     g[1] = 2.0 * ( x[1] + x[2] );
     g[2] = 2.0 * ( x[0] + x[2] + p ) + 2.0 * ( x[1] + x[2] );
@@ -109,7 +109,7 @@ int grad(int n, const double x[], double g[], const void *userdata){
 }
 
 // Hessian-vector product
-int hessprod(int n, const double x[], double u[], const double v[], 
+int hessprod(int n, const double x[], double u[], const double v[],
              bool got_h, const void *userdata){
     u[0] = u[0] + 2.0 * ( v[0] + v[2] ) - cos( x[0] ) * v[0];
     u[1] = u[1] + 2.0 * ( v[1] + v[2] );
@@ -118,8 +118,8 @@ int hessprod(int n, const double x[], double u[], const double v[],
 }
 
 // Sparse Hessian-vector product
[…]
