### 2022-08-17 08216c6
further generic install ideas for python interfaces
