+++ src/bgo/python/bgo_pyiface.c
@@ -1,7 +1,7 @@
 //* \file bgo_pyiface.c */
 
 /*
- * THIS VERSION: GALAHAD 4.0 - 2022-03-13 AT 11:30 GMT.
+ * THIS VERSION: GALAHAD 4.1 - 2022-08-17 AT 14:30 GMT.
  *
  *-*-*-*-*-*-*-*-*-  GALAHAD_BGO PYTHON INTERFACE  *-*-*-*-*-*-*-*-*-*-
  *
@@ -9,15 +9,14 @@
  *  Principal author: Jaroslav Fowkes & Nick Gould
  *
  *  History -
- *   originally released GALAHAD Version 3.3. July 27th 2021
+ *   originally released GALAHAD Version 4.1. August 17th 2022
  *
  *  For full documentation, see
  *   http://galahad.rl.ac.uk/galahad-www/specs.html
  */
+
 #include "galahad_python.h"
-
 #include "galahad_bgo.h"
-
 
 /* Nested UGO control and inform prototypes */
 bool ugo_update_control(struct ugo_control_type *control, PyObject *py_options);
@@ -71,7 +70,8 @@ static int eval_g(int n, const double x[], double g[], const void *userdata){
 
     // Wrap input array as NumPy array
     npy_intp xdim[] = {n};
-    PyArrayObject *py_x = (PyArrayObject*) PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+    PyArrayObject *py_x = (PyArrayObject*) 
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
 
     // Build Python argument list
     PyObject *arglist = Py_BuildValue("(O)", py_x);
@@ -105,8 +105,9 @@ static int eval_g(int n, const double x[], double g[], const void *userdata){
 
 //  *-*-*-*-*-*-*-*-*-*-   UPDATE CONTROL    -*-*-*-*-*-*-*-*-*-*
 
-/* Update the control options: use C defaults but update any passed via Python */
-static bool bgo_update_control(struct bgo_control_type *control, PyObject *py_options){
+/* Update the control options: use C defaults but update any passed via Python*/
+static bool bgo_update_control(struct bgo_control_type *control, 
+                               PyObject *py_options){
 
     // Use C defaults if Python options not passed
     if(!py_options) return true;
@@ -132,7 +133,8 @@ static bool bgo_update_control(struct bgo_control_type *control, PyObject *py_op
 
         // Parse each float/double option
         if(strcmp(key_name, "obj_unbounded") == 0){
-            if(!parse_double_option(value, "obj_unbounded", &control->obj_unbounded))
+            if(!parse_double_option(value, "obj_unbounded", 
+                                    &control->obj_unbounded))
                 return false;
             continue;
         }
@@ -140,7 +142,8 @@ static bool bgo_update_control(struct bgo_control_type *control, PyObject *py_op
 
         // Parse each bool option
         if(strcmp(key_name, "space_critical") == 0){
-            if(!parse_bool_option(value, "space_critical", &control->space_critical))
+            if(!parse_bool_option(value, "space_critical", 
+                                  &control->space_critical))
                 return false;
             continue;
         }
@@ -148,7 +151,8 @@ static bool bgo_update_control(struct bgo_control_type *control, PyObject *py_op
 
         // Parse each char option
         if(strcmp(key_name, "prefix") == 0){
-            if(!parse_char_option(value, "prefix", control->prefix))
+            if(!parse_char_option(value, "prefix", 
[…]
