### 2022-08-18 #4291 @nimgould
Hi Melissa, that is disappointing. I already looked through your speadsheet, and there seems to be a vacuum of people whose interests lie in optimization and who program in fortran.

The main journal in the general area of software for optimization is Mathematical Programming Computation. https://www.springer.com/journal/12532 I looked through the list of associate and technical editors, and aside for those who I already suggested, the following might also be helpful -

Andreas Waechter, Northwestern University,
 Hande Y. Benson, Drexel University,
Stefan Wild, Argonne National Laboratory,
Elizabeth Wong, University of California San Diego,

And while I don't suggest this seriously, the most knowledgeable person would be
Dominique Orban, Ecole Polytechnique de Montreal
who was a previous author of GALAHAD, and certainly has the Fortran/C skills
