### 2022-08-18 e49b55b
more work on generic python installs
