+++ src/sec/C/sec_ciface.f90
@@ -60,12 +60,12 @@
 
 !  copy C control parameters to fortran
 
-    SUBROUTINE copy_control_in( ccontrol, fcontrol, f_indexing ) 
+    SUBROUTINE copy_control_in( ccontrol, fcontrol, f_indexing )
     TYPE ( sec_control_type ), INTENT( IN ) :: ccontrol
     TYPE ( f_sec_control_type ), INTENT( OUT ) :: fcontrol
     LOGICAL, optional, INTENT( OUT ) :: f_indexing
     INTEGER :: i
-    
+
     ! C or Fortran sparse matrix indexing
     IF ( PRESENT( f_indexing ) ) f_indexing = ccontrol%f_indexing
 
@@ -89,12 +89,12 @@
 
 !  copy fortran control parameters to C
 
-    SUBROUTINE copy_control_out( fcontrol, ccontrol, f_indexing ) 
+    SUBROUTINE copy_control_out( fcontrol, ccontrol, f_indexing )
     TYPE ( f_sec_control_type ), INTENT( IN ) :: fcontrol
     TYPE ( sec_control_type ), INTENT( OUT ) :: ccontrol
     LOGICAL, OPTIONAL, INTENT( IN ) :: f_indexing
     INTEGER :: i, l
-    
+
     ! C or Fortran sparse matrix indexing
     IF ( PRESENT( f_indexing ) ) ccontrol%f_indexing = f_indexing
 
@@ -119,10 +119,9 @@
 
 !  copy C inform parameters to fortran
 
-    SUBROUTINE copy_inform_in( cinform, finform ) 
+    SUBROUTINE copy_inform_in( cinform, finform )
     TYPE ( sec_inform_type ), INTENT( IN ) :: cinform
     TYPE ( f_sec_inform_type ), INTENT( OUT ) :: finform
-    INTEGER :: i
 
     ! Integers
     finform%status = cinform%status
@@ -132,10 +131,9 @@
 
 !  copy fortran inform parameters to C
 
-    SUBROUTINE copy_inform_out( finform, cinform ) 
+    SUBROUTINE copy_inform_out( finform, cinform )
     TYPE ( f_sec_inform_type ), INTENT( IN ) :: finform
     TYPE ( sec_inform_type ), INTENT( OUT ) :: cinform
-    INTEGER :: i, l
 
     ! Integers
     cinform%status = finform%status
@@ -149,7 +147,7 @@
 !  C interface to fortran sec_initialize
 !  -------------------------------------
 
-  SUBROUTINE sec_initialize( cdata, ccontrol, status ) BIND( C ) 
+  SUBROUTINE sec_initialize( cdata, ccontrol, status ) BIND( C )
   USE GALAHAD_SEC_double_ciface
   IMPLICIT NONE
 
@@ -164,7 +162,7 @@
 ! TYPE ( f_sec_full_data_type ), POINTER :: fdata
   TYPE ( f_sec_control_type ) :: fcontrol
   TYPE ( f_sec_inform_type ) :: finform
-  LOGICAL :: f_indexing 
+  LOGICAL :: f_indexing
 
 !  allocate fdata
 
@@ -180,7 +178,7 @@
   f_indexing = .FALSE.
 ! fdata%f_indexing = f_indexing
 
-!  copy control out 
+!  copy control out
 
   CALL copy_control_out( fcontrol, ccontrol, f_indexing )
   RETURN
@@ -217,11 +215,11 @@
 !  copy control in
 
   CALL copy_control_in( ccontrol, fcontrol, f_indexing )
-  
+
 !  open specfile for reading
 
   OPEN( UNIT = device, FILE = fspecfile )
-  
+
[…]
