+++ src/bgo/python/bgo_pyiface.c
@@ -34,6 +34,8 @@ static int status = 0;                   // exit status
 /* Python eval_* function pointers */
 static PyObject *py_eval_f = NULL;
 static PyObject *py_eval_g = NULL;
+static PyObject *py_eval_h = NULL;
+static PyObject *py_eval_hprod = NULL;
 
 /* C eval_* function wrappers */
 static int eval_f(int n, const double x[], double *f, const void *userdata){
@@ -103,6 +105,86 @@ static int eval_g(int n, const double x[], double g[], const void *userdata){
     return 0;
 }
 
+static int eval_h(int n, int ne, const double x[], double hval[], const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyArrayObject *py_x = (PyArrayObject*)
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+
+    // Build Python argument list
+    PyObject *arglist = Py_BuildValue("(O)", py_x);
+
+    // Call Python eval_h
+    PyObject *result = PyObject_CallObject(py_eval_h, arglist);
+    Py_DECREF(py_x);    // Free py_x memory
+    Py_DECREF(arglist); // Free arglist memory
+
+    // Check that eval was successful
+    if(!result)
+        return -1;
+
+    // Check return value is of correct type, size, and shape
+    if(!check_array_double("eval_h return value", (PyArrayObject*) result, ne)){
+        Py_DECREF(result); // Free result memory
+        return -1;
+    }
+
+    // Get return value data pointer and copy data into hval
+    const double *val = (double *) PyArray_DATA((PyArrayObject*) result);
+    for(int i=0; i<ne; i++) {
+        hval[i] = val[i];
+    }
+
+    // Free result memory
+    Py_DECREF(result);
+
+    return 0;
+}
+
+static int eval_hprod(int n, const double x[], double u[], const double v[], bool goth, const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyArrayObject *py_x = (PyArrayObject*)
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+    PyArrayObject *py_u = (PyArrayObject*)
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) u);
+    PyArrayObject *py_v = (PyArrayObject*)
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) v);
+
+    // Build Python argument list
+    PyObject *arglist = Py_BuildValue("(OOO)", py_x, py_u, py_v);
+
+    // Call Python eval_hprod
+    PyObject *result = PyObject_CallObject(py_eval_hprod, arglist);
+    Py_DECREF(py_x);    // Free py_x memory
+    Py_DECREF(py_u);    // Free py_u memory
+    Py_DECREF(py_v);    // Free py_v memory
+    Py_DECREF(arglist); // Free arglist memory
+
+    // Check that eval was successful
+    if(!result)
+        return -1;
+
+    // Check return value is of correct type, size, and shape
+    if(!check_array_double("eval_hprod return value", (PyArrayObject*) result, n)){
+        Py_DECREF(result); // Free result memory
+        return -1;
+    }
+
+    // Get return value data pointer and copy data into u
+    const double *uval = (double *) PyArray_DATA((PyArrayObject*) result);
+    for(int i=0; i<n; i++) {
+        u[i] = uval[i];
+    }
+
+    // Free result memory
+    Py_DECREF(result);
+
+    return 0;
+}
+
[…]
