### 2022-08-24 0a5f6e0
simplify 32bit integer array conversion in bgo python interface
