+++ src/bgo/python/bgo_pyiface.c
@@ -207,49 +207,49 @@ static bool bgo_update_control(struct bgo_control_type *control,
 
         // Parse each int option
         if(strcmp(key_name, "error") == 0){
-            if(!parse_int_option(value, "error", 
+            if(!parse_int_option(value, "error",
                                  &control->error))
                 return false;
             continue;
         }
         if(strcmp(key_name, "out") == 0){
-            if(!parse_int_option(value, "out", 
+            if(!parse_int_option(value, "out",
                                  &control->out))
                 return false;
             continue;
         }
         if(strcmp(key_name, "print_level") == 0){
-            if(!parse_int_option(value, "print_level", 
+            if(!parse_int_option(value, "print_level",
                                  &control->print_level))
                 return false;
             continue;
         }
         if(strcmp(key_name, "attempts_max") == 0){
-            if(!parse_int_option(value, "attempts_max", 
+            if(!parse_int_option(value, "attempts_max",
                                  &control->attempts_max))
                 return false;
             continue;
         }
         if(strcmp(key_name, "max_evals") == 0){
-            if(!parse_int_option(value, "max_evals", 
+            if(!parse_int_option(value, "max_evals",
                                  &control->max_evals))
                 return false;
             continue;
         }
         if(strcmp(key_name, "sampling_strategy") == 0){
-            if(!parse_int_option(value, "sampling_strategy", 
+            if(!parse_int_option(value, "sampling_strategy",
                                  &control->sampling_strategy))
                 return false;
             continue;
         }
         if(strcmp(key_name, "hypercube_discretization") == 0){
-            if(!parse_int_option(value, "hypercube_discretization", 
+            if(!parse_int_option(value, "hypercube_discretization",
                                  &control->hypercube_discretization))
                 return false;
             continue;
         }
         if(strcmp(key_name, "alive_unit") == 0){
-            if(!parse_int_option(value, "alive_unit", 
+            if(!parse_int_option(value, "alive_unit",
                                  &control->alive_unit))
                 return false;
             continue;
@@ -377,22 +377,22 @@ static PyObject* bgo_make_time_dict(const struct bgo_time_type *time){
 static PyObject* bgo_make_inform_dict(const struct bgo_inform_type *inform){
     PyObject *py_inform = PyDict_New();
 
-    // Set int inform entries      
-    PyDict_SetItemString(py_inform, "status", 
+    // Set int inform entries
+    PyDict_SetItemString(py_inform, "status",
                          PyLong_FromLong(inform->status));
-    PyDict_SetItemString(py_inform, "alloc_status", 
+    PyDict_SetItemString(py_inform, "alloc_status",
                          PyLong_FromLong(inform->alloc_status));
-    PyDict_SetItemString(py_inform, "f_eval", 
+    PyDict_SetItemString(py_inform, "f_eval",
                          PyLong_FromLong(inform->f_eval));
[…]
