### 2022-08-25 53fc5ba
updated install for python possibilities
