+++ src/bgo/python/test_bgo.py
@@ -8,18 +8,20 @@ bgo.initialize()
 options = {'print_level' : 3, 'ugo_options' : {'print_level' : 4}}
 print(options)
 
-# set parameter
+# set parameters
 p = 4
+freq = 10
+mag = 1000
 # set bounds
 n = 3
-x_l = np.array([-np.inf,-np.inf,0.])
-x_u = np.array([1.,1.,1.])
+x_l = np.array([-10.0,-10.0,-10.0])
+x_u = np.array([0.5,0.5,0.5])
 
 # set Hessian sparsity
 H_type = 'coordinate'
 ne = 5
-H_row = np.array([0,2,1,2,2])
-H_col = np.array([0,0,1,1,2])
+H_row = np.array([0,1,2,2,2])
+H_col = np.array([0,1,0,1,2])
 H_ptr = None
 
 # load data (and optionally non-default options)
@@ -29,21 +31,24 @@ bgo.load(n, x_l, x_u, H_type, ne, H_row, H_col, H_ptr, options=options)
 # NB python functions have access to external variables
 # So no need for userdata like in C or Fortran
 def eval_f(x):
-    return (x[0] + x[2] + p)**2 + (x[1] + x[2])**2 + np.cos(x[0]);
+    return (x[0] + x[2] + p)**2 + (x[1] + x[2])**2 + mag * np.cos(freq * x[0]) + x[0] + x[1] + x[2]
 def eval_g(x):
-    return np.array([2. * ( x[0] + x[2] + p ) - np.sin(x[0]),
-                     2. * ( x[1] + x[2] ),
-                     2. * ( x[0] + x[2] + p ) + 2.0 * ( x[1] + x[2] )])
+    return np.array([2. * ( x[0] + x[2] + p ) 
+                     - mag * freq * np.sin(freq * x[0]) + 1.,
+                     2. * ( x[1] + x[2] ) + 1.,
+                     2. * ( x[0] + x[2] + p ) + 2.0 * ( x[1] + x[2] ) + 1.])
 def eval_h(x):
-    return np.array([2.0 - np.cos(x[0]),2.,2.,2.,2.])
+    return np.array([2. - mag * freq * freq * np.cos(freq * x[0]),2.,2.,2.,2.])
 
 def eval_hprod(x,u,v):
-    return np.array([u[0] + 2.0 * ( v[0] + v[2] ) - np.cos(x[0]) * v[0],
-                     u[1] + 2.0 * ( v[1] + v[2]),
-                     u[2] + 2.0 * ( v[0] + v[1] + 2.0 * v[2] )])
+    return np.array([u[0] + 2. * ( v[0] + v[2] ) 
+                     - mag * freq * freq * np.cos(freq * x[0]) * v[0],
+                     u[1] + 2. * ( v[1] + v[2]),
+                     u[2] + 2. * ( v[0] + v[1] + 2. * v[2] )])
 
-# starting value
-x = np.array([1.,1.,1.])
+# set starting point
+
+x = np.array([0.,0.,0.])
 
 # starting gradient
 g = eval_g(x)
