### 2022-08-25 f5e5cde
fix for dummy hsl_ma57
