+++ src/dum/hsl_ma57d.f90
@@ -151,9 +151,9 @@ module hsl_ma57_double
       module procedure ma57_part_solve1,ma57_part_solve2
    end interface
 
-   interface ma57_get_n__
-      module procedure ma57_get_n_double
-   end interface ma57_get_n__
+!  interface ma57_get_n__
+!     module procedure ma57_get_n_double
+!  end interface ma57_get_n__
 
 contains
 
