+++ src/bgo/python/bgo_pyiface.c
@@ -931,7 +931,7 @@ PyDoc_STRVAR(bgo_module_doc,
 "matrix storage\n"
 "--------------\n"
 "\n"
-"The symmetric :math:`n` by :math:`n` matrix :math:`H = \nabla_{xx}f` may\n"
+"The symmetric :math:`n` by :math:`n` matrix :math:`H = \\nabla_{xx}f` may\n"
 "be presented and stored in a variety of formats. But crucially symmetry \n"
 "is exploited by only storing values from the lower triangular part \n"
 "(i.e, those entries that lie on or below the leading diagonal).\n"
@@ -974,7 +974,7 @@ PyDoc_STRVAR(bgo_module_doc,
 "call order\n"
 "----------\n"
 "The functions should be called in the following order, with\n"
-"[] indicating an optional call\n" 
+"[] indicating an optional call\n"
 "\n"
 "  ``bgo.initialize``\n"
 "\n"
