+++ src/ugo/python/ugo_pyiface.c
@@ -81,31 +81,31 @@ bool ugo_update_control(struct ugo_control_type *control, PyObject *py_options){
 
         // Parse each int option
         if(strcmp(key_name, "error") == 0){
-            if(!parse_int_option(value, "error", 
+            if(!parse_int_option(value, "error",
                                  &control->error))
                 return false;
             continue;
         }
         if(strcmp(key_name, "out") == 0){
-            if(!parse_int_option(value, "out", 
+            if(!parse_int_option(value, "out",
                                  &control->out))
                 return false;
             continue;
         }
         if(strcmp(key_name, "print_level") == 0){
-            if(!parse_int_option(value, "print_level", 
+            if(!parse_int_option(value, "print_level",
                                  &control->print_level))
                 return false;
             continue;
         }
         if(strcmp(key_name, "start_print") == 0){
-            if(!parse_int_option(value, "start_print", 
+            if(!parse_int_option(value, "start_print",
                                  &control->start_print))
                 return false;
             continue;
         }
         if(strcmp(key_name, "stop_print") == 0){
-            if(!parse_int_option(value, "stop_print", 
+            if(!parse_int_option(value, "stop_print",
                                  &control->stop_print))
                 return false;
             continue;
@@ -117,49 +117,49 @@ bool ugo_update_control(struct ugo_control_type *control, PyObject *py_options){
             continue;
         }
         if(strcmp(key_name, "maxit") == 0){
-            if(!parse_int_option(value, "maxit", 
+            if(!parse_int_option(value, "maxit",
                                  &control->maxit))
                 return false;
             continue;
         }
         if(strcmp(key_name, "initial_points") == 0){
-            if(!parse_int_option(value, "initial_points", 
+            if(!parse_int_option(value, "initial_points",
                                  &control->initial_points))
                 return false;
             continue;
         }
         if(strcmp(key_name, "storage_increment") == 0){
-            if(!parse_int_option(value, "storage_increment", 
+            if(!parse_int_option(value, "storage_increment",
                                  &control->storage_increment))
                 return false;
             continue;
         }
         if(strcmp(key_name, "buffer") == 0){
-            if(!parse_int_option(value, "buffer", 
+            if(!parse_int_option(value, "buffer",
                                  &control->buffer))
                 return false;
             continue;
         }
         if(strcmp(key_name, "lipschitz_estimate_used") == 0){
-            if(!parse_int_option(value, "lipschitz_estimate_used", 
+            if(!parse_int_option(value, "lipschitz_estimate_used",
                                  &control->lipschitz_estimate_used))
                 return false;
             continue;
         }
         if(strcmp(key_name, "next_interval_selection") == 0){
-            if(!parse_int_option(value, "next_interval_selection", 
+            if(!parse_int_option(value, "next_interval_selection",
                                  &control->next_interval_selection))
                 return false;
             continue;
         }
         if(strcmp(key_name, "refine_with_newton") == 0){
-            if(!parse_int_option(value, "refine_with_newton", 
+            if(!parse_int_option(value, "refine_with_newton",
                                  &control->refine_with_newton))
                 return false;
             continue;
         }
         if(strcmp(key_name, "alive_unit") == 0){
-            if(!parse_int_option(value, "alive_unit", 
[…]
