+++ src/matlab/galahad_matlab.F90
@@ -287,8 +287,10 @@
 !  ------------------------------------------------------
 
       mwPointer :: mxCreateStructMatrix
+      INTEGER * 4  :: ninform_i4
 
-      pr = mxCreateStructMatrix( 1_mws_, 1_mws_, ninform, finform )
+      ninform_i4 = ninform
+      pr = mxCreateStructMatrix( 1_mws_, 1_mws_, ninform_i4, finform )
       CALL mxSetField( struct, 1_mwi_, name, pr )
 
       RETURN
@@ -975,19 +977,8 @@
       mwSize :: nn
       INTEGER * 4 :: Y_array( 1 )
 
-!     CALL mexWarnMsgTxt( ' 4 ' )
-
       nn = 1
-      SELECT CASE ( di_ )
-      CASE ( int8_ )
-        CALL mxCopyPtrToInteger8( px, Y_array, nn )
-      CASE ( int2_ )
-        CALL mxCopyPtrToInteger2( px, Y_array, nn )
-      CASE ( int1_ )
-        CALL mxCopyPtrToInteger1( px, Y_array, nn )
-      CASE default
-        CALL mxCopyPtrToInteger4( px, Y_array, nn )
-      END SELECT
+      CALL mxCopyPtrToInteger4( px, Y_array, nn )
       y = Y_array( 1 )
 
       RETURN
@@ -1013,19 +1004,8 @@
       mwSize :: nn
       INTEGER * 8 :: Y_array( 1 )
 
-!     CALL mexWarnMsgTxt( ' 4 ' )
-
       nn = 1
-      SELECT CASE ( di_ )
-      CASE ( int8_ )
-        CALL mxCopyPtrToInteger8( px, Y_array, nn )
-      CASE ( int2_ )
-        CALL mxCopyPtrToInteger2( px, Y_array, nn )
-      CASE ( int1_ )
-        CALL mxCopyPtrToInteger1( px, Y_array, nn )
-      CASE default
-        CALL mxCopyPtrToInteger4( px, Y_array, nn )
-      END SELECT
+      CALL mxCopyPtrToInteger8( px, Y_array, nn )
       y = Y_array( 1 )
 
       RETURN
@@ -1033,11 +1013,10 @@
 
 !  -*-*-*-*-*-*- g a l m x  C o p y  P t r  T o  I n t e g e r 4 4 -*-*-*-*-*-
 
-      SUBROUTINE galmxCopyPtrToInteger44( px, Y, n, sparse )
+      SUBROUTINE galmxCopyPtrToInteger44( px, Y, n )
       mwPointer :: px
       INTEGER * 4 :: n
       INTEGER * 4, DIMENSION( n ) :: Y
-      LOGICAL, OPTIONAL :: sparse
 
 !  --------------------------------------------------------------
 
@@ -1055,58 +1034,28 @@
       INTEGER :: alloc_stat
       mwIndex, DIMENSION(:), ALLOCATABLE :: temp_mwi
 
-!     CHARACTER ( len = 80 ) :: debug = REPEAT( ' ', 80 )
-!     CALL mexWarnMsgTxt( ' 44 ' )
-
       nn = n
-      IF ( PRESENT( sparse ) ) THEN
-        SELECT CASE( mwi_ )
-        CASE( kind( Y ) )
-          CALL mxCopyPtrToInteger4( px, Y, nn )
-        CASE default
-          ALLOCATE( temp_mwi( n ), STAT = alloc_stat )
-          IF ( alloc_stat /= 0 ) CALL mexErrMsgTxt( ' allocation failure ' )
-          SELECT CASE ( mwi_ )
-          CASE ( int8_ )
[…]
