+++ src/lsrt/lsrt.f90
@@ -316,7 +316,7 @@
 
       END SUBROUTINE LSRT_initialize
 
-! G A L A H A D - L S R T _ F U L L _ I N I T I A L I Z E   S U B R O U T I N E 
+! G A L A H A D - L S R T _ F U L L _ I N I T I A L I Z E   S U B R O U T I N E
 
      SUBROUTINE LSRT_full_initialize( data, control, inform )
 
@@ -1570,12 +1570,6 @@
      TYPE ( LSRT_control_type ), INTENT( IN ) :: control
      TYPE ( LSRT_inform_type ), INTENT( INOUT ) :: inform
 
-!-----------------------------------------------
-!   L o c a l   V a r i a b l e s
-!-----------------------------------------------
-
-     CHARACTER ( LEN = 80 ) :: array_name
-
 !  deallocate workspace
 
      CALL LSRT_terminate( data%lsrt_data, control, inform )
@@ -1907,13 +1901,13 @@
 ! =============================================================================
 ! -----------------------------------------------------------------------------
 
-!-  G A L A H A D -  L S R T _ i m p o r t _ c o n t r o l  S U B R O U T I N E 
+!-  G A L A H A D -  L S R T _ i m p o r t _ c o n t r o l  S U B R O U T I N E
 
      SUBROUTINE LSRT_import_control( control, data, status )
 
 !  import control parameters prior to solution. Arguments are as follows:
 
-!  control is a derived type whose components are described in the leading 
+!  control is a derived type whose components are described in the leading
 !   comments to LSRT_solve
 !
 !  data is a scalar variable of type LSRT_full_data_type used for internal data
@@ -1953,20 +1947,20 @@
 !  status is a scalar variable of type default intege that indicates the
 !   success or otherwise of the solve.
 !
-!   This must be set to 
+!   This must be set to
 !      1 on initial entry. Set U (below) to b for this entry
 !
 !   Possible exit values are:
 !      0 the solution has been found
-!      2 The user must perform the operation u := u + A v, and re-enter the 
-!        subroutine. The vectors u and v are available in the arrays U and V 
+!      2 The user must perform the operation u := u + A v, and re-enter the
+!        subroutine. The vectors u and v are available in the arrays U and V
 !        (below) respectively, and the result u must overwrite the content of
-!        U. No argument except U should be altered before re-entering the 
+!        U. No argument except U should be altered before re-entering the
 !        subroutine
-!      3 The user must perform the operation v := v + A^T u, and re-enter the 
-!        subroutine. The vectors u and v are available in the arrays U and V 
+!      3 The user must perform the operation v := v + A^T u, and re-enter the
+!        subroutine. The vectors u and v are available in the arrays U and V
 !        (below) respectively, and the result v must overwrite the content of
-!        V. No argument except V should be altered before re-entering the 
+!        V. No argument except V should be altered before re-entering the
 !        subroutine
 !      4 The user must reset U (below) to b, and re-eneter the subroutine
 !        without changing the other arguments
@@ -1976,16 +1970,16 @@
 !    -18 the iteration limit has been exceeded
 !    -25 status is not between 1 and 5 on entry
 !
-!  m is a scalar variable of type default intege that holds the number 
+!  m is a scalar variable of type default intege that holds the number
 !   of equations (i.e., rows of A) . m must be positive
 !
-!  n is a scalar variable of type default intege that holds the number 
+!  n is a scalar variable of type default intege that holds the number
 !   of variables (i.e., columns of A) . n must be positive
 !
-!  power is a scalar of type default real, that holds the positive value 
[…]

[…]
