### 2022-09-09 #4291 @frankecurtis
@melissawm Following up here on a couple of emails that I sent to you.  I would be grateful for a pointer to some guidance on reviewing for this journal, since I'm unfamiliar with the process.  Are there particular review questions that I should be answering, or should I simply comment on anything that I find relevant?  I see an earlier comment with a paper (a few pages).  I suppose I should be reviewing that in addition to the code?  Also, is there a particular deadline?  I should be able to manage with any deadline, but having a deadline would help me prioritize my time and set expectations relative to other reviewers as well.  Guidance would be greatly appreciated.
