### 2022-09-10 #4291 @arfon
:wave: @frankecurtis – EiC of JOSS here. You can read more about our reviewing process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
