### 2022-10-04 #4291 @melissawm
Welcome @jajhall - please review the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for JOSS and let us know if you have further questions. Thank you!
