### 2022-10-13 c16a104
improve python documentation
