+++ src/bgo/Python/bgo_pyiface.c
@@ -1,7 +1,7 @@
 //* \file bgo_pyiface.c */
 
 /*
- * THIS VERSION: GALAHAD 4.1 - 2022-08-17 AT 14:30 GMT.
+ * THIS VERSION: GALAHAD 4.1 - 2022-10-13 AT 11:10 GMT.
  *
  *-*-*-*-*-*-*-*-*-  GALAHAD_BGO PYTHON INTERFACE  *-*-*-*-*-*-*-*-*-*-
  *
@@ -379,95 +379,6 @@ static PyObject* bgo_make_inform_dict(const struct bgo_inform_type *inform){
 
 //  *-*-*-*-*-*-*-*-*-*-   BGO_INITIALIZE    -*-*-*-*-*-*-*-*-*-*
 
-PyDoc_STRVAR(py_bgo_initialize_doc,
-"bgo.initialize()\n"
-"\n"
-"Set default option values and initialize private data\n"
-"\n"
-"Returns\n"
-"-------\n"
-"options : dict\n"
-"    dictionary containing default control options:\n"
-"     error : int\n"
-"       error and warning diagnostics occur on stream error.\n"
-"     out : int\n"
-"       general output occurs on stream out.\n"
-"     print_level : int\n"
-"       the level of output required. Possible values are:\n"
-"\n"
-"       * <= 0\n"
-"\n"
-"         no output\n"
-"       * 1\n"
-"\n"
-"         a one-line summary for every improvement\n"
-"       * 2\n"
-"\n"
-"         a summary of each iteration\n"
-"       * >= 3\n"
-"\n"
-"         increasingly verbose (debugging) output.\n"
-"     attempts_max : int\n"
-"       the maximum number of random searches from the best point\n"
-"       found so far.\n"
-"     max_evals : int\n"
-"       the maximum number of function evaluations made.\n"
-"     sampling_strategy : int\n"
-"       sampling strategy used. Possible values are\n"
-"\n"
-"       * 1\n"
-"\n"
-"         uniformly spread\n"
-"       * 2\n"
-"\n"
-"         Latin hypercube sampling\n"
-"       * 3\n"
-"\n"
-"         uniformly spread within a Latin hypercube.\n"
-"     hypercube_discretization : int\n"
-"       hyper-cube discretization (for sampling stategies 2 and 3).\n"
-"     alive_unit : int\n"
-"       removal of the file alive_file from unit alive_unit\n"
-"       terminates execution.\n"
-"     alive_file : str\n"
-"       see alive_unit.\n"
-"     infinity : float\n"
-"       any bound larger than infinity in modulus will be regarded as\n"
-"       infinite.\n"
-"     obj_unbounded : float\n"
-"       the smallest value the objective function may take before the\n"
-"       problem is marked as unbounded.\n"
-"     cpu_time_limit : float\n"
-"       the maximum CPU time allowed (-ve means infinite).\n"
-"     clock_time_limit : float\n"
-"       the maximum elapsed clock time allowed (-ve means infinite).\n"
-"     random_multistart : bool\n"
-"       perform random-multistart as opposed to local minimize and\n"
-"       probe.\n"
-"     hessian_available : bool\n"
-"       is the Hessian matrix of second derivatives available or is\n"
-"       access only via matrix-vector products?.\n"
-"     space_critical : bool\n"
-"       if ``space_critical`` is True, every effort will be made to\n"
-"       use as little space as possible. This may result in longer\n"
-"       computation time.\n"
-"     deallocate_error_fatal : bool\n"
-"       if ``deallocate_error_fatal`` is True, any array/pointer\n"
-"       deallocation error will terminate execution. Otherwise,\n"
-"       computation will continue.\n"
[…]
