+++ src/ugo/Python/ugo_pyiface.c
@@ -1,7 +1,7 @@
 //* \file ugo_pyiface.c */
 
 /*
- * THIS VERSION: GALAHAD 4.1 - 2022-08-28 AT 09:35 GMT.
+ * THIS VERSION: GALAHAD 4.1 - 2022-10-13 AT 10:30 GMT.
  *
  *-*-*-*-*-*-*-*-*-  GALAHAD_UGO PYTHON INTERFACE  *-*-*-*-*-*-*-*-*-*-
  *
@@ -308,129 +308,6 @@ PyObject* ugo_make_inform_dict(const struct ugo_inform_type *inform){
 
 //  *-*-*-*-*-*-*-*-*-*-   UGO_INITIALIZE    -*-*-*-*-*-*-*-*-*-*
 
-PyDoc_STRVAR(py_ugo_initialize_doc,
-"ugo.initialize()\n"
-"\n"
-"Set default option values and initialize private data.\n"
-"\n"
-"Returns\n"
-"-------\n"
-"options : dict\n"
-"    dictionary containing default control options:\n"
-"     error : int\n"
-"       error and warning diagnostics occur on stream error.\n"
-"     out :  int\n"
-"       general output occurs on stream out.\n"
-"     print_level : int\n"
-"       the level of output required. Possible values are:\n"
-"\n"
-"       * <= 0\n"
-"\n"
-"         no output\n"
-"       * 1\n"
-"\n"
-"         a one-line summary for every improvement\n"
-"       * 2\n"
-"\n"
-"         a summary of each iteration\n"
-"       * >= 3\n"
-"\n"
-"         increasingly verbose (debugging) output.\n"
-"\n"
-"     start_print : int\n"
-"       any printing will start on this iteration.\n"
-"     stop_print : int\n"
-"       any printing will stop on this iteration.\n"
-"     print_gap : int\n"
-"       the number of iterations between printing.\n"
-"     maxit : int\n"
-"       the maximum number of iterations allowed.\n"
-"     initial_point : int\n"
-"       the number of initial (uniformly-spaced) evaluation points \n"
-"       (<2 reset to 2).\n"
-"     storage_increment : int\n"
-"       incremenets of storage allocated (less that 1000 will be\n"
-"       reset to 1000).\n"
-"     buffer : int\n"
-"       unit for any out-of-core writing when expanding arrays.\n"
-"     lipschitz_estimate_used : int\n"
-"       what sort of Lipschitz constant estimate will be used:\n"
-"\n"
-"       * 1\n"
-"\n"
-"         the global contant provided \n"
-"\n"
-"       * 2\n"
-"\n"
-"         an estimated global contant\n"
-"\n"
-"       * 3\n"
-"\n"
-"         estimated local costants.\n"
-"\n"
-"     next_interval_selection : int\n"
-"       how is the next interval for examination chosen:\n"
-"\n"
-"       * 1\n"
-"\n"
-"         traditional\n"
-"       * 2 \n"
-"\n"
-"         local_improvement.\n"
-"     refine_with_newton : int\n"
-"       try refine_with_newton Newton steps from the vacinity of\n"
-"       the global minimizer to try to improve the estimate.\n"
-"     alive_unit : int\n"
-"       removal of the file alive_file from unit alive_unit\n"
-"       terminates execution.\n"
-"     alive_file : str\n"
-"       see alive_unit.\n"
-"     stop_length : float\n"
-"       overall convergence tolerances. The iteration will terminate\n"
-"       when the step is less than ``stop_length``.\n"
-"     small_g_for_newton : float\n"
-"       if the absolute value of the gradient is smaller than \n"
-"       small_g_for_newton, the next evaluation point may be at a \n"
-"       Newton estimate of a local minimizer.\n"
-"     small_g : float\n"
[…]
