+++ doc/Python/go.rst
@@ -2,8 +2,8 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-GALAHAD Python Interfaces
-=========================
+Global Optimization
+-------------------
 
 **Release:** 1.0
 
@@ -12,24 +12,19 @@ GALAHAD Python Interfaces
 **Author:** `Jaroslav Fowkes <jaroslav.fowkes@stfc.ac.uk>`_ and `Nick Gould <nick.gould@stfc.ac.uk>`_
 
 GALAHAD [1]_ is a suite of open-source routines for large-scale continuous
-optimization. GALAHAD 4.1 and above provides Python modules that link
-transparently to the underlying fortran.
+optimization.  Currently there are three packages designed to find global
+optima, or at least good approximations to them.
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
    :caption: Contents
    :titlesonly:
 
-   bco
-   go
-
-* :ref:`genindex`
-
-   * :ref:`modindex`
-   * :ref:`search`
+   UGO - univariate global optimization of a smooth function over an interval <ugo>
+   BGO - multivariate global optimization of a smooth function within a box using stochastic methods<bgo>
 
 References
-----------
+^^^^^^^^^^
 
 .. [1]
    Gould, N. I. M., Orban, D., & Toint, Ph. L. (2003). GALAHAD, a library of thread-safe Fortran 90 packages for large-scale nonlinear optimization. ACM Transactions on Mathematical Software (TOMS), 29(4), 353-372.

[…]
