+++ src/dgo/Python/dgo_pyiface.c
@@ -1,0 +1,1100 @@
+//* \file dgo_pyiface.c */
+
+/*
+ * THIS VERSION: GALAHAD 4.1 - 2022-09-19 AT 13:50 GMT.
+ *
+ *-*-*-*-*-*-*-*-*-  GALAHAD_DGO PYTHON INTERFACE  *-*-*-*-*-*-*-*-*-*-
+ *
+ *  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
+ *  Principal author: Jaroslav Fowkes & Nick Gould
+ *
+ *  History -
+ *   originally released GALAHAD Version 4.1. September 19th 2022
+ *
+ *  For full documentation, see
+ *   http://galahad.rl.ac.uk/galahad-www/specs.html
+ */
+
+#include "galahad_python.h"
+#include "galahad_dgo.h"
+
+/* Nested UGO control and inform prototypes */
+bool ugo_update_control(struct ugo_control_type *control, 
+                        PyObject *py_options);
+PyObject* ugo_make_inform_dict(const struct ugo_inform_type *inform);
+//bool trb_update_control(struct trb_control_type *control,
+//                        PyObject *py_options);
+//PyObject* trb_make_inform_dict(const struct trb_inform_type *inform);
+//bool hash_update_control(struct hash_control_type *control,
+//                         PyObject *py_options);
+//PyObject* hash_make_inform_dict(const struct hash_inform_type *inform);
+
+/* Module global variables */
+static void *data;                       // private internal data
+static struct dgo_control_type control;  // control struct
+static struct dgo_inform_type inform;    // inform struct
+static bool init_called = false;         // record if initialise was called
+static int status = 0;                   // exit status
+
+//  *-*-*-*-*-*-*-*-*-*-   CALLBACK FUNCTIONS    -*-*-*-*-*-*-*-*-*-*
+
+/* Python eval_* function pointers */
+static PyObject *py_eval_f = NULL;
+static PyObject *py_eval_g = NULL;
+static PyObject *py_eval_h = NULL;
+
+/* C eval_* function wrappers */
+static int eval_f(int n, const double x[], double *f, const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyObject *py_x = PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+
+    // Build Python argument list
+    PyObject *arglist = Py_BuildValue("(O)", py_x);
+
+    // Call Python eval_f
+    PyObject *result = PyObject_CallObject(py_eval_f, arglist);
+    Py_DECREF(py_x);    // Free py_x memory
+    Py_DECREF(arglist); // Free arglist memory
+
+    // Check that eval was successful
+    if(!result)
+        return -1;
+
+    // Extract single double return value
+    if(!parse_double("eval_f return value", result, f)){
+        Py_DECREF(result); // Free result memory
+        return -1;
+    }
+
+    // Free result memory
+    Py_DECREF(result);
+
+    return 0;
+}
+
+static int eval_g(int n, const double x[], double g[], const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyArrayObject *py_x = (PyArrayObject*)
+       PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+
+    // Build Python argument list
+    PyObject *arglist = Py_BuildValue("(O)", py_x);
+
+    // Call Python eval_g
+    PyObject *result = PyObject_CallObject(py_eval_g, arglist);
+    Py_DECREF(py_x);    // Free py_x memory
+    Py_DECREF(arglist); // Free arglist memory
+
+    // Check that eval was successful
[…]
