### 2022-10-16 2720541
updates to sls to allow for checking that optional dependencies are present, and changes of defaults for dependent packages
