+++ src/trs/trs.f90
@@ -241,12 +241,12 @@
 !  symmetric (indefinite) linear equation solver
 
         CHARACTER ( LEN = 30 ) :: symmetric_linear_solver =                    &
-           "sils" // REPEAT( ' ', 26 )
+           "ssids" // REPEAT( ' ', 25 )
 
 !  definite linear equation solver
 
         CHARACTER ( LEN = 30 ) :: definite_linear_solver =                     &
-           "sils" // REPEAT( ' ', 26 )
+           "ssids" // REPEAT( ' ', 25 )
 
 !  all output lines will be prefixed by
 !    prefix(2:LEN(TRIM(%prefix))-1)
@@ -477,7 +477,7 @@
 
       CALL SLS_initialize( control%symmetric_linear_solver,                    &
                            data%SLS_data, control%SLS_control,                 &
-                           inform%SLS_inform )
+                           inform%SLS_inform, check = .TRUE. )
 
 !  Set initial values for factorization controls and data
 
@@ -1735,10 +1735,12 @@
         CALL CPU_time( time_record ) ; CALL CLOCK_time( clock_record )
         IF ( constrained ) THEN
           CALL SLS_initialize_solver( control%symmetric_linear_solver,         &
-                                      data%SLS_data, inform%SLS_inform )
+                                      data%SLS_data, inform%SLS_inform,        &
+                                      check = .TRUE. )
         ELSE
           CALL SLS_initialize_solver( control%definite_linear_solver,          &
-                                      data%SLS_data, inform%SLS_inform )
+                                      data%SLS_data, inform%SLS_inform,        &
+                                      check = .TRUE. )
         END IF
         inform%max_entries_factors = 0
 
