+++ src/uls/uls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-09-28 AT 11:50 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-10-16 AT 11:50 GMT.
 
 !-*-*-*-*-*-*-*-*- G A L A H A D _ U L S    M O D U L E  -*-*-*-*-*-*-*-*-*-
 
@@ -20,6 +20,12 @@
 !     |                                     |
 !     |     Unsymmetric Linear Systems      |
 !     |                                     |
+!     |  Packages currently supported:      |
+!     |                                     |
+!     |          MA28/GLS                   |
+!     |          MA48                       |
+!     |          GETRF from LAPACK          |
+!     |                                     |
 !     ---------------------------------------
 
      USE GALAHAD_SYMBOLS
@@ -29,6 +35,7 @@
      USE GALAHAD_SMT_double
      USE GALAHAD_STRING, ONLY: STRING_put, STRING_get, STRING_lower_word
      USE GALAHAD_GLS_double
+     USE GALAHAD_LAPACK_interface, ONLY : GETRF, GETRS
      USE HSL_ZD11_double
      USE HSL_MA48_double
 
@@ -278,6 +285,10 @@
 
        LOGICAL :: alternative = .FALSE.
 
+!  name of linear solver used
+
+       CHARACTER ( LEN = len_solver ) :: solver = REPEAT( ' ', len_solver )
+
 !  the output array from
 
        TYPE ( GLS_ainfo ) :: gls_ainfo
@@ -289,6 +300,10 @@
        TYPE ( MA48_ainfo ) :: ma48_ainfo
        TYPE ( MA48_finfo ) :: ma48_finfo
        TYPE ( MA48_sinfo ) :: ma48_sinfo
+
+!  the output scalars and arrays from LAPACK routines
+
+       INTEGER :: lapack_error = 0
 
      END TYPE ULS_inform_type
 
@@ -305,15 +320,18 @@
 
        INTEGER, DIMENSION( 64 ) :: PARDISO_PT
        INTEGER, DIMENSION( 64 ) :: pardiso_iparm = - 1
-       INTEGER, ALLOCATABLE, DIMENSION( : ) :: ORDER, MAPS, PIVOTS
+       INTEGER, ALLOCATABLE, DIMENSION( : ) :: ROWS, MAPS, PIVOTS
        INTEGER, ALLOCATABLE, DIMENSION( : , : ) :: MAP
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: RESIDUALS
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: RESIDUALS_zero
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: SOL
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: RES
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: SCALE
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: X2
+       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: RHS
        REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: D
+       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: B2
+       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: RES2
[…]
