### 2022-10-19 a4964c4
start to interface to pastix, mumps and umfpack ... early days!
