+++ src/dum/mumps.f90
@@ -1,0 +1,2 @@
+      SUBROUTINE GALAHAD_mumps( )
+      END SUBROUTINE GALAHAD_mumps
