### 2022-10-20 dc1e3c3
improvements to cqp and its friends
