+++ src/sls/sls.f90
@@ -852,7 +852,7 @@
 
     INTEGER, PARAMETER :: n_dummy = 2
     INTEGER, DIMENSION( n_dummy + 1 )  :: PTR = (/ 1, 2, 3 /)
-    INTEGER, DIMENSION( n_dummy ) :: ROW = (/ 2, 1 /)
+    INTEGER, DIMENSION( n_dummy ) :: ROW = (/ 1, 2 /)
     INTEGER, DIMENSION( 8 ) :: ICNTL_metis
     INTEGER, DIMENSION( n_dummy ) :: PERM, INVP
     TYPE ( mc68_control ) :: control_mc68
