+++ src/clls/clls.f90
@@ -1141,13 +1141,13 @@
 !  Read the specfiles for SLS and SLS-POUNCE
 
       IF ( PRESENT( alt_specname ) ) THEN
-        CALL SLS_read_specfile( control%SLS_control, device,                 &
+        CALL SLS_read_specfile( control%SLS_control, device,                   &
                        alt_specname = TRIM( alt_specname ) // '-SLS')
-        CALL SLS_read_specfile( control%SLS_pounce_control, device,          &
+        CALL SLS_read_specfile( control%SLS_pounce_control, device,            &
                        alt_specname = TRIM( alt_specname ) // '-SLS-POUNCE' )
       ELSE
         CALL SLS_read_specfile( control%SLS_control, device )
-        CALL SLS_read_specfile( control%SLS_pounce_control, device,          &
+        CALL SLS_read_specfile( control%SLS_pounce_control, device,            &
                        alt_specname = 'SLS-POUNCE' )
       END IF
 
@@ -1850,8 +1850,8 @@
         IF ( control%out > 0 .AND. control%print_level >= 1 )                  &
           WRITE( control%out,                                                  &
            "( /, A, 1X, I0, ' equalit', A, ' from ', I0, ' constraint', A )" ) &
-              prefix, data%dims%c_equality,                                    &
-              TRIM( STRING_ies( data%dims%c_equality ) ),                      &
+              prefix, data%LSP_dims%c_equality,                                    &
+              TRIM( STRING_ies( data%LSP_dims%c_equality ) ),                      &
               prob%m, TRIM( STRING_pleural( prob%m ) )
 
 !  set control parameters
@@ -1861,12 +1861,12 @@
 
 !  find any dependent rows
 
-        nzc = prob%A%ptr( data%dims%c_equality + 1 ) - 1
+        nzc = prob%A%ptr( data%LSP_dims%c_equality + 1 ) - 1
         CALL CPU_TIME( time_record ) ; CALL CLOCK_time( clock_record )
-        CALL FDC_find_dependent( prob%n, data%dims%c_equality,                 &
+        CALL FDC_find_dependent( prob%n, data%LSP_dims%c_equality,                 &
                                  prob%A%val( : nzc ),                          &
                                  prob%A%col( : nzc ),                          &
-                                 prob%A%ptr( : data%dims%c_equality + 1 ),     &
+                                 prob%A%ptr( : data%LSP_dims%c_equality + 1 ),     &
                                  prob%C_l, n_depen, data%Index_C_freed,        &
                                  data%FDC_data, data%FDC_control,              &
                                  inform%FDC_inform )
@@ -1983,7 +1983,7 @@
 
 !  store the problem dimensions
 
-        data%dims_save_freed = data%dims
+        data%LSP_dims_save_freed = data%LSP_dims
         data%a_ne = prob%A%ne
 
         IF ( printi ) WRITE( control%out,                                      &
@@ -2002,15 +2002,15 @@
         inform%time%clock_preprocess =                                         &
           inform%time%clock_preprocess + clock_now - clock_record
 
[…]
