### 2022-10-26 #4882 @editorialbot
[REVIEW]: GALAHAD 4.0: an open source library of Fortran packages with C and Matlab interfaces for continuous optimization
**Submitting author:** @nimgould (<a href="http://orcid.org/0000-0002-1031-1588">Nicholas Gould</a>)
**Repository:** https://github.com/ralna/GALAHAD
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.0
