### 2022-10-26 0237ceb
move towards pastix integration install now asks to install shared libraries
