### 2022-10-31 6275f7c
bug fixes for ssids
