+++ src/dum/pastixf_interfaces.f90
@@ -1,119 +1,146 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-10-25 AT 16:25 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-10-26 AT 16:25 GMT.
 
 ! G A L A H A D  -  D U M M Y   P A S T I X F _ I N T E R F A C E S  M O D U L E
 
+!  Extracted from
+!> @file pastixf_interfaces.f90
+!>
+!> PaStiX Fortran 90 wrapper
+!>
+!> @copyright 2017-2022 Bordeaux INP, CNRS (LaBRI UMR 5800), Inria,
+!>                      Univ. Bordeaux. All rights reserved.
+!>
+!> @version 6.2.0
+!> @author Mathieu Faverge
+!> @author Tony Delarue
+!> @author Selmane Lebdaoui
+!> @date 2022-01-13
+
  MODULE pastixf_interfaces
 
-   USE pastixf_enums
-   USE iso_c_binding, ONLY : c_double, c_int, c_ptr, c_int32_t, c_int64_t
-
    INTERFACE pastixInitParam
-     MODULE PROCEDURE pastixInitParam_f08
+     SUBROUTINE pastixInitParam_f08( iparm, dparm )
+       USE iso_c_binding, ONLY : c_double
+       USE pastixf_enums, ONLY : pastix_int_t
+       IMPLICIT NONE
+       INTEGER ( KIND = pastix_int_t ), INTENT( INOUT ), target :: iparm( : )
+       REAL ( KIND = c_double ), INTENT( INOUT ), target :: dparm( : )
+     END SUBROUTINE pastixInitParam_f08
    END INTERFACE pastixInitParam
 
    INTERFACE pastixInit
-     MODULE PROCEDURE pastixInit_f08
+     SUBROUTINE pastixInit_f08( pastix_data, pastix_comm, iparm, dparm )
+       USE iso_c_binding, ONLY : c_double
+       USE spmf_enums, ONLY : MPI_Comm
+       USE pastixf_enums, ONLY : pastix_data_t, pastix_int_t
+       IMPLICIT NONE
+       TYPE ( pastix_data_t ), INTENT( INOUT ), pointer :: pastix_data
+       TYPE ( MPI_Comm ), INTENT( IN ) :: pastix_comm
+       INTEGER ( KIND = pastix_int_t ), INTENT( INOUT ), target  :: iparm( : )
+       REAL ( KIND = c_double ), INTENT( INOUT ), target  :: dparm( : )
+     END SUBROUTINE pastixInit_f08
    END INTERFACE pastixInit
 
    INTERFACE pastix_task_analyze
-     MODULE PROCEDURE pastix_task_analyze_f08
+     SUBROUTINE pastix_task_analyze_f08( pastix_data, spm, info )
+       USE iso_c_binding, ONLY : c_int
+       USE spmf_enums, ONLY : spmatrix_t
+       USE pastixf_enums, ONLY : pastix_data_t
+       IMPLICIT NONE
+       TYPE ( pastix_data_t ), INTENT( INOUT ), TARGET :: pastix_data
+       TYPE ( spmatrix_t ), INTENT( IN ), TARGET :: spm
+       INTEGER ( KIND = c_int ), INTENT( OUT ), OPTIONAL :: info
+     END SUBROUTINE pastix_task_analyze_f08
    END INTERFACE pastix_task_analyze
 
    INTERFACE pastix_task_numfact
-     MODULE PROCEDURE pastix_task_numfact_f08
+     SUBROUTINE pastix_task_numfact_f08( pastix_data, spm, info )
+       USE iso_c_binding, ONLY : c_int
[…]
