+++ src/dum/cutest_dummy.f90
@@ -1,6 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-08-14 AT 09:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-10-29 AT 10:45 GMT.
 
-!-*-*-  P Y T H O N _ D U M M Y  P A C K A G E S / S U B P R O G R A M S  -*-*-
+!-*-*-  C U T E S T _ D U M M Y  P A C K A G E S / S U B P R O G R A M S  -*-*-
 
 !  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
 !  Principal authors: Daniel Robinson and Nick Gould
