+++ src/dum/spmf_interfaces.f90
@@ -1,97 +1,119 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-10-25 AT 16:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-10-26 AT 15:30 GMT.
 
 !-  G A L A H A D  -  D U M M Y   S P M F _ I N T E R F A C E S   M O D U L E  -
+
+!  Extracted from
+!> @file spmf_interfaces.f90
+!>
+!> SPM Fortran 90 wrapper
+!>
+!> @copyright 2017-2022 Bordeaux INP, CNRS (LaBRI UMR 5800), Inria,
+!>                      Univ. Bordeaux. All rights reserved.
+!>
+!> @version 1.2.0
+!> @author Mathieu Faverge
+!> @author Tony Delarue
+!> @date 2022-02-22
 
  MODULE spmf_interfaces
 
    USE iso_c_binding, ONLY : c_float, c_double, c_ptr,                         &
                              c_int, c_int32_t, c_int64_t
-   USE spmf_enums
-
    INTERFACE spmInit
-     MODULE PROCEDURE spmInit_f08 
+     SUBROUTINE spmInit_f08( spm )
+      USE spmf_enums, ONLY : spmatrix_t
+      IMPLICIT NONE
+      TYPE ( spmatrix_t ), INTENT( INOUT ), TARGET :: spm
+     END SUBROUTINE spmInit_f08
    END INTERFACE spmInit
 
    INTERFACE spmUpdateComputedFields
-     MODULE PROCEDURE spmUpdateComputedFields_f08
+     SUBROUTINE spmUpdateComputedFields_f08( spm )
+       USE spmf_enums, ONLY : spmatrix_t
+       IMPLICIT NONE
+       TYPE ( spmatrix_t ), INTENT( INOUT ), TARGET :: spm
+     END SUBROUTINE spmUpdateComputedFields_f08
    END INTERFACE spmUpdateComputedFields
 
    INTERFACE spmAlloc
-     MODULE PROCEDURE spmAlloc_f08
+     SUBROUTINE spmAlloc_f08( spm )
+       USE spmf_enums, ONLY : spmatrix_t
+       IMPLICIT NONE
+       TYPE ( spmatrix_t ), INTENT( INOUT ), TARGET :: spm
+     END SUBROUTINE spmAlloc_f08
    END INTERFACE spmAlloc
 
    INTERFACE spmCheckAndCorrect
-     MODULE PROCEDURE spmCheckAndCorrect_f08
+     SUBROUTINE spmCheckAndCorrect_f08( spm_in, spm_out, info )
+       USE iso_c_binding, ONLY : c_int
+       USE spmf_enums, ONLY : spmatrix_t
+       IMPLICIT NONE
+       TYPE ( spmatrix_t ), INTENT( IN ), TARGET :: spm_in
+       TYPE ( spmatrix_t ), INTENT( INOUT ), TARGET :: spm_out
+       INTEGER( KIND = c_int ), INTENT( OUT ), OPTIONAL :: info
+     END SUBROUTINE spmCheckAndCorrect_f08
    END INTERFACE spmCheckAndCorrect
 
    INTERFACE spmGetArray
-     MODULE PROCEDURE spmGetArray_f08
+     SUBROUTINE spmGetArray_f08( spm, colptr, rowptr, zvalues, cvalues,        &
+                                 dvalues, svalues, dofs, loc2glob, glob2loc )
+       USE iso_c_binding, ONLY : c_double_complex, c_float_complex, c_double,  &
+                                 c_float
+       USE spmf_enums, ONLY : spmatrix_t, spm_int_t
+       IMPLICIT NONE
[…]
