### 2022-10-31 a6b2ba2
updated install script
