### 2022-10-31 b8d102c
update sls to accept pastix
