+++ src/sls/sls.f90
@@ -841,34 +841,6 @@
        TYPE ( SMT_type ) :: matrix
      END TYPE SLS_full_data_type
 
-!  interfaces
-
-!     INTERFACE
-!       SUBROUTINE pastixOrderGetArray( order, permtab )
-!       IMPORT :: pastix_order_t, pastix_int_t 
-!       IMPLICIT NONE
-!       TYPE( pastix_order_t ), INTENT( IN ), target  :: order
-!       INTEGER ( pastix_int_t ), DIMENSION( : ), INTENT( OUT ), OPTIONAL,     &
-!                                 POINTER :: permtab
-!       END SUBROUTINE pastixOrderGetArray
-!    END INTERFACE
-
-!    INTERFACE
-!      SUBROUTINE spmGetArray( spm, colptr, rowptr, dvalues, svalues )
-!      IMPORT :: spmatrix_t, spm_int_t, c_double, c_float
-!      IMPLICIT NONE
-!      TYPE ( spmatrix_t ), INTENT( IN ), TARGET :: spm
-!      INTEGER ( spm_int_t ), DIMENSION(:), INTENT( OUT ), OPTIONAL,           &
-!                          POINTER :: colptr
-!      INTEGER ( spm_int_t ), DIMENSION( : ), INTENT( OUT ), OPTIONAL,         &
-!                          POINTER :: rowptr
-!      REAL ( c_double ), DIMENSION( : ), INTENT( OUT ), OPTIONAL,             &
-!                         POINTER :: dvalues
-!      REAL ( c_float ),  DIMENSION( : ), INTENT( OUT ), OPTIONAL,             &
-!                         POINTER :: svalues
-!      END SUBROUTINE spmGetArray
-!    END INTERFACE
-
    CONTAINS
 
 !-*-*-*-*-*-   S L S _ I N I T I A L I Z E   S U B R O U T I N E   -*-*-*-*-*-
@@ -889,13 +861,6 @@
      TYPE ( SLS_control_type ), INTENT( INOUT ) :: control
      TYPE ( SLS_inform_type ), INTENT( OUT ) :: inform
      LOGICAL, OPTIONAL, INTENT( IN ) :: check
-
-!  local variables
-
-     INTEGER, DIMENSION( 30 ) :: ICNTL_ma27
-     REAL ( KIND = wp ), DIMENSION( 5 ) :: CNTL_ma27
-     TYPE ( MA57_control ) :: control_ma57
-     TYPE ( ssids_akeep ) :: akeep_ssids
 
 !  local variables
 
@@ -1289,8 +1254,11 @@
 !      ELSE
          data%iparm_pastix( 44 ) = 1
 !      END IF
-       CALL pastixInit( data%pastix_data, MPI_COMM_WORLD,                  &
-                            data%iparm_pastix, data%dparm_pastix )
+!      OPEN( 2, FILE = "/dev/null", STATUS = "OLD" ) ! try to get rid of msgs
+       CALL pastixInit( data%pastix_data, MPI_COMM_WORLD,                      &
+                        data%iparm_pastix, data%dparm_pastix )
+!      CLOSE( 2 )
+!      OPEN( 2, FILE = "/dev/stdout", STATUS = "OLD" )
        data%must_be_definite = .FALSE.
 
 !  = unavailable solver =
@@ -2008,19 +1976,16 @@
        inform%flops_elimination = info_ssids%num_flops
        inform%max_front_size  = info_ssids%maxfront
        inform%max_depth_assembly_tree = info_ssids%maxdepth
-     CASE ( - 30  )
+     CASE ( - 50  )
        inform%status = GALAHAD_error_allocate
[…]
