+++ src/qpa/qpa.f90
@@ -7903,12 +7903,16 @@
 
       m_working = 0
       K%ne = n_free
+      a_abs = one
       DO i = 1, m
         IF ( C_stat( i ) /= 0 ) THEN
           m_working = m_working + 1
           WORKING( m_working ) = i
           DO j =  A_ptr( i ), A_ptr( i + 1 ) - 1
-            IF ( B_stat( A_col( j ) ) == 0 ) K%ne = K%ne + 1
+            IF ( B_stat( A_col( j ) ) == 0 ) THEN
+              K%ne = K%ne + 1
+              a_abs = MAX( a_abs, ABS( A_val( j ) ) )
+            END IF
           END DO
         END IF
       END DO
@@ -8038,6 +8042,7 @@
       DO i = 1, n_free
 !       K%row( i ) = i ; K%col( i ) = i ; K%val( i ) = one
         K%row( i ) = i ; K%col( i ) = i ; K%val( i ) = diag
+!       K%row( i ) = i ; K%col( i ) = i ; K%val( i ) = 100.0_wp * a_abs
       END DO
 
 !     write(6,*) K%n, K%ne
@@ -8136,6 +8141,9 @@
 !  Determine the block diagonal part of the factors and the pivot order
 
       CALL SLS_enquire( SLS_data, inform%SLS_inform, PIVOTS = P, D = D )
+!write(6,"( ' P ', ( 20I4 ) )" ) P( : K%n )
+!write(6,"( ' dia ', 4ES12.4 )" ) D( 1, 1 : K%n )
+!write(6,"( ' off ', 4ES12.4 )" ) D( 2, 1 : K%n-1 )
 
 !  Find the largest diagonal (of the inverse)
 
@@ -8189,10 +8197,10 @@
           twobytwo = .FALSE.
           CYCLE
         END IF
-        IF ( i < inform%SLS_inform%rank ) THEN
 
 !  A 2x2 block
 
+        IF ( i < inform%SLS_inform%rank ) THEN
           IF ( P( i ) < 0 ) THEN
             twobytwo = .TRUE.
 
@@ -8209,18 +8217,22 @@
 
             IF ( pmin <= n_free ) THEN
               IF ( rmax >= big ) THEN
+                IF ( pmax > n_free ) THEN
                 n_depen = n_depen + 1
                 C_stat( WORKING( pmax - n_free ) ) = 0
                 IF ( out > 0 .AND. control%print_level >= 3 ) THEN
                   WRITE(  out, "( '2x2 block ', 2i7, ' eval = ', ES12.4 )" )   &
                    pmax - n_free, pmin - n_free, one / rmax
                 END IF
+                END IF
               ELSE IF ( rmin == zero ) THEN
+                IF ( pmax > n_free ) THEN
                 n_depen = n_depen + 1
                 C_stat( WORKING( pmax - n_free ) ) = 0
                 IF ( out > 0 .AND. control%print_level >= 3 ) THEN
                   WRITE( out, "( '2x2 block ', 2i7, ' eval = infinity' )" )    &
                    pmax - n_free, pmin - n_free
+                END IF
                 END IF
               ELSE
                 IF ( out > 0 .AND. control%print_level >= 5 )                  &
@@ -8229,18 +8241,22 @@
               END IF
 
               IF ( rmin >= big ) THEN
+                IF ( pmin > n_free ) THEN
                 n_depen = n_depen + 1
                 C_stat( WORKING( pmin - n_free ) ) = 0
                 IF ( out > 0 .AND. control%print_level >= 3 ) THEN
                   WRITE( out, "( '2x2 block ', 2i7, ' eval = ', ES12.4 )" )    &
                     pmin - n_free, pmax - n_free, one / rmin
                 END IF
+                END IF
               ELSE IF ( rmax == zero ) THEN
[…]
