+++ src/sec/sec.f90
@@ -545,12 +545,6 @@
      TYPE ( SEC_control_type ), INTENT( IN ) :: control
      TYPE ( SEC_inform_type ), INTENT( INOUT ) :: inform
 
-!-----------------------------------------------
-!   L o c a l   V a r i a b l e s
-!-----------------------------------------------
-
-     CHARACTER ( LEN = 80 ) :: array_name
-
 !  deallocate workspace
 
      CALL SEC_terminate( data%sec_data, control, inform )
