+++ src/qp/useqp.f90
@@ -20,7 +20,6 @@
 !    ------------------------------------------------------
 
       USE CUTEst_interface_double
-!NOT95USE GALAHAD_CPU_time
       USE GALAHAD_CLOCK
       USE GALAHAD_QPT_double
       USE GALAHAD_SORT_double, only: SORT_reorder_by_rows
