+++ src/dlp/dlp.f90
@@ -35,7 +35,6 @@
 !      ------------------------------------------------------
 
 !$    USE omp_lib
-!NOT95USE GALAHAD_CPU_time
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
       USE GALAHAD_STRING, ONLY: STRING_pleural, STRING_verb_pleural,           &
