+++ src/oblivion/lcf/lcf.f90
@@ -25,7 +25,6 @@
 !     |                                            |
 !     ----------------------------------------------
 
-!NOT95USE GALAHAD_CPU_time
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
       USE GALAHAD_SPACE_double
