### 2022-11-01 3f615ae
fix to enable build of single precision ssids
