+++ src/spral/blas_iface.f90
@@ -13,50 +13,58 @@ module spral_blas_iface
   interface
     subroutine daxpy( n, a, x, incx, y, incy )
       implicit none
+      integer, parameter :: PRECISION = kind(1.0D+0)
       integer, intent(in) :: n, incx, incy
-      double precision, intent(in) :: a
-      double precision, intent(in   ), dimension(*) :: x
-      double precision, intent(inout), dimension(*) :: y
+      real(PRECISION), intent(in) :: a
+      real(PRECISION), intent(in   ), dimension(*) :: x
+      real(PRECISION), intent(inout), dimension(*) :: y
     end subroutine daxpy
     subroutine dcopy( n, x, incx, y, incy )
       implicit none
+      integer, parameter :: PRECISION = kind(1.0D+0)
       integer, intent(in) :: n, incx, incy
-      double precision, intent(in ), dimension(*) :: x
-      double precision, intent(out), dimension(*) :: y
+      real(PRECISION), intent(in ), dimension(*) :: x
+      real(PRECISION), intent(out), dimension(*) :: y
     end subroutine dcopy
-    double precision function ddot( n, x, incx, y, incy )
+    function ddot( n, x, incx, y, incy )
       implicit none
+      integer, parameter :: PRECISION = kind(1.0D+0)
+      real(PRECISION) :: ddot
       integer, intent(in) :: n, incx, incy
-      double precision, intent(in), dimension(*) :: x
-      double precision, intent(in), dimension(*) :: y
+      real(PRECISION), intent(in), dimension(*) :: x
+      real(PRECISION), intent(in), dimension(*) :: y
     end function ddot
-    double precision function dnrm2( n, x, incx )
+    function dnrm2( n, x, incx )
       implicit none
+      integer, parameter :: PRECISION = kind(1.0D+0)
+      real(PRECISION) :: dnrm2
       integer, intent(in) :: n, incx
-      double precision, intent(in), dimension(*) :: x
+      real(PRECISION), intent(in), dimension(*) :: x
     end function dnrm2
     subroutine dscal( n, a, x, incx )
       implicit none
+      integer, parameter :: PRECISION = kind(1.0D+0)
       integer, intent(in) :: n, incx
-      double precision, intent(in) :: a
-      double precision, intent(in), dimension(*) :: x
+      real(PRECISION), intent(in) :: a
+      real(PRECISION), intent(in), dimension(*) :: x
     end subroutine dscal
     subroutine zcopy( n, x, incx, y, incy )
       implicit none
-      integer, parameter :: PRECISION = kind(1.0D0)
+      integer, parameter :: PRECISION = kind(1.0D+0)
       integer, intent(in) :: n, incx, incy
       complex(PRECISION), intent(in ), dimension(*) :: x
       complex(PRECISION), intent(out), dimension(*) :: y
     end subroutine zcopy
-    double precision function dznrm2( n, x, incx )
+    function dznrm2( n, x, incx )
       implicit none
-      integer, parameter :: PRECISION = kind(1.0D0)
+      integer, parameter :: PRECISION = kind(1.0D+0)
+      real(PRECISION) :: dznrm2
       integer, intent(in) :: n, incx
       complex(PRECISION), intent(in), dimension(*) :: x
     end function dznrm2
     function zdotc( n, x, incx, y, incy )
       implicit none
-      integer, parameter :: PRECISION = kind(1.0D0)
+      integer, parameter :: PRECISION = kind(1.0D+0)
       integer, intent(in) :: n, incx, incy
       complex(PRECISION), intent(in), dimension(*) :: x
[…]
