### 2022-11-03 72c4d93
lancelot fix to check for non-existent solvers
