### 2022-11-03 2dba98d
dummy ma27 control values modified
