### 2022-11-06 6b8ef27
additions to sls fixes for tests when standalone and with a minimal install
