+++ src/sls/sls.f90
@@ -878,6 +878,7 @@
 
      LOGICAL :: check_available, hsl_available, metis_available
 
+!    write(6,*) ' omp cancellation = ', omp_get_cancellation()
 !  initialize the solver-specific data
 
      CALL SLS_initialize_solver( solver, data, inform, check )
@@ -1047,7 +1048,7 @@
          ELSE
            control%ordering = 7
          END IF
-       END IF 
+       END IF
      END SELECT
 
      RETURN
@@ -3688,7 +3689,7 @@
          data%spm%nnz = data%ne
          data%spm%dof = 1
 
-!  set space for the matrix in the SPM structure 
+!  set space for the matrix in the SPM structure
 
          CALL spmUpdateComputedFields( data%spm )
          CALL spmAlloc( data%spm )
@@ -5946,7 +5947,7 @@
            CALL SYTRS( 'L', data%n, 1, data%matrix_dense, data%n,              &
                        data%PIVOTS, data%X2, data%n, inform%lapack_error )
          ELSE
-           tiny = 100.0_wp * control%absolute_pivot_tolerance 
+           tiny = 100.0_wp * control%absolute_pivot_tolerance
            CALL SLS_sytr_singular_solve( data%n, 1, data%matrix_dense,         &
                                          data%n, data%PIVOTS, data%X2,         &
                                          data%n, tiny, inform%lapack_error )
@@ -6328,7 +6329,7 @@
              CALL SYTRS( 'L', data%n, nrhs, data%matrix_dense, data%n,         &
                          data%PIVOTS, data%X2, data%n, inform%lapack_error )
            ELSE
-             tiny = 100.0_wp * control%absolute_pivot_tolerance 
+             tiny = 100.0_wp * control%absolute_pivot_tolerance
              CALL SLS_sytr_singular_solve( data%n, nrhs, data%matrix_dense,    &
                                            data%n, data%PIVOTS, data%X2,       &
                                            data%n, tiny, inform%lapack_error )
@@ -6363,7 +6364,7 @@
              CALL SYTRS( 'L', data%n, nrhs, data%matrix_dense, data%n,         &
                          data%PIVOTS, X, data%n, inform%lapack_error )
            ELSE
-             tiny = 100.0_wp * control%absolute_pivot_tolerance 
+             tiny = 100.0_wp * control%absolute_pivot_tolerance
              CALL SLS_sytr_singular_solve( data%n, nrhs, data%matrix_dense,    &
                                            data%n, data%PIVOTS, X,             &
                                            data%n, tiny, inform%lapack_error )
@@ -8733,7 +8734,7 @@
 
 ! =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
-!  Solve the systems A X = B, where the factoriization A = L D L^T is found 
+!  Solve the systems A X = B, where the factoriization A = L D L^T is found
 !  by the LAPACK routine SYTRF, A is reported singular, and B is input in X
 
 ! =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
@@ -8752,8 +8753,8 @@
      INTEGER :: j, k, kp
      REAL ( KIND = wp ) :: ak, akm1, akm1k, bk, bkm1, denom
 
-!  first solve L*D*X = B, overwriting B with X. k is the main loop index, 
-!  increasing from 1 to n in steps of 1 or 2, depending on the size of 
[…]
