### 2022-11-08 9c335d1
updated shared library conversion order. worked around character copy gcc compiler issue
