### 2022-11-10 288e3ff
temporary change to ssids config file
