### 2022-11-10 122ea7b
updates to instructions so that valgrind can be used on test programs
