+++ src/qpb/qpbs2.f90
@@ -1,0 +1,370 @@
+! THIS VERSION: GALAHAD 2.4 - 08/04/2010 AT 08:00 GMT.
+   PROGRAM GALAHAD_QPB_EXAMPLE
+   USE GALAHAD_QPB_double                            ! double precision version
+   USE GALAHAD_SYMBOLS
+   IMPLICIT NONE
+   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
+   REAL ( KIND = wp ), PARAMETER :: infbnd = 10.0_wp ** 9
+   REAL ( KIND = wp ), PARAMETER :: infty = 10.0_wp * infbnd
+   TYPE ( QPT_problem_type ) :: p
+   TYPE ( QPB_data_type ) :: data
+   TYPE ( QPB_control_type ) :: control
+   TYPE ( QPB_inform_type ) :: info
+   INTEGER :: n, m, h_ne, a_ne, smt_stat, alloc_stat
+   INTEGER :: data_storage_type, i, status, scratch_out = 56
+   CHARACTER ( len = 1 ) :: st
+
+   n = 2 ; m = 1 ; h_ne = 3 ; a_ne = 2
+   ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
+   ALLOCATE( p%C( m ), p%C_l( m ), p%C_u( m ) )
+   ALLOCATE( p%X( n ), p%Y( m ), p%Z( n ) )
+   ALLOCATE( p%H%ptr( n + 1 ), p%A%ptr( m + 1 ) )
+
+   p%n = n ; p%m = m ; p%f = 0.05_wp
+   p%G = (/ 0.0_wp, 0.0_wp /)
+   p%C_l = (/ 1.0_wp /)
+   p%C_u = (/ 1.0_wp /)
+   p%X_l = (/ 0.0_wp, 0.0_wp /)
+   p%X_u = (/ 2.0_wp, 3.0_wp /)
+
+   p%new_problem_structure = .TRUE.
+   ALLOCATE( p%H%val( h_ne ), p%H%row( 0 ), p%H%col( h_ne ) )
+   ALLOCATE( p%A%val( a_ne ), p%A%row( 0 ), p%A%col( a_ne ) )
+   IF ( ALLOCATED( p%H%type ) ) DEALLOCATE( p%H%type )
+   CALL SMT_put( p%H%type, 'SPARSE_BY_ROWS', smt_stat )
+   p%H%col = (/ 1, 2, 1 /)
+   p%H%ptr = (/ 1, 2, 4 /)
+   IF ( ALLOCATED( p%A%type ) ) DEALLOCATE( p%A%type )
+   CALL SMT_put( p%A%type, 'SPARSE_BY_ROWS', smt_stat )
+   p%A%col = (/ 1, 2 /)
+   p%A%ptr = (/ 1, 3 /)
+go to 7
+   CALL QPB_initialize( data, control, info )
+   control%infinity = infbnd
+   control%restore_problem = 2
+! control%print_level = 1
+! control%SBLS_control%print_level = 1
+! control%LSQP_control%print_level = 1
+
+!  test with new and existing data
+
+   DO i = 0, 20
+!  DO i = 0, 0
+     IF ( i == 0 ) THEN
+       control%precon = 0
+     ELSE IF ( i == 1 ) THEN
+       control%SBLS_control%preconditioner = 1
+     ELSE IF ( i == 2 ) THEN
+       control%SBLS_control%preconditioner = 2
+     ELSE IF ( i == 3 ) THEN
+       control%SBLS_control%preconditioner = 3
+     ELSE IF ( i == 4 ) THEN
+       control%SBLS_control%preconditioner = 4
+     ELSE IF ( i == 5 ) THEN
[…]
