### 2022-11-11 e4af5df
attempt Intel compilers (ifort/icc)
