+++ .github/workflows/standalone.yml
@@ -8,31 +8,34 @@ on:
     types: [opened, synchronize, reopened]
 jobs:
   build:
-    name: GALAHAD ${{ matrix.os }}
+    name: GALAHAD ${{ matrix.os }} ${{ matrix.compiler }}
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        include:
+          - os: ubuntu-latest
+            compiler: gnu
+            input: 'y1\nnnnyn4\n1\nn2\nn4\nnyby'
+          - os: macos-latest
+            compiler: gnu
+            input: 'y1\nnnnyn6\nn1\nn3\nnyby'
+          - os: ubuntu-latest
+            compiler: intel
+            input: 'y1\nnnnyn4\n1\nn3\nn5\nnyby'
+            hpckit: 'https://registrationcenter-download.intel.com/akdlm/irc_nas/18975/l_HPCKit_p_2022.3.1.16997_offline.sh'
+            script: 'install_linux.sh'
+          - os: macos-latest
+            compiler: intel
+            input: 'y1\nnnnyn6\nn2\nn4\nnyby'
+            hpckit: 'https://registrationcenter-download.intel.com/akdlm/irc_nas/18977/m_HPCKit_p_2022.3.1.15344_offline.dmg'
+            script: 'install_macos.sh'
     runs-on: ${{ matrix.os }}
     env:
       ARCHDEFS: ${{ github.workspace }}/ARCHDefs
       GALAHAD: ${{ github.workspace }}
-      GCCVER: 12
     steps:
       - name: Set up Homebrew
         uses: Homebrew/actions/setup-homebrew@master
-      - name: Install C and Fortran compilers
-        run: |
-          brew update
-          brew install gcc
-          if [[ "${{ matrix.os }}" == "ubuntu-latest" ]]
-          then
-            # TODO: make brew work on linux
-            sudo apt install -y libhwloc-dev
-          else
-            brew install hwloc
-          fi
-          eval $(brew shellenv)
       - name: Check out GALAHAD
         uses: actions/checkout@v3
       - name: Download ARCHDefs
@@ -47,35 +50,54 @@ jobs:
         run: |
           tar zxf ARCHDefs-${{ steps.download-archdefs.outputs.tag_name }}.tar.gz
           mv ralna-ARCHDefs-* ARCHDefs
+      - name: Install GNU C and Fortran compilers
+        if: matrix.compiler == 'gnu'
+        run: |
+          brew update
+          eval $(brew shellenv)
+          brew install gcc
+      - name: Install Intel OneAPI C and Fortran compilers
+        if: matrix.compiler == 'intel'
+        run: |
+          ci-scripts/${{ matrix.script }} "${{ matrix.hpckit }}" all
+      - name: Install hwloc
+        run: |
+          if [[ "${{ matrix.os }}" == "ubuntu-latest" ]]
+          then
+            # TODO: make brew work on linux
+            sudo apt install -y libhwloc-dev
+          else
+            brew install hwloc
+          fi
       # - name: Setup tmate session
       #   uses: mxschmitt/action-tmate@v3
       - name: Build GALAHAD
         run: |
-          cd $GALAHAD
-          if [[ "${{ matrix.os }}" == "ubuntu-latest" ]]
+          [[ "${{ matrix.compiler }}" == "intel" ]] && source /opt/intel/oneapi/setvars.sh
+          if [[ "${{ matrix.os }}" == "macos-latest" ]]
           then
-            printf "y1\nnnnyn4\n1\nn2\nn4\nnyby" > install_config
[…]
