### 2022-11-12 22ffda8
added script to complie a single shared library from multiple ststic ones. Various missing deallocations added to stop memory leaks.
