### 2022-11-17 301d1fd
try hwloc with brew on linux again
