### 2022-11-17 5a2d28c
numerous changes to remove memory leaks found by valgrind
