### 2022-11-18 5934fb3
disable compilation of ssids in single precision. "Final" memory leaks fixed (except for ssids)
