+++ src/sbls/sbls.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-11 AT 16:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-11-18 AT 14:15 GMT.
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S B L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -263,18 +263,18 @@
 
 !  indefinite linear equation solver
 
-        CHARACTER ( LEN = 30 ) :: symmetric_linear_solver =                    &
-           "ssids" // REPEAT( ' ', 25 )
+        CHARACTER ( LEN = 30 ) :: symmetric_linear_solver = "ssids" //         &
+                                                             REPEAT( ' ', 25 )
 
 !  definite linear equation solver
 
-        CHARACTER ( LEN = 30 ) :: definite_linear_solver =                     &
-           "ssids" // REPEAT( ' ', 25 )
+        CHARACTER ( LEN = 30 ) :: definite_linear_solver = "ssids" //          &
+                                                            REPEAT( ' ', 25 )
 
 !  unsymmetric linear equation solver
 
-        CHARACTER ( LEN = 30 ) :: unsymmetric_linear_solver =                  &
-           "gls" // REPEAT( ' ', 27 )
+        CHARACTER ( LEN = 30 ) :: unsymmetric_linear_solver = "gls" //         &
+                                                               REPEAT( ' ', 27 )
 
 !  all output lines will be prefixed by
 !    prefix(2:LEN(TRIM(%prefix))-1)
