### 2022-11-19 444ba1a
remove unused variable from amd
