### 2022-11-20 2d3ffae
added -lhwloc to creation of joint shared library. removed unused variables from some c interfaces
