+++ src/arc/C/arc_ciface.f90
@@ -690,7 +690,6 @@
   CHARACTER ( KIND = C_CHAR, LEN = opt_strlen( ctype ) ) :: ftype
   TYPE ( f_arc_control_type ) :: fcontrol
   TYPE ( f_arc_full_data_type ), POINTER :: fdata
-  INTEGER, DIMENSION( : ), ALLOCATABLE :: row_find, col_find, ptr_find
   LOGICAL :: f_indexing
 
 !  copy control and inform in
