### 2022-11-20 ce05126
name of directory non-free changed to external. -lhwloc added to create_one_shared script variable X in lhs.h changed to make compatible with clang.jl
