+++ src/external/pardiso/pardisot.f90
@@ -1,0 +1,157 @@
+PROGRAM TEST_PARDISO
+
+  USE GALAHAD_SYMBOLS
+  IMPLICIT NONE
+
+!  precision
+
+  INTEGER, PARAMETER :: wp = KIND( 1.0D0 )
+  INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
+
+!  variables
+
+  INTEGER, PARAMETER :: n = 8
+  INTEGER, PARAMETER :: nrhs = 1
+  INTEGER:: nz, maxfct, mnum, mtype, phase, error, msglvl
+  INTEGER, ALLOCATABLE, DIMENSION ( : ) :: IA, JA, IPARM
+  INTEGER ( KIND = long ), ALLOCATABLE, DIMENSION( : ) :: PT
+  REAL( KIND = wp ), ALLOCATABLE, DIMENSION ( : ) :: A, DPARM
+  REAL( KIND = wp ), ALLOCATABLE, DIMENSION ( : , : ) :: B, X
+  INTEGER :: i, j, l, idum( 1 )
+  REAL( KIND = wp ) :: ddum( 1 )
+  LOGICAL :: all_in_one = .FALSE.
+  LOGICAL :: easy_problem = .FALSE.
+
+  ALLOCATE( IPARM( 64 ), DPARM( 64 ), PT( 64 ) )
+  ALLOCATE( IA( n + 1 ), B( n, nrhs ), X( n, nrhs ) )
+
+!  fill all arrays containing matrix data
+
+  IF ( easy_problem ) THEN
+    nz = n
+    ALLOCATE( A( nz ), JA( nz ) )
+    DO i = 1, n
+      IA( i ) = i ; JA( i ) = i ; A( i ) = 1.0_wp
+    END DO
+    IA( n + 1 ) = n + 1
+  ELSE 
+    nz = 18
+    ALLOCATE( A( nz ), JA( nz ) )
+    IA = (/ 1, 5, 8, 10, 12, 15, 17, 18, 19 /)
+    JA = (/ 1, 3, 6, 7, 2, 3, 5, 3, 8, 4, 7, 5, 6, 7, 6, 8, 7, 8 /)
+    A = (/ 7.0_wp, 1.0_wp, 2.0_wp, 7.0_wp, -4.0_wp, 8.0_wp, 2.0_wp, 1.0_wp,    &
+           5.0_wp, 7.0_wp, 9.0_wp, 5.0_wp, 1.0_wp, 5.0_wp,-1.0_wp, 5.0_wp,     &
+           11.0_wp, 5.0_wp /)
+  END IF
+
+!  set the RHS so that the required solution is a vector of ones
+
+! B = 1.0_wp ! set right-hand side
+  B = 0.0_wp
+  DO i = 1, n
+    DO l = IA( i ), IA( i + 1 ) - 1
+      j = JA( l )
+      B( i, 1 ) = B( i, 1 ) + A( l )
+      IF ( i /= j ) B( j, 1 ) = B( j, 1 ) + A( l )
+    END DO
+  END DO
+
+!  set defaults
+
+  mtype = - 2 ! symmetric, indefinite
+! mtype = 1 ! symmetric
+! mtype = 2 ! symmetric, positive definite
+  CALL PARDISOINIT( PT, mtype, 0, IPARM, DPARM, error )
+
+!  check for error returns
+
+  IF ( error == GALAHAD_unavailable_option ) THEN
+    WRITE( 6, "( ' PARDISO is not available' )" )
+    GO TO 1
+  ELSE IF ( error /= 0 ) THEN
+    WRITE( 6, "( ' the following ERROR was detected: ', I0 )" ) error
+    GO TO 1
+  ELSE
[…]
