+++ src/trs/C/trs_ciface.f90
@@ -116,7 +116,7 @@
       INTEGER ( KIND = C_INT ) :: status
       INTEGER ( KIND = C_INT ) :: alloc_status
       INTEGER ( KIND = C_INT ) :: factorizations
-      INTEGER ( KIND = C_INT ) :: max_entries_factors
+      INTEGER ( KIND = C_LONG_LONG ) :: max_entries_factors
       INTEGER ( KIND = C_INT ) :: len_history
       REAL ( KIND = wp ) :: obj
       REAL ( KIND = wp ) :: x_norm
