+++ include/galahad_nls.h
@@ -385,6 +385,7 @@ for i,j=1,...,n\endmanonly
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
@@ -1024,7 +1025,7 @@ struct nls_subproblem_inform_type {
 
     /// \brief
     /// the maximum number of entries in the factors
-    long int max_entries_factors;
+    int64_t max_entries_factors;
 
     /// \brief
     /// the total integer workspace required for the factorization
