+++ include/galahad_sbls.h
@@ -302,6 +302,7 @@ may be obtained by the package.
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
@@ -596,11 +597,11 @@ struct sbls_inform_type {
 
     /// \brief
     /// the total integer workspace required for the factorization
-    long int factorization_integer;
+    int64_t factorization_integer;
 
     /// \brief
     /// the total real workspace required for the factorization
-    long int factorization_real;
+    int64_t factorization_real;
 
     /// \brief
     /// the preconditioner used
