+++ include/galahad_tru.h
@@ -200,6 +200,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
@@ -537,7 +538,7 @@ struct tru_inform_type {
     int factorization_status;
 
     /// the maximum number of entries in the factors
-    long int max_entries_factors;
+    int64_t max_entries_factors;
 
     /// the total integer workspace required for the factorization
     int factorization_integer;
