+++ include/galahad_bqpb.h
@@ -273,6 +273,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
@@ -685,11 +686,11 @@ struct bqpb_inform_type {
 
     /// \brief
     /// the total integer workspace required for the factorization
-    long int factorization_integer;
+    int64_t factorization_integer;
 
     /// \brief
     /// the total real workspace required for the factorization
-    long int factorization_real;
+    int64_t factorization_real;
 
     /// \brief
     /// the total number of factorizations performed
