+++ include/galahad_trb.h
@@ -610,7 +610,7 @@ struct trb_inform_type {
 
     /// \brief
     /// the maximum number of entries in the factors
-    long max_entries_factors;
+    int64_t max_entries_factors;
 
     /// \brief
     /// the total integer workspace required for the factorization

[…]
