+++ include/galahad_trb.h
@@ -215,6 +215,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
