+++ include/hsl_ma77.h
@@ -44,7 +44,6 @@
 extern "C" {
 #else
 #include <stdbool.h>
-#include <stdint.h>
 #endif
 
 // include guard
