+++ include/spral_ssids.h
@@ -4,6 +4,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
