### 2022-11-21 ffbcf1a
Fix wrong type in header files
