+++ include/galahad_trs.h
@@ -502,7 +502,7 @@ struct trs_inform_type {
 
     /// \brief
     /// the maximum number of entries in the factors
-    int max_entries_factors;
+    int64_t max_entries_factors;
 
     /// \brief
     /// the number of \f$(||x||_M,\lambda)\f$ pairs in the history
