+++ include/galahad_nls.h
@@ -1134,7 +1134,7 @@ struct nls_inform_type {
 
     /// \brief
     /// the maximum number of entries in the factors
-    int max_entries_factors;
+    int64_t max_entries_factors;
 
     /// \brief
     /// the total integer workspace required for the factorization
