+++ include/galahad_bsc.h
@@ -130,7 +130,6 @@
 extern "C" {
 #else
 #include <stdbool.h>
-#include <stdint.h>
 #endif
 
 // include guard
