+++ include/galahad_trs.h
@@ -243,6 +243,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
