### 2022-11-21 f2f75e4
more changes to correctly specify long integers in C. factorization_integer/real varrables now properly filled
