+++ include/galahad_trb.h
@@ -16,17 +16,17 @@
  */
 
 /*! \mainpage GALAHAD C package trb
- 
+
   \section trb_intro Introduction
 
   \subsection trb_purpose Purpose
 
   The trb package uses a <b>trust-region method to find a (local)
-   minimizer of a differentiable objective function 
+   minimizer of a differentiable objective function
   \f$\mathbf{f(x)}\f$ of many variables \f$\mathbf{x}\f$,
-  where the variables satisfy the simple bounds 
-  \f$\mathbf{x^l \leq x \leq x^u}\f$.</b> 
-  The method offers the choice of 
+  where the variables satisfy the simple bounds
+  \f$\mathbf{x^l \leq x \leq x^u}\f$.</b>
+  The method offers the choice of
   direct and iterative solution of the key subproblems, and
   is most suitable for large problems. First derivatives are required,
   and if second derivatives can be calculated, they will be exploited---if
@@ -44,30 +44,30 @@
 
   \subsection trb_terminology Terminology
 
-  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose 
+  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose
   \f$i\f$-th component is \f$\partial f(x)/\partial x_i\f$.
-  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix 
+  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix
   whose \f$i,j\f$-th entry is \f$\partial^2 f(x)/\partial x_i \partial x_j\f$.
-  The Hessian is \e sparse if a significant and useful proportion of the 
+  The Hessian is \e sparse if a significant and useful proportion of the
   entries are universally zero.
 
   \subsection trb_method Method
 
-  A trust-region method is used. In this, an improvement to a current estimate 
-  of the required minimizer, \f$x_k\f$ is sought by computing a step \f$s_k\f$. 
-  The step is chosen to approximately minimize a model \f$m_k(s)\f$ of 
- \f$f(x_k + s)\f$ within the intersection of the bound  constraints 
-  \f$x^l \leq x \leq x^u\f$ and a trust region \f$\|s_k\| \leq \Delta_k\f$ 
+  A trust-region method is used. In this, an improvement to a current estimate
+  of the required minimizer, \f$x_k\f$ is sought by computing a step \f$s_k\f$.
+  The step is chosen to approximately minimize a model \f$m_k(s)\f$ of
+ \f$f(x_k + s)\f$ within the intersection of the bound  constraints
+  \f$x^l \leq x \leq x^u\f$ and a trust region \f$\|s_k\| \leq \Delta_k\f$
   for some specified positive "radius" \f$\Delta_k\f$. The quality of the
   resulting step \f$s_k\f$ is assessed by computing the "ratio"
-  \f$(f(x_k) - f(x_k + s_k))/ (m_k(0) - m_k(s_k))\f$. The step is deemed 
+  \f$(f(x_k) - f(x_k + s_k))/ (m_k(0) - m_k(s_k))\f$. The step is deemed
   to have succeeded if the ratio exceeds a given \f$\eta_s > 0\f$,
   and in this case \f$x_{k+1} = x_k + s_k\f$. Otherwise
   \f$x_{k+1} = x_k\f$, and the radius is reduced by powers of a given
   reduction factor until it is smaller than \f$\|s_k\|\f$. If the ratio is
   larger than  \f$\eta_v \geq \eta_d\f$, the radius will be increased so that
[…]
