+++ src/dum/mpi.f90
@@ -2,11 +2,12 @@
 
 !-*-*-*-  G A L A H A D  -  D U M M Y   M P I   S U B R O U T I N E  -*-*-*-*-
 
-  SUBROUTINE MPI_INIT( ierr )
-  INTEGER ierr
-  ierr = - 999
-  END SUBROUTINE MPI_INIT
+   SUBROUTINE MPI_INIT( ierr )
+   INTEGER ierr
+   ierr = - 1 ! error code
+   END SUBROUTINE MPI_INIT
 
-  SUBROUTINE MPI_FINALIZE( ierr )
-  INTEGER ierr
-  END SUBROUTINE MPI_FINALIZE
+   SUBROUTINE MPI_FINALIZE( ierr )
+   INTEGER ierr
+   ierr = - 1 ! error code
+   END SUBROUTINE MPI_FINALIZE
