+++ meson_options.txt
@@ -1,0 +1,1 @@
+option('ciface', type : 'boolean', value : false)
