+++ include/galahad_hash.h
@@ -46,6 +46,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
