+++ include/galahad_qpb.h
@@ -367,6 +367,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
