+++ include/galahad_l2rt.h
@@ -198,6 +198,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
