+++ include/galahad_convert.h
@@ -40,6 +40,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
