+++ include/galahad_sils.h
@@ -105,6 +105,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
