+++ include/galahad_fit.h
@@ -43,6 +43,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
