+++ include/galahad_bgo.h
@@ -235,6 +235,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
