+++ include/galahad_wcp.h
@@ -444,6 +444,7 @@ $$
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
