+++ include/galahad_lms.h
@@ -87,6 +87,7 @@ but the crucial observation is that it is possible to construct
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
