+++ include/galahad_cro.h
@@ -225,6 +225,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
