+++ include/galahad_lpb.h
@@ -255,6 +255,7 @@
 extern "C" {
 #else
 #include <stdbool.h>
+#include <stdint.h>
 #endif
 
 // include guard
