### 2022-11-23 2b94d2c
updates to C comments
