+++ include/galahad_tru.h
@@ -224,40 +224,50 @@ extern "C" {
  */
 struct tru_control_type {
 
+    /// \brief 
     /// use C or Fortran sparse matrix indexing
     bool f_indexing;
 
+    /// \brief 
     /// error and warning diagnostics occur on stream error
     int error;
 
+    /// \brief 
     /// general output occurs on stream out
     int out;
 
-    /// \brief the level of output required.
+    /// \brief 
+    /// the level of output required.
     /// \li \f$\leq\f$ 0 gives no output,
     /// \li  = 1 gives a one-line summary for every iteration,
     /// \li  = 2 gives a summary of the inner iteration for each iteration,
     /// \li \f$\geq\f$ 3 gives increasingly verbose (debugging) output
     int print_level;
 
+    /// \brief 
     /// any printing will start on this iteration
     int start_print;
 
+    /// \brief 
     /// any printing will stop on this iteration
     int stop_print;
 
+    /// \brief 
     /// the number of iterations between printing
     int print_gap;
 
+    /// \brief 
     /// the maximum number of iterations allowed
     int maxit;
 
+    /// \brief 
     /// removal of the file alive_file from unit alive_unit terminates execution
     int alive_unit;
     /// see alive_unit
     char alive_file[31];
 
-    /// \brief the descent strategy used.
+    /// \brief 
+    /// the descent strategy used.
     ///
     /// Possible values are
     /// \li <= 0 a monotone strategy is used.
@@ -265,7 +275,8 @@ struct tru_control_type {
     /// .non_monotine is used.
     int non_monotone;
 
-    ///\brief  the model used.
+    /// \brief  
+    /// the model used.
     ///
     /// Possible values are
     /// \li 0  dynamic (*not yet implemented*)
@@ -306,17 +317,20 @@ struct tru_control_type {
     /// \li 10  diagonalizing norm from GALAHAD_DPS (*subproblem_direct only*)
     int norm;
 
-    /// \brief specify the semi-bandwidth of the band matrix \f$P\f$ if required
+    /// \brief 
+    /// specify the semi-bandwidth of the band matrix \f$P\f$ if required
     int semi_bandwidth;
 
-    /// \brief number of vectors used by the L-BFGS matrix \f$P\f$ if required
+    /// \brief 
+    /// number of vectors used by the L-BFGS matrix \f$P\f$ if required
     int lbfgs_vectors;
 
     /// \brief
     /// number of vectors used by the sparsity-based secant Hessian if required
     int max_dxg;
 
-    /// \brief number of vectors used by the Lin-More' incomplete factorization
+    /// \brief 
+    /// number of vectors used by the Lin-More' incomplete factorization
     /// matrix \f$P\f$ if required
     int icfs_vectors;
 
@@ -350,15 +364,18 @@ struct tru_control_type {
     /// see stop_g_absolute
     real_wp_ stop_s;
 
-    /// \brief try to pick a good initial trust-region radius using
+    /// \brief 
+    /// try to pick a good initial trust-region radius using
     /// .advanced_start iterates of a variant on the strategy of
     /// Sartenaer SISC 18(6) 1990:1788-1803
     int advanced_start;
 
-    /// \brief initial value for the trust-region radius
+    /// \brief 
+    /// initial value for the trust-region radius
     real_wp_ initial_radius;
 
-    /// \brief maximum permitted trust-region radius
+    /// \brief 
+    /// maximum permitted trust-region radius
     real_wp_ maximum_radius;
 
     /// \brief a potential iterate will only be accepted if the actual decrease
@@ -388,10 +405,12 @@ struct tru_control_type {
     /// is marked as unbounded
     real_wp_ obj_unbounded;
 
-    /// \brief the maximum CPU time allowed (-ve means infinite)
+    /// \brief 
[…]
