+++ doc/Python/translate/time.f90
@@ -16,6 +16,7 @@ program time
 ! read in next line
     line =  REPEAT( ' ', 100 )
     read( in, "( A )", end = 9, err = 9 ) line
+    line = ADJUSTL( line ) 
     line_length = len( trim( line ) ) 
 ! interpret line
     if ( len( trim( line ) ) == 0 ) cycle ! blank line
@@ -31,6 +32,12 @@ program time
       else if ( line( i : i + 3 ) == 'TYPE' ) THEN
         type = 't'
       else if ( line( i : i + 2 ) == ':: ' ) THEN
+        do j = i + 3, line_length
+          if ( line( j : j ) == ' ' ) then
+            line_length = j - 1
+            exit
+          end if
+        end do
         name =  REPEAT( ' ', 120 )
         name( 1 : line_length - i - 2 ) = line( i + 3 : line_length )
 !       write( out, "( A, ' ', A )" ) type, trim( name )
