### 2022-11-23 #4882 @frankecurtis
@melissawm @nimgould I am done with my review.

I previously communicated to @nimgould about an issue that I had during the installation process (since I was following the Homebrew instructions, which are apparently out of date).  I did not submit an issue about this to the GALAHAD repository, but @nimgould is aware of it nonetheless.

In terms of the review section "Documentation > A statement of need", I suppose I was expecting to see the statement of need written more upfront in the documentation; instead, I only found it here: https://github.com/ralna/GALAHAD/blob/master/doc/paper_v4/paper.md
Still, it was present, so I checked the box.

Finally, I did encounter one issue when running the general tests for my installation.  I submitted an issue on the GALAHAD repository.  But it doesn't seem necessary to block the review for this.  Overall, my review is complete.  Thanks.
