### 2022-11-23 f11bc9b
improvements to workflow
- specify compilers as env vars
- upload log as artifact in case of error
