### 2022-11-23 480454a
improvements to build script
- specify version and default options
- begin HSL integration
