### 2022-11-24 2b4ca52
solver cdqp added to dgal. mumps dummy interface improved
