+++ src/nls/Python/nls_pyiface.c
@@ -1,0 +1,943 @@
+//* \file nls_pyiface.c */
+
+/*
+ * THIS VERSION: GALAHAD 4.1 - 2022-10-13 AT 14:30 GMT.
+ *
+ *-*-*-*-*-*-*-*-*-  GALAHAD_NLS PYTHON INTERFACE  *-*-*-*-*-*-*-*-*-*-
+ *
+ *  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
+ *  Principal author: Jaroslav Fowkes & Nick Gould
+ *
+ *  History -
+ *   originally released GALAHAD Version 4.1. September 21st 2022
+ *
+ *  For full documentation, see
+ *   http://galahad.rl.ac.uk/galahad-www/specs.html
+ */
+
+#include "galahad_python.h"
+#include "galahad_nls.h"
+
+/* Nested TRS, GLTR, PSLS, LMS and SHA control and inform prototypes */
+//bool trs_update_control(struct trs_control_type *control,
+//                        PyObject *py_options);
+//PyObject* trs_make_inform_dict(const struct trs_inform_type *inform);
+//bool gltr_update_control(struct gltr_control_type *control,
+//                         PyObject *py_options);
+//PyObject* gltr_make_inform_dict(const struct gltr_inform_type *inform);
+//bool psls_update_control(struct psls_control_type *control,
+//                         PyObject *py_options);
+//PyObject* psls_make_inform_dict(const struct psls_inform_type *inform);
+//bool lms_update_control(struct lms_control_type *control,
+//                        PyObject *py_options);
+//PyObject* lms_make_inform_dict(const struct lms_inform_type *inform);
+//bool sha_update_control(struct sha_control_type *control,
+//                        PyObject *py_options);
+//PyObject* sha_make_inform_dict(const struct sha_inform_type *inform);
+
+/* Module global variables */
+static void *data;                       // private internal data
+static struct nls_control_type control;  // control struct
+static struct nls_inform_type inform;    // inform struct
+static bool init_called = false;         // record if initialise was called
+static int status = 0;                   // exit status
+
+//  *-*-*-*-*-*-*-*-*-*-   CALLBACK FUNCTIONS    -*-*-*-*-*-*-*-*-*-*
+
+/* Python eval_* function pointers */
+static PyObject *py_eval_f = NULL;
+static PyObject *py_eval_g = NULL;
+static PyObject *py_eval_h = NULL;
+
+/* C eval_* function wrappers */
+static int eval_f(int n, const double x[], double *f, const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyObject *py_x = PyArray_SimpleNewFromData(1, xdim, NPY_DOUBLE, (void *) x);
+
+    // Build Python argument list
+    PyObject *arglist = Py_BuildValue("(O)", py_x);
+
+    // Call Python eval_f
+    PyObject *result = PyObject_CallObject(py_eval_f, arglist);
+    Py_DECREF(py_x);    // Free py_x memory
+    Py_DECREF(arglist); // Free arglist memory
+
+    // Check that eval was successful
+    if(!result)
+        return -1;
+
+    // Extract single double return value
+    if(!parse_double("eval_f return value", result, f)){
+        Py_DECREF(result); // Free result memory
+        return -1;
+    }
+
+    // Free result memory
+    Py_DECREF(result);
+
+    return 0;
+}
+
+static int eval_g(int n, const double x[], double g[], const void *userdata){
+
+    // Wrap input array as NumPy array
+    npy_intp xdim[] = {n};
+    PyArrayObject *py_x = (PyArrayObject*)
[…]
