### 2022-11-24 8859972
added arc and tru python interfaces plus docimentation
