+++ src/trb/trb.f90
@@ -266,7 +266,7 @@
 
 !   overall convergence tolerances. The iteration will terminate when the
 !     norm of the gradient of the objective function is smaller than
-!       MAX( %stop_pg_absolute, %stop_pg_relative * norm of the initial gradient
+!      MAX( %stop_pg_absolute, %stop_pg_relative * norm of the initial gradient)
 !     or if the step is less than %stop_s
 
        REAL ( KIND = wp ) :: stop_pg_absolute = tenm5
