+++ src/arc/arc.f90
@@ -260,7 +260,7 @@
 
 !   overall convergence tolerances. The iteration will terminate when the
 !     norm of the gradient of the objective function is smaller than
-!       MAX( %stop_g_absolute, %stop_g_relative * norm of the initial gradient
+!       MAX( %stop_g_absolute, %stop_g_relative * norm of the initial gradient )
 !     or if the step is less than %stop_s
 
        REAL ( KIND = wp ) :: stop_g_absolute = tenm5
