### 2022-11-24 ae620ab
updated hsl c interfaces to account for hsl name repairs
