### 2022-11-24 87f5c52
removed any reference to minpacks dgt. fixed wrong integer types for eqp
