+++ src/dum/cutest_dummy.f90
@@ -20,9 +20,6 @@
    SUBROUTINE GROUP()
    END SUBROUTINE GROUP
 
-   SUBROUTINE dgqt()
-   END SUBROUTINE dgqt
-
    MODULE GALAHAD_CUTEST_FUNCTIONS_double
 
      IMPLICIT NONE
