+++ src/trs/usetrs.f90
@@ -96,8 +96,6 @@
      CHARACTER ( LEN = 30 ) :: ms_rfilename = 'MSRES.d'
      CHARACTER ( LEN = 30 ) :: ms_sfilename = 'MSSOL.d'
      REAL ( KIND = wp ) ::  radius = 1.0_wp
-     LOGICAL :: more_sorensen = .FALSE.
-     LOGICAL :: trs = .TRUE.
 !    LOGICAL :: one_norm = .TRSE.
 
 !  Output file characteristics
@@ -136,8 +134,6 @@
        spec( 11 )%keyword  = 'ms-solution-file-name'
        spec( 12 )%keyword  = 'ms-solution-file-device'
        spec( 13 )%keyword = 'radius'
-       spec( 14 )%keyword = 'use-more-sorensen'
-       spec( 15 )%keyword = 'use-trs'
 
 !   Read the specfile
 
@@ -158,8 +154,6 @@
        CALL SPECFILE_assign_string ( spec( 11 ), ms_sfilename, errout )
        CALL SPECFILE_assign_integer( spec( 12 ), ms_sfiledevice, errout )
        CALL SPECFILE_assign_real( spec( 13 ), radius, errout )
-       CALL SPECFILE_assign_logical( spec( 14 ), more_sorensen, errout )
-       CALL SPECFILE_assign_logical( spec( 15 ), trs, errout )
      END IF
 
 !  Set copyright
@@ -176,10 +170,6 @@
 
      CALL CUTEST_udimen( cutest_status, input, n )
      IF ( cutest_status /= 0 ) GO TO 910
-     IF ( n > 2000 .AND. more_sorensen ) THEN
-       WRITE( errout, "( ' more than 2000 variables ... stopping ' )" )
-       STOP
-     END IF
 
      ALLOCATE( X( n ), X0( n ), X_l( n ), X_u( n ), G( n ), VNAMES( n ) )
      CALL CUTEST_usetup( cutest_status, input, control%error, io_buffer,       &
@@ -201,253 +191,121 @@
      CALL CUTEST_ugr( cutest_status, n, X0, G )
      IF ( cutest_status /= 0 ) GO TO 910
 
-!  Use TRS
-
-     IF ( trs ) THEN
-       solv = 'TRS   '
+     solv = 'TRS   '
 
 !  Evaluate the Hessian
 
-       CALL CUTEST_udimsh( cutest_status, nnzh )
-       IF ( cutest_status /= 0 ) GO TO 910
-       H%ne = nnzh
-       CALL SMT_put( H%type, 'COORDINATE', smt_stat )
-       ALLOCATE( H%row( nnzh ), H%col( nnzh ), H%val( nnzh ) )
-       CALL CUTEST_ush( cutest_status, n, X0, H%ne, nnzh, H%val, H%row, H%col )
-       IF ( cutest_status /= 0 ) GO TO 910
+     CALL CUTEST_udimsh( cutest_status, nnzh )
+     IF ( cutest_status /= 0 ) GO TO 910
+     H%ne = nnzh
+     CALL SMT_put( H%type, 'COORDINATE', smt_stat )
+     ALLOCATE( H%row( nnzh ), H%col( nnzh ), H%val( nnzh ) )
+     CALL CUTEST_ush( cutest_status, n, X0, H%ne, nnzh, H%val, H%row, H%col )
+     IF ( cutest_status /= 0 ) GO TO 910
 
 !  If required, open a file for the results
 
-       IF ( write_result_summary ) THEN
-         INQUIRE( FILE = trs_rfilename, EXIST = filexx )
-         IF ( filexx ) THEN
-            OPEN( trs_rfiledevice, FILE = trs_rfilename, FORM = 'FORMATTED',   &
[…]
