### 2022-11-24 acaf366
options to specify BLAS/LAPACK libraries
