### 2022-11-25 43d23fa
unused variables removed
