+++ src/gls/glss.f90
@@ -1,8 +1,8 @@
-   PROGRAM GALAHAD_GLS_example  ! GALAHAD 3.3 - 06/05/2021 AT 10:15 GMT.
+   PROGRAM GALAHAD_GLS_example  ! GALAHAD 4.1 - 2022-11-25 AT 09:20 GMT.
    USE GALAHAD_GLS_DOUBLE
    IMPLICIT NONE
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) 
-   INTEGER :: i, info, rank
+   INTEGER :: info, rank
    INTEGER, PARAMETER :: m = 3
    INTEGER, PARAMETER :: n = 3
    INTEGER, PARAMETER :: ne = 7
