### 2022-11-25 f7e34e4
another unused variable removed
