### 2022-11-26 0015af5
CI: compile and test separately
