### 2022-11-26 7dfa588
added missing variable definition
