+++ src/dum/pardiso.f90
@@ -6,7 +6,7 @@
    USE GALAHAD_SYMBOLS
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )  
    INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
-   INTEGER ( KIND = long ), INTENT( INOUT ), DIMENSION( 64 ) :: PT
+   INTEGER ( KIND = C_INT64_T ), INTENT( INOUT ), DIMENSION( 64 ) :: PT
    INTEGER, INTENT( IN ) :: mtype, solver
    INTEGER, INTENT( INOUT ), DIMENSION( 64 ) :: IPARM
    REAL ( KIND = wp ), INTENT( INOUT ), DIMENSION( 64 ) :: DPARM
@@ -19,7 +19,7 @@
    USE GALAHAD_SYMBOLS
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )  
    INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
-   INTEGER ( KIND = long ), INTENT( INOUT ), DIMENSION( 64 ) :: PT
+   INTEGER ( KIND = C_INT64_T ), INTENT( INOUT ), DIMENSION( 64 ) :: PT
    INTEGER, INTENT( IN ) :: maxfct, mnum, mtype, phase, n, nrhs, msglvl
    INTEGER, INTENT( OUT ) :: error
    INTEGER, INTENT( INOUT ), DIMENSION( 64 ) :: IPARM
