### 2022-11-27 9a96907
many unused variables removed some progress on mumps interface to sls
