+++ src/external/mumps/mumps_types.f90
@@ -12,7 +12,8 @@
 
      IMPLICIT NONE
      PUBLIC
-     INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
+     INTEGER, PRIVATE, PARAMETER :: long = SELECTED_INT_KIND( 18 )
+     INTEGER, PARAMETER :: MPI_COMM_WORLD = 0
 
 !  include the current mumps derived types
 
