+++ src/presolve/presolve.f90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-09-28 AT 15:25 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:45 GMT.
 
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -426,8 +426,7 @@
       REAL ( KIND = wp ), PRIVATE, PARAMETER :: ZERO    = 0.0_wp
       REAL ( KIND = wp ), PRIVATE, PARAMETER :: HALF    = 0.5_wp
       REAL ( KIND = wp ), PRIVATE, PARAMETER :: ONE     = 1.0_wp
-      REAL ( KIND = wp ), PRIVATE, PARAMETER :: TWO     = 2.0_wp
-      REAL ( KIND = wp ), PRIVATE, PARAMETER :: TEN = 10.0_wp
+      REAL ( KIND = wp ), PRIVATE, PARAMETER :: TEN     = 10.0_wp
       REAL ( KIND = sp ), PRIVATE, PARAMETER :: TEN_sp  = 10.0_sp
       REAL ( KIND = dp ), PRIVATE, PARAMETER :: TEN_dp  = 10.0_dp
       REAL ( KIND = wp ), PRIVATE, PARAMETER :: HUNDRED = TEN * TEN
@@ -441,18 +440,9 @@
 
       INTEGER, PRIVATE, PARAMETER :: DEF_WRITE_UNIT            = 6
 
-!     Default error output unit number
-
-      INTEGER, PRIVATE, PARAMETER :: DEF_ERROR_UNIT            = 6
-
 !     Default printout level
 
       INTEGER, PRIVATE, PARAMETER :: DEF_PRINT_LEVEL           = SILENT
-
-!     Value beyond which a number is equal to plus infinity
-
-      REAL( KIND = sp ), PRIVATE, PARAMETER :: DEF_INFINITY_sp = TEN_sp ** 19
-      REAL( KIND = dp ), PRIVATE, PARAMETER :: DEF_INFINITY_dp = TEN_dp ** 19
 
 !     Default maximum number of analysis passes for a single call to PRESOLVE
 
@@ -484,28 +474,22 @@
 
       REAL ( KIND = sp ), PRIVATE, PARAMETER ::                                &
                                    DEF_PIVOT_TOL_sp = TEN_sp ** ( -6 )
-      REAL ( KIND = dp ), PRIVATE, PARAMETER ::                                &
-                                   DEF_PIVOT_TOL_dp = TEN_dp ** (-10 )
 
 !     Default minimum relative bound improvement
 
       REAL ( KIND = sp ), PRIVATE, PARAMETER :: DEF_MRBI_sp = TEN_sp ** ( -6 )
-      REAL ( KIND = dp ), PRIVATE, PARAMETER :: DEF_MRBI_dp = TEN_dp ** ( -10 )
 
 !     Default maximum growth factor between original and reduced problems
 
       REAL ( KIND = sp ), PRIVATE, PARAMETER :: DEF_MAX_GROWTH_sp = TEN_sp ** 4
-      REAL ( KIND = dp ), PRIVATE, PARAMETER :: DEF_MAX_GROWTH_dp = TEN_dp ** 8
 
 !     Default relative accuracy for the linear constraints
 
[…]
