+++ src/sls/sls.f90
@@ -30,8 +30,10 @@
 !     |               MA87                        |
 !     |               MA97                        |
 !     |               SSIDS from SPRAL            |
+!     |               MUMPS                       |
 !     |               PARDISO                     |
 !     |               MKL PARDISO                 |
+!     |               PASTIX                      |
 !     |               WSMP                        |
 !     |               POTR from LAPACK            |
 !     |               SYTR from LAPACK            |
@@ -62,8 +64,7 @@
      USE HSL_MC68_integer
      USE MKL_PARDISO
      USE SPRAL_SSIDS
-!    USE spmf
-!    USE pastixf, MPI_COMM_WORLD_duplicate => MPI_COMM_WORLD
+     USE GALAHAD_MUMPS_TYPES_double, MPI_COMM_WORLD_mumps => MPI_COMM_WORLD
      USE spmf_enums, MPI_COMM_WORLD_pastix => MPI_COMM_WORLD
      USE spmf_interfaces
      USE pastixf_enums, MPI_COMM_WORLD_pastix_duplic8 => MPI_COMM_WORLD
@@ -716,6 +717,10 @@
 
        INTEGER :: pastix_info = 0
 
+!  the output scalar from mpi
+
+       INTEGER :: mpi_ierr = 0
+
 !  the output scalars and arrays from LAPACK routines
 
        INTEGER :: lapack_error = 0
@@ -736,6 +741,8 @@
        INTEGER :: set_res = - 1
        INTEGER :: set_res2 = - 1
        LOGICAL :: got_maps_scale = .FALSE.
+       LOGICAL :: no_mpi = .FALSE.
+       LOGICAL :: no_mumps = .FALSE.
        LOGICAL :: no_pastix = .FALSE.
        LOGICAL :: no_sils = .FALSE.
        LOGICAL :: no_ma57 = .FALSE.
@@ -828,6 +835,7 @@
        INTEGER, POINTER, DIMENSION( : ) :: ROW, COL, PTR
        REAL ( KIND = c_double ), POINTER, DIMENSION( : ) :: VAL
        TYPE ( pastix_data_t ), POINTER :: pastix_data
+       TYPE ( DMUMPS_STRUC ) :: mumps_par
        TYPE ( spmatrix_t ), POINTER :: spm, spm_check
        TYPE ( pastix_order_t ), POINTER :: order_pastix => NULL( )
        INTEGER ( kind = pastix_int_t ), DIMENSION( : ), POINTER :: PERMTAB
@@ -1114,7 +1122,7 @@
 !  local variables
 
      INTEGER :: flag_ssids
-     LOGICAL :: check_available
+     LOGICAL :: check_available, mpi_initialzed_flag
      INTEGER, DIMENSION( 30 ) :: ICNTL_ma27
      REAL ( KIND = wp ), DIMENSION( 5 ) :: CNTL_ma27
      TYPE ( MA57_control ) :: control_ma57
@@ -1150,7 +1158,7 @@
            data%solver( 1 : data%len_solver ) = 'ssids'
            GO TO 10
          ELSE
-           inform%status = GALAHAD_error_unknown_solver ; RETURN
+           inform%status = GALAHAD_unavailable_option ; RETURN
          END IF
        END IF
        data%must_be_definite = .FALSE.
@@ -1169,7 +1177,7 @@
            data%solver( 1 : data%len_solver ) = 'sils'
            GO TO 10
          ELSE
-           inform%status = GALAHAD_error_unknown_solver ; RETURN
+           inform%status = GALAHAD_unavailable_option ; RETURN
          END IF
        END IF
        data%must_be_definite = .FALSE.
@@ -1200,7 +1208,7 @@
 
      CASE ( 'ssids' )
[…]

[…]
