### 2022-12-03 a8720dc
icfs added to GALAHAD by pernission, and extended to modern fortran. mumps stub now provided in sls. documentation updated to mention MPI .
