+++ src/sls/sls.f90
@@ -713,6 +713,12 @@
        INTEGER, DIMENSION( 64 ) :: wsmp_iparm = - 1
        REAL ( KIND = wp ), DIMENSION( 64 ) :: wsmp_dparm = - 1.0_wp
 
+!  the output scalars and arrays from mumps
+
+       INTEGER :: mumps_error = 0
+       INTEGER, DIMENSION( 80 ) :: mumps_info = - 1
+       REAL ( KIND = wp ), DIMENSION( 40 ) :: mumps_rinfo = - 1.0_wp
+
 !  the output scalar from pastix
 
        INTEGER :: pastix_info = 0
@@ -835,13 +841,13 @@
        INTEGER, POINTER, DIMENSION( : ) :: ROW, COL, PTR
        REAL ( KIND = c_double ), POINTER, DIMENSION( : ) :: VAL
        TYPE ( pastix_data_t ), POINTER :: pastix_data
-       TYPE ( DMUMPS_STRUC ) :: mumps_par
        TYPE ( spmatrix_t ), POINTER :: spm, spm_check
        TYPE ( pastix_order_t ), POINTER :: order_pastix => NULL( )
        INTEGER ( kind = pastix_int_t ), DIMENSION( : ), POINTER :: PERMTAB
        INTEGER ( KIND = pastix_int_t ) :: iparm_pastix( 75 )
 !      REAL ( KIND = c_double ) :: dparm_pastix( 24 )
        REAL ( KIND = KIND( 1.0D+2 ) ) :: dparm_pastix( 24 )
+       TYPE ( DMUMPS_STRUC ) :: mumps_par
 
      END TYPE SLS_data_type
 
@@ -1038,6 +1044,10 @@
 !  = PaStiX =
 
      CASE ( 'pastix' )
+
+!  = MUMPS =
+
+     CASE ( 'mumps' )
 
 !  = POTR =
 
@@ -1283,25 +1293,22 @@
 
      CASE ( 'mumps' )
        CALL MPI_INITIALIZED( mpi_initialzed_flag, inform%mpi_ierr )
-!write(6,*) ' mpi initialised? ', mpi_initialzed_flag
        IF ( mpi_initialzed_flag ) THEN
          data%no_mpi = .FALSE.
        ELSE
          CALL MPI_INIT( inform%mpi_ierr )
          data%no_mpi = inform%mpi_ierr < 0
        END IF 
-!write(6,*) ' mpi? ', .NOT. data%no_mpi
 
        IF ( data%no_mpi ) THEN
          data%no_mumps = .TRUE.
        ELSE
          data%mumps_par%COMM = MPI_COMM_WORLD_mumps
          data%mumps_par%JOB = - 1
-         data%mumps_par%SYM = 1
-         data%mumps_par%PAR = 1
+         data%mumps_par%SYM = 2 ! symmetric
+         data%mumps_par%PAR = 1 ! parallel solve
          CALL DMUMPS( data%mumps_par )
          data%no_mumps = data%mumps_par%INFOG( 1 ) == - 999
-!        IF ( data%no_mumps ) CALL MPI_FINALIZE( inform%mpi_ierr )
        END IF
 
        IF ( data%no_mumps ) THEN
@@ -1906,6 +1913,97 @@
 !  End of SLS_copy_control_to_wsmp
 
      END SUBROUTINE SLS_copy_control_to_wsmp
+
+!-*-  S L S _ C O P Y _ C O N T R O L _ T O _ M U M P S  S U B R O U T I N E -*-
+
+     SUBROUTINE SLS_copy_control_to_mumps( control, ICNTL, CNTL )
+
+!  copy control parameters to their MUMPS equivalents
+
+!  Dummy arguments
+
[…]
