+++ src/dum/include/dmumps_struc.h
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-24 AT 09:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-01 AT 13:00 GMT.
 
       TYPE DMUMPS_STRUC( long )
       INTEGER, KIND :: long
@@ -14,11 +14,17 @@
         INTEGER :: MYID
         INTEGER :: N
         INTEGER :: NZ
+        INTEGER :: NRHS
+        INTEGER :: LRHS
         INTEGER( KIND = long ) :: NNZ
         DOUBLE PRECISION, DIMENSION( : ), POINTER :: A
         INTEGER, DIMENSION( : ), POINTER :: IRN, JCN
         DOUBLE PRECISION, DIMENSION( : ), POINTER :: RHS
         INTEGER, DIMENSION( : ), POINTER :: PERM_IN
         DOUBLE PRECISION, DIMENSION( : ), POINTER :: COLSCA, ROWSCA
-        INTEGER :: INFOG( 80 )
+        INTEGER, DIMENSION( : ), POINTER :: SYM_PERM, UNS_PERM
+        INTEGER, DIMENSION( 60 ) ::  ICNTL
+        DOUBLE PRECISION, DIMENSION( 15 ) :: CNTL
+        INTEGER, DIMENSION( 80 ) :: INFOG
+        DOUBLE PRECISION, DIMENSION( 40 ) ::  RINFOG
       END TYPE DMUMPS_STRUC
