+++ src/dum/mumps.f90
@@ -5,6 +5,7 @@
       SUBROUTINE DMUMPS( mumps_par )
       USE GALAHAD_MUMPS_TYPES_double
       IMPLICIT NONE
+      INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
       TYPE ( DMUMPS_STRUC( long = long ) ) :: mumps_par
       mumps_par%INFOG( 1 ) = - 999  ! error code
       RETURN

[…]
