### 2022-12-03 98eb98a
missing kind for DMUMPS_STRUC added
