### 2022-12-03 9c9cbc3
gfortran misbehaving with parameterised derived types ... another try ...
