+++ src/dum/include/dmumps_struc.h
@@ -1,7 +1,5 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-01 AT 13:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-03 AT 14:20 GMT.
 
-!     TYPE DMUMPS_STRUC( long )
-!       INTEGER, KIND :: long
       TYPE DMUMPS_STRUC
 
 !  This dummy structure contains a subset of the parameters for the
@@ -17,7 +15,6 @@
         INTEGER :: NZ
         INTEGER :: NRHS
         INTEGER :: LRHS
-!       INTEGER( KIND = long ) :: NNZ
         INTEGER( KIND = SELECTED_INT_KIND( 18 ) ) :: NNZ
         DOUBLE PRECISION, DIMENSION( : ), POINTER :: A
         INTEGER, DIMENSION( : ), POINTER :: IRN, JCN
