+++ src/sls/sls.f90
@@ -847,7 +847,6 @@
        INTEGER ( KIND = pastix_int_t ) :: iparm_pastix( 75 )
 !      REAL ( KIND = c_double ) :: dparm_pastix( 24 )
        REAL ( KIND = KIND( 1.0D+2 ) ) :: dparm_pastix( 24 )
-!      TYPE ( DMUMPS_STRUC( long = long ) ) :: mumps_par
        TYPE ( DMUMPS_STRUC ) :: mumps_par
 
      END TYPE SLS_data_type
@@ -864,7 +863,6 @@
        SUBROUTINE DMUMPS( mumps_par )
        USE GALAHAD_MUMPS_TYPES_double
        INTEGER, PARAMETER :: long_kind = SELECTED_INT_KIND( 18 )
-!      TYPE ( DMUMPS_STRUC( long = long_kind ) ) :: mumps_par
        TYPE ( DMUMPS_STRUC ) :: mumps_par
        END SUBROUTINE DMUMPS
      END INTERFACE
