+++ src/icfs/icfs.f90
@@ -63,7 +63,9 @@
       PRIVATE
       PUBLIC :: ICFS_initialize, ICFS_terminate, ICFS_read_specfile,           &
                 ICFS_factorize, ICFS_triangular_solve, DICFS, DSTRSOL,         &
-                ICFS_full_initialize, ICFS_full_terminate
+                ICFS_full_initialize, ICFS_full_terminate,                     &
+                ICFS_reset_control, ICFS_information,                          &
+                ICFS_factorize_matrix, ICFS_solve_system
 
 !--------------------
 !   P r e c i s i o n
@@ -105,7 +107,7 @@
 
         INTEGER :: print_level = 0
 
-!  number of extra vectors of length n required by the incomplete Cholesky 
+!  number of extra vectors of length n required by the incomplete Cholesky
 !  factorization
 
         INTEGER :: icfs_vectors = 10
@@ -218,6 +220,8 @@
         TYPE ( ICFS_data_type ) :: ICFS_data
         TYPE ( ICFS_control_type ) :: ICFS_control
         TYPE ( ICFS_inform_type ) :: ICFS_inform
+        INTEGER, ALLOCATABLE, DIMENSION( : ) :: ptr
+        INTEGER, ALLOCATABLE, DIMENSION( : ) :: row
       END TYPE ICFS_full_data_type
 
   CONTAINS
@@ -413,23 +417,23 @@
 !  n is an integer variable that gives the order of A (number of rows/columns)
 !
 !  PTR is an integer array of dimension n + 1, whose j-th component gives the
-!   starting address for list of nonzero values and their corresponding row 
-!   indices in column j of the strict lower triangular part of A. That is, 
-!   the nonzeros in column j of the strict lower triangle of A must be in 
+!   starting address for list of nonzero values and their corresponding row
+!   indices in column j of the strict lower triangular part of A. That is,
+!   the nonzeros in column j of the strict lower triangle of A must be in
 !   positions
 !            PTR(j), ... , PTR(j+1) - 1.
 !   Note that PTR(n+1) points to the first position beyond that needed to
 !   store A
 !
 !  ROW is an integer array of dimension at least PTR(n+1)-1 that contains the
-!   row indices of the strict lower triangular part of A in the compressed 
+!   row indices of the strict lower triangular part of A in the compressed
 !   column storage format
 !
-!  DIAG is a real array of dimension at least n whose j-th component 
+!  DIAG is a real array of dimension at least n whose j-th component
 !   contains the value of the j-th diagonal of A
 !
 !  VAL is a real array of dimension at least PTR(n+1)-1 that contains the
-!   values of the strict lower triangular part of A in the compressed 
+!   values of the strict lower triangular part of A in the compressed
 !   column storage format, in the same order as ROW
 !
 !  data is a structure of type ICFS_data_type which holds private internal data
@@ -595,7 +599,7 @@
 
 !  n is an integer variable that gives the order of A (number of rows/columns)
 !
-!  R is a real array of dimension at least n whose j-th component on input 
+!  R is a real array of dimension at least n whose j-th component on input
 !   contains the value of the j-th component of the vector r, and that will
 !   be replaced on output by that of the j-th component of the solution x
 !
@@ -826,7 +830,7 @@
 !         On exit acol_ptr is unchanged.
 !
 !       arow_ind is an integer array of dimension nnz.
-!         On entry arow_ind must contain row indices for the strict 
+!         On entry arow_ind must contain row indices for the strict
 !            lower triangular part of A in compressed column storage.
 !         On exit arow_ind is unchanged.
 !
@@ -848,7 +852,7 @@
 !       lrow_ind is an integer array of dimension nnz+n*p.
[…]
