+++ doc/src/icfs.tex
@@ -14,8 +14,8 @@
 %%%%%%%%%%%%%%%%%%%%%% SUMMARY %%%%%%%%%%%%%%%%%%%%%%%%
 
 \galsummary
-Given a symmetric matrix $\bmA$, this package 
-{\bf computes a symmetric, positive-definite approximation 
+Given a symmetric matrix $\bmA$, this package
+{\bf computes a symmetric, positive-definite approximation
 $\bmL \bmL^T$ using an incomplete Cholesky factorization}; the
 resulting matrix $\bmL$ is lower triangular.
 Subsequently, the solution $\bmx$ to the either of the linear systems
@@ -26,14 +26,14 @@ may be found for a given vector $\bmb$.
 
 \galattributes
 \galversions{\tt  \fullpackagename\_single, \fullpackagename\_double}.
-\galuses 
-{\tt GALAHAD\_SY\-M\-BOLS}, 
+\galuses
+{\tt GALAHAD\_SY\-M\-BOLS},
 {\tt GALAH\-AD\-\_\-SP\-ECFILE} and
 {\tt GALAHAD\_SPACE}.
 \galdate May 1998/December 2022.
 \galorigin C.-J, Lin and J. J. Mor\'{e}, Argonne National Laboratory,
 enhanced for modern fortran by N. I. M. Gould, Rutherford Appleton Laboratory.
-\gallanguage Fortran~2003. 
+\gallanguage Fortran~2003.
 
 %%%%%%%%%%%%%%%%%%%%%% HOW TO USE %%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -57,20 +57,20 @@ Access to the package requires a {\tt USE} statement such as
 The user is {\bf strongly advised} to use the double
 precision version unless single precision corresponds to 8-byte arithmetic.
 
-If it is required to use both modules at the same time, the derived types 
-{\tt \packagename\_control\_type}, 
+If it is required to use both modules at the same time, the derived types
+{\tt \packagename\_control\_type},
 {\tt \packagename\_inform\_type},
 {\tt \packagename\_data\_type}
 and
 {\tt NLPT\_userdata\_type},
 (Section~\ref{galtypes})
 and the subroutines
-{\tt \packagename\_initialize}, 
+{\tt \packagename\_initialize},
 {\tt \packagename\_\-factorize},
 {\tt \packagename\_\-solve},
 {\tt \packagename\_terminate},
 (Section~\ref{galarguments})
-and 
+and
 {\tt \packagename\_read\_specfile}
 (Section~\ref{galfeatures})
 must be renamed on one of the {\tt USE} statements.
@@ -82,41 +82,41 @@ Four derived data types are accessible from the package.
 
 %%%%%%%%%%% control type %%%%%%%%%%%
 
-\subsubsection{The derived data type for holding control 
+\subsubsection{The derived data type for holding control
  parameters}\label{typecontrol}
-The derived data type 
-{\tt \packagename\_control\_type} 
-is used to hold controlling data. Default values may be obtained by calling 
+The derived data type
+{\tt \packagename\_control\_type}
+is used to hold controlling data. Default values may be obtained by calling
 {\tt \packagename\_initialize}
 (see Section~\ref{subinit}),
-while components may also be changed by calling 
+while components may also be changed by calling
 {\tt \fullpackagename\_read\-\_spec}
-(see Section~\ref{readspec}). 
-The components of 
-{\tt \packagename\_control\_type} 
+(see Section~\ref{readspec}).
+The components of
+{\tt \packagename\_control\_type}
 are:
 
 
 \begin{description}
 
 \itt{error} is a scalar variable of type default \integer, that holds the
[…]
