+++ include/galahad_icfs.h
@@ -1,0 +1,370 @@
+//* \file galahad_icfs.h */
+
+/*
+ * THIS VERSION: GALAHAD 4.1 - 2022-12-06 AT 09:20 GMT.
+ *
+ *-*-*-*-*-*-*-*-*-  GALAHAD_ICFS C INTERFACE  *-*-*-*-*-*-*-*-*-*-
+ *
+ *  Copyright reserved, GALAHAD productions
+ *  Principal authors: Jaroslav Fowkes & Nick Gould
+ *
+ *  History -
+ *   originally released GALAHAD Version 4.1. Decemeber 6th 2022
+ *
+ *  For full documentation, see
+ *   http://galahad.rl.ac.uk/galahad-www/specs.html
+ */
+
+/*! \mainpage GALAHAD C package icfs
+
+  \section icfs_intro Introduction
+
+  \subsection icfs_purpose Purpose
+
+  Given a symmetric matrix \f$\bmA\f$, this package
+  <b> computes a symmetric, positive-definite approximation
+  \f$\bmL \bmL^T\f$ using an incomplete Cholesky factorization</b>;
+  the resulting matrix \f$\bmL\f$ is lower triangular.
+  Subsequently, the solution \f$\bmx\f$ to the either of the linear systems
+  \f$\bmL \bmx = \bmb\f$ and \f$\bmL^T \bmx = \bmb\f$
+  may be found for a given vector \f$\bmb\f$.
+
+  \subsection icfs_authors Authors
+  C.-J, Lin and J. J. More', Argonne National Laboratory,
+
+  C interface, additionally N. I. M. Gould and J. Fowkes, STFC-Rutherford
+  Appleton Laboratory.
+
+ \subsection icfs_date Originally released
+
+  May 1998, C interface December 2022.
+ */
+
+#ifdef __cplusplus
+extern "C" {
+#else
+#include <stdbool.h>
+#include <stdint.h>
+#endif
+
+// include guard
+#ifndef GALAHAD_ICFS_H
+#define GALAHAD_ICFS_H
+
+// precision
+#include "galahad_precision.h"
+
+/**
+ * control derived type as a C struct
+ */
+struct icfs_control_type {
+
+    /// \brief
+    /// use C or Fortran sparse matrix indexing
+    bool f_indexing;
+
+    /// \brief
+    /// unit for error messages
+    int error;
+
+    /// \brief
+    /// unit for monitor output
+    int out;
+
+    /// \brief
+    /// controls level of diagnostic output
+    int print_level;
+
+    /// \brief
+    /// number of extra vectors of length n required by the incomplete
+    /// Cholesky preconditioner
+    int icfs_vectors;
+
+    /// \brief
+    /// an initial estimate of the shift $\alpha$ used so that the incomplete
+    /// factorization of \f$A + D\f$ is positive definite, where \f$D\f$ is
+    /// a diagonal matrix whose entries are no larger than \f$\alpha\f$.
+    /// This value \f$\alpha\f$ may subsequently be increased, as necessary,
+    /// by the package, see icfs_inform_type.shift.
+    real_wp_ shift;
+
+    /// \brief
+    /// if .space_critical true, every effort will be made to use as little
+    /// space as possible. This may result in longer computation time
+    bool space_critical;
+
+    /// \brief
+    /// if .deallocate_error_fatal is true, any array/pointer deallocation error
+    /// will terminate execution. Otherwise, computation will continue
+    bool deallocate_error_fatal;
+
+    /// \brief
+    /// all output lines will be prefixed by
+    /// prefix(2:LEN(TRIM(.prefix))-1)
+    /// where prefix contains the required string enclosed in quotes,
+    /// e.g. "string" or 'string'
[…]
