+++ include/galahad_sils.h
@@ -36,6 +36,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection sils_date Originally released
 
   April 2001, C interface December 2021.
