+++ include/galahad_dps.h
@@ -43,6 +43,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection dps_date Originally released
 
   August 2011, C interface December 2021.
