+++ include/galahad_trs.h
@@ -58,6 +58,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection trs_date Originally released
 
   October 2008, C interface December 2021.
