+++ include/galahad_fit.h
@@ -33,6 +33,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection fit_date Originally released
 
   March 2010, C interface January 2022.
