+++ include/galahad_qpb.h
@@ -62,6 +62,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection qpb_date Originally released
   December 1999, C interface January 2022.
 
