+++ include/galahad_ir.h
@@ -34,6 +34,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection ir_date Originally released
 
   October 2008, C interface January 2022
