+++ include/galahad_arc.h
@@ -38,6 +38,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection arc_date Originally released
 
   May 2011, C interface August 2021.
