+++ include/galahad_lpa.h
@@ -66,6 +66,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection lpa_date Originally released
   October 2018, C interface September 2021.
 
