+++ include/galahad_bqpb.h
@@ -55,6 +55,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection bqpb_date Originally released
 
   July 2021, C interface December 2021.
