+++ include/galahad_rqs.h
@@ -55,6 +55,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection rqs_date Originally released
 
   November 2008, C interface December 2021.
