+++ include/galahad_lhs.h
@@ -32,6 +32,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection lhs_date Originally released
 
   June 2016, C interface March 2022.
