+++ include/galahad_sbls.h
@@ -80,6 +80,8 @@ may be obtained by the package.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection sbls_date Originally released
   April 2006, C interface November 2021.
 
