+++ include/galahad_tru.h
@@ -37,6 +37,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection tru_date Originally released
 
   July 2008, C interface August 2021.
