+++ include/galahad_sec.h
@@ -32,6 +32,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection sec_date Originally released
 
   May 2008, C interface January 2022.
