+++ include/galahad_cro.h
@@ -62,6 +62,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection cro_date Originally released
 
   August 2010, C interface January 2022.
