+++ include/galahad_rpd.h
@@ -90,6 +90,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection rpd_date Originally released
 
   January 2006, C interface January 2022.
