### 2022-12-06 2bfcbe9
Update authors in header files
