+++ include/galahad_scu.h
@@ -64,6 +64,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection scu_date Originally released
 
   March 2005, C interface January 2022.
