+++ include/galahad_lpb.h
@@ -57,6 +57,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection lpb_date Originally released
   August 2018, C interface September 2021.
 
