+++ include/galahad_dqp.h
@@ -65,6 +65,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection dqp_date Originally released
   August 2012, C interface December 2021.
 
