+++ include/galahad_blls.h
@@ -50,6 +50,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection blls_date Originally released
 
   October 2019, C interface March 2022.
