+++ include/galahad_sha.h
@@ -32,6 +32,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection sha_date Originally released
 
   April 2013, C interface January 2022.
