+++ include/galahad_fdc.h
@@ -36,6 +36,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection fdc_date Originally released
 
   August 2006, C interface January 2021
