+++ include/galahad_bqp.h
@@ -49,6 +49,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection bqp_date Originally released
 
   November 2009, C interface February 2022.
