+++ include/galahad_clls.h
@@ -58,6 +58,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection clls_date Originally released
 
   July 2022.
