+++ include/galahad_gls.h
@@ -48,6 +48,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection gls_date Originally released
 
   March 2006, C interface December 2021.
