+++ include/galahad_ccqp.h
@@ -66,6 +66,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection ccqp_date Originally released
 
   November 2010, C interface September 2021.
