+++ include/galahad_slls.h
@@ -50,6 +50,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection slls_date Originally released
 
   October 2019, C interface July 2022.
