+++ include/galahad_sls.h
@@ -48,6 +48,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection sls_date Originally released
 
   August 2009, C interface December 2021.
