+++ include/galahad_ugo.h
@@ -33,6 +33,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection ugo_date Originally released
 
   July 2016, C interface August 2021.
