+++ include/galahad_trb.h
@@ -38,6 +38,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection trb_date Originally released
 
   July 2021, C interface August 2021.
