+++ include/galahad_bgo.h
@@ -43,6 +43,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection bgo_date Originally released
 
   July 2016, C interface August 2021.
