+++ include/galahad_bsc.h
@@ -35,6 +35,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection bsc_date Originally released
 
   October 2013, C interface January 2022.
