+++ include/galahad_lstr.h
@@ -39,6 +39,8 @@
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
 
+  Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
+
   \subsection lstr_date Originally released
 
   November 2007, C interface December 2021.
