### 2022-12-06 c6a5724
Update include/galahad.h
