+++ doc/paper_v4/paper.bib
@@ -36,7 +36,7 @@
 
 @article{GoulOrbaToin15,
     AUTHOR = {N.~I.~M. Gould and D. Orban and {\relax Ph}. L. Toint},
-    TITLE = {{CUTEst}: A Constrained and Unconstrained Testing Environment 
+    TITLE = {{CUTEst}: A Constrained and Unconstrained Testing Environment
              with Safe Threads for Mathematical Optimization},
    JOURNAL = {Computational Optimization and Applications},
     VOLUME = {60},
