+++ src/ssids/gpu_factor.f90
@@ -22,7 +22,7 @@ module spral_ssids_gpu_factor
   public :: parfactor ! Performs factorization phase using multiple streams
   ! C interfaces
   public :: spral_ssids_assign_nodes_to_levels
-  
+
   type :: ntype
      integer :: level = 1
      integer :: subtree = 0
@@ -82,7 +82,7 @@ contains
     type(C_PTR), optional, intent(in) :: ptr_scale
 
     type(C_PTR) :: gpu_LDLT
-   
+
     integer :: st
 
     type(cuda_settings_type) :: user_settings
@@ -109,7 +109,7 @@ contains
       stream_data, alloc, options, stats, ptr_scale)
     if (stats%flag .lt. 0) return
     if (stats%cuda_error .ne. 0) goto 200
-    
+
     ! Extract contribution to parent of subtree (if any)
     if (C_ASSOCIATED(gpu_LDLT)) then
        call transfer_contrib(nnodes, sptr, rptr, rlist, nodes, gpu_contribs, &
@@ -293,11 +293,11 @@ contains
     integer :: LDLT_size, max_LDLT_size
     integer(long) :: pc_size
     integer :: ncb, max_ncb
-  
+
     integer :: nch
     logical :: free_contrib
     integer :: p, q ! general purpose indices
-  
+
     ! per-node assembly info
     type(asmtype), dimension(:), allocatable :: asminf
 
@@ -330,7 +330,7 @@ contains
     eps = tiny(1.0_wp)
 
     if (gpu%num_levels .eq. 0) return ! Shortcut empty streams (v. small matrices)
-  
+
     gpu%n = n
     gpu%nnodes = nnodes
 
@@ -374,11 +374,11 @@ contains
        max_ncb = max(max_ncb, ncb)
        max_idata_size = max(max_idata_size, idata_size)
     end do
-  
+
     ii = nptr(nnodes + 1) - 1
     stats%cuda_error = cudaMalloc(gpu_LDLT, aligned_size(2*max_LDLT_size*C_SIZEOF(dummy_real)))
     if (stats%cuda_error .ne. 0) goto 200
-  
+
     allocate(gpu%values_L(gpu%num_levels), gpu%off_L(nnodes), &
          off_LDLT(nnodes), asminf(nnodes), stat=stats%st)
     if (stats%st .ne. 0) goto 100
@@ -437,7 +437,7 @@ contains
           m = blkm - blkn
           pc_size = pc_size + m*(m+0_long)
        end do
-      
+
        !
        ! Generate pointers for this level
        !
@@ -462,7 +462,7 @@ contains
           allocate(gpu_ldcol(gpu%lvlptr(lev+1)-gpu%lvlptr(lev)), stat=stats%st)
           if (stats%st .ne. 0) goto 100
           do llist = gpu%lvlptr(lev), gpu%lvlptr(lev + 1) - 1
-             node = gpu%lvllist(llist)        
+             node = gpu%lvllist(llist)
              i = llist - gpu%lvlptr(lev) + 1
              gpu_ldcol(i) = &
                   c_ptr_plus(ptr_levLD, gpu%off_L(node)*C_SIZEOF(dummy_real))
@@ -476,7 +476,7 @@ contains
           ndelay = nodes(node)%ndelay
           blkn = sptr(node+1) - sptr(node) + ndelay
           blkm = int(rptr(node+1) - rptr(node)) + ndelay
-      
+
           nodes(node)%gpu_lcol = c_ptr_plus(gpu%values_L(lev)%ptr_levL, &
                level_size*C_SIZEOF(dummy_real))
 
@@ -490,7 +490,7 @@ contains
           ndelay = nodes(node)%ndelay
[…]
