+++ src/arc/C/arctf.c
@@ -21,7 +21,7 @@ int prec( int n, const double x[], double u[], const double v[], const void * );
 int fun_diag( int n, const double x[], double *f, const void * );
 int grad_diag( int n, const double x[], double g[], const void * );
 int hess_diag( int n, int ne, const double x[], double hval[], const void * );
-int hessprod_diag( int n, const double x[], double u[], const double v[], 
+int hessprod_diag( int n, const double x[], double u[], const double v[],
                    bool got_h, const void * );
 
 int main(void) {
@@ -61,40 +61,40 @@ int main(void) {
         //control.print_level = 1;
 
         // Start from 1.5
-        double x[] = {1.5,1.5,1.5}; 
+        double x[] = {1.5,1.5,1.5};
 
         switch(d){
             case 1: // sparse co-ordinate storage
                 st = 'C';
-                arc_import( &control, &data, &status, n, "coordinate", 
+                arc_import( &control, &data, &status, n, "coordinate",
                             ne, H_row, H_col, NULL );
                 arc_solve_with_mat( &data, &userdata, &status,
                                   n, x, g, ne, fun, grad, hess, prec );
                 break;
-            case 2: // sparse by rows  
+            case 2: // sparse by rows
                 st = 'R';
-                arc_import( &control, &data, &status, n, "sparse_by_rows", 
+                arc_import( &control, &data, &status, n, "sparse_by_rows",
                             ne, NULL, H_col, H_ptr);
                 arc_solve_with_mat( &data, &userdata, &status,
                                   n, x, g, ne, fun, grad, hess, prec );
                 break;
             case 3: // dense
                 st = 'D';
-                arc_import(&control, &data, &status, n, "dense", 
+                arc_import(&control, &data, &status, n, "dense",
                            ne, NULL, NULL, NULL );
                 arc_solve_with_mat( &data, &userdata, &status,
                                   n, x, g, ne, fun, grad, hess_dense, prec );
                 break;
             case 4: // diagonal
                 st = 'I';
-                arc_import( &control, &data, &status, n, "diagonal", 
+                arc_import( &control, &data, &status, n, "diagonal",
                            ne, NULL, NULL, NULL );
-                arc_solve_with_mat( &data, &userdata, &status, n, x, g, 
+                arc_solve_with_mat( &data, &userdata, &status, n, x, g,
                                   ne, fun_diag, grad_diag, hess_diag, prec );
                 break;
             case 5: // access by products
                 st = 'P';
-                arc_import( &control, &data, &status, n, "absent", 
+                arc_import( &control, &data, &status, n, "absent",
                            ne, NULL, NULL, NULL );
                 arc_solve_without_mat( &data, &userdata, &status,
                                      n, x, g, fun, grad, hessprod, prec );
@@ -103,7 +103,7 @@ int main(void) {
         arc_information( &data, &inform, &status );
 
         if(inform.status == 0){
-            printf("%c:%6i iterations. Optimal objective value = %5.2f status = %1i\n", 
+            printf("%c:%6i iterations. Optimal objective value = %5.2f status = %1i\n",
                    st, inform.iter, inform.obj, inform.status);
         }else{
             printf("%c: ARC_solve exit status = %1i\n", st, inform.status);
@@ -127,7 +127,7 @@ int main(void) {
     double u[n], v[n];
     int index_nz_u[n], index_nz_v[n];
     double H_val[ne], H_dense[n*(n+1)/2], H_diag[n];
- 
+
     for(int d=1; d <= 5; d++){
 
         // Initialize ARC
@@ -138,15 +138,15 @@ int main(void) {
         //control.print_level = 1;
 
         // Start from 1.5
[…]
