+++ include/galahad_lpa.h
@@ -61,6 +61,7 @@
   package LPB is an alternative.
 
   \subsection lpa_authors Authors
+
   N. I. M. Gould and J. K. Reid, STFC-Rutherford Appleton Laboratory, 
   England.
 
@@ -69,6 +70,7 @@
   Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
 
   \subsection lpa_date Originally released
+
   October 2018, C interface September 2021.
 
   \subsection lpa_terminology Terminology
@@ -214,6 +216,7 @@
   Mathematical Programming <b>24</b> 55-69.
 
   \subsection lpa_call_order Call order
+
   To solve a given problem, functions from the lpa package must be called 
   in the following order:
 
@@ -256,6 +259,7 @@
   by supplying data that is already stored using 1-based indexing. 
 
   \subsubsection unsymmetric_matrix_dense Dense storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is, 
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -264,6 +268,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j 
@@ -274,6 +279,7 @@
   is recorded as A_ne = \f$ne\f$.
 
   \subsubsection unsymmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
