+++ include/galahad_blls.h
@@ -46,6 +46,7 @@
   by reverse communication or from a user-provided subroutine.
 
   \subsection blls_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
@@ -122,6 +123,7 @@
   In preparation.
 
   \subsection blls_call_order Call order
+
   To solve a given problem, functions from the blls package must be called 
   in the following order:
 
@@ -172,6 +174,7 @@
   by supplying data that is already stored using 1-based indexing. 
 
   \subsubsection unsymmetric_matrix_dense_row Dense row storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is, 
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -180,6 +183,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_dense_column Dense column storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by columns, that is, 
   the values of the entries of each column in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -188,6 +192,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j 
@@ -198,6 +203,7 @@
   is recorded as A_ne = \f$ne\f$.
 
   \subsubsection unsymmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
@@ -211,6 +217,7 @@
   its predecessors.
 
   \subsubsection unsymmetric_matrix_column_wise Sparse column-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in column j appear directly before those
   in column j+1. For the j-th column of \f$A\f$ the j-th component of the
