+++ include/galahad_bqpb.h
@@ -51,6 +51,7 @@
   Full advantage is taken of any zero coefficients in the matrix \f$H\f$.
 
   \subsection bqpb_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
@@ -213,6 +214,7 @@
   by supplying data that is already stored using 1-based indexing. 
 
   \subsubsection symmetric_matrix_dense Dense storage format
+
   The matrix \f$H\f$ is stored as a compact  dense matrix by rows, that is, 
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -224,6 +226,7 @@
   for \f$0 \leq j \leq i \leq n-1\f$.
 
   \subsubsection symmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$H\f$,
   its row index i, column index j 
@@ -234,6 +237,7 @@
   Note that only the entries in the lower triangle should be stored.
 
   \subsubsection symmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$H\f$ the i-th component of the
@@ -248,6 +252,7 @@
   its predecessor.
 
   \subsubsection symmetric_matrix_diagonal Diagonal storage format
+
   If \f$H\f$ is diagonal (i.e., \f$H_{ij} = 0\f$ for all 
   \f$0 \leq i \neq j \leq n-1\f$) only the diagonals entries 
   \f$H_{ii}\f$, \f$0 \leq i \leq n-1\f$ need
