+++ include/galahad_roots.h
@@ -28,6 +28,7 @@
   these are provided and used by other GALAHAD packages with C interfaces.
   
   \subsection roots_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
