+++ include/galahad_nls.h
@@ -42,6 +42,7 @@
   derivatives themselves, that can also be used to advantage.
 
   \subsection nls_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
@@ -266,6 +267,7 @@ for i,j=1,...,n\endmanonly
   by supplying data that is already stored using 1-based indexing.
 
   \subsubsection unsymmetric_matrix_dense Dense storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -274,6 +276,7 @@ for i,j=1,...,n\endmanonly
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_dense_cols Dense by columns storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by columns, that is,
   the values of the entries of each column in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -282,6 +285,7 @@ for i,j=1,...,n\endmanonly
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j
@@ -292,6 +296,7 @@ for i,j=1,...,n\endmanonly
   is recorded as A_ne = \f$ne\f$.
 
   \subsubsection unsymmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
@@ -305,6 +310,7 @@ for i,j=1,...,n\endmanonly
   its predecessor.
 
   \subsubsection unsymmetric_matrix_column_wise Sparse column-wise storage format
+
   Once again only the nonzero entries are stored, but this time
   they are ordered so that those in column j appear directly before those
   in column j+1. For the j-th column of \f$A\f$ the j-th component of the
@@ -326,6 +332,7 @@ for i,j=1,...,n\endmanonly
   (i.e, those entries that lie on or below the leading diagonal).
 
   \subsubsection symmetric_matrix_dense Dense storage format
+
   The matrix \f$H\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -337,6 +344,7 @@ for i,j=1,...,n\endmanonly
   for \f$0 \leq j \leq i \leq n-1\f$.
 
   \subsubsection symmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$H\f$,
   its row index i, column index j
@@ -347,6 +355,7 @@ for i,j=1,...,n\endmanonly
   Note that only the entries in the lower triangle should be stored.
 
   \subsubsection symmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$H\f$ the i-th component of the
[…]
