+++ include/galahad_lstr.h
@@ -35,6 +35,7 @@
   \f$v + A^T u\f$.
 
   \subsection lstr_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
