+++ include/galahad_qpb.h
@@ -57,6 +57,7 @@
   that is not the global solution to the given problem.
 
   \subsection qpb_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England, and
   Philippe L. Toint, University of Namur, Belgium.
 
@@ -65,6 +66,7 @@
   Julia interface, additionally A. Montoison and D. Orban, Polytechnique Montréal.
 
   \subsection qpb_date Originally released
+
   December 1999, C interface January 2022.
 
   \subsection qpb_terminology Terminology
@@ -227,6 +229,7 @@
   Mathematical Programming <b>87</b> 215-249.
 
   \subsection qpb_call_order Call order
+
   To solve a given problem, functions from the qpb package must be called
   in the following order:
 
@@ -269,6 +272,7 @@
   by supplying data that is already stored using 1-based indexing.
 
   \subsubsection unsymmetric_matrix_dense Dense storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -277,6 +281,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j
@@ -287,6 +292,7 @@
   is recorded as A_ne = \f$ne\f$.
 
   \subsubsection unsymmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
@@ -308,6 +314,7 @@
   (i.e, those entries that lie on or below the leading diagonal).
 
   \subsubsection symmetric_matrix_dense Dense storage format
+
   The matrix \f$H\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -319,6 +326,7 @@
   for \f$0 \leq j \leq i \leq n-1\f$.
 
   \subsubsection symmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$H\f$,
   its row index i, column index j
@@ -329,6 +337,7 @@
   Note that only the entries in the lower triangle should be stored.
 
   \subsubsection symmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$H\f$ the i-th component of the
@@ -343,6 +352,7 @@
   its predecessor.
 
   \subsubsection symmetric_matrix_diagonal Diagonal storage format
+
   If \f$H\f$ is diagonal (i.e., \f$H_{ij} = 0\f$ for all
   \f$0 \leq i \neq j \leq n-1\f$) only the diagonals entries
   \f$H_{ii}\f$, \f$0 \leq i \leq n-1\f$ need
