+++ include/galahad_sls.h
@@ -44,6 +44,7 @@
   solvers directly rather that via this package.
 
   \subsection sls_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
@@ -202,6 +203,7 @@ NY 10598, USA (2010).
 
 
   \subsection sls_call_order Call order
+
   To solve a given problem, functions from the sls package must be called
   in the following order:
 
@@ -252,6 +254,7 @@ NY 10598, USA (2010).
   by supplying data that is already stored using 1-based indexing.
 
   \subsubsection symmetric_matrix_dense Dense storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -263,6 +266,7 @@ NY 10598, USA (2010).
   for \f$0 \leq j \leq i \leq n-1\f$.
 
   \subsubsection symmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j
@@ -273,6 +277,7 @@ NY 10598, USA (2010).
   Note that only the entries in the lower triangle should be stored.
 
   \subsubsection symmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
