+++ include/galahad_bsc.h
@@ -31,6 +31,7 @@
   these are provided and used by other GALAHAD packages with C interfaces.
 
   \subsection bsc_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England.
 
   C interface, additionally J. Fowkes, STFC-Rutherford Appleton Laboratory.
@@ -42,6 +43,7 @@
   October 2013, C interface January 2022.
 
  \subsection bsc_call_order Call order
+
   To solve a given problem, functions from the bsc package must be called
   in the following order:
 
@@ -75,6 +77,7 @@
   by supplying data that is already stored using 1-based indexing. 
 
   \subsubsection unsymmetric_matrix_dense Dense storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by rows, that is, 
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -83,6 +86,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_dense Dense by columns storage format
+
   The matrix \f$A\f$ is stored as a compact  dense matrix by columns, that is, 
   the values of the entries of each column in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -91,6 +95,7 @@
   \f$0 \leq j \leq n-1\f$.
 
   \subsubsection unsymmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$A\f$,
   its row index i, column index j 
@@ -101,6 +106,7 @@
   is recorded as A_ne = \f$ne\f$.
 
   \subsubsection unsymmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$A\f$ the i-th component of the
@@ -114,6 +120,7 @@
   its predecessor.
 
   \subsubsection unsymmetric_matrix_column_wise Sparse column-wise storage format
+
   Once again only the nonzero entries are stored, but this time
   they are ordered so that those in column j appear directly before those
   in column j+1. For the j-th column of \f$A\f$ the j-th component of the
