+++ include/galahad_tru.h
@@ -32,6 +32,7 @@
   not the derivatives themselves, that may also be exploited.
 
   \subsection tru_authors Authors
+
   N. I. M. Gould, STFC-Rutherford Appleton Laboratory, England,
   and Ph. L. Toint, The University of Namur, Belgium.
 
@@ -163,6 +164,7 @@
   by supplying data that is already stored using 1-based indexing.
 
   \subsection symmetric_matrix_dense Dense storage format
+
   The matrix \f$H\f$ is stored as a compact  dense matrix by rows, that is,
   the values of the entries of each row in turn are
   stored in order within an appropriate real one-dimensional array.
@@ -174,6 +176,7 @@
   for \f$0 \leq j \leq i \leq n-1\f$.
 
   \subsection symmetric_matrix_coordinate Sparse co-ordinate storage format
+
   Only the nonzero entries of the matrices are stored.
   For the \f$l\f$-th entry, \f$0 \leq l \leq ne-1\f$, of \f$H\f$,
   its row index i, column index j
@@ -184,6 +187,7 @@
   Note that only the entries in the lower triangle should be stored.
 
   \subsection symmetric_matrix_row_wise Sparse row-wise storage format
+
   Again only the nonzero entries are stored, but this time
   they are ordered so that those in row i appear directly before those
   in row i+1. For the i-th row of \f$H\f$ the i-th component of the
