### 2022-12-07 80dcdb8
Remove white blanks at the end of the lines in header files
