+++ include/galahad.h
@@ -7,7 +7,7 @@
   GALAHAD is foremost a modern fortran library of packages designed
   to solve continuous optimization problems, with a particular
   emphasis on those that involve a large number of unknowns.
-  Since many application programs or applications are written in 
+  Since many application programs or applications are written in
   other languages, of late there has been a considerable effort to
   provide interfaces to GALAHAD. Thus there are Matlab interfaces,
   and here we provide details of those to C using the standardized
@@ -24,7 +24,7 @@
 
   GALAHAD provides packages as named for the following problems:
 
-  \li \link fdc \endlink - determine consistency and redundancy of linear 
+  \li \link fdc \endlink - determine consistency and redundancy of linear
       systems
       \latexonly \href{fdc_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/fdc/fdc.html">(link)</a> \endhtmlonly
@@ -51,83 +51,83 @@
       \latexonly \href{presolve_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/presolve/presolve.html">(link)</a>
       \endhtmlonly
-  \li \link bqp \endlink - bound-constrained convex quadratic programming 
+  \li \link bqp \endlink - bound-constrained convex quadratic programming
       using a gradient-projection method
       \latexonly \href{bqp_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/bqp/bqp.html">(link)</a> \endhtmlonly
-  \li \link bqpb \endlink - bound-constrained convex quadratic programming 
+  \li \link bqpb \endlink - bound-constrained convex quadratic programming
       using an interior-point method
       \latexonly \href{bqpb_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/bqpb/bqpb.html">(link)</a> \endhtmlonly
-  \li \link lsqp \endlink - linear and separable quadratic programming using 
+  \li \link lsqp \endlink - linear and separable quadratic programming using
       an interior-point method
       \latexonly \href{lsqp_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/lsqp/lsqp.html">(link)</a> \endhtmlonly
-  \li \link cqp \endlink - convex quadratic programming using 
+  \li \link cqp \endlink - convex quadratic programming using
       an interior-point method
       \latexonly \href{cqp_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/cqp/cqp.html">(link)</a> \endhtmlonly
-  \li \link dqp \endlink - convex quadratic programming using a 
+  \li \link dqp \endlink - convex quadratic programming using a
        dual active-set method
       \latexonly \href{dqp_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/dqp/dqp.html">(link)</a> \endhtmlonly
-  \li \link eqp \endlink - equality-constrained quadratic programming using 
+  \li \link eqp \endlink - equality-constrained quadratic programming using
       an iterative method
       \latexonly \href{eqp_c.pdf}{(link)} \endlatexonly
       \htmlonly <a href="../../../html/C/eqp/eqp.html">(link)</a> \endhtmlonly
-  \li \link trs \endlink - the trust-region subproblem using 
+  \li \link trs \endlink - the trust-region subproblem using
        matrix factorization
       \latexonly \href{trs_c.pdf}{(link)} \endlatexonly
[…]
