+++ include/galahad_lstr.h
@@ -16,7 +16,7 @@
  */
 
 /*! \mainpage GALAHAD C package lstr
- 
+
   \section lstr_intro Introduction
 
   \subsection lstr_purpose Purpose
@@ -29,8 +29,8 @@
   This problem commonly occurs as a trust-region subproblem in nonlinear
   optimization calculations, and may be used to regularize the solution
   of under-determined or ill-conditioned linear least-squares problems.
-  The method may be suitable for large \f$m\f$ and/or \f$n\f$ as no 
-  factorization involving \f$A\f$ is required. Reverse communication is used 
+  The method may be suitable for large \f$m\f$ and/or \f$n\f$ as no
+  factorization involving \f$A\f$ is required. Reverse communication is used
   to obtain matrix-vector products of the form \f$u + A v\f$ and
   \f$v + A^T u\f$.
 
@@ -86,7 +86,7 @@
      min || B_k y - ||b|| e_1 ||_2
 \n
 \endmanonly
-  using a QR factorization of \f$B_k\f$. Only the most recent 
+  using a QR factorization of \f$B_k\f$. Only the most recent
   \f$v_k\f$ and \f$u_{k+1}\f$
   are required, and their predecessors discarded, to compute \f$x_k\f$ from
   \f$x_{k-1}\f$. This method has the important property that the iterates
@@ -162,19 +162,19 @@
 
   \subsection lstr_call_order Call order
 
-  To solve a given problem, functions from the lstr package must be called 
+  To solve a given problem, functions from the lstr package must be called
   in the following order:
 
   - \link lstr_initialize \endlink - provide default control parameters and
       set up initial data structures
-  - \link lstr_read_specfile \endlink (optional) - override control values 
+  - \link lstr_read_specfile \endlink (optional) - override control values
       by reading replacement values from a file
-  - \link lstr_import_control \endlink - import control parameters prior to 
+  - \link lstr_import_control \endlink - import control parameters prior to
       solution
-  - \link lstr_solve_problem \endlink - solve the problem by reverse 
-      communication, a sequence of calls are made under control of a status 
-      parameter, each exit either asks the user to provide additional 
-      informaton and to re-enter, or reports that either the solution has 
+  - \link lstr_solve_problem \endlink - solve the problem by reverse
+      communication, a sequence of calls are made under control of a status
+      parameter, each exit either asks the user to provide additional
+      informaton and to re-enter, or reports that either the solution has
       been found or that an error has occurred
   - \link lstr_information \endlink (optional) - recover information about
     the solution and solution process
@@ -200,7 +200,7 @@ extern "C" {
 #endif
 
 // include guard
-#ifndef GALAHAD_LSTR_H 
+#ifndef GALAHAD_LSTR_H
 #define GALAHAD_LSTR_H
 
 // precision
@@ -329,7 +329,7 @@ struct lstr_inform_type {
     int biters;
 
     /// \brief
-    /// the smallest number of inner iterations performed during an 
+    /// the smallest number of inner iterations performed during an
     /// outer iteration
     int biter_min;
 
@@ -339,7 +339,7 @@ struct lstr_inform_type {
     int biter_max;
 
     /// \brief
-    /// the Lagrange multiplier, \f$\lambda\f$, corresponding to the 
+    /// the Lagrange multiplier, \f$\lambda\f$, corresponding to the
     /// trust-region constraint
     real_wp_ multiplier;
 
@@ -356,14 +356,14 @@ struct lstr_inform_type {
     real_wp_ Atr_norm;
 
     /// \brief
-    /// the average number of inner iterations performed during an outer 
[…]
