### 2022-12-07 9dabb99
Remove white blanks at the end of the lines in header files
