+++ include/galahad_bgo.h
@@ -16,16 +16,16 @@
  */
 
 /*! \mainpage GALAHAD C package bgo
- 
+
   \section bgo_intro Introduction
 
   \subsection bgo_purpose Purpose
 
-  The bgo package uses a <b>multi-start trust-region method to find an 
-  approximation to the global minimizer of a differentiable objective 
-  function \f$f(x)\f$ of \f$n\f$ variables \f$x\f$, subject to simple bounds 
+  The bgo package uses a <b>multi-start trust-region method to find an
+  approximation to the global minimizer of a differentiable objective
+  function \f$f(x)\f$ of \f$n\f$ variables \f$x\f$, subject to simple bounds
   \f$x^l \leq x \leq x^u\f$ on the variables.</b>
-  Here, any of the components of the vectors of bounds \f$x^l\f$ and \f$x^u\f$ 
+  Here, any of the components of the vectors of bounds \f$x^l\f$ and \f$x^u\f$
   may be infinite. The method offers the choice of direct
   and iterative solution of the key trust-region subproblems, and
   is suitable for large problems. First derivatives are required,
@@ -33,9 +33,9 @@
   the product of second derivatives with a vector may be found but
   not the derivatives themselves, that may also be exploited.
 
-  The package offers both random multi-start and local-minimize-and probe 
-  methods to try to locate the global minimizer. There are no theoretical 
-  guarantees unless the sampling is huge, and realistically the success of 
+  The package offers both random multi-start and local-minimize-and probe
+  methods to try to locate the global minimizer. There are no theoretical
+  guarantees unless the sampling is huge, and realistically the success of
   the methods decreases as the dimension and nonconvexity increase.
 
   \subsection bgo_authors Authors
@@ -52,25 +52,25 @@
 
   \subsection bgo_terminology Terminology
 
-  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose 
+  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose
   \f$i\f$-th component is \f$\partial f(x)/\partial x_i\f$.
-  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix 
+  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix
   whose \f$i,j\f$-th entry is \f$\partial^2 f(x)/\partial x_i \partial x_j\f$.
-  The Hessian is \e sparse if a significant and useful proportion of the 
+  The Hessian is \e sparse if a significant and useful proportion of the
   entries are universally zero.
 
   \subsection bgo_method Method
 
-  A choice of two methods is available. 
+  A choice of two methods is available.
   In the first, local-minimization-and-probe, approach, local minimization
-  and univariate global minimization are intermixed. Given a current 
+  and univariate global minimization are intermixed. Given a current
   champion \f$x^S_k\f$, a local minimizer \f$x_k\f$ of \f$f(x)\f$ within the
   feasible box \f$x^l \leq x \leq x^u\f$ is found using the GALAHAD package trb.
   Thereafter \f$m\f$ random directions \f$p\f$ are generated, and univariate
   local minimizer of \f$f(x_k + \alpha p)\f$ as a function of the scalar
   \f$\alpha\f$ along each \f$p\f$ within the interval \f$[\alpha^L,\alpha^u]\f$,
-  where \f$\alpha^L\f$ and \f$\alpha^u\f$ are the smallest and largest 
[…]
