### 2022-12-07 e71d50e
Remove white blanks at the end of the lines in header files
