+++ include/galahad_gls.h
@@ -16,17 +16,17 @@
  */
 
 /*! \mainpage GALAHAD C package gls
- 
+
   \section gls_intro Introduction
 
   \subsection gls_purpose Purpose
 
-  This package <b>solves sparse unsymmetric system of linear equations.</b>. 
+  This package <b>solves sparse unsymmetric system of linear equations.</b>.
   Given an \f$m\f$ by \f$n\f$ sparse matrix \f$A = a_{ij}\f$, the package
   solves the system \f$A x = b\f$ (or optionally \f$A^T x = b\f$).
-  The matrix \f$A\f$ can be rectangular. 
+  The matrix \f$A\f$ can be rectangular.
 
-  <b>N.B.</b> The package is simply a sophisticated interface to the 
+  <b>N.B.</b> The package is simply a sophisticated interface to the
   HSL package MA33, and requires that a user has obtained the latter.
   <b> MA33 is not included in GALAHAD</b>
   but is available without charge to recognised academics, see
@@ -36,7 +36,7 @@
   automatically and, if there is insufficient space for the factorization,
   more space is allocated and the factorization is repeated.  The package
   also returns the number of entries in the factors and has facilities
-  for identifying the rows and columns that are treated specially 
+  for identifying the rows and columns that are treated specially
   when the matrix is singular or rectangular.
 
   Currently, only the control and inform parameters are exposed;
@@ -114,7 +114,7 @@ extern "C" {
 #endif
 
 // include guard
-#ifndef GALAHAD_GLS_H 
+#ifndef GALAHAD_GLS_H
 #define GALAHAD_GLS_H
 
 // precision
@@ -146,7 +146,7 @@ struct gls_control {
     int ldiag;
 
     /// \brief
-    /// Minimum block size for block-triangular form (BTF). Set to \f$n\f$ 
+    /// Minimum block size for block-triangular form (BTF). Set to \f$n\f$
     /// to avoid
     int btf;
 
@@ -345,7 +345,7 @@ struct gls_sinfo {
 
 // *-*-*-*-*-*-*-*-*-*-    G L S  _ I N I T I A L I Z E    -*-*-*-*-*-*-*-*-*
 
-void gls_initialize( void **data, 
+void gls_initialize( void **data,
                      struct gls_control *control );
 
 /*!<
@@ -353,24 +353,24 @@ void gls_initialize( void **data,
 
   @param[in,out] data  holds private internal data
 
-  @param[out] control  is a struct containing control information 
+  @param[out] control  is a struct containing control information
               (see gls_control)
 */
 
 // *-*-*-*-*-*-*-*-*-    G L S  _ R E A D _ S P E C F I L E   -*-*-*-*-*-*-*
 
-void gls_read_specfile( struct gls_control *control, 
+void gls_read_specfile( struct gls_control *control,
                         const char specfile[] );
 
 /*!<
-  Read the content of a specification file, and assign values associated 
+  Read the content of a specification file, and assign values associated
   with given keywords to the corresponding control parameters.
   By default, the spcification file will be named RUNGLS.SPC and
   lie in the current directory.
   Refer to Table 2.1 in the fortran documentation provided in
   $GALAHAD/doc/gls.pdf for a list of keywords that may be set.
 
-  @param[in,out]  control  is a struct containing control information 
+  @param[in,out]  control  is a struct containing control information
               (see gls_control)
 
   @param[in]  specfile  is a character string containing the name of
@@ -384,7 +384,7 @@ void gls_import( struct gls_control *control,
                  int *status );
 
 /*!<
- Import problem data into internal storage prior to solution. 
+ Import problem data into internal storage prior to solution.
 
  @param[in] control is a struct whose members provide control
[…]
