+++ include/galahad_sils.h
@@ -16,16 +16,16 @@
  */
 
 /*! \mainpage GALAHAD C package sils
- 
+
   \section sils_intro Introduction
 
   \subsection sils_purpose Purpose
- 
-  This package <b>solves sparse symmetric system of linear equations.</b>. 
+
+  This package <b>solves sparse symmetric system of linear equations.</b>.
   Given an \f$n\f$ by \f$n\f$ sparse matrix \f$A = {a_{ij}}\f$, and an
   \f$n\f$ vector \f$b\f$, the package solves the system \f$A x = b\f$.
   The matrix \f$A\f$ need not be definite. There is an option for iterative
-  refinement. 
+  refinement.
 
   Currently, only the control and inform parameters are exposed;
   these are provided and used by other GALAHAD packages with C interfaces.
@@ -45,11 +45,11 @@
 
   \subsection sils_method Method
 
-  The method used is a direct method based on a sparse variant 
-  of Gaussian elimination and is discussed further by 
+  The method used is a direct method based on a sparse variant
+  of Gaussian elimination and is discussed further by
 
-  I. S. Duff and J. K. Reid (1983), 
-  ACM Trans. Math. Software <b>9</b> pp.302-325. 
+  I. S. Duff and J. K. Reid (1983),
+  ACM Trans. Math. Software <b>9</b> pp.302-325.
 
   \subsection main_symmetric_matrices Symmetric matrix storage formats
 
@@ -115,7 +115,7 @@ extern "C" {
 #endif
 
 // include guard
-#ifndef GALAHAD_SILS_H 
+#ifndef GALAHAD_SILS_H
 #define GALAHAD_SILS_H
 
 // precision
@@ -226,8 +226,8 @@ struct sils_control_type {
     real_wp_ multiplier;
 
     /// \brief
-    /// If previously allocated internal workspace arrays are greater than 
-    /// reduce times the currently required sizes, they are reset to current 
+    /// If previously allocated internal workspace arrays are greater than
+    /// reduce times the currently required sizes, they are reset to current
     /// requirment
     real_wp_ reduce;
 
@@ -400,7 +400,7 @@ struct sils_finfo_type {
     int nstatic;
 
     /// \brief
-    /// First pivot modification when control.pivoting=4 
+    /// First pivot modification when control.pivoting=4
     int modstep;
 
     /// \brief
@@ -480,7 +480,7 @@ struct sils_sinfo_type {
 
 // *-*-*-*-*-*-*-*-*-*-    S I L S  _ I N I T I A L I Z E    -*-*-*-*-*-*-*-*-*
 
-void sils_initialize( void **data, 
+void sils_initialize( void **data,
                       struct sils_control_type *control,
                      int *status );
 
@@ -488,7 +488,7 @@ void sils_initialize( void **data,
  Set default control values and initialize private data
 
   @param[in,out] data  holds private internal data
-  @param[out] control  is a struct containing control information 
+  @param[out] control  is a struct containing control information
               (see sils_control_type)
   @param[out] status is a scalar variable of type int, that gives
               the exit status from the package.
@@ -498,18 +498,18 @@ void sils_initialize( void **data,
 
 // *-*-*-*-*-*-*-*-*-    S I L S  _ R E A D _ S P E C F I L E   -*-*-*-*-*-*-*
 
-void sils_read_specfile( struct sils_control_type *control, 
+void sils_read_specfile( struct sils_control_type *control,
                          const char specfile[] );
 
 /*!<
-  Read the content of a specification file, and assign values associated 
+  Read the content of a specification file, and assign values associated
   with given keywords to the corresponding control parameters.
   By default, the spcification file will be named RUNSILS.SPC and
   lie in the current directory.
[…]
