### 2022-12-07 6614657
Remove white blanks at the end of the lines in header files
