+++ include/galahad_dgo.h
@@ -16,15 +16,15 @@
  */
 
 /*! \mainpage GALAHAD C package dgo
- 
+
   \section dgo_intro Introduction
 
   \subsection dgo_purpose Purpose
 
-  The dgo package uses a <b>deterministic partition-and-bound trust-region 
-  method to find an approximation to the global minimizer of a differentiable 
+  The dgo package uses a <b>deterministic partition-and-bound trust-region
+  method to find an approximation to the global minimizer of a differentiable
   objective function \f$f(x)\f$ of \f$n\f$ variables \f$x\f$, subject to a
-  finite set of simple bounds \f$x^l \leq x \leq x^u\f$ on the 
+  finite set of simple bounds \f$x^l \leq x \leq x^u\f$ on the
   variables.</b>
   The method offers the choice of direct
   and iterative solution of the key trust-region subproblems, and
@@ -50,34 +50,34 @@
 
   \subsection dgo_terminology Terminology
 
-  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose 
+  The \e gradient \f$\nabla_x f(x)\f$ of \f$f(x)\f$ is the vector whose
   \f$i\f$-th component is \f$\partial f(x)/\partial x_i\f$.
-  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix 
+  The \e Hessian \f$\nabla_{xx} f(x)\f$ of \f$f(x)\f$ is the symmetric matrix
   whose \f$i,j\f$-th entry is \f$\partial^2 f(x)/\partial x_i \partial x_j\f$.
-  The Hessian is \e sparse if a significant and useful proportion of the 
+  The Hessian is \e sparse if a significant and useful proportion of the
   entries are universally zero.
 
   \subsection dgo_method Method
 
-  Starting with the initial box \f$x^l \leq x \leq x^u\f$, a sequence of 
+  Starting with the initial box \f$x^l \leq x \leq x^u\f$, a sequence of
   boxes is generated by considering the current set, and partitioning
-  a promising candidate into three equally-sized sub-boxes by splitting 
-  along one of the box dimensions. Each partition requires only a pair of 
-  new function and derivative evaluations, and these values, together with 
-  estimates of Lipschitz constants, makes it possible to remove other boxes 
-  from further consideration as soon as they cannot contain a global minimizer. 
+  a promising candidate into three equally-sized sub-boxes by splitting
+  along one of the box dimensions. Each partition requires only a pair of
+  new function and derivative evaluations, and these values, together with
+  estimates of Lipschitz constants, makes it possible to remove other boxes
+  from further consideration as soon as they cannot contain a global minimizer.
   Efficient control of the dictionary of vertices of the sub-boxes
-  is handled using a suitable hashing procedure provided by the GALAHAD 
-  package hash; each sub-box is indexed by the concatenated 
+  is handled using a suitable hashing procedure provided by the GALAHAD
+  package hash; each sub-box is indexed by the concatenated
   coordinates of a pair of opposite vertices. At various
   stages, local minimization in a promising sub-box, using the GALAHAD
-  package trb, may be used to improve the best-known upper bound 
+  package trb, may be used to improve the best-known upper bound
   on the global minimizer.
-  If \f$n=1\f$, the specialised GALAHAD univariate global minimization package 
+  If \f$n=1\f$, the specialised GALAHAD univariate global minimization package
   ugo is called directly.
 
-  We reiterate that although there are theoretical guarantees, 
-  these may require a large number of evaluations as the dimension 
+  We reiterate that although there are theoretical guarantees,
+  these may require a large number of evaluations as the dimension
   and nonconvexity increase.
[…]
