+++ include/galahad_l2rt.h
@@ -16,13 +16,13 @@
  */
 
 /*! \mainpage GALAHAD C package l2rt
- 
+
   \section l2rt_intro Introduction
 
   \subsection l2rt_purpose Purpose
 
   Given a real \f$m\f$ by \f$n\f$ matrix \f$A\f$,
-  a real \f$m\f$ vector \f$b\f$ and scalars \f$\sigma>0\f$, 
+  a real \f$m\f$ vector \f$b\f$ and scalars \f$\sigma>0\f$,
   \f$\mu\geq 0\f$ and \f$p \geq 2\f$,  this package finds an
   <b> approximate minimizer of the regularised linear-least-\f$\ell_2\f$-norm
   objective function
@@ -37,8 +37,8 @@
   optimization calculations involving quadratic or cubic regularisation,
   and may be used to regularise the solution
   of under-determined or ill-conditioned linear least-squares problems.
-  The method may be suitable for large \f$m\f$ and/or \f$n\f$ as no 
-  factorization involving \f$A\f$ is required. Reverse communication is 
+  The method may be suitable for large \f$m\f$ and/or \f$n\f$ as no
+  factorization involving \f$A\f$ is required. Reverse communication is
   used to obtain matrix-vector products of the form \f$u + A v\f$ and
   \f$v + A^T u\f$.
 
@@ -70,7 +70,7 @@
 
   The method is iterative. Starting  with the vector \f$u_1 = b\f$, a
   bi-diagonalisation process is used to generate the vectors \f$v_k\f$ and
-  \f$u_k+1\f$ so that the \f$n\f$ by \f$k\f$ matrix 
+  \f$u_k+1\f$ so that the \f$n\f$ by \f$k\f$ matrix
   \f$V_k = ( v_1 \ldots v_k)\f$
   and the \f$m\f$ by \f$(k+1)\f$ matrix \f$U_k = ( u_1 \ldots u_{k+1})\f$
   together satisfy
@@ -98,8 +98,8 @@
      ( B_k^T ( B_k y(lambda) - ||b|| e_1 ) + lambda y(lambda) = 0,
 \n
 \endmanonly
-  where 
-\f[\lambda = \mu + \sigma \|y(\lambda)\|_2^{p-2} 
+  where
+\f[\lambda = \mu + \sigma \|y(\lambda)\|_2^{p-2}
 \sqrt{ \| B_k y(\lambda) - \|b\| e_1 \|_2^2 + \mu \|y(\lambda)\|_2^2},\f]
 \manonly
 \n
@@ -128,7 +128,7 @@
                      mu ||y(lambda)||^2_2} + mu - lambda = 0.
 \n
 \endmanonly
-  In practice (3) is reformulated, and a more rapidly converging 
+  In practice (3) is reformulated, and a more rapidly converging
   iteration is used. Having found  \f$y_k\f$, a second pass in which
   \f$x_k = V_k y_k\f$ is regenerated is needed---this need only be done
   once \f$x_k\f$ has implicitly deemed to be sufficiently close to optimality.
@@ -157,7 +157,7 @@
     sparse least squares.
   ACM Transactions on Mathematical Software, 8(2):195--209, 1982.
 
-  Additional details on the Newton-like process needed to determine 
+  Additional details on the Newton-like process needed to determine
   \f$\lambda\f$ and other details are described in
 
   C. Cartis, N. I. M. Gould and Ph. L. Toint,
@@ -167,19 +167,19 @@
 
   \subsection l2rt_call_order Call order
 
-  To solve a given problem, functions from the l2rt package must be called 
+  To solve a given problem, functions from the l2rt package must be called
   in the following order:
 
   - \link l2rt_initialize \endlink - provide default control parameters and
[…]
