### 2022-12-07 fa2b32b
modified all makefiles to allow .F90 instead of .f90 where appropriate. added galahad_modules.h to allow for command-line selection of double/single changed licence to BSD from LGPL with added contributors
