+++ include/galahad_modules.h
@@ -1,0 +1,5 @@
+#ifdef GALAHAD_SINGLE
+#define GALAHAD_USERDATA_MODULE GALAHAD_USERDATA_single
+#else
+#define GALAHAD_USERDATA_MODULE GALAHAD_USERDATA_double
+#endif
