+++ include/galahad_wcp.h
@@ -24,23 +24,23 @@
   This package uses a primal-dual interior-point method
   to <b>find a well-centered interior point</b> \f$x\f$ for a set of
   general linear constraints
-  \f[\mbox{(1)} \;\; c_i^l  \leq  a_i^Tx  \leq c_i^u, \;\;\; i = 1, \ldots , m,\f]
+  \f[\mbox{(1)} \;\; c_i^l \leq a_i^Tx \leq c_i^u, \;\;\; i = 1, \ldots , m,\f]
 \manonly
   \n
-   (1)  c_i^l \[<=] a_i^Tx \[<=] c_i^u, i = 1, ... , m,
+   (1) c_i^l \[<=] a_i^Tx \[<=] c_i^u, i = 1, ... , m,
   \n
 \endmanonly
   and the simple bound constraints
-  \f[\mbox{(2)} \;\; x_j^l  \leq  x_j \leq x_j^u, \;\;\; j = 1, \ldots , n,\f]
+  \f[\mbox{(2)} \;\; x_j^l \leq x_j \leq x_j^u, \;\;\; j = 1, \ldots , n,\f]
 \manonly
   \n
-   (2)   x_j^l \[<=] x_j \[<=] x_j^u, j = 1, ... , n,
+   (2) x_j^l \[<=] x_j \[<=] x_j^u, j = 1, ... , n,
   \n
 \endmanonly
   where the vectors
   \f$a_{i}\f$, \f$c^l\f$, \f$c^u\f$, \f$x^l\f$ and \f$x^u\f$ are given.
   More specifically, if possible, the package finds a solution to the
-  system of  primal optimality equations
+  system of primal optimality equations
   \f[\mbox{(3)} \;\; A x = c,\f]
 \manonly
   \n
@@ -58,7 +58,7 @@
   \f[\mbox{(5)} \;\;
   ( c_i - c^l_i ) y^l_i = (\mu_c^l)_i \;\mbox{and}\;
   ( c_i - c_i^u ) y^u_i = (\mu_c^u)_i, \;\;\;
-   i = 1, \ldots , m,  \f]
+   i = 1, \ldots , m, \f]
 \manonly
   \n
   (c_i - c^l_i) y^l_i = (mu_c^l)_i and
@@ -67,9 +67,9 @@
 \endmanonly
   and
   \f[\mbox{(6)} \;\;
-  ((x_j - x^l_j ) z_j^l = (\mu_x^l)_j  \;\mbox{and}\;
+  ((x_j - x^l_j ) z_j^l = (\mu_x^l)_j \;\mbox{and}\;
   ( x_j - x^u_j ) z_j^u = (\mu_x^u)_j, \;\;\;
-   j = 1, \ldots , n,  \f]
+   j = 1, \ldots , n, \f]
 \manonly
   \n
   (x_j - c^l_j) z^l_j = (mu_x^l)_j and
@@ -80,13 +80,13 @@
 \latexonly
   \[
 \mbox{(7)} \;\; c^l \leq c \leq c^u, \;\; x^l \leq x \leq x^u, \;\;
-y^l \geq 0 , \;\;  y^u \leq 0 , \;\; z^l \geq 0 \;\; \mbox{and} \;\; z^u \leq 0
+y^l \geq 0 , \;\; y^u \leq 0 , \;\; z^l \geq 0 \;\; \mbox{and} \;\; z^u \leq 0
   \]
 \endlatexonly
 \htmlonly
   $$
 \mbox{(7)} \;\; c^l \leq c \leq c^u, \;\; x^l \leq x \leq x^u, \;\;
[…]
