+++ src/tools/tools.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 08:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-08 AT 13:15 GMT.
+
+#include "galahad_modules.h"
+
 ! 30/06/2003: procedures _is_are and _s moved from QPA.
 ! 23/07/2003: invalid "END INTERFACE" arguments removed
 ! 17/11/2009: procedures _char_integer and _char_real imported from OTHERS.
@@ -25,6 +28,8 @@
 !
 
    MODULE GALAHAD_TOOLS
+
+      USE GALAHAD_PRECISION
 
 !-------------------------------------------------------------------------------
 !   A c c e s s
@@ -77,13 +82,6 @@
          MODULE PROCEDURE TOOLS_char_real_sp, TOOLS_char_real_dp
       END INTERFACE
 
-!-------------------------------------------------------------------------------
-!   P r e c i s i o n
-!-------------------------------------------------------------------------------
-
-      INTEGER, PRIVATE, PARAMETER :: sp = KIND( 1.0E+0 )
-      INTEGER, PRIVATE, PARAMETER :: dp = KIND( 1.0D+0 )
-
    CONTAINS
 
 !==============================================================================
@@ -96,15 +94,15 @@
 
 !     Arguments:
 
-      INTEGER, INTENT( IN ) :: n
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: n
 
 !             The dimension of x.
 
-      REAL ( KIND = sp ), DIMENSION( n ), INTENT( IN ) :: x
+      REAL ( KIND = sp_ ), DIMENSION( n ), INTENT( IN ) :: x
 
 !             The vector to print.
 
-      INTEGER, INTENT( IN ) :: out
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: out
 
 !             The output device number
 
@@ -112,7 +110,7 @@
 
 !     Local variable
 
-      INTEGER :: j, i
+      INTEGER ( KIND = ip_ ) :: j, i
 
       WRITE( out, 101 )
       j = 1
@@ -138,15 +136,15 @@
 
 !     Arguments:
 
-      INTEGER, INTENT( IN ) :: n
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: n
 
 !             The dimension of x.
 
-      REAL ( KIND = dp ), DIMENSION( n ), INTENT( IN ) :: x
+      REAL ( KIND = dp_ ), DIMENSION( n ), INTENT( IN ) :: x
 
 !             The vector to print.
 
-      INTEGER, INTENT( IN ) :: out
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: out
 
 !             The output device number
 
@@ -154,7 +152,7 @@
 
 !     Local variable
 
-      INTEGER :: j, i
+      INTEGER ( KIND = ip_ ) :: j, i
 
       WRITE( out, 101 )
       j = 1
@@ -181,15 +179,15 @@
 
 !     Arguments:
 
-      INTEGER, INTENT( IN ) :: n
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: n
 
 !             The dimension of ix.
 
-      INTEGER, DIMENSION( n ), INTENT( IN ) :: ix
[…]
