+++ src/specfile/specfile.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 09:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-08 AT 07:05 GMT.
 
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -213,8 +213,11 @@
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 
-   MODULE GALAHAD_SPECFILE_double
+#include "galahad_modules.h"
 
+   MODULE GALAHAD_SPECFILE_precision
+
+      USE GALAHAD_PRECISION
       USE GALAHAD_SYMBOLS  ! to make the symbols known for translation
 
 !-------------------------------------------------------------------------------
@@ -240,13 +243,6 @@
       END INTERFACE
 
 !-------------------------------------------------------------------------------
-!   P r e c i s i o n
-!-------------------------------------------------------------------------------
-
-      INTEGER, PRIVATE, PARAMETER :: wp = KIND( 1.0D+0 )
-      INTEGER, PRIVATE, PARAMETER :: long = SELECTED_INT_KIND( 18 )
-
-!-------------------------------------------------------------------------------
 !-------------------------------------------------------------------------------
 !
 !                      PUBLIC TYPES AND DEFINITIONS
@@ -268,17 +264,17 @@
 
       TYPE, PUBLIC :: SPECFILE_item_type
 
-         CHARACTER( LEN = 50 ) :: keyword
+         CHARACTER ( LEN = 50 ) :: keyword
 
 !              the string that specifies the control parameter whose value
 !              is to be assigned.
 
-         CHARACTER( LEN = 30 ) :: value
+         CHARACTER ( LEN = 30 ) :: value
 
 !              the value assigned to the control parameter, expressed as
 !              a string.
 
-         INTEGER :: line
+         INTEGER ( KIND = ip_ ) :: line
 
 !              0   : if no specfile command has been read so far that assigns a
 !                    value to the considered control parameter
@@ -304,7 +300,7 @@
 
 !     Argument
 
-      INTEGER              , INTENT( OUT )   :: new_length, break
+      INTEGER ( KIND = ip_ ), INTENT( OUT )   :: new_length, break
       CHARACTER( LEN = 80 ), INTENT( INOUT ) :: string
 
 !     Programming: Ph. Toint, December 2001.
@@ -313,7 +309,7 @@
 
 !     Local variables
 
-      INTEGER              :: i, j, nl, nb, last
+      INTEGER ( KIND = ip_ ) :: i, j, nl, nb, last
       CHARACTER( LEN = 1 ) :: c
 
 !     Remove the leading and trailing blanks
@@ -407,11 +403,11 @@
 
 !        The specification item
 
-      INTEGER, INTENT( INOUT ) :: iparam
+      INTEGER ( KIND = ip_ ), INTENT( INOUT ) :: iparam
 
 !        The control parameter to be assigned.
 
[…]
