+++ include/galahad_modules.h
@@ -1,5 +1,423 @@
 #ifdef GALAHAD_SINGLE
-#define GALAHAD_USERDATA_MODULE GALAHAD_USERDATA_single
+#define GALAHAD_ARC_precision GALAHAD_ARC_single
+#define GALAHAD_BARC_precision GALAHAD_BARC_single
+#define GALAHAD_BGO_precision GALAHAD_BGO_single
+#define GALAHAD_BLLS_precision GALAHAD_BLLS_single
+#define GALAHAD_BQPB_precision GALAHAD_BQPB_single
+#define GALAHAD_BQP_precision GALAHAD_BQP_single
+#define GALAHAD_BSC_precision GALAHAD_BSC_single
+#define GALAHAD_CCQP_precision GALAHAD_CCQP_single
+#define GALAHAD_CDQP_precision GALAHAD_CDQP_single
+#define GALAHAD_CHECK_precision GALAHAD_CHECK_single
+#define GALAHAD_CHECKPOINT_precision GALAHAD_CHECKPOINT_single
+#define GALAHAD_CLLS_precision GALAHAD_CLLS_single
+#define GALAHAD_CONVERT_precision GALAHAD_CONVERT_single
+#define GALAHAD_CQP_precision GALAHAD_CQP_single
+#define GALAHAD_CQPS_precision GALAHAD_CQPS_single
+#define GALAHAD_CRO_precision GALAHAD_CRO_single
+#define GALAHAD_CUTEST_FUNCTIONS_precision GALAHAD_CUTEST_FUNCTIONS_single
+#define GALAHAD_DEMO_precision GALAHAD_DEMO_single
+#define GALAHAD_DGO_precision GALAHAD_DGO_single
+#define GALAHAD_DLP_precision GALAHAD_DLP_single
+#define GALAHAD_DPS_PRECISION GALAHAD_DPS_single
+#define GALAHAD_DQP_precision GALAHAD_DQP_single
+#define GALAHAD_EQP_precision GALAHAD_EQP_single
+#define GALAHAD_ERMO_precision GALAHAD_ERMO_single
+#define GALAHAD_EXTEND_precision GALAHAD_EXTEND_single
+#define GALAHAD_FASTR_precision GALAHAD_FASTR_single
+#define GALAHAD_FDC_precision GALAHAD_FDC_single
+#define GALAHAD_FDH_precision GALAHAD_FDH_single
+#define GALAHAD_FILTER_precision GALAHAD_FILTER_single
+#define GALAHAD_FILTRANE_precision GALAHAD_FILTRANE_single
+#define GALAHAD_FISQP_precision GALAHAD_FISQP_single
+#define GALAHAD_FIT_precision GALAHAD_FIT_single
+#define GALAHAD_FUNNEL_precision GALAHAD_FUNNEL_single
+#define GALAHAD_FUNNEL_equality_precision GALAHAD_FUNNEL_equality_single
+#define GALAHAD_GLRT_precision GALAHAD_GLRT_single
+#define GALAHAD_GLS_precision GALAHAD_GLS_single
+#define GALAHAD_GLTR_precision GALAHAD_GLTR_single
+#define GALAHAD_ICFS_precision GALAHAD_ICFS_single
+#define GALAHAD_IR_precision GALAHAD_IR_single
+#define GALAHAD_ISM_precision GALAHAD_ISM_single
+#define GALAHAD_L1QP_precision GALAHAD_L1QP_single
+#define GALAHAD_L2RT_precision GALAHAD_L2RT_single
+#define GALAHAD_LCF_precision GALAHAD_LCF_single
+#define GALAHAD_LHS_precision GALAHAD_LHS_single
+#define GALAHAD_LLS_precision GALAHAD_LLS_single
[…]
