+++ src/ugo/ugo.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-08 AT 07:10 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ U G O   M O D U L E  *-*-*-*-*-*-*-*-*-*-
 
@@ -11,7 +13,7 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_UGO_double
+   MODULE GALAHAD_UGO_precision
 
 !     -------------------------------------------------------
 !    |                                                       |
@@ -22,11 +24,12 @@
 !    |                                                       |
 !     -------------------------------------------------------
 
+     USE GALAHAD_PRECISION
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_USERDATA_double
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_USERDATA_precision
 
      IMPLICIT NONE
 
@@ -48,38 +51,32 @@
         MODULE PROCEDURE UGO_terminate, UGO_full_terminate
       END INTERFACE UGO_terminate
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: quarter = 0.25_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-     REAL ( KIND = wp ), PARAMETER :: epsmch = EPSILON( one )
-     REAL ( KIND = wp ), PARAMETER :: infinity = HUGE( one )
+     REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: quarter = 0.25_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: half = 0.5_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: two = 2.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: epsmch = EPSILON( one )
+     REAL ( KIND = rp_ ), PARAMETER :: infinity = HUGE( one )
 
-     INTEGER, PARAMETER :: header_interval = 50
-     REAL ( KIND = wp ), PARAMETER :: midxdg_min = ten * epsmch
+     INTEGER ( KIND = ip_ ), PARAMETER :: header_interval = 50
+     REAL ( KIND = rp_ ), PARAMETER :: midxdg_min = ten * epsmch
 
 !  Lipschitz constant estimate used (step 2)
 
-     INTEGER, PARAMETER  :: global_lipschitz_available = 1
[…]
