+++ src/precision/precision.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-06 AT 07:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-08 AT 07:00 GMT.
 
 !-*-*-*-*-*-*-*- G A L A H A D _ P R E C I S I O N   M O D U L E -*-*-*-*-*-*-*-
 
@@ -8,7 +8,7 @@
 !  History -
 !   originally released, GALAHAD Version 4.1, December 6th 2022
 
-MODULE GALAHAD_precision
+MODULE GALAHAD_PRECISION
   USE ISO_C_BINDING
   USE ISO_FORTRAN_ENV
 
@@ -22,39 +22,41 @@ MODULE GALAHAD_precision
 
 !  basic kinds used
 
- INTEGER, PARAMETER :: i4 = INT32
- INTEGER, PARAMETER :: i8 = INT64
- INTEGER, PARAMETER :: r4 = REAL32
- INTEGER, PARAMETER :: r8 = REAL64
- INTEGER, PARAMETER :: c4 = KIND( ( 1.0_r4, 1.0_r4 ) )
- INTEGER, PARAMETER :: c8 = KIND( ( 1.0_r8, 1.0_r8 ) )
+ INTEGER, PARAMETER :: i4_ = INT32
+ INTEGER, PARAMETER :: i8_ = INT64
+ INTEGER, PARAMETER :: r4_ = REAL32
+ INTEGER, PARAMETER :: r8_ = REAL64
+ INTEGER, PARAMETER :: c4_ = KIND( ( 1.0_r4_, 1.0_r4_ ) )
+ INTEGER, PARAMETER :: c8_ = KIND( ( 1.0_r8_, 1.0_r8_ ) )
 
 !  common aliases
 
- INTEGER, PARAMETER :: sp = r4
- INTEGER, PARAMETER :: dp = r8
- INTEGER, PARAMETER :: long = i8
+ INTEGER, PARAMETER :: sp_ = r4_
+ INTEGER, PARAMETER :: dp_ = r8_
+ INTEGER, PARAMETER :: long_ = i8_
 
 !--------------------------------
 !   P r e c i s i o n s  u s e d
 !--------------------------------
 
 #ifdef GALAHAD_SINGLE
- INTEGER, PARAMETER :: wp = r4
- INTEGER, PARAMETER :: cp = c4
- INTEGER, PARAMETER :: wpc = C_FLOAT
+ INTEGER, PARAMETER :: real_bytes_ = 4
+ INTEGER, PARAMETER :: rp_ = r4_
+ INTEGER, PARAMETER :: cp_ = c4_
+ INTEGER, PARAMETER :: rpc_ = C_FLOAT
 #else
- INTEGER, PARAMETER :: wp = r8
- INTEGER, PARAMETER :: cp = c8
- INTEGER, PARAMETER :: wpc = C_DOUBLE
+ INTEGER, PARAMETER :: real_bytes_ = 8
+ INTEGER, PARAMETER :: rp_ = r8_
+ INTEGER, PARAMETER :: cp_ = c8_
+ INTEGER, PARAMETER :: rpc_ = C_DOUBLE
 #endif
 
 #ifdef GALAHAD_64BIT_INTEGER
- INTEGER, PARAMETER :: ip = i8
- INTEGER, PARAMETER :: ipc = C_INT64_T
+ INTEGER, PARAMETER :: ip_ = i8_
+ INTEGER, PARAMETER :: ipc_ = C_INT64_T
 #else
- INTEGER, PARAMETER :: ip = i4
[…]

[…]
