+++ src/space/space.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 15/05/2014 AT 15:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-08 AT 07:10 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _ S P A C E   M O D U L E  *-*-*-*-*-*-*-*-*-*
 
@@ -9,10 +11,11 @@
 !  Copyright reserved
 !  January 27th 2005
 
-   MODULE GALAHAD_SPACE_double
+   MODULE GALAHAD_SPACE_precision
 
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SMT_double
+     USE GALAHAD_PRECISION
+     USE GALAHAD_SMT_precision
 
      IMPLICIT NONE
 
@@ -22,11 +25,6 @@
                SPACE_extend_array, SPACE_extend_carray,                        &
                SPACE_dealloc_pointer, SPACE_dealloc_array,                     &
                SPACE_dealloc_smt_type
-
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-     INTEGER, PARAMETER :: wcp = KIND( ( 1.0D+0, 1.0D+0 ) )
 
 !  generic interfaces to cope with both pointer and allocatable
 !  real and integer arrays
@@ -120,10 +118,10 @@
 
 !  Dummy arguments
 
-     INTEGER, INTENT( IN ) :: len
-     INTEGER, INTENT( OUT ) :: status, alloc_status
-     REAL ( KIND = wp ), POINTER, DIMENSION( : ) :: point
-     INTEGER, OPTIONAL :: out
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: len
+     INTEGER ( KIND = ip_ ), INTENT( OUT ) :: status, alloc_status
+     REAL ( KIND = rp_ ), POINTER, DIMENSION( : ) :: point
+     INTEGER ( KIND = ip_ ), OPTIONAL :: out
      LOGICAL, OPTIONAL :: deallocate_error_fatal, exact_size
      CHARACTER ( LEN = 80 ), OPTIONAL :: point_name
      CHARACTER ( LEN = 80 ), OPTIONAL :: bad_alloc
@@ -211,10 +209,10 @@
 
 !  Dummy arguments
 
-     INTEGER, INTENT( IN ) :: l, u
-     INTEGER, INTENT( OUT ) :: status, alloc_status
-     REAL ( KIND = wp ), POINTER, DIMENSION( : ) :: point
-     INTEGER, OPTIONAL :: out
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: l, u
+     INTEGER ( KIND = ip_ ), INTENT( OUT ) :: status, alloc_status
+     REAL ( KIND = rp_ ), POINTER, DIMENSION( : ) :: point
+     INTEGER ( KIND = ip_ ), OPTIONAL :: out
      LOGICAL, OPTIONAL :: deallocate_error_fatal, exact_size
      CHARACTER ( LEN = 80 ), OPTIONAL :: point_name
      CHARACTER ( LEN = 80 ), OPTIONAL :: bad_alloc
@@ -304,10 +302,10 @@
 
 !  Dummy arguments
 
-     INTEGER, INTENT( IN ) :: len1, len2
-     INTEGER, INTENT( OUT ) :: status, alloc_status
-     REAL ( KIND = wp ), POINTER, DIMENSION( : , : ) :: point
-     INTEGER, OPTIONAL :: out
[…]
