+++ src/dum/hsl_zb01s.f90
@@ -1,5 +1,5 @@
-    module zb01_single
+    module hsl_zb01_single
     contains
       subroutine zb01s( )
       end subroutine zb01s
-    end module zb01_single
+    end module hsl_zb01_single
