+++ src/ugo/ugot2.F90
@@ -1,19 +1,19 @@
-   PROGRAM GALAHAD_UGO_EXAMPLE  !  GALAHAD 2.8 - 03/06/2016 AT 08:35 GMT
-   USE GALAHAD_UGO_double                       ! double precision version
-   USE GALAHAD_NLPT_double, ONLY: NLPT_userdata_type
+#include "galahad_modules.h"
+   PROGRAM GALAHAD_UGO_TEST_PROGRAM2  !  GALAHAD 4.1 - 2022-12-09 AT 07:45 GMT
+   USE GALAHAD_USERDATA_precision
+   USE GALAHAD_UGO_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( UGO_control_type ) :: control
    TYPE ( UGO_inform_type ) :: inform
    TYPE ( UGO_data_type ) :: data
-   TYPE ( NLPT_userdata_type ) :: userdata
+   TYPE ( GALAHAD_userdata_type ) :: userdata
    EXTERNAL :: FGH
-   INTEGER :: prob
-   REAL ( KIND = wp ) :: x_l, x_u, x, f, g, h, f_min, x_min
-!  REAL ( KIND = wp ), PARAMETER :: p = 4.0_wp
-   REAL ( KIND = wp ), PARAMETER :: p = 0.0_wp
-   REAL ( KIND = wp ), PARAMETER :: pi = 4.0_wp * ATAN( 1.0_wp )
-   x_l = - 1.0_wp; x_u = 2.0_wp                 ! bounds on x
+   INTEGER ( KIND = ip_ ) :: prob
+   REAL ( KIND = rp_ ) :: x_l, x_u, x, f, g, h, f_min, x_min
+!  REAL ( KIND = rp_ ), PARAMETER :: p = 4.0_rp_
+   REAL ( KIND = rp_ ), PARAMETER :: p = 0.0_rp_
+   REAL ( KIND = rp_ ), PARAMETER :: pi = 4.0_rp_ * ATAN( 1.0_rp_ )
+   x_l = - 1.0_rp_; x_u = 2.0_rp_                 ! bounds on x
    DO
      WRITE( 6, "( ' Input test problem number [1-23]:' )" )
      READ( 5, * ) prob
@@ -29,140 +29,140 @@
 !  problem 01
 
    CASE ( 01 )
-     x_l = - 1.5_wp ; x_u = 11.0_wp
-     f_min = - 29763.233_wp ; x_min = 10.0_wp
+     x_l = - 1.5_rp_ ; x_u = 11.0_rp_
+     f_min = - 29763.233_rp_ ; x_min = 10.0_rp_
 
 !  problem 02
 
    CASE ( 02 )
-     x_l = 2.7_wp ; x_u = 7.5_wp
-     f_min = - 1.899599_wp ; x_min = 5.145735_wp
+     x_l = 2.7_rp_ ; x_u = 7.5_rp_
+     f_min = - 1.899599_rp_ ; x_min = 5.145735_rp_
 
 !  problem 03
 
    CASE ( 03 )
-     x_l = - 10.0_wp ; x_u = 10.0_wp
-     f_min = - 12.03124_wp ; x_min = -6.7745761_wp
+     x_l = - 10.0_rp_ ; x_u = 10.0_rp_
+     f_min = - 12.03124_rp_ ; x_min = -6.7745761_rp_
 
 !  problem 04
 
    CASE ( 04 )
-     x_l = 1.9_wp ; x_u = 3.9_wp
[…]
