+++ src/ugo/ugot.F90
@@ -1,19 +1,19 @@
-   PROGRAM GALAHAD_UGO_TEST_PROGRAM  !  GALAHAD 4.1 - 2022-11-27 AT 14:05 GMT
-   USE GALAHAD_UGO_double                       ! double precision version
-   USE GALAHAD_USERDATA_double
+#include "galahad_modules.h"
+   PROGRAM GALAHAD_UGO_TEST_PROGRAM  !  GALAHAD 4.1 - 2022-12-09 AT 07:45 GMT
+   USE GALAHAD_USERDATA_precision
+   USE GALAHAD_UGO_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( UGO_control_type ) :: control
    TYPE ( UGO_inform_type ) :: inform
    TYPE ( UGO_data_type ) :: data
    TYPE ( GALAHAD_userdata_type ) :: userdata
    EXTERNAL :: FGH
-   INTEGER :: prob
-   REAL ( KIND = wp ) :: x_l, x_u, x, f, g, h, f_min, x_min
-   REAL ( KIND = wp ), PARAMETER :: pi = 4.0_wp * ATAN( 1.0_wp )
-   REAL ( KIND = wp ), PARAMETER :: accurate_x = ( 10.0_wp ) ** ( - 5 )
-   REAL ( KIND = wp ), PARAMETER :: accurate_f = ( 10.0_wp ) ** ( - 5 )
-   x_l = - 1.0_wp; x_u = 2.0_wp                 ! bounds on x
+   INTEGER ( KIND = ip_ ) :: prob
+   REAL ( KIND = rp_ ) :: x_l, x_u, x, f, g, h, f_min, x_min
+   REAL ( KIND = rp_ ), PARAMETER :: pi = 4.0_rp_ * ATAN( 1.0_rp_ )
+   REAL ( KIND = rp_ ), PARAMETER :: accurate_x = ( 10.0_rp_ ) ** ( - 5 )
+   REAL ( KIND = rp_ ), PARAMETER :: accurate_f = ( 10.0_rp_ ) ** ( - 5 )
+   x_l = - 1.0_rp_; x_u = 2.0_rp_                 ! bounds on x
 
    ALLOCATE( userdata%integer( 1 ) )           ! Allocate space for parameter
    CALL UGO_initialize( data, control, inform )
@@ -29,134 +29,134 @@
 !    problem 01
 
      CASE ( 01 )
-       x_l = - 1.5_wp ; x_u = 11.0_wp
-       f_min = - 29763.233_wp ; x_min = 10.0_wp
+       x_l = - 1.5_rp_ ; x_u = 11.0_rp_
+       f_min = - 29763.233_rp_ ; x_min = 10.0_rp_
 
 !    problem 02
 
      CASE ( 02 )
-       x_l = 2.7_wp ; x_u = 7.5_wp
-       f_min = - 1.899599_wp ; x_min = 5.145735_wp
+       x_l = 2.7_rp_ ; x_u = 7.5_rp_
+       f_min = - 1.899599_rp_ ; x_min = 5.145735_rp_
 
 !    problem 03
 
      CASE ( 03 )
-       x_l = - 10.0_wp ; x_u = 10.0_wp
-       f_min = - 12.03124_wp ; x_min = -6.7745761_wp
+       x_l = - 10.0_rp_ ; x_u = 10.0_rp_
+       f_min = - 12.03124_rp_ ; x_min = -6.7745761_rp_
 
 !    problem 04
 
      CASE ( 04 )
-       x_l = 1.9_wp ; x_u = 3.9_wp
[…]
