+++ src/ugo/C/ugo_ciface.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-08-14 AT 11:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-09 AT 08:45 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ U G O   C   I N T E R F A C E  -*-*-*-*-*-*-
 
@@ -11,10 +13,11 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_UGO_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_UGO_precision_ciface
+!   USE ISO_C_BINDING
+    USE GALAHAD_PRECISION
     USE GALAHAD_common_ciface
-    USE GALAHAD_UGO_double, only:                                              &
+    USE GALAHAD_UGO_double, ONLY:                                              &
         f_ugo_time_type      => UGO_time_type,                                 &
         f_ugo_inform_type    => UGO_inform_type,                               &
         f_ugo_control_type   => UGO_control_type,                              &
@@ -28,66 +31,59 @@
         f_ugo_information    => UGO_information,                               &
         f_ugo_terminate      => UGO_terminate
 
-    USE GALAHAD_NLPT_double, only:                                             &
+    USE GALAHAD_NLPT_double, ONLY:                                             &
         f_nlpt_userdata_type => NLPT_userdata_type
 
     IMPLICIT NONE
-
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: ugo_control_type
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: out
-      INTEGER ( KIND = C_INT ) :: print_level
-      INTEGER ( KIND = C_INT ) :: start_print
-      INTEGER ( KIND = C_INT ) :: stop_print
-      INTEGER ( KIND = C_INT ) :: print_gap
-      INTEGER ( KIND = C_INT ) :: maxit
-      INTEGER ( KIND = C_INT ) :: initial_points
-      INTEGER ( KIND = C_INT ) :: storage_increment
-      INTEGER ( KIND = C_INT ) :: buffer
-      INTEGER ( KIND = C_INT ) :: lipschitz_estimate_used
-      INTEGER ( KIND = C_INT ) :: next_interval_selection
-      INTEGER ( KIND = C_INT ) :: refine_with_newton
-      INTEGER ( KIND = C_INT ) :: alive_unit
+      INTEGER ( KIND = ipc_ ) :: error
+      INTEGER ( KIND = ipc_ ) :: out
+      INTEGER ( KIND = ipc_ ) :: print_level
+      INTEGER ( KIND = ipc_ ) :: start_print
[…]
