### 2022-12-09 c17a1a2
created c galahad precision file
