### 2022-12-09 3774147
make c tests precsision dependent
