### 2022-12-09 3dd8b6b
make c tests precsision dependent
