+++ src/ugo/C/ugos.c
@@ -3,20 +3,21 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
 #include "galahad_ugo.h"
 
 struct userdata_type {
-   double a;
+   real_wp_ a;
 };
 
-// Evaluate test problem objective, first and second derivatives 
-int fgh(double x, double *f, double *g, double *h, const void *userdata){
+// Evaluate test problem objective, first and second derivatives
+int fgh(real_wp_ x, real_wp_ *f, real_wp_ *g, real_wp_ *h, const void *userdata){
    struct userdata_type *myuserdata = (struct userdata_type *) userdata;
-   double a = myuserdata->a;
+   real_wp_ a = myuserdata->a;
 
    *f = x * x * cos( a*x );
    *g = - a * x * x * sin( a*x ) + 2.0 * x * cos( a*x );
-   *h = - a * a* x * x * cos( a*x ) - 4.0 * a * x * sin( a*x ) 
+   *h = - a * a* x * x * cos( a*x ) - 4.0 * a * x * sin( a*x )
         + 2.0 * cos( a*x );
    return 0;
 }
@@ -43,20 +44,20 @@ int main(void) {
     userdata.a = 10.0;
 
     // Test problem bounds
-    double x_l = -1.0; 
-    double x_u = 2.0;     
+    real_wp_ x_l = -1.0;
+    real_wp_ x_u = 2.0;
 
     // Test problem objective, gradient, Hessian values
-    double x, f, g, h;
+    real_wp_ x, f, g, h;
 
     // import problem data
     ugo_import( &control, &data, &status, &x_l, &x_u );
 
     // Set for initial entry
-    status = 1; 
-    
+    status = 1;
+
     // Call UGO_solve
-    ugo_solve_direct( &data, &userdata, &status, &x, &f, &g, &h, fgh );    
+    ugo_solve_direct( &data, &userdata, &status, &x, &f, &g, &h, fgh );
 
     // Record solution information
     ugo_information( &data, &inform, &status );
