### 2022-12-11 de3414d
Add a Meson script for UGO to generate single and double precision libraries
