+++ meson_options.txt
@@ -1,0 +1,14 @@
+option('ciface',
+       type : 'boolean',
+       value : false,
+       description : 'whether to build the C interfaces')
+
+option('libblas',
+       type : 'array',
+       value : [],
+       description : 'array [path, libname] with a path to search for BLAS library named libname')
+
+option('liblapack',
+       type : 'array',
+       value : [],
+       description : 'array [path, libname] with a path to search for LAPACK library named libname')
