### 2022-12-15 3aa0b71
linear and subproblem solvers updated to new precision format
