### 2022-12-15 c5588de
linear and subproblem solvers updated to new precision format
