+++ src/ir/irt.F90
@@ -1,10 +1,11 @@
-! THIS VERSION: GALAHAD 3.3 - 29/05/2021 AT 11:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-14 AT 15:30 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_IR_TEST  !  further work needed!!
-   USE GALAHAD_IR_double                           ! double precision version
-   USE GALAHAD_SMT_double
-   USE GALAHAD_SLS_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_IR_precision                           ! double precision version
+   USE GALAHAD_SMT_precision
+   USE GALAHAD_SLS_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )      ! set precision
    TYPE ( SMT_type ) :: matrix
    TYPE ( SLS_data_type ) :: SLS_data
    TYPE ( SLS_control_type ) SLS_control
@@ -12,10 +13,10 @@
    TYPE ( IR_data_type ) :: data
    TYPE ( IR_control_type ) :: control
    TYPE ( IR_inform_type ) :: inform
-   INTEGER, PARAMETER :: n = 5
-   INTEGER, PARAMETER :: ne = 7
-   REAL ( KIND = wp ) :: B( n ), X( n )
-   INTEGER :: i, s
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 5
+   INTEGER ( KIND = ip_ ), PARAMETER :: ne = 7
+   REAL ( KIND = rp_ ) :: B( n ), X( n )
+   INTEGER ( KIND = ip_ ) :: i, s
 ! Read matrix order and number of entries
    matrix%n = n
    matrix%ne = ne
@@ -23,11 +24,11 @@
    ALLOCATE( matrix%val( ne ), matrix%row( ne ), matrix%col( ne ) )
    matrix%row( : ne ) = (/ 1, 1, 2, 2, 3, 3, 5 /)
    matrix%col( : ne ) = (/ 1, 2, 3, 5, 3, 4, 5 /)
-   matrix%val( : ne ) = (/ 2.0_wp, 3.0_wp, 4.0_wp, 6.0_wp, 1.0_wp,             &
-                           5.0_wp, 1.0_wp /)
+   matrix%val( : ne ) = (/ 2.0_rp_, 3.0_rp_, 4.0_rp_, 6.0_rp_, 1.0_rp_,        &
+                           5.0_rp_, 1.0_rp_ /)
    CALL SMT_put( matrix%type, 'COORDINATE', s )     ! Specify co-ordinate
 ! Set right-hand side
-   B( : n ) = (/ 8.0_wp, 45.0_wp, 31.0_wp, 15.0_wp, 17.0_wp /)
+   B( : n ) = (/ 8.0_rp_, 45.0_rp_, 31.0_rp_, 15.0_rp_, 17.0_rp_ /)
 ! Specify the solver (in this case sils)
    CALL SLS_initialize( 'sils', SLS_data, SLS_control, SLS_inform )
 ! Analyse
