### 2022-12-15 4c6691f
linear and subproblem solvers updated to new precision format
