+++ src/sils/old_sils.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-11 AT 09:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*- G A L A H A D _ S I L S    M O D U L E  -*-*-*-*-*-*-*-*-*-
 
@@ -12,7 +14,9 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_SILS_double
+   MODULE GALAHAD_SILS_precision
+            
+     USE GALAHAD_PRECISION
 
 !     ---------------------------------------------
 !     |                                           |
@@ -25,7 +29,7 @@
 !     |                                           |
 !     ---------------------------------------------
 
-     USE GALAHAD_SMT_double
+     USE GALAHAD_SMT_precision
 
      IMPLICIT NONE
  
@@ -35,16 +39,13 @@
                SMT_type
    
 !--------------------
-!   P r e c i s i o n
-!--------------------
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )  
 
 !  Set other parameters
 
-     REAL ( KIND = wp ), PRIVATE, PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PRIVATE, PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PRIVATE, PARAMETER :: one = 1.0_wp
+     REAL ( KIND = rp_ ), PRIVATE, PARAMETER :: zero = 0.0_rp_
+     REAL ( KIND = rp_ ), PRIVATE, PARAMETER :: half = 0.5_rp_
+     REAL ( KIND = rp_ ), PRIVATE, PARAMETER :: one = 1.0_rp_
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -52,54 +53,54 @@
 
      TYPE, PUBLIC :: SILS_factors
        PRIVATE 
-       INTEGER, ALLOCATABLE, DIMENSION( : )  :: keep 
-       INTEGER, ALLOCATABLE, DIMENSION( : )  :: iw 
-       INTEGER, ALLOCATABLE, DIMENSION( : )  :: iw1 
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : )  :: val 
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : )  :: w ! len maxfrt 
-       REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : )  :: r ! length n 
-       INTEGER :: n       ! Matrix order
-       INTEGER :: nrltot  ! Size for val without compression
-       INTEGER :: nirtot  ! Size for iw without compression
-       INTEGER :: nrlnec  ! Size for val with compression
-       INTEGER :: nirnec  ! Size for iw with compression
-       INTEGER :: nsteps  ! Number of elimination steps
-       INTEGER :: maxfrt  ! Largest front size
-       INTEGER :: latop   ! Position of final entry of val
-       INTEGER :: dim_iw1 ! Size of iw1 for solves
-       INTEGER :: pivoting ! type of pivoting used
-       REAL ( KIND = wp ) :: ops
+       INTEGER ( KIND = ip_ ) ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : )  :: keep 
[…]
