+++ src/slls/slls.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-09-28 AT 13:55 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-11 AT 09:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S L L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -11,8 +13,8 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_SLLS_double
-
+   MODULE GALAHAD_SLLS_precision
+            
 !        --------------------------------------------------------------------
 !        |                                                                  |
 !        | Solve the simplex-constrained linear-least-squares problem       |
@@ -26,21 +28,22 @@
 !        |                                                                  |
 !        --------------------------------------------------------------------
 
+     USE GALAHAD_PRECISION
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
      USE GALAHAD_STRING
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_SORT_double, ONLY: SORT_heapsort_build,                       &
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_SORT_precision, ONLY: SORT_heapsort_build,                    &
                                     SORT_heapsort_smallest, SORT_quicksort
-     USE GALAHAD_SBLS_double
-     USE GALAHAD_NORMS_double
-     USE GALAHAD_QPT_double
-     USE GALAHAD_QPD_double, ONLY: QPD_SIF
-     USE GALAHAD_USERDATA_double
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_CONVERT_double, ONLY: CONVERT_control_type,                   &
-                                       CONVERT_inform_type,                    &
-                                       CONVERT_to_sparse_column_format
+     USE GALAHAD_SBLS_precision
+     USE GALAHAD_NORMS_precision
+     USE GALAHAD_QPT_precision
+     USE GALAHAD_QPD_precision, ONLY: QPD_SIF
+     USE GALAHAD_USERDATA_precision
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_CONVERT_precision, ONLY: CONVERT_control_type,                &
+                                          CONVERT_inform_type,                 &
+                                          CONVERT_to_sparse_column_format
      IMPLICIT NONE
 
      PRIVATE
@@ -63,39 +66,40 @@
 
      INTERFACE SLLS_initialize
        MODULE PROCEDURE SLLS_initialize, SLLS_full_initialize
+            
+     USE GALAHAD_PRECISION
      END INTERFACE SLLS_initialize
 
      INTERFACE SLLS_terminate
        MODULE PROCEDURE SLLS_terminate, SLLS_full_terminate
+            
+     USE GALAHAD_PRECISION
      END INTERFACE SLLS_terminate
 
 !--------------------
-!   P r e c i s i o n
-!--------------------
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
[…]
