### 2022-12-15 a8f845c
linear and subproblem solvers updated to new precision format
