+++ src/lstr/runlstr_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 27/05/2014 AT 14:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-15 AT 15:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D   R U N L S T R _ S I F  *-*-*-*-*-*-*-*-*-*-
 
@@ -19,11 +21,12 @@
 !    | regularized linear least-squares              |
 !     -----------------------------------------------
 
-   USE GALAHAD_USELSTR_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USELSTR_precision
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
 
 !  Open the data input file

[…]
