### 2022-12-15 90a2570
linear and subproblem solvers updated to new precision format
