+++ src/symbols/symbols.F90
@@ -1,4 +1,5 @@
 ! THIS VERSION: GALAHAD 3.3 - 28/09/2021 AT 11:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-10 AT 11:15 GMT.
 
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -30,6 +31,7 @@
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 
+      USE GALAHAD_PRECISION, ONLY : ip_
       IMPLICIT NONE
       PUBLIC :: SYMBOLS_status
 
@@ -47,206 +49,359 @@
 
 !  New exit conditions (1 to -99; others will be package specific)
 
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_ok                      = 0
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_ready_to_solve          = 1
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_allocate          = - 1
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_deallocate        = - 2
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_restrictions      = - 3
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_bad_bounds        = - 4
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_primal_infeasible = - 5
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_dual_infeasible   = - 6
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_unbounded         = - 7
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_no_center         = - 8
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_analysis          = - 9
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_factorization     = - 10
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_solve             = - 11
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_uls_analysis      = - 12
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_uls_factorization = - 13
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_uls_solve         = - 14
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_preconditioner    = - 15
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_ill_conditioned   = - 16
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_tiny_step         = - 17
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_no_progress             =          &
-                                    GALAHAD_error_tiny_step
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_max_iterations    = - 18
-      INTEGER, PUBLIC, PARAMETER :: GALAHAD_error_time_limit        = - 19
[…]
