### 2022-12-15 ecffa3f
Add a Julia interface for single and double precision UGO
