### 2022-12-16 4a53979
added missing second_derivative_available control to ugo_ciface
