### 2022-12-16 1512e72
added missing second_derivative_available control to ugo_ciface
