### 2022-12-16 9ccb8de
nonlinear and structured subproblem solvers updated to new precision format
