+++ src/rand/rand.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 03/07/2014 AT 12:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 10:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*  G A L A H A D _ R A N D   M O D U L E  *-*-*-*-*-*-*-*-*-*-
 
@@ -12,9 +14,11 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-      MODULE GALAHAD_RAND_double
-
+      MODULE GALAHAD_RAND_precision
+            
 !  Portable random number generator by Linus Schrange, TOMS 5, 1979, pp 132-138
+
+         USE GALAHAD_PRECISION
 
          IMPLICIT NONE
 
@@ -24,15 +28,14 @@
 
 !  Define the working precision to be double
 
-         INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
-         INTEGER, PARAMETER :: a = 16807, b15 = 32768
-         INTEGER, PARAMETER :: b16 = 65536, p = 2147483647
-         INTEGER, PARAMETER :: b30 = 1073741824, q = 1073741823 
+         INTEGER ( KIND = ip_ ), PARAMETER :: a = 16807, b15 = 32768
+         INTEGER ( KIND = ip_ ), PARAMETER :: b16 = 65536, p = 2147483647
+         INTEGER ( KIND = ip_ ), PARAMETER :: b30 = 1073741824, q = 1073741823 
 
          TYPE, PUBLIC :: RAND_seed
            PRIVATE
-           INTEGER :: ix =  b16 - 1
+           INTEGER ( KIND = ip_ ) :: ix =  b16 - 1
          END TYPE
 
          INTERFACE RAND_random_real
@@ -75,16 +78,16 @@
 
          TYPE ( RAND_seed ), INTENT( INOUT ) :: seed
          LOGICAL, INTENT( IN ) :: positive
-         REAL ( KIND = wp ), INTENT( OUT ) :: random
+         REAL ( KIND = rp_ ), INTENT( OUT ) :: random
 
 !-----------------------------------------------
 !   L o c a l   V a r i a b l e s
 !-----------------------------------------------
 
-         INTEGER :: fhi, k, leftlo, xalo, xhi 
-         REAL ( KIND = wp ) :: x 
-         REAL ( KIND = wp ), PARAMETER ::                                      &
-                one = 1.0_wp, two = 2.0_wp, rb16 = two ** 16,                  &
+         INTEGER ( KIND = ip_ ) :: fhi, k, leftlo, xalo, xhi 
+         REAL ( KIND = rp_ ) :: x 
+         REAL ( KIND = rp_ ), PARAMETER ::                                     &
+                one = 1.0_rp_, two = 2.0_rp_, rb16 = two ** 16,                &
                 big  = one / ( two ** 31 - one ), big2 = two * big
 
          xhi = seed%ix / b16 
@@ -139,13 +142,13 @@
 
          TYPE ( RAND_seed ), INTENT( INOUT ) :: seed
          LOGICAL, INTENT( IN ) :: positive
-         REAL ( KIND = wp ), INTENT( OUT ), DIMENSION( : ) :: RANDOM
+         REAL ( KIND = rp_ ), INTENT( OUT ), DIMENSION( : ) :: RANDOM
 
 !-----------------------------------------------
 !   L o c a l   V a r i a b l e s
 !-----------------------------------------------
 
-         INTEGER :: i, dim1
[…]
