+++ src/lmt/lmt.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 15/12/2014 AT 09:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 10:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-*- G A L A H A D _ L M T   M O D U L E -*-*-*-*-*-*-*-*-*-
 
@@ -12,23 +14,19 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-    MODULE GALAHAD_LMT_double
-
+    MODULE GALAHAD_LMT_precision
+            
 !    ----------------------------------------------------------------
 !   |                                                                |
 !   | Derived types for limited-memory secant Hessian approximations |
 !   |                                                                |
 !    ----------------------------------------------------------------
 
-     IMPLICIT NONE
+      USE GALAHAD_PRECISION
+
+      IMPLICIT NONE
 
       PRIVATE
-
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -42,19 +40,19 @@
 
 !  unit for error messages
 
-        INTEGER :: error = 6
+        INTEGER ( KIND = ip_ ) :: error = 6
 
 !  unit for monitor output
 
-        INTEGER :: out = 6
+        INTEGER ( KIND = ip_ ) :: out = 6
 
 !  controls level of diagnostic output
 
-        INTEGER :: print_level = 0
+        INTEGER ( KIND = ip_ ) :: print_level = 0
 
 !  limited memory length
 
-        INTEGER :: memory_length = 10
+        INTEGER ( KIND = ip_ ) :: memory_length = 10
 
 !  limited-memory method required (others may be added in due course):
 !    1 BFGS (default)
@@ -62,7 +60,7 @@
 !    3 BFGS inverse
 !    4 shifted BFGS inverse
 
-        INTEGER :: method = 1
+        INTEGER ( KIND = ip_ ) :: method = 1
 
 !  allow space to permit different methods if required (less efficient)
 
@@ -92,35 +90,35 @@
 
 !  total cpu time spent in the package
 
-        REAL ( KIND = wp ) :: total = 0.0
+        REAL ( KIND = rp_ ) :: total = 0.0
 
 !  cpu time spent setting up space for the secant approximation
 
-        REAL ( KIND = wp ) :: setup = 0.0
+        REAL ( KIND = rp_ ) :: setup = 0.0
 
 !  cpu time spent updating the secant approximation
 
-        REAL ( KIND = wp ) :: form = 0.0
+        REAL ( KIND = rp_ ) :: form = 0.0
 
 !  cpu time spent applying the secant approximation
 
-        REAL ( KIND = wp ) :: apply = 0.0
+        REAL ( KIND = rp_ ) :: apply = 0.0
 
 !  total clock time spent in the package
 
-        REAL ( KIND = wp ) :: clock_total = 0.0
+        REAL ( KIND = rp_ ) :: clock_total = 0.0
 
 !  clock time spent setting up space for the secant approximation
 
-        REAL ( KIND = wp ) :: clock_setup = 0.0
[…]
