### 2022-12-16 e3d3e72
nonlinear and structured subproblem solvers updated to new precision format
