+++ src/lms/lmst.F90
@@ -1,17 +1,23 @@
-! THIS VERSION: GALAHAD 2.6 - 12/06/2014 AT 15:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 10:30 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_LMS_test
-   USE GALAHAD_LMS_double                   ! double precision version
-   USE GALAHAD_rand_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_LMS_precision
+   USE GALAHAD_RAND_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   INTEGER, PARAMETER :: n = 5, m = 3
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 5, m = 3
    TYPE ( LMS_data_type ) :: data, data2
    TYPE ( LMS_control_type ) :: control, control2       
    TYPE ( LMS_inform_type ) :: inform, inform2
-   REAL ( KIND = wp ), DIMENSION( n ) :: S, Y, U, V
-   INTEGER :: iter, fail, method, mem
-   REAL ( KIND = wp ) :: delta, lambda
+   REAL ( KIND = rp_ ), DIMENSION( n ) :: S, Y, U, V
+   INTEGER ( KIND = ip_ ) :: iter, fail, method, mem
+   REAL ( KIND = rp_ ) :: delta, lambda
    TYPE ( RAND_seed ) :: seed
+#ifdef GALAHAD_SINGLE
+   REAL ( KIND = sp_ ) :: error_tol = 1.0_sp_
+#else
+   REAL ( KIND = dp_ ) :: error_tol = 0.00001_dp_ 
+#endif
 
 ! ==================== tests error returns ======================
 
@@ -26,7 +32,7 @@
    CALL LMS_initialize( data, control, inform ) !  initialize data
    control%print_level = 1
    CALL LMS_setup( n, data, control, inform )
-   delta = 0.0_wp ; S = 1.0_wp ; Y = S 
+   delta = 0.0_rp_ ; S = 1.0_rp_ ; Y = S 
    CALL LMS_form( S, Y, delta, data, control, inform ) ! delta <= 0
    WRITE( 6, "( ' exit status = ', I0 )" ) inform%status
    CALL LMS_terminate( data, control, inform )  !  delete internal workspace
@@ -35,8 +41,8 @@
    control%print_level = 1
    control%method = 4
    CALL LMS_setup( n, data, control, inform )
-   delta = 1.0_wp ; S = 1.0_wp ; Y = S 
-   CALL LMS_form( S, Y, delta, data, control, inform, lambda = - 1.0_wp )
+   delta = 1.0_rp_ ; S = 1.0_rp_ ; Y = S 
+   CALL LMS_form( S, Y, delta, data, control, inform, lambda = - 1.0_rp_ )
    WRITE( 6, "( ' exit status = ', I0 )" ) inform%status
    CALL LMS_terminate( data, control, inform )  !  delete internal workspace
 
@@ -44,16 +50,16 @@
    control%print_level = 1
    control%method = 4
    CALL LMS_setup( n, data, control, inform )
-   delta = 0.0_wp ; S = 1.0_wp ; Y = S
+   delta = 0.0_rp_ ; S = 1.0_rp_ ; Y = S
    CALL LMS_form( S, Y, delta, data, control, inform )
-   CALL LMS_form_shift( - 1.0_wp, data, control, inform ) ! lambda < 0
+   CALL LMS_form_shift( - 1.0_rp_, data, control, inform ) ! lambda < 0
[…]
