+++ src/roots/rootst.F90
@@ -1,13 +1,14 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 11:50 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_ROOTS_test_deck
-   USE GALAHAD_ROOTS_double                       ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_ROOTS_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+   REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
 
-   INTEGER :: i, order, type, nroots
-   REAL ( KIND = wp ) :: root, A( 0 : 4 ), ROOTS( 4 ), C( 0 : 10 )
-   REAL ( KIND = wp ) :: A1( 0 : 5 ), A2( 0: 2 ), ROOTS2( 1 )
+   INTEGER ( KIND = ip_ ) :: i, order, type, nroots
+   REAL ( KIND = rp_ ) :: root, A( 0 : 4 ), ROOTS( 4 ), C( 0 : 10 )
+   REAL ( KIND = rp_ ) :: A1( 0 : 5 ), A2( 0: 2 ), ROOTS2( 1 )
    LOGICAL :: debug = .FALSE.
    TYPE ( ROOTS_data_type ) :: data
    TYPE ( ROOTS_control_type ) :: control        
@@ -31,28 +32,28 @@
 
 !  STOP
 
-   A( 0 ) = 3.0_wp / 16.0_wp
-   A( 1 ) = - 1.0_wp
-   A( 2 ) = 19.0_wp / 16.0_wp
-   A( 3 ) = - 1.0_wp
-   A( 4 ) = 1.0_wp
-!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_wp, 99.0_wp,         &
+   A( 0 ) = 3.0_rp_ / 16.0_rp_
+   A( 1 ) = - 1.0_rp_
+   A( 2 ) = 19.0_rp_ / 16.0_rp_
+   A( 3 ) = - 1.0_rp_
+   A( 4 ) = 1.0_rp_
+!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_rp_, 99.0_rp_,       &
 !                                           data, control, inform )
-!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_wp, 1.0_wp,          &
+!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_rp_, 1.0_rp_,        &
 !                                           data, control, inform )
-!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_wp, 0.5_wp,          &
+!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_rp_, 0.5_rp_,        &
 !                                           data, control, inform )
-!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_wp, 0.25_wp,         &
+!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_rp_, 0.25_rp_,       &
 !                                           data, control, inform )
-!   root = ROOTS_smallest_root_in_interval( A( : 4 ), 0.0_wp, 0.24_wp,         &
[…]
