+++ src/rqs/runrqs_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.3 - 05/01/2009 AT 16:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 15:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D   R U N R Q S _ S I F  *-*-*-*-*-*-*-*-
 
@@ -7,17 +9,18 @@
 !  November 24th 2008
 
    PROGRAM RUNRQS_SIF
-   USE GALAHAD_USERQS_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USERQS_precision
 
 !  Main program for the SIF interface to RQS, a solver for the regularised
 !  quadratic mninimization subproblem
 
 !  Problem insif characteristics
 
-   INTEGER, PARAMETER :: errout = 6
-   INTEGER, PARAMETER :: insif = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: errout = 6
+   INTEGER ( KIND = ip_ ), PARAMETER :: insif = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
-   INTEGER :: iostat
+   INTEGER ( KIND = ip_ ) :: iostat
 
 !  Open the data input file
 

[…]
