### 2022-12-16 0098931
nonlinear and structured subproblem solvers updated to new precision format
