+++ src/trs/trs_paper.F90
@@ -1,60 +1,61 @@
-   PROGRAM GALAHAD_TRS_EXAMPLE  !  GALAHAD 2.2 - 05/06/2008 AT 13:30 GMT.
-   USE GALAHAD_TRS_DOUBLE                          ! double precision version
+   PROGRAM GALAHAD_TRS_EXAMPLE  !  GALAHAD 4.1 - 2022-12-16 AT 14:10 GMT.
+#include "galahad_modules.h"
+   USE GALAHAD_PRECISION
+   USE GALAHAD_TRS_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )       ! set precision
-   REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-   INTEGER, PARAMETER :: n = 3, h_ne = 4           ! problem dimensions
-   INTEGER :: i, s
-   REAL ( KIND = wp ), DIMENSION( n ) :: C, X
+   REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 3, h_ne = 4 ! problem dimensions
+   INTEGER ( KIND = ip_ ) :: i, s
+   REAL ( KIND = rp_ ), DIMENSION( n ) :: C, X
    TYPE ( SMT_type ) :: H, M
    TYPE ( TRS_data_type ) :: data
    TYPE ( TRS_control_type ) :: control        
    TYPE ( TRS_inform_type ) :: inform
    LOGICAL, PARAMETER :: use_m = .FALSE.
-   REAL ( KIND = wp ) :: f = 0.0_wp           ! constant term, f
-!  REAL ( KIND = wp ) :: radius = 0.001_wp    ! trust-region radius (small)
-   REAL ( KIND = wp ) :: radius = 1.0_wp      ! trust-region radius (medium)
-!  REAL ( KIND = wp ) :: radius = 100000.0_wp ! trust-region radius (large)
+   REAL ( KIND = rp_ ) :: f = 0.0_rp_           ! constant term, f
+!  REAL ( KIND = rp_ ) :: radius = 0.001_rp_    ! trust-region radius (small)
+   REAL ( KIND = rp_ ) :: radius = 1.0_rp_      ! trust-region radius (medium)
+!  REAL ( KIND = rp_ ) :: radius = 100000.0_rp_ ! trust-region radius (large)
    IF ( use_m ) THEN
 !    CALL SMT_put( M%type, 'DIAGONAL', s )      ! Specify diagonal for M
 !    ALLOCATE( M%val( n ) )                     ! 
-!    M%val = (/ 1.0_wp, 1.0_wp, 1.0_wp /)       ! M-norm, M
+!    M%val = (/ 1.0_rp_, 1.0_rp_, 1.0_rp_ /)       ! M-norm, M
      CALL SMT_put( M%type, 'COORDINATE', s )    ! Specify co-ordinate for H
      ALLOCATE( M%val( h_ne ), M%row( h_ne ), M%col( h_ne ) )
-     M%val = (/ 1.0_wp, 1.0_wp, 1.0_wp, 0.9_wp /) ! M-norm, M
+     M%val = (/ 1.0_rp_, 1.0_rp_, 1.0_rp_, 0.9_rp_ /) ! M-norm, M
      M%row = (/ 1, 2, 3, 2 /)                     ! NB lower triangle
      M%col = (/ 1, 2, 3, 1 /) ; M%ne = h_ne
    END IF
    CALL SMT_put( H%type, 'COORDINATE', s )    ! Specify co-ordinate for H
    ALLOCATE( H%val( h_ne ), H%row( h_ne ), H%col( h_ne ) )
[…]
