+++ src/rqs/userqs.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.5 - 09/02/2013 AT 17:55 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 15:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D   U S E _ R Q S  -*-*-*-*-*-*-*-*-*-*-
 
@@ -6,18 +8,19 @@
 !  Copyright reserved
 !  November 24th 2008
 
-   MODULE GALAHAD_USERQS_double
-
-!  This is the driver program for running RQS for a variety of computing
+   MODULE GALAHAD_USERQS_precision
+            
+ !  This is the driver program for running RQS for a variety of computing
 !  systems. It opens and closes all the files, allocate arrays, reads and
 !  checks data, and calls the appropriate minimizers
 
-     USE CUTEst_interface_double
+     USE GALAHAD_PRECISION
+     USE CUTEst_interface_precision
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_RQS_double
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_RQS_precision
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_COPYRIGHT
-     USE GALAHAD_SPACE_double
+     USE GALAHAD_SPACE_precision
      IMPLICIT NONE
 
      PRIVATE
@@ -31,11 +34,10 @@
 
 !  Dummy argument
 
-     INTEGER, INTENT( IN ) :: input
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
 !  Set precision
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !-------------------------------
 !   D e r i v e d   T y p e s
@@ -49,22 +51,22 @@
 !   L o c a l   P a r a m e t e r s
 !------------------------------------
 
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-!    REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
+     REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+!    REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
 
 !----------------------------------
 !   L o c a l   V a r i a b l e s
 !----------------------------------
 
-     INTEGER :: iores, i, j, ir, ic, l, smt_stat, cutest_status
+     INTEGER ( KIND = ip_ ) :: iores, i, j, ir, ic, l, smt_stat, cutest_status
 
 !  Problem characteristics
 
-     INTEGER :: n, nnzh
-     INTEGER :: n_threads = 1
-     REAL ( KIND = wp ) ::  f
+     INTEGER ( KIND = ip_ ) :: n, nnzh
+     INTEGER ( KIND = ip_ ) :: n_threads = 1
+     REAL ( KIND = rp_ ) ::  f
      CHARACTER ( LEN = 10 ) :: pname
-     REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, X0, X_l, X_u, G
+     REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, X0, X_l, X_u, G
      CHARACTER ( LEN = 10 ), ALLOCATABLE, DIMENSION( : ) :: VNAMES
      TYPE ( SMT_type ) :: H
 
@@ -74,8 +76,8 @@
 
[…]
