+++ src/sbls/sbls.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-18 AT 14:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 12:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ S B L S   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -13,9 +15,9 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-    MODULE GALAHAD_SBLS_double
-
-!      ---------------------------------------------------------------
+    MODULE GALAHAD_SBLS_precision
+            
+ !      ---------------------------------------------------------------
 !     |                                                               |
 !     | Given matrices A and (symmetric) H and C, provide and apply   |
 !     | preconditioners for the symmetric block linear system         |
@@ -32,19 +34,20 @@
 !     |                                                               |
 !      ---------------------------------------------------------------
 
+     USE GALAHAD_PRECISION
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
-      USE GALAHAD_SPACE_double
-      USE GALAHAD_SMT_double
-      USE GALAHAD_QPT_double, ONLY: QPT_keyword_H, QPT_keyword_A
-      USE GALAHAD_LMT_double, LMS_control_type => LMT_control_type,            &
-                              LMS_data_type => LMT_data_type
-      USE GALAHAD_SORT_double, ONLY: SORT_reorder_by_rows
-      USE GALAHAD_ROOTS_double, ONLY : ROOTS_quadratic
-      USE GALAHAD_SLS_double
-      USE GALAHAD_ULS_double
-      USE GALAHAD_SPECFILE_double
-      USE GALAHAD_NORMS_double, ONLY: TWO_norm
+      USE GALAHAD_SPACE_precision
+      USE GALAHAD_SMT_precision
+      USE GALAHAD_QPT_precision, ONLY: QPT_keyword_H, QPT_keyword_A
+      USE GALAHAD_LMT_precision, LMS_control_type => LMT_control_type,         &
+                                 LMS_data_type => LMT_data_type
+      USE GALAHAD_SORT_precision, ONLY: SORT_reorder_by_rows
+      USE GALAHAD_ROOTS_precision, ONLY : ROOTS_quadratic
+      USE GALAHAD_SLS_precision
+      USE GALAHAD_ULS_precision
+      USE GALAHAD_SPECFILE_precision
+      USE GALAHAD_NORMS_precision, ONLY: TWO_norm
       USE GALAHAD_BLAS_interface, ONLY : GEMV
       USE GALAHAD_LAPACK_interface, ONLY : POTRF, POTRS, SYTRF, SYTRS
 
@@ -73,26 +76,18 @@
        MODULE PROCEDURE SBLS_terminate, SBLS_full_terminate
      END INTERFACE SBLS_terminate
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-      INTEGER, PARAMETER :: real_bytes = 8
-      INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
[…]

[…]
