### 2022-12-16 f76859c
nonlinear and structured subproblem solvers updated to new precision format
