+++ src/trs/trst.F90
@@ -1,14 +1,15 @@
-! THIS VERSION: GALAHAD 2.3 - 03/11/2008 AT 11:45 GMT.
-   PROGRAM GALAHAD_TRS_test_deck
-   USE GALAHAD_TRS_DOUBLE                            ! double precision version
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 14:10 GMT.
+#include "galahad_modules.h"
+   PROGRAM GALAHAD_TRS_test_program
+   USE GALAHAD_PRECISION
+   USE GALAHAD_TRS_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp, two = 2.0_wp
-   INTEGER :: i, smt_stat, n, nn, pass, h_ne, m_ne, a_ne, data_storage_type
-   INTEGER :: ia, im, ifa
-   REAL ( KIND = wp ) :: f, radius
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, C
+   REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_, two = 2.0_rp_
+   INTEGER ( KIND = ip_ ) :: i, smt_stat, n, nn, pass, h_ne, m_ne, a_ne
+   INTEGER ( KIND = ip_ ) :: ia, im, ifa, data_storage_type
+   REAL ( KIND = rp_ ) :: f, radius
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, C
    TYPE ( SMT_type ) :: H, M, A
    TYPE ( TRS_data_type ) :: data
    TYPE ( TRS_control_type ) :: control
@@ -16,14 +17,13 @@
 
    CHARACTER ( len = 1 ) :: st, afa
    CHARACTER ( len = 2 ) :: ma
-   INTEGER, PARAMETER :: n_errors = 5
-   INTEGER, DIMENSION( n_errors ) :: errors = (/                               &
+   INTEGER ( KIND = ip_ ), PARAMETER :: n_errors = 5
+   INTEGER ( KIND = ip_ ), DIMENSION( n_errors ) :: errors = (/                &
        GALAHAD_error_restrictions,                                             &
        GALAHAD_error_restrictions,                                             &
        GALAHAD_error_preconditioner,                                           &
        GALAHAD_error_ill_conditioned,                                          &
-       GALAHAD_error_max_iterations                                            &
-         /)
+       GALAHAD_error_max_iterations /)
 
 ! Initialize output unit
 
@@ -35,18 +35,18 @@
 !  =============
 
    n = 5
-   f = 1.0_wp
+   f = 1.0_rp_
    CALL SMT_put( H%type, 'COORDINATE', smt_stat )    ! Specify co-ordinate for H
    H%ne = 2 * n - 1
    ALLOCATE( H%val( H%ne ), H%row( H%ne ), H%col( H%ne ) )
    DO i = 1, n
-    H%row( i ) = i ; H%col( i ) = i ; H%val( i ) = - 2.0_wp
+    H%row( i ) = i ; H%col( i ) = i ; H%val( i ) = - 2.0_rp_
    END DO
    DO i = 1, n - 1
[…]
