### 2022-12-16 bfec9f4
nonlinear and structured subproblem solvers updated to new precision format
