+++ src/rqs/rqst.F90
@@ -1,15 +1,16 @@
-! THIS VERSION: GALAHAD 2.3 - 03/11/2008 AT 11:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 14:40 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_RQS_test_deck
-   USE GALAHAD_RQS_DOUBLE                            ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_RQS_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp, two = 2.0_wp
-   REAL ( KIND = wp ), PARAMETER :: epsmch = EPSILON( one )
-   INTEGER :: i, smt_stat, n, nn, pass, h_ne, m_ne, a_ne, data_storage_type
-   INTEGER :: ia, im, ifa
-   REAL ( KIND = wp ) :: f, sigma, p
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, C
+   REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_, two = 2.0_rp_
+   REAL ( KIND = rp_ ), PARAMETER :: epsmch = EPSILON( one )
+   INTEGER ( KIND = ip_ ) :: i, smt_stat, n, nn, pass, h_ne, m_ne, a_ne
+   INTEGER ( KIND = ip_ ) :: ia, im, ifa, data_storage_type
+   REAL ( KIND = rp_ ) :: f, sigma, p
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, C
    TYPE ( SMT_type ) :: H, M, A
    TYPE ( RQS_data_type ) :: data
    TYPE ( RQS_control_type ) :: control
@@ -17,15 +18,14 @@
 
    CHARACTER ( len = 1 ) :: st, afa
    CHARACTER ( len = 2 ) :: ma
-   INTEGER, PARAMETER :: n_errors = 6
-   INTEGER, DIMENSION( n_errors ) :: errors = (/                               &
+   INTEGER ( KIND = ip_ ), PARAMETER :: n_errors = 6
+   INTEGER ( KIND = ip_ ), DIMENSION( n_errors ) :: errors = (/                &
        GALAHAD_error_restrictions,                                             &
        GALAHAD_error_restrictions,                                             &
        GALAHAD_error_restrictions,                                             &
        GALAHAD_error_preconditioner,                                           &
        GALAHAD_error_ill_conditioned,                                          &
-       GALAHAD_error_max_iterations                                            &
-         /)
+       GALAHAD_error_max_iterations /)
 
 ! Initialize output unit
 
@@ -37,18 +37,18 @@
 !  =============
 
    n = 5
-   f = 1.0_wp
+   f = 1.0_rp_
    CALL SMT_put( H%type, 'COORDINATE', smt_stat )    ! Specify co-ordinate for H
    H%ne = 2 * n - 1
[…]
