+++ src/roots/roots.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 11:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D _ R O O T S   M O D U L E  -*-*-*-*-*-*-*-*-
 
@@ -12,7 +14,9 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_ROOTS_double
+   MODULE GALAHAD_ROOTS_precision
+            
+     USE GALAHAD_PRECISION
 
 !     --------------------------------------------------------------------
 !     |                                                                  |
@@ -21,10 +25,10 @@
 !     --------------------------------------------------------------------
 
       USE GALAHAD_SYMBOLS
-      USE GALAHAD_SPACE_double
-      USE GALAHAD_SORT_double, ONLY : SORT_quicksort
+      USE GALAHAD_SPACE_precision
+      USE GALAHAD_SORT_precision, ONLY : SORT_quicksort
       USE GALAHAD_LAPACK_interface, ONLY : HSEQR, GELS
-      USE GALAHAD_SPECFILE_double
+      USE GALAHAD_SPECFILE_precision
 
       IMPLICIT NONE
 
@@ -48,11 +52,8 @@
      END INTERFACE ROOTS_terminate
 
 !--------------------
-!   P r e c i s i o n
-!--------------------
 
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-      INTEGER, PARAMETER :: wcp = KIND( ( 1.0D+0, 1.0D+0 ) )
+      INTEGER ( KIND = ip_ ), PARAMETER :: wcp = KIND( ( 1.0D+0, 1.0D+0 ) )
 
 !------------------------------------
 !   G e n e r i c   I n t e r f a c e
@@ -66,28 +67,28 @@
 !   P a r a m e t e r s
 !----------------------
 
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-      REAL ( KIND = wp ), PARAMETER :: three = 3.0_wp
-      REAL ( KIND = wp ), PARAMETER :: four = 4.0_wp
-      REAL ( KIND = wp ), PARAMETER :: six = 6.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-      REAL ( KIND = wp ), PARAMETER :: quarter = 0.25_wp
-      REAL ( KIND = wp ), PARAMETER :: threequarters = 0.75_wp
-      REAL ( KIND = wp ), PARAMETER :: onesixth = one / six
-      REAL ( KIND = wp ), PARAMETER :: onethird = one / three
-      REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-      REAL ( KIND = wp ), PARAMETER :: twothirds = two / three
-      REAL ( KIND = wp ), PARAMETER :: pi = 3.1415926535897931_wp
-!     REAL ( KIND = wp ), PARAMETER :: magic = twothirds * pi
-      REAL ( KIND = wp ), PARAMETER :: magic = 2.0943951023931953_wp  !! 2 pi/3
[…]
