+++ src/sha/sha.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ S H A   M O D U L E  *-*-*-*-*-*-*-*-*-*-
 
@@ -11,8 +13,8 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_SHA_double
-
+   MODULE GALAHAD_SHA_precision
+            
 !    ------------------------------------------------
 !   |                                                |
 !   | SHA: find an approximation to a sparse Hessian |
@@ -20,9 +22,10 @@
 !   |                                                |
 !    ------------------------------------------------
 
+     USE GALAHAD_PRECISION
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SPACE_double
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SPACE_precision
      USE GALAHAD_LAPACK_interface, ONLY : GETRF, GETRS, GELSS, GELSD, GELSY
 
      IMPLICIT NONE
@@ -43,17 +46,11 @@
        MODULE PROCEDURE SHA_terminate, SHA_full_terminate
      END INTERFACE SHA_terminate
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -67,17 +64,17 @@
 
 !   error and warning diagnostics occur on stream error
 
-       INTEGER :: error = 6
+       INTEGER ( KIND = ip_ ) :: error = 6
 
 !   general output occurs on stream out
 
-       INTEGER :: out = 6
+       INTEGER ( KIND = ip_ ) :: out = 6
 
 !   the level of output required. <= 0 gives no output, = 1 gives a one-line
 !    summary for every iteration, = 2 gives a summary of the inner iteration
 !    for each iteration, >= 3 gives increasingly verbose (debugging) output
 
-       INTEGER :: print_level = 0
+       INTEGER ( KIND = ip_ ) :: print_level = 0
 
 !  which approximation algorithm should be used?
 !    0 : unsymmetric (alg 2.1 in paper)
@@ -85,7 +82,7 @@
 !    2 : composite (alg 2.3 in paper)
 !    3 : composite 2 (alg 2.2/3 in paper)
 
-       INTEGER :: approximation_algorithm = 2
+       INTEGER ( KIND = ip_ ) :: approximation_algorithm = 2
 
 !  which dense linear equation solver should be used?
 !    1 : Gaussian elimination
@@ -93,11 +90,11 @@
 !    3 : singular-value decomposition
[…]
