+++ src/trb/trb.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-21 AT 13:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 07:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ T R B   M O D U L E  *-*-*-*-*-*-*-*-*-*-
 
@@ -11,7 +13,9 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_TRB_double
+   MODULE GALAHAD_TRB_precision
+            
+     USE GALAHAD_PRECISION
 
 !     ------------------------------------------------------------------
 !    |                                                                  |
@@ -24,19 +28,20 @@
 
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type, NLPT_userdata_type
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SLS_double, ONLY: SLS_coord_to_extended_csr
-     USE GALAHAD_PSLS_double
-     USE GALAHAD_GLTR_double
-     USE GALAHAD_TRS_double
-     USE LANCELOT_CAUCHY_double
-     USE LANCELOT_CG_double
-     USE GALAHAD_LMS_double
-     USE GALAHAD_SHA_double
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_MOP_double, ONLY: mop_Ax
-     USE GALAHAD_NORMS_double, ONLY: TWO_NORM, INFINITY_NORM
+     USE GALAHAD_USERDATA_precision
+     USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SLS_precision, ONLY: SLS_coord_to_extended_csr
+     USE GALAHAD_PSLS_precision
+     USE GALAHAD_GLTR_precision
+     USE GALAHAD_TRS_precision
+     USE LANCELOT_CAUCHY_precision
+     USE LANCELOT_CG_precision
+     USE GALAHAD_LMS_precision
+     USE GALAHAD_SHA_precision
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_MOP_precision, ONLY: mop_Ax
+     USE GALAHAD_NORMS_precision, ONLY: TWO_NORM, INFINITY_NORM
      USE GALAHAD_STRING, ONLY: STRING_integer_6
      USE GALAHAD_BLAS_interface, ONLY: SWAP
      USE GALAHAD_LAPACK_interface, ONLY : GESVD
@@ -47,12 +52,11 @@
 
      PRIVATE
      PUBLIC :: TRB_initialize, TRB_read_specfile, TRB_solve,                   &
-               TRB_projection, TRB_terminate, NLPT_problem_type,               &
-               NLPT_userdata_type, SMT_type, SMT_put,                          &
+               TRB_projection, TRB_terminate, TRB_information,                 &
                TRB_import, TRB_solve_with_mat, TRB_solve_without_mat,          &
                TRB_solve_reverse_with_mat, TRB_solve_reverse_without_mat,      &
                TRB_full_initialize, TRB_full_terminate, TRB_reset_control,     &
-               TRB_information
+               NLPT_problem_type, GALAHAD_userdata_type, SMT_type, SMT_put
 
 !----------------------
 !   I n t e r f a c e s
@@ -66,76 +70,69 @@
[…]
