+++ src/nls/usenls.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 05/05/2021 AT 14:15 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D   U S E _ N L S  -*-*-*-*-*-*-*-*-*-*-
 
@@ -6,19 +8,21 @@
 !  Copyright reserved
 !  October 27th 2015
 
-   MODULE GALAHAD_USENLS_double
+   MODULE GALAHAD_USENLS_precision
+            
+     USE GALAHAD_PRECISION
 
 !  This is the driver program for running NLS for a variety of computing
 !  systems. It opens and closes all the files, allocate arrays, reads and
 !  checks data, and calls the appropriate minimizers
 
 !    USE GALAHAD_CLOCK
-     USE GALAHAD_NLS_double
+     USE GALAHAD_NLS_precision
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_COPYRIGHT
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_CUTEST_FUNCTIONS_double
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_CUTEST_FUNCTIONS_precision
      USE GALAHAD_STRING
 
      IMPLICIT NONE
@@ -34,11 +38,10 @@
 
 !  Dummy argument
 
-     INTEGER, INTENT( IN ) :: input
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
 !  Set precision
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !-------------------------------
 !   D e r i v e d   T y p e s
@@ -58,31 +61,31 @@
 
 !  Problem input characteristics
 
-     INTEGER :: iores, i, j, ir, ic, l, status
+     INTEGER ( KIND = ip_ ) :: iores, i, j, ir, ic, l, status
      LOGICAL :: filexx, is_specfile, hessian_pattern_required
 !    REAL :: timeo, timet
-!    REAL ( KIND = wp ) :: clocko, clockt
-     REAL ( KIND = wp ) :: cutest_time
+!    REAL ( KIND = rp_ ) :: clocko, clockt
+     REAL ( KIND = rp_ ) :: cutest_time
      CHARACTER ( LEN = 10 ) :: name
 
 !  Functions
 
-!$    INTEGER :: OMP_GET_MAX_THREADS
+!$    INTEGER ( KIND = ip_ ) :: OMP_GET_MAX_THREADS
 
 !  Specfile characteristics
 
-     INTEGER, PARAMETER :: input_specfile = 34
-     INTEGER, PARAMETER :: lspec = 29
+     INTEGER ( KIND = ip_ ), PARAMETER :: input_specfile = 34
+     INTEGER ( KIND = ip_ ), PARAMETER :: lspec = 29
      CHARACTER ( LEN = 16 ) :: specname = 'RUNNLS'
      TYPE ( SPECFILE_item_type ), DIMENSION( lspec ) :: spec
      CHARACTER ( LEN = 16 ) :: runspec = 'RUNNLS.SPC'
 
 !  Default values for specfile-defined parameters
 
-     INTEGER :: dfiledevice = 26
-     INTEGER :: rfiledevice = 47
-     INTEGER :: sfiledevice = 62
-     INTEGER :: wfiledevice = 59
+     INTEGER ( KIND = ip_ ) :: dfiledevice = 26
[…]

[…]
