### 2022-12-17 de16589
nonlinear optimization solvers pdated to new precision format
