+++ src/bgo/bgot.F90
@@ -1,25 +1,26 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-03-07 AT 14:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 15:50 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_BGO_TEST  !! far from complete
-   USE GALAHAD_BGO_double                       ! double precision version
+   USE GALAHAD_USERDATA_precision
+   USE GALAHAD_BGO_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( NLPT_problem_type ):: nlp
    TYPE ( BGO_control_type ) :: control
    TYPE ( BGO_inform_type ) :: inform
    TYPE ( BGO_data_type ) :: data
    TYPE ( GALAHAD_userdata_type ) :: userdata
    EXTERNAL :: FUN, GRAD, HESS, HPROD
-   INTEGER :: s
-   INTEGER, PARAMETER :: n = 3, h_ne = 5
-   REAL ( KIND = wp ), PARAMETER :: p = 4.0_wp
-   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20    ! infinity
+   INTEGER ( KIND = ip_ ) :: s
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 3, h_ne = 5
+   REAL ( KIND = rp_ ), PARAMETER :: p = 4.0_rp_
+   REAL ( KIND = rp_ ), PARAMETER :: infinity = 10.0_rp_ ** 20    ! infinity
 ! start problem data
    nlp%pname = 'BGOSPEC'                        ! name
    nlp%n = n ; nlp%H%ne = h_ne                  ! dimensions
    ALLOCATE( nlp%X( n ), nlp%G( n ), nlp%X_l( n ), nlp%X_u( n ) )
-   nlp%X_l = -10.0_wp ; nlp%X_u = 0.5_wp ! search in [-10,1/2]
-   nlp%X = 0.0_wp  ! start from 1.0
+   nlp%X_l = -10.0_rp_ ; nlp%X_u = 0.5_rp_ ! search in [-10,1/2]
+   nlp%X = 0.0_rp_  ! start from 1.0
 !  sparse co-ordinate storage format
    CALL SMT_put( nlp%H%type, 'COORDINATE', s )  ! Specify co-ordinate storage
    ALLOCATE( nlp%H%val( h_ne ), nlp%H%row( h_ne ), nlp%H%col( h_ne ) )
@@ -57,14 +58,13 @@
    END PROGRAM GALAHAD_BGO_TEST
 
    SUBROUTINE FUN( status, X, userdata, f )     ! Objective function
-   USE GALAHAD_USERDATA_double, ONLY: GALAHAD_userdata_type
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-   INTEGER, INTENT( OUT ) :: status
-   REAL ( KIND = wp ), INTENT( OUT ) :: f
-   REAL ( KIND = wp ), DIMENSION( : ),INTENT( IN ) :: X
+   USE GALAHAD_USERDATA_precision
+   INTEGER ( KIND = ip_ ), INTENT( OUT ) :: status
+   REAL ( KIND = rp_ ), INTENT( OUT ) :: f
[…]
