+++ src/nls/nlsti.F90
@@ -1,32 +1,33 @@
-! THIS VERSION: GALAHAD 3.3 - 26/07/2021 AT 13:50 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 15:45 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_NLS_interface_test
-   USE GALAHAD_NLS_double                       ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_NLS_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( NLS_control_type ) :: control
    TYPE ( NLS_inform_type ) :: inform
    TYPE ( NLS_full_data_type ) :: data
    TYPE ( GALAHAD_userdata_type ) :: userdata
-   INTEGER :: n, m, J_ne, H_ne, P_ne
-   INTEGER :: data_storage_type, model, status, eval_status
+   INTEGER ( KIND = ip_ ) :: n, m, J_ne, H_ne, P_ne
+   INTEGER ( KIND = ip_ ) :: data_storage_type, model, status, eval_status
    LOGICAL :: transpose
-   INTEGER, DIMENSION( 0 ) :: null
-   REAL ( KIND = wp ), PARAMETER :: p = 1.0_wp
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, G, C, Y, W, U, V
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: J_row, J_col, J_ptr
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: J_val, J_dense
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: H_row, H_col, H_ptr
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: H_val, H_dense, H_diag
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: P_row, P_col, P_ptr
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: P_val, P_dense
+   INTEGER ( KIND = ip_ ), DIMENSION( 0 ) :: null
+   REAL ( KIND = rp_ ), PARAMETER :: p = 1.0_rp_
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, G, C, Y, W, U, V
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: J_row, J_col, J_ptr
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: J_val, J_dense
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: H_row, H_col, H_ptr
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: H_val, H_dense, H_diag
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: P_row, P_col, P_ptr
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: P_val, P_dense
    CHARACTER ( len = 2 ) :: st
 
 ! set up problem data
 
    n = 2 ;  m = 3 ; J_ne = 5 ; H_ne = 2 ; P_ne = 2
    ALLOCATE( X( n ), G( n ), C( m ), Y( m ), W( m ), H_diag( n ) )
-   W = 1.0_wp
+   W = 1.0_rp_
[…]
