+++ src/sha/usesha.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.5 - 10/04/2013 AT 16:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D   U S E _ S H A  -*-*-*-*-*-*-*-*-*-*-
 
@@ -6,22 +8,23 @@
 !  Copyright reserved
 !  April 10th 2013
 
-   MODULE GALAHAD_USESHA_double
-
+   MODULE GALAHAD_USESHA_precision
+            
 !  This is the driver program for running SHA for a variety of computing
 !  systems. It opens and closes all the files, allocate arrays, reads and
 !  checks data, and calls the appropriate minimizers
 
-     USE GALAHAD_SHA_double
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_PRECISION
+     USE GALAHAD_SHA_precision
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_COPYRIGHT
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_RAND_double
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_RAND_precision
      USE GALAHAD_CLOCK
-     USE GALAHAD_CUTEST_FUNCTIONS_double
-     USE CUTEst_interface_double
+     USE GALAHAD_CUTEST_FUNCTIONS_precision
+     USE CUTEst_interface_precision
 
-     USE GALAHAD_NORMS_double, ONLY : TWO_NORM
+     USE GALAHAD_NORMS_precision, ONLY : TWO_NORM
      IMPLICIT NONE
 
      PRIVATE
@@ -35,11 +38,7 @@
 
 !  Dummy argument
 
-     INTEGER, INTENT( IN ) :: input
-
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
 !-------------------------------
 !   D e r i v e d   T y p e s
@@ -55,45 +54,47 @@
 
 !  problem input characteristics
 
-     INTEGER :: iores, difs, difs_max, i, k, lh, m, n, nnzh, alloc_stat, status
-     REAL ( KIND = wp ) :: pert, clocka, clocke, clocks, level, reduce
-     REAL ( KIND = wp ) :: clock( 2 ), error( 2 )
+     INTEGER ( KIND = ip_ ) :: iores, difs, difs_max, i, k, lh
+     INTEGER ( KIND = ip_ ) :: m, n, nnzh, alloc_stat, status
+     REAL ( KIND = rp_ ) :: pert, clocka, clocke, clocks, level, reduce
+     REAL ( KIND = rp_ ) :: clock( 2 ), error( 2 )
      LOGICAL :: filexx, is_specfile
      CHARACTER ( LEN =  1 ) :: ptype
      CHARACTER ( LEN =  10 ) :: pname
      TYPE ( RAND_seed ) :: seed
 
-     INTEGER, ALLOCATABLE, DIMENSION( : ) :: ROW, COL, RD, ROW_COUNT
-     REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: VAL, VAL_est
-     REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, X_l, X_u
[…]
