+++ src/lhs/lhs.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 02/08/2021 AT 11:50 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ L H S   M O D U L E  *-*-*-*-*-*-*-*-*-*-
 
@@ -10,7 +12,9 @@
 !   see http://people.sc.fsu.edu/~jburkardt/f_src/ihs/ihs.html
 !   originally released GALAHAD Version 2.8. June 29th 2016
 
-  MODULE GALAHAD_LHS_double
+  MODULE GALAHAD_LHS_precision
+            
+     USE GALAHAD_PRECISION
 
 !     --------------------------------------------------
 !    |                                                  |
@@ -19,9 +23,9 @@
 !     --------------------------------------------------
 
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_NORMS_double, ONLY: TWO_NORM
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_NORMS_precision, ONLY: TWO_NORM
 
      IMPLICIT NONE
 
@@ -29,20 +33,12 @@
      PUBLIC :: LHS_initialize, LHS_read_specfile, LHS_ihs, LHS_get_seed,       &
                LHS_terminate
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-     INTEGER, PARAMETER :: sp = KIND( 1.0 )
-     INTEGER, PARAMETER :: ip = KIND( 1 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-     INTEGER, PARAMETER :: i4_huge = HUGE( 1 )
-     REAL ( KIND = wp ), PARAMETER :: r8_huge = HUGE( 1.0_wp )
+     INTEGER ( KIND = ip_ ), PARAMETER :: i4_huge = HUGE( 1 )
+     REAL ( KIND = rp_ ), PARAMETER :: r8_huge = HUGE( 1.0_rp_ )
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -56,19 +52,19 @@
 
 !   error and warning diagnostics occur on stream error
 
-       INTEGER :: error = 6
+       INTEGER ( KIND = ip_ ) :: error = 6
 
 !   general output occurs on stream out
 
-       INTEGER :: out = 6
+       INTEGER ( KIND = ip_ ) :: out = 6
 
 !   the level of output required. <= 0 gives no output, >0 debugging
 
-       INTEGER :: print_level = 0
+       INTEGER ( KIND = ip_ ) :: print_level = 0
 
 !   the duplication factor. This must be at least 1, a value of 5 is reasonable
 !
-       INTEGER :: duplication = 5
+       INTEGER ( KIND = ip_ ) :: duplication = 5
 
 !   if %space_critical true, every effort will be made to use as little
 !    space as possible. This may result in longer computation time
@@ -96,11 +92,11 @@
 
 !  return status. 0 = success, /= 0 failure
 
-       INTEGER :: status = 0
+       INTEGER ( KIND = ip_ ) :: status = 0
 
 !  the status of the last attempted allocation/deallocation
 
-       INTEGER :: alloc_status = 0
[…]
