+++ src/sha/runsha_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.5 - 10/04/2013 AT 15:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D   R U N S H A _ S I F  *-*-*-*-*-*-*-*-
 
@@ -7,17 +9,18 @@
 !  April 10th 2013
 
    PROGRAM RUNSHA_SIF
-   USE GALAHAD_USESHA_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USESHA_precision
 
 !  Main program for the SIF interface to SHA, a method for estimating
 !  sparse Hessian matrices
 
 !  Problem insif characteristics
 
-   INTEGER, PARAMETER :: errout = 6
-   INTEGER, PARAMETER :: insif = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: errout = 6
+   INTEGER ( KIND = ip_ ), PARAMETER :: insif = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
-   INTEGER :: iostat
+   INTEGER ( KIND = ip_ ) :: iostat
 
 !  Open the data input file
 

[…]
