+++ src/tru/trut.F90
@@ -1,19 +1,20 @@
-! THIS VERSION: GALAHAD 3.3 - 27/04/2021 AT 14:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 15:15 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_TRU_test_deck
-   USE GALAHAD_TRU_double                       ! double precision version
+   USE GALAHAD_USERDATA_precision
+   USE GALAHAD_TRU_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( NLPT_problem_type ):: nlp
    TYPE ( TRU_control_type ) :: control
    TYPE ( TRU_inform_type ) :: inform
    TYPE ( TRU_data_type ) :: data
-   TYPE ( NLPT_userdata_type ) :: userdata
+   TYPE ( GALAHAD_userdata_type ) :: userdata
    EXTERNAL :: FUN, GRAD, HESS, HESSPROD, PREC
-   INTEGER :: i, s, scratch_out
+   INTEGER ( KIND = ip_ ) :: i, s, scratch_out
    logical :: alive
-   REAL ( KIND = wp ), PARAMETER :: p = 4.0_wp
-   REAL ( KIND = wp ) :: dum
+   REAL ( KIND = rp_ ), PARAMETER :: p = 4.0_rp_
+   REAL ( KIND = rp_ ) :: dum
 ! start problem data
    nlp%n = 1 ; nlp%H%ne = 1                     ! dimensions
    ALLOCATE( nlp%X( nlp%n ), nlp%G( nlp%n ) )
@@ -65,19 +66,19 @@
 !     control%GLTR_control%print_level = 4
      inform%status = 1                            ! set for initial entry
      nlp%n = 1
-     nlp%X = 1.0_wp                               ! start from one
+     nlp%X = 1.0_rp_                               ! start from one
      control%hessian_available = .FALSE.          ! Hessian prods will be used
 
      IF ( s == - GALAHAD_error_restrictions ) THEN
        nlp%n = 0
      ELSE IF ( s == - GALAHAD_error_unbounded ) THEN
-       control%obj_unbounded = - ( 10.0_wp ) ** 10
+       control%obj_unbounded = - ( 10.0_rp_ ) ** 10
      ELSE IF ( s == - GALAHAD_error_preconditioner ) THEN
        control%norm = - 3               ! User's preconditioner
      ELSE IF ( s == - GALAHAD_error_max_iterations ) THEN
        control%maxit = 0
      ELSE IF ( s == - GALAHAD_error_cpu_limit ) THEN
-       control%cpu_time_limit = 0.0_wp
+       control%cpu_time_limit = 0.0_rp_
      END IF
      DO                                           ! Loop to solve problem
 !      write(6,*) 'in ', inform%status
@@ -97,17 +98,17 @@
            END IF
          END IF
          IF ( s == - GALAHAD_error_cpu_limit ) THEN
-           dum = 0.0_wp
+           dum = 0.0_rp_
            DO i = 1, 10000000
-             dum = dum + 0.0000001_wp * i / ( i + 1 )
[…]
