+++ src/arc/arc.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-21 AT 12:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 07:30 GMT.
+
+#include "galahad_modules.h"
 
 !  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
 !  Principal authors: Nick Gould and Margherita Porcelli
@@ -9,8 +11,8 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_ARC_double
-
+   MODULE GALAHAD_ARC_precision
+            
 !     ----------------------------------------------------
 !    |                                                    |
 !    | ARC, an adaptive regularised cubic model algorithm |
@@ -22,20 +24,22 @@
 !    |                                                    |
 !     ----------------------------------------------------
 
+!    USE GALAHAD_PRECISION
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type, NLPT_userdata_type
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_ROOTS_double
-     USE GALAHAD_PSLS_double
-     USE GALAHAD_GLRT_double
-     USE GALAHAD_RQS_double
-     USE GALAHAD_DPS_double
-     USE GALAHAD_LMS_double
-     USE GALAHAD_SHA_double
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_MOP_double, ONLY: mop_Ax
-     USE GALAHAD_NORMS_double, ONLY: TWO_NORM
+     USE GALAHAD_USERDATA_precision
+     USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_ROOTS_precision
+     USE GALAHAD_PSLS_precision
+     USE GALAHAD_GLRT_precision
+     USE GALAHAD_RQS_precision
+     USE GALAHAD_DPS_precision
+     USE GALAHAD_LMS_precision
+     USE GALAHAD_SHA_precision
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_MOP_precision, ONLY: mop_Ax
+     USE GALAHAD_NORMS_precision, ONLY: TWO_NORM
      USE GALAHAD_STRING, ONLY: STRING_integer_6
      USE GALAHAD_BLAS_interface, ONLY: SWAP
      USE GALAHAD_LAPACK_interface, ONLY : GESVD
@@ -44,12 +48,11 @@
 
      PRIVATE
      PUBLIC :: ARC_initialize, ARC_read_specfile, ARC_solve,                   &
-               ARC_adjust_weight, ARC_terminate, NLPT_problem_type,            &
-               NLPT_userdata_type, SMT_type, SMT_put,                          &
+               ARC_adjust_weight, ARC_terminate, ARC_information,              &
                ARC_import, ARC_solve_with_mat, ARC_solve_without_mat,          &
                ARC_solve_reverse_with_mat, ARC_solve_reverse_without_mat,      &
                ARC_full_initialize, ARC_full_terminate, ARC_reset_control,     &
-               ARC_information
+               NLPT_problem_type, GALAHAD_userdata_type, SMT_type, SMT_put
 
 !----------------------
 !   I n t e r f a c e s
@@ -63,77 +66,70 @@
[…]
