+++ src/arc/arcti.F90
@@ -1,22 +1,23 @@
-! THIS VERSION: GALAHAD 3.4 - 05/01/2022 AT 10:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 08:30 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_ARC_interface_test
-   USE GALAHAD_ARC_double                       ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_ARC_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
    TYPE ( NLPT_problem_type ):: nlp
    TYPE ( ARC_control_type ) :: control
    TYPE ( ARC_inform_type ) :: inform
    TYPE ( ARC_full_data_type ) :: data
    TYPE ( NLPT_userdata_type ) :: userdata
-   INTEGER :: n, ne
-   INTEGER :: i, s, data_storage_type, status, eval_status
+   INTEGER ( KIND = ip_ ) :: n, ne
+   INTEGER ( KIND = ip_ ) :: i, s, data_storage_type, status, eval_status
    LOGICAL :: alive
-   REAL ( KIND = wp ), PARAMETER :: p = 4.0_wp
-   REAL ( KIND = wp ) :: dum, f
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, G, U, V
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: H_row, H_col, H_ptr
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: H_val, H_dense, H_diag
+   REAL ( KIND = rp_ ), PARAMETER :: p = 4.0_rp_
+   REAL ( KIND = rp_ ) :: dum, f
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, G, U, V
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: H_row, H_col, H_ptr
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: H_val, H_dense, H_diag
    CHARACTER ( len = 1 ) :: st
 
 ! problem data complete   
@@ -43,7 +44,7 @@
    DO data_storage_type = 1, 5
      CALL ARC_initialize( data, control, inform )
 !    control%print_level = 1
-     X = 1.5_wp  ! start from 1.5
+     X = 1.5_rp_  ! start from 1.5
      SELECT CASE ( data_storage_type )
      CASE ( 1 ) ! sparse co-ordinate storage
        st = 'C'
@@ -90,12 +91,12 @@
 
    WRITE( 6, "( /, ' tests reverse-communication options ', / )" )
 
-   f = 0.0_wp
+   f = 0.0_rp_
    ALLOCATE( U( n ), V( n ) ) ! reverse-communication input/output
    DO data_storage_type = 1, 5
      CALL ARC_initialize( data, control, inform )
 !    control%print_level = 1
-     X = 1.5_wp  ! start from 1.5
+     X = 1.5_rp_  ! start from 1.5
[…]

[…]
