+++ src/lancelot_simple/lancelot_simplet.F90
@@ -1,5 +1,9 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 15:00 GMT.
+
+#include "galahad_modules.h"
+
       PROGRAM LANCELOT_simplet
-!
+            
 !-------------------------------------------------------------------------------
 !                                                                              !
 !    This programs provides a test for lancelot_simple.f90                     !
@@ -12,14 +16,15 @@
 !                                                                              !
 !------------------------------------------------------------------------------!
 !
-       USE LANCELOT_simple_double
+       USE GALAHAD_PRECISION
+       USE LANCELOT_simple_precision
        IMPLICIT NONE
-       INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-       REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
-       INTEGER :: n, neq, nin, iters, maxit, print_level, exit_code
-       REAL ( KIND = wp ) :: gradtol, feastol, fx
+       REAL ( KIND = rp_ ), PARAMETER :: infinity = 10.0_rp_ ** 20
+       INTEGER ( KIND = ip_ ) :: n, neq, nin, iters, maxit
+       INTEGER ( KIND = ip_ ) :: print_level, exit_code
+       REAL ( KIND = rp_ ) :: gradtol, feastol, fx
        CHARACTER ( LEN = 10 ), ALLOCATABLE, DIMENSION(:) :: VNAMES, CNAMES
-       REAL ( KIND = wp ),     ALLOCATABLE, DIMENSION(:) :: BL, BU, X, CX, Y
+       REAL ( KIND = rp_ ),     ALLOCATABLE, DIMENSION(:) :: BL, BU, X, CX, Y
        EXTERNAL :: FUN, GRAD, HESS
 !
 ! THE TEST PROBLEM DIMENSIONS (user defined)
@@ -35,15 +40,15 @@
 !
 ! starting point
 !
-       X(1) = -1.2_wp                 ! starting point (componentwise)
-       X(2) =  1.0_wp
+       X(1) = -1.2_rp_                 ! starting point (componentwise)
+       X(2) =  1.0_rp_
 !
 ! bounds on the variables
 !
-       BL(1) =  0.0_wp                ! lower bounds (componentwise)
+       BL(1) =  0.0_rp_                ! lower bounds (componentwise)
        BL(2) = -infinity
        BU(1) =  infinity              ! upper bounds (componentwise)
-       BU(2) =  3.0_wp
+       BU(2) =  3.0_rp_
 !
 ! names
 !
@@ -55,81 +60,81 @@
 ! algorithmic parameters
 !
        maxit       = 100
-       gradtol     = 0.00001_wp
-       feastol     = 0.00001_wp
+       gradtol     = 0.00001_rp_
+       feastol     = 0.00001_rp_
        print_level = 1
 !
 ! call LANCELOT simple for a number of test cases
 !
-       WRITE(6,*) '====================================================', &
+       WRITE(6,*) '====================================================',      &
                   ' ================'
        CALL LANCELOT_simple( 0, X, FUN, fx, exit_code )
-       WRITE(6,*) '==================== TEST 1: EXIT_CODE = ', exit_code, &
+       WRITE(6,*) '==================== TEST 1: EXIT_CODE = ', exit_code,      &
                   '  ======================'
 !
-       X(1) = -1.2_wp
-       X(2) =  1.0_wp
+       X(1) = -1.2_rp_
+       X(2) =  1.0_rp_
        CALL LANCELOT_simple( 2, X, FUN, fx, exit_code, NEQ = -1 )
[…]
