+++ src/lancelot/initw.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 10:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-  L A N C E L O T  -B-  INITW  M O D U L E  *-*-*-*-*-*-*-*
 
@@ -6,10 +8,11 @@
 !  Copyright reserved
 !  February 1st 1995
 
-     MODULE LANCELOT_INITW_double
-  
-       USE GALAHAD_EXTEND_double, ONLY: EXTEND_arrays
-       USE LANCELOT_OTHERS_double
+     MODULE LANCELOT_INITW_precision
+            
+       USE GALAHAD_PRECISION
+       USE GALAHAD_EXTEND_precision, ONLY: EXTEND_arrays
+       USE LANCELOT_OTHERS_precision
        IMPLICIT NONE
        
        PRIVATE
@@ -17,12 +20,11 @@
        
 !  Set precision
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !  Set other parameters
 
-       REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-       REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+       REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+       REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
   
      CONTAINS
 
@@ -57,63 +59,62 @@
 !   D u m m y   A r g u m e n t s
 !-----------------------------------------------
 
-       INTEGER, INTENT( IN ) :: n, ng, nel, ntotel, nvrels, nnza, numvar
-       INTEGER, INTENT( IN ) :: iprint, iout, buffer
-       INTEGER, INTENT( OUT ) :: lfxi  , lgxi  , lhxi  , lggfx , ldx  
-       INTEGER, INTENT( OUT ) :: lnguvl, lnhuvl, nvargp, status, alloc_status
-       INTEGER, INTENT( OUT ) :: ntotin, ntype , nsets , maxsel
+       INTEGER ( KIND = ip_ ), INTENT( IN ) :: n, ng, nel, ntotel, nvrels, nnza
+       INTEGER ( KIND = ip_ ), INTENT( IN ) :: numvar, iprint, iout, buffer
+       INTEGER ( KIND = ip_ ), INTENT( OUT ) :: lfxi, lgxi, lhxi, lggfx, ldx  
+       INTEGER ( KIND = ip_ ), INTENT( OUT ) :: lnguvl, lnhuvl, nvargp
+       INTEGER ( KIND = ip_ ), INTENT( OUT ) :: status, alloc_status
+       INTEGER ( KIND = ip_ ), INTENT( OUT ) :: ntotin, ntype , nsets , maxsel
        LOGICAL, INTENT( IN ) :: direct, fdgrad, skipg
        LOGICAL, INTENT( OUT ) :: altriv
        CHARACTER ( LEN = 80 ), INTENT( OUT ) :: bad_alloc
-       INTEGER, INTENT( IN ), DIMENSION( ntotel  ) :: IELING
-       INTEGER, INTENT( IN ), DIMENSION( ng  + 1 ) :: ISTADA, ISTADG
-       INTEGER, INTENT( IN ), DIMENSION( nel + 1 ) :: ISTAEV
[…]
