+++ src/lancelot/frntl.F90
@@ -1,4 +1,27 @@
-! THIS VERSION: GALAHAD 2.6 - 23/06/2013 AT 13:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 09:40 GMT.
+
+#ifdef LANCELOT_USE_MA57
+#define SILS_control MA57_control
+#define SILS_factors MA57_factors
+#define SILS_ainfo MA57_ainfo
+#define SILS_finfo MA57_finfo
+#define SILS_sinfo MA57_sinfo
+#define SILS_data MA57_data
+#define SILS_cntl MA57_cntl
+#define SILS_infoa MA57_infoa
+#define SILS_infof MA57_infof
+#define SILS_infos MA57_infof
+#define SILS_analyse MA57_analyse
+#define SILS_factorize MA57_factorize
+#define SILS_solve MA57_solve
+#endif
+
+#include "galahad_modules.h"
+
+#ifdef LANCELOT_USE_MA57
+#define GALAHAD_SILS_double HSL_MA57_double
+#define GALAHAD_SILS_single HSL_MA57_single
+#endif
 
 !-*-*-*-*-*-*-*-  L A N C E L O T  -B-  FRNTL  M O D U L E  *-*-*-*-*-*-*-*-*
 
@@ -6,32 +29,30 @@
 !  Copyright reserved
 !  February 2nd 1995
 
-   MODULE LANCELOT_FRNTL_double
-
-     USE GALAHAD_SMT_double
-     USE GALAHAD_SILS_double
-     USE GALAHAD_SCU_double, ONLY : SCU_matrix_type, SCU_data_type,            &
-       SCU_inform_type, SCU_restart_m_eq_0, SCU_solve, SCU_append
-     USE LANCELOT_ASMBL_double
-     USE LANCELOT_MDCHL_double
+   MODULE LANCELOT_FRNTL_precision
+            
+     USE GALAHAD_PRECISION
+     USE GALAHAD_SMT_precision
+     USE GALAHAD_SILS_precision
+     USE GALAHAD_SCU_precision, ONLY : SCU_matrix_type, SCU_data_type,         &
+                                       SCU_inform_type, SCU_restart_m_eq_0,    &
+                                       SCU_solve, SCU_append
+     USE LANCELOT_ASMBL_precision
+     USE LANCELOT_MDCHL_precision
    
      IMPLICIT NONE
    
      PRIVATE
      PUBLIC :: FRNTL_get_search_direction
 
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !  Set other parameters
 
-     INTEGER, PARAMETER :: liwmin = 1, lwmin = 1
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: four = 4.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
+     INTEGER ( KIND = ip_ ), PARAMETER :: liwmin = 1, lwmin = 1
+     REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: four = 4.0_rp_
[…]
