+++ src/lancelot/runlancelot_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.0 - 24/10/2017 AT 14:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D   R U N L A N C E L O T _ S I F  *-*-*-*-*-*-*-*-
 
@@ -7,14 +9,15 @@
 !  March 14th 2003
 
    PROGRAM RUNLANCELOT_SIF
-   USE GALAHAD_USELANCELOT_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USELANCELOT_precision
 
 !  Main program for the SIF interface to LANCELOT B, an augmented Lagrangian
 !  algorithm for nonlinear programming
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
 
 !  Open the data input file
