+++ src/lancelot/drchg.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 12/03/2014 AT 13:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 09:55 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-  L A N C E L O T  -B-  D R C H G   M O D U L E  -*-*-*-*-*-*-*-
 
@@ -10,40 +12,37 @@
 !   March 14th 2003 as uselanb
 !   update released with GALAHAD Version 2.0. May 11th 2006
 
-   MODULE LANCELOT_DRCHG_double
-
-     USE LANCELOT_types_double, ONLY: LANCELOT_problem_type
+   MODULE LANCELOT_DRCHG_precision
+            
+     USE GALAHAD_PRECISION
+     USE LANCELOT_TYPES_precision, ONLY: LANCELOT_problem_type
 
      IMPLICIT NONE
 
      PRIVATE
      PUBLIC ::  DRCHG_save_type, DRCHG_check_group_derivatives
 
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !  Set other parameters
 
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: tenp2 = 100.0_wp
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: tenp2 = 100.0_rp_
 
 !   The DRCHG_save_type derived type
 
      TYPE :: DRCHG_save_type
-      REAL ( KIND = wp ) :: epsqrt
+      REAL ( KIND = rp_ ) :: epsqrt
       LOGICAL :: warning
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: XT
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: GTVALS
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: XT
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: GTVALS
      END TYPE DRCHG_save_type
 
    CONTAINS
 
 !-  L A N C E L O T -B- DRCHG_check_group_derivatives  S U B R O U T I N E -
 
-     SUBROUTINE DRCHG_check_group_derivatives(                                 &
-                      prob, X, GVALS , ITESTG, ntestg, relpr, iprint, iout,    &
-                      status, S, GROUP )
+     SUBROUTINE DRCHG_check_group_derivatives( prob, X, GVALS, ITESTG, ntestg, &
+                                               relpr, iprint, iout, status,    &
+                                               S, GROUP )
 
 !  Given a vector of functionals GVALS, check their analytical derivatives
 !  against approximations by differences at the given point X
@@ -53,12 +52,12 @@
 !-----------------------------------------------
 
      TYPE ( LANCELOT_problem_type ), INTENT( INOUT ) :: prob
-     INTEGER, INTENT( IN    ) :: ntestg, iprint, iout
-     INTEGER, INTENT( INOUT ) :: status
-     REAL ( KIND = wp ), INTENT( IN ) :: relpr
-     INTEGER, INTENT( IN    ), DIMENSION( prob%ng ) :: ITESTG
-     REAL ( KIND = wp ), INTENT( INOUT ), DIMENSION( prob%ng ) :: X
[…]
