+++ src/lancelot/lancelot_steering.F90
@@ -1,4 +1,20 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 09:45 GMT.
+
+#ifdef LANCELOT_USE_MA57
+#define SILS_initialize MA57_initialize
+#define SILS_data MA57_data
+#define SILS_cntl MA57_cntl
+#define SILS_control MA57_control
+#define SILS_factors MA57_factors
+#define SILS_finalize MA57_finalize
+#endif
+
+#include "galahad_modules.h"
+
+#ifdef LANCELOT_USE_MA57
+#define GALAHAD_SILS_double HSL_MA57_double
+#define GALAHAD_SILS_single HSL_MA57_single
+#endif
 
 !-*-*-  L A N C E L O T  -B-  LANCELOT _ S T E E R I N G _  M O D U L E  -*-*-
 
@@ -13,8 +29,8 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE LANCELOT_steering_double
-
+   MODULE LANCELOT_steering_precision
+            
 !  |------------------------------------------------------------------|
 !  |                                                                  |
 !  |  Find a local minimizer of a smooth (group partially separable)  |
@@ -25,23 +41,24 @@
 !  |                                                                  |
 !  |------------------------------------------------------------------|
 
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SPACE_double
-     USE LANCELOT_TYPES_double
-     USE LANCELOT_INITW_double
-     USE LANCELOT_OTHERS_double
-     USE LANCELOT_HSPRD_double
-     USE LANCELOT_CAUCHY_double
-     USE LANCELOT_CG_double
-     USE LANCELOT_PRECN_double
-     USE LANCELOT_FRNTL_double
-     USE LANCELOT_STRUTR_double
-     USE GALAHAD_SMT_double
-     USE GALAHAD_SILS_double
-     USE GALAHAD_SCU_double, ONLY : SCU_matrix_type, SCU_data_type,            &
+     USE GALAHAD_PRECISION
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SPACE_precision
+     USE LANCELOT_TYPES_precision
+     USE LANCELOT_INITW_precision
+     USE LANCELOT_OTHERS_precision
+     USE LANCELOT_HSPRD_precision
+     USE LANCELOT_CAUCHY_precision
+     USE LANCELOT_CG_precision
+     USE LANCELOT_PRECN_precision
+     USE LANCELOT_FRNTL_precision
+     USE LANCELOT_STRUTR_precision
+     USE GALAHAD_SMT_precision
+     USE GALAHAD_SILS_precision
+     USE GALAHAD_SCU_precision, ONLY : SCU_matrix_type, SCU_data_type,         &
        SCU_inform_type, SCU_factorize, SCU_terminate
-     USE LANCELOT_ASMBL_double, ONLY : ASMBL_save_type
-     USE GALAHAD_EXTEND_double, ONLY : EXTEND_save_type
+     USE LANCELOT_ASMBL_precision, ONLY : ASMBL_save_type
+     USE GALAHAD_EXTEND_precision, ONLY : EXTEND_save_type
 
      IMPLICIT NONE
 
@@ -52,29 +69,28 @@
 
 !  Set precision
 
[…]

[…]
