+++ src/lancelot/asmbl.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 09:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-  L A N C E L O T  -B-  ASMBL  M O D U L E  *-*-*-*-*-*-*-*
 
@@ -6,24 +8,21 @@
 !  Copyright reserved
 !  January 25th 1995
 
-   MODULE LANCELOT_ASMBL_double
-
+   MODULE LANCELOT_ASMBL_precision
+            
+     USE GALAHAD_PRECISION
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_EXTEND_double, ONLY: EXTEND_arrays
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_EXTEND_precision, ONLY: EXTEND_arrays
      IMPLICIT NONE
 
      PRIVATE
      PUBLIC :: ASMBL_save_type, ASMBL_assemble_hessian
 
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !  Set other parameters
 
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+     REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
 
 !  ================================
 !  The ASMBL_save_type derived type
@@ -31,9 +30,9 @@
 
      TYPE :: ASMBL_save_type
        LOGICAL :: ptr_status
-       INTEGER, DIMENSION( 30 ) :: ICNTL
-       INTEGER, DIMENSION( 20 ) :: INFO
-       REAL ( KIND = wp ), DIMENSION( 5 ) :: CNTL
+       INTEGER ( KIND = ip_ ), DIMENSION( 30 ) :: ICNTL
+       INTEGER ( KIND = ip_ ), DIMENSION( 20 ) :: INFO
+       REAL ( KIND = rp_ ), DIMENSION( 5 ) :: CNTL
      END TYPE ASMBL_save_type
 
    CONTAINS
@@ -68,27 +67,31 @@
 !   D u m m y   A r g u m e n t s
 !-----------------------------------------------
 
-      INTEGER, INTENT( IN ) :: n, nel, ng, maxsel, nsemib, nvargp, nnza
-      INTEGER, INTENT( IN ) :: nvrels, ntotel, n_free
-      INTEGER, INTENT( IN ) :: lnguvl, lnhuvl, iprint, error, out
-      INTEGER, INTENT( OUT ) :: status, alloc_status
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: n, nel, ng, maxsel, nsemib, nvargp
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: nvrels, ntotel, n_free, nnza
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: lnguvl, lnhuvl, iprint, error, out
+      INTEGER ( KIND = ip_ ), INTENT( OUT ) :: status, alloc_status
       LOGICAL, INTENT( IN ) :: fixed_structure, no_zeros, use_band, skipg
       CHARACTER ( LEN = 80 ) :: bad_alloc
-      INTEGER, INTENT( IN  ), DIMENSION( n ) :: I_free
-      INTEGER, INTENT( IN ), DIMENSION( nnza ) :: ICNA
[…]
