+++ src/lancelot/drche.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.6 - 12/03/2014 AT 13:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:45 GMT.
+
+#include "galahad_modules.h"
 
 !-*--*-*-*-*-  L A N C E L O T  -B-  D R C H E   M O D U L E  -*-*-*-*-*-*-*-
 
@@ -13,9 +15,10 @@
 !  Copyright reserved
 !  January 31st 1995
 
-   MODULE LANCELOT_DRCHE_double
-
-     USE LANCELOT_types_double, ONLY : LANCELOT_problem_type
+   MODULE LANCELOT_DRCHE_precision
+            
+     USE GALAHAD_PRECISION
+     USE LANCELOT_TYPES_precision, ONLY : LANCELOT_problem_type
 
      IMPLICIT NONE
 
@@ -24,25 +27,25 @@
 
 !  Set precision
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !  Set other parameters
 
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: tenp2 = 100.0_wp
+     REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: half = 0.5_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: tenp2 = 100.0_rp_
 
 !   The DRCHE_save_type derived type
 
     TYPE :: DRCHE_save_type
-      INTEGER :: itest, ielf, j, kk, lg, lh, lend, nelvar, ninvar, ntestl
+      INTEGER ( KIND = ip_ ) :: itest, ielf, j, kk, lg, lh, lend
+      INTEGER ( KIND = ip_ ) :: nelvar, ninvar, ntestl
       LOGICAL :: intre, warning
-      REAL ( KIND = wp ) :: epsqrt
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: XT
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: XINT
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: FTUVAL
-      REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : , : ) :: WINV
+      REAL ( KIND = rp_ ) :: epsqrt
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: XT
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: XINT
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: FTUVAL
+      REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: WINV
     END TYPE DRCHE_save_type
 
       CONTAINS
@@ -63,18 +66,18 @@
 !-----------------------------------------------
 
      TYPE ( LANCELOT_problem_type ), INTENT( INOUT ) :: prob
-     INTEGER, INTENT( IN ) :: lfuval, nelmax, ninmax, iprint, iout
-     INTEGER, INTENT( INOUT ) :: status, ncalcf
-     REAL ( KIND = wp ), INTENT( IN ) :: relpr
[…]

[…]
