+++ src/lancelot/lancelot_steerings.F90
@@ -1,10 +1,12 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 13:00 GMT.
+#include "galahad_modules.h"
    SUBROUTINE RANGE( ielemn, transp, W1, W2, nelvar, ninvar, ieltyp, lw1, lw2 )
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   INTEGER, INTENT( IN ) :: ielemn, nelvar, ninvar, ieltyp, lw1, lw2
+   USE GALAHAD_PRECISION
+   INTEGER ( KIND = ip_ ), INTENT( IN ) :: ielemn, nelvar, ninvar, ieltyp
+   INTEGER ( KIND = ip_ ), INTENT( IN ) :: lw1, lw2
    LOGICAL, INTENT( IN ) :: transp
-   REAL ( KIND = wp ), INTENT( IN ), DIMENSION ( lw1 ) :: W1
-   REAL ( KIND = wp ), INTENT( OUT ), DIMENSION ( lw2 ) :: W2
+   REAL ( KIND = rp_ ), INTENT( IN ), DIMENSION( lw1 ) :: W1
+   REAL ( KIND = rp_ ), INTENT( OUT ), DIMENSION( lw2 ) :: W2
    SELECT CASE ( ieltyp )
    CASE ( 1 ) ! Element type 1 has a non-trivial transformation
      IF ( transp ) THEN
@@ -24,18 +26,21 @@
                       IELVAR, INTVAR, ISTADH, ISTEPA, ICALCF, ltypee,          &
                       lstaev, lelvar, lntvar, lstadh, lstepa, lcalcf,          &
                       lfuval, lxvalu, lepvlu, ifflag, ifstat )
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   INTEGER, INTENT( IN ) :: ncalcf, ifflag, ltypee, lstaev, lelvar, lntvar
-   INTEGER, INTENT( IN ) :: lstadh, lstepa, lcalcf, lfuval, lxvalu, lepvlu
-   INTEGER, INTENT( OUT ) :: ifstat
-   INTEGER, INTENT( IN ) :: ITYPEE( LTYPEE ), ISTAEV( LSTAEV ), IELVAR( LELVAR )
-   INTEGER, INTENT( IN ) :: INTVAR( LNTVAR ), ISTADH( LSTADH ), ISTEPA( LSTEPA )
-   INTEGER, INTENT( IN ) :: ICALCF( LCALCF )
-   REAL ( KIND = wp ), INTENT( IN ) :: XVALUE( LXVALU )
-   REAL ( KIND = wp ), INTENT( IN ) :: EPVALU( LEPVLU )
-   REAL ( KIND = wp ), INTENT( INOUT ) :: FUVALS( LFUVAL )
-   INTEGER :: ielemn, ieltyp, ihstrt, ilstrt, igstrt, ipstrt, jcalcf
-   REAL ( KIND = wp ) :: v1, v2, u1, u2, u3, cs, sn
+   USE GALAHAD_PRECISION
+   INTEGER ( KIND = ip_ ), INTENT( IN ) :: ncalcf, ifflag, ltypee, lstaev
+   INTEGER ( KIND = ip_ ), INTENT( IN ) :: lelvar, lntvar, lstadh, lstepa
+   INTEGER ( KIND = ip_ ), INTENT( IN ) :: lcalcf, lfuval, lxvalu, lepvlu
[…]
