+++ src/lancelot/lancelot_types.F90
@@ -1,4 +1,21 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-18 AT 13:30 GMT.
+
+#ifdef LANCELOT_USE_MA57
+#define SILS_control MA57_control
+#define SILS_factors MA57_factors
+#define SILS_ainfo MA57_ainfo
+#define SILS_finfo MA57_finfo
+#define SILS_sinfo MA57_sinfo
+#define SILS_data MA57_data
+#define SILS_cntl MA57_cntl
+#endif
+
+#include "galahad_modules.h"
+
+#ifdef LANCELOT_USE_MA57
+#define GALAHAD_SILS_double HSL_MA57_double
+#define GALAHAD_SILS_single HSL_MA57_single
+#endif
 
 !-*-*-  L A N C E L O T  -B-  L A N C E L O T _ T Y P E S _  M O D U L E  -*-*-
 
@@ -11,72 +28,69 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE LANCELOT_types_double
-
+   MODULE LANCELOT_TYPES_precision
+            
 !  |-------------------------------------------------------|
 !  |                                                       |
 !  |  Derived types used by LANCELOT B and its associates  |
 !  |                                                       |
 !  |-------------------------------------------------------|
 
-     USE GALAHAD_EXTEND_double, ONLY: EXTEND_save_type
-     USE LANCELOT_ASMBL_double, ONLY: ASMBL_save_type
-     USE LANCELOT_CAUCHY_double, ONLY: CAUCHY_save_type
-     USE LANCELOT_CG_double, ONLY: CG_save_type
-     USE LANCELOT_OTHERS_double, ONLY: OTHERS_fdgrad_save_type
-     USE LANCELOT_PRECN_double, ONLY: PRECN_save_type
-     USE GALAHAD_SCU_double, ONLY: SCU_matrix_type, SCU_data_type,             &
-                                   SCU_inform_type
-     USE GALAHAD_SILS_double, ONLY: SILS_control, SILS_factors,                &
-                                    SILS_ainfo, SILS_finfo, SILS_sinfo
-     USE GALAHAD_SMT_double, ONLY: SMT_type
+     USE GALAHAD_PRECISION
+     USE GALAHAD_EXTEND_precision, ONLY: EXTEND_save_type
+     USE LANCELOT_ASMBL_precision, ONLY: ASMBL_save_type
+     USE LANCELOT_CAUCHY_precision, ONLY: CAUCHY_save_type
+     USE LANCELOT_CG_precision, ONLY: CG_save_type
+     USE LANCELOT_OTHERS_precision, ONLY: OTHERS_fdgrad_save_type
+     USE LANCELOT_PRECN_precision, ONLY: PRECN_save_type
+     USE GALAHAD_SCU_precision, ONLY: SCU_matrix_type, SCU_data_type,          &
+                                      SCU_inform_type
[…]
