### 2022-12-19 c1de4ee
filtrane and various utlities updated to new precision format
