+++ src/miqr/miqr.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ M I Q R    M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -13,8 +15,8 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-    MODULE GALAHAD_MIQR_double
-
+    MODULE GALAHAD_MIQR_precision
+            
 !     ---------------------------------------------------
 !     |                                                 |
 !     | Given a real matrix A, compute a multilevel     |
@@ -29,13 +31,14 @@
 !     |                                                 |
 !     ---------------------------------------------------
 
+      USE GALAHAD_PRECISION
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
-      USE GALAHAD_SPACE_double
-      USE GALAHAD_SPECFILE_double
-      USE GALAHAD_NORMS_double
-      USE GALAHAD_SMT_double
-      USE GALAHAD_CONVERT_double
+      USE GALAHAD_SPACE_precision
+      USE GALAHAD_SPECFILE_precision
+      USE GALAHAD_NORMS_precision
+      USE GALAHAD_SMT_precision
+      USE GALAHAD_CONVERT_precision
 
       IMPLICIT NONE
 
@@ -43,23 +46,16 @@
       PUBLIC :: MIQR_initialize, MIQR_read_specfile, MIQR_form, MIQR_apply,    &
                 MIQR_terminate, SMT_type, SMT_put, SMT_get
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-      INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-!     INTEGER, PARAMETER :: max_miqr_levels = 10
-      INTEGER, PARAMETER :: max_miqr_levels = 100
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-!     REAL ( KIND = wp ), PARAMETER :: epsmch = EPSILON( one )
+!     INTEGER ( KIND = ip_ ), PARAMETER :: max_miqr_levels = 10
+      INTEGER ( KIND = ip_ ), PARAMETER :: max_miqr_levels = 100
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+!     REAL ( KIND = rp_ ), PARAMETER :: epsmch = EPSILON( one )
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -73,87 +69,87 @@
 
 !  unit for error messages
 
-        INTEGER :: error = 6
+        INTEGER ( KIND = ip_ ) :: error = 6
 
 !  unit for monitor output
 
-        INTEGER :: out = 6
+        INTEGER ( KIND = ip_ ) :: out = 6
 
 !  controls level of diagnostic output
 
-        INTEGER :: print_level = 0
[…]
