+++ src/apps/sas/sase.F90
@@ -1,4 +1,5 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-07-06 AT 07:45 GMT
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 11:30 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_SAS_EXAMPLE
 
 ! use slls to solve a multilinear fitting problem
@@ -7,46 +8,46 @@
 ! where l, r, theta and phi lie in regular simplexes {z : e^T z = 1, z >= 0},
 ! using alternating solves with l free and r, theta, phi fixed etc.
 
-   USE GALAHAD_SLLS_double         ! double precision version
-   USE GALAHAD_SLS_double
-   USE GALAHAD_RAND_double
-   USE GALAHAD_SPACE_double
-   USE GALAHAD_NORMS_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_SLLS_precision
+   USE GALAHAD_SLS_precision
+   USE GALAHAD_RAND_precision
+   USE GALAHAD_SPACE_precision
+   USE GALAHAD_NORMS_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-   REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+   REAL ( KIND = rp_ ), PARAMETER :: infinity = ten ** 20
    TYPE ( QPT_problem_type ) :: p_l, p_r, p_t, p_p
    TYPE ( SLLS_data_type ) :: data_l, data_r, data_t, data_p
    TYPE ( SLLS_control_type ) :: control_l, control_r, control_t, control_p
    TYPE ( SLLS_inform_type ) :: inform_l, inform_r, inform_t, inform_p
    TYPE ( GALAHAD_userdata_type ) :: userdata
-   INTEGER :: i, il, ir, it, ip, ix, iy, j, k, ll, ne, status, alloc_status
-   INTEGER :: n, nl_free, nr_free, nt_free, np_free, iter, pass
-   INTEGER :: repeat
-   REAL ( KIND = wp ) :: val, f, xi, b, dxi, db, a_11, a_12, a_22, r_1, r_2
-   REAL ( KIND = wp ) :: alpha, beta, f_best, prod1, prod2, prod3, prod4
-   REAL ( KIND = wp ) :: step, f_step, gts
+   INTEGER ( KIND = ip_ ) :: i, il, ir, it, ip, ix, iy, j, k, ll, ne
+   INTEGER ( KIND = ip_ ) :: n, nl_free, nr_free, nt_free, np_free, iter, pass
+   INTEGER ( KIND = ip_ ) :: repeat, status, alloc_status
+   REAL ( KIND = rp_ ) :: val, f, xi, b, dxi, db, a_11, a_12, a_22, r_1, r_2
+   REAL ( KIND = rp_ ) :: alpha, beta, f_best, prod1, prod2, prod3, prod4
+   REAL ( KIND = rp_ ) :: step, f_step, gts
    REAL :: time, time_start, time_total
-!   INTEGER, PARAMETER :: nl = 5
[…]

[…]
