+++ src/fdc/fdcti.F90
@@ -1,21 +1,22 @@
-! THIS VERSION: GALAHAD 4.0 - 20/01/2022 AT 09:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 14:00 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_FDC_test    !! to be expanded
-   USE GALAHAD_FDC_double                      ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_FDC_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )   ! set precision
-   INTEGER, PARAMETER :: n = 4, m = 3, a_ne = 10
-   INTEGER :: n_depen, status
-   INTEGER :: A_ptr( m + 1 ), A_col( a_ne )
-   REAL ( KIND = wp ) :: A_val( a_ne ), B( m )
-   INTEGER :: DEPEN( m )
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 4, m = 3, a_ne = 10
+   INTEGER ( KIND = ip_ ) :: n_depen, status
+   INTEGER ( KIND = ip_ ) :: A_ptr( m + 1 ), A_col( a_ne )
+   REAL ( KIND = rp_ ) :: A_val( a_ne ), B( m )
+   INTEGER ( KIND = ip_ ) :: DEPEN( m )
    TYPE ( FDC_full_data_type ) :: data
    TYPE ( FDC_control_type ) :: control        
    TYPE ( FDC_inform_type ) :: inform
-   A_val = (/ 1.0_wp, 2.0_wp, 3.0_wp, 4.0_wp, 2.0_wp, -4.0_wp, 6.0_wp,         &
-             -8.0_wp, 5.0_wp, 10.0_wp /)
+   A_val = (/ 1.0_rp_, 2.0_rp_, 3.0_rp_, 4.0_rp_, 2.0_rp_, -4.0_rp_, 6.0_rp_,  &
+             -8.0_rp_, 5.0_rp_, 10.0_rp_ /)
    A_col = (/ 1, 2, 3, 4, 1, 2, 3, 4, 2, 4 /)
    A_ptr = (/ 1, 5, 9, 11 /)
-   B = (/ 5.0_wp, 10.0_wp, 0.0_wp /)
+   B = (/ 5.0_rp_, 10.0_rp_, 0.0_rp_ /)
    CALL FDC_initialize( data, control, inform )  ! Initialize control parameters
    CALL FDC_find_dependent_rows( control, data, inform, status, m, n,          &
                                  A_col, A_ptr, A_val, B, n_depen, DEPEN )
