### 2022-12-19 89efae1
filtrane and various utlities updated to new precision format
