### 2022-12-19 aaad554
filtrane and various utlities updated to new precision format
