+++ src/string/string.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 03/06/2021 AT 11:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 12:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _ S T R I N G   M O D U L E  *-*-*-*-*-*-*-*-*
 
@@ -12,6 +14,8 @@
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
    MODULE GALAHAD_STRING
+            
+     USE GALAHAD_PRECISION
 
 !     ----------------------------
 !    |  Set strings appropriate   |
@@ -34,8 +38,8 @@
 
 !  Set precision
 
-     INTEGER, PARAMETER :: sp = KIND( 1.0E+0 )
-     INTEGER, PARAMETER :: dp = KIND( 1.0D+0 )
+     INTEGER ( KIND = ip_ ), PARAMETER :: sp = KIND( 1.0E+0 )
+     INTEGER ( KIND = ip_ ), PARAMETER :: dp = KIND( 1.0D+0 )
 
      REAL( KIND = sp ), PARAMETER :: teneps_s = 10.0_sp * EPSILON( 1.0_sp )
      REAL( KIND = dp ), PARAMETER :: teneps_d = 10.0_dp * EPSILON( 1.0_dp )
@@ -45,31 +49,32 @@
 !---------------------------------
 
      INTERFACE STRING_sign
-       MODULE PROCEDURE STRING_sign_single, STRING_sign_double
+       MODULE PROCEDURE STRING_sign_single, STRING_sign_precision
      END INTERFACE
 
      INTERFACE STRING_real_7
-       MODULE PROCEDURE STRING_real_7_single, STRING_real_7_double
+       MODULE PROCEDURE STRING_real_7_single, STRING_real_7_precision
       END INTERFACE
 
      INTERFACE STRING_real_12
-       MODULE PROCEDURE STRING_real_12_single, STRING_real_12_double
+       MODULE PROCEDURE STRING_real_12_single, STRING_real_12_precision
      END INTERFACE
 
      INTERFACE STRING_trim_real_24
-       MODULE PROCEDURE STRING_trim_real_24_single, STRING_trim_real_24_double
+       MODULE PROCEDURE STRING_trim_real_24_single,                            &
+                        STRING_trim_real_24_precision
      END INTERFACE
 
      INTERFACE STRING_es
-       MODULE PROCEDURE STRING_es_single, STRING_es_double
+       MODULE PROCEDURE STRING_es_single, STRING_es_precision
      END INTERFACE
 
      INTERFACE STRING_es12
-       MODULE PROCEDURE STRING_es12_single, STRING_es12_double
+       MODULE PROCEDURE STRING_es12_single, STRING_es12_precision
      END INTERFACE
 
      INTERFACE STRING_exponent
-       MODULE PROCEDURE STRING_exponent_single, STRING_exponent_double
+       MODULE PROCEDURE STRING_exponent_single, STRING_exponent_precision
      END INTERFACE
 
      INTERFACE STRING_lower
@@ -94,7 +99,7 @@
 !   D u m m y   A r g u m e n t
 !--------------------------------
 
-     INTEGER, INTENT( IN ) :: val
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: val
 
      IF ( val /= 1 ) THEN
        STRING_pleural = "s"
@@ -120,7 +125,7 @@
 !   D u m m y   A r g u m e n t
 !--------------------------------
 
[…]
