+++ src/dum/umfpack.F90
@@ -1,2 +1,3 @@
-      SUBROUTINE GALAHAD_umfpack( )
-      END SUBROUTINE GALAHAD_umfpack
+   SUBROUTINE GALAHAD_umfpack( )
+   USE GALAHAD_PRECISION
+   END SUBROUTINE GALAHAD_umfpack

[…]
