### 2022-12-19 cc8ae2d
filtrane and various utlities updated to new precision format
