+++ src/scale/scalet.F90
@@ -1,21 +1,22 @@
-! THIS VERSION: GALAHAD 2.4 - 17/01/2011 AT 15:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 16:40 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_SCALE_TEST
-   USE GALAHAD_SCALE_double         ! double precision version
-   USE GALAHAD_SMT_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_SCALE_precision
+   USE GALAHAD_SMT_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infty = 10.0_wp ** 20
-   REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
+   REAL ( KIND = rp_ ), PARAMETER :: infty = 10.0_rp_ ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
    TYPE ( QPT_problem_type ) :: p
    TYPE ( SCALE_trans_type ) :: trans
    TYPE ( SCALE_data_type ) :: data
    TYPE ( SCALE_control_type ) :: control        
    TYPE ( SCALE_inform_type ) :: info
-   INTEGER :: n, m, h_ne, a_ne, tests, smt_stat
-   INTEGER :: data_storage_type, i, status, scratch_out = 56
+   INTEGER ( KIND = ip_ ) :: n, m, h_ne, a_ne, tests, smt_stat
+   INTEGER ( KIND = ip_ ) :: data_storage_type, i, status, scratch_out = 56
    CHARACTER ( len = 1 ) :: st
-   INTEGER :: scale
+   INTEGER ( KIND = ip_ ) :: scale
 
    n = 3 ; m = 2 ; h_ne = 4 ; a_ne = 4 
    ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
@@ -62,33 +63,33 @@
      control%infinity = infty
 
      p%new_problem_structure = .TRUE.
-     p%n = n ; p%m = m ; p%f = 1.0_wp
-     p%G = (/ 0.0_wp, 2.0_wp, 0.0_wp /)
-     p%C_l = (/ 1.0_wp, 2.0_wp /)
-     p%C_u = (/ 4.0_wp, infty /)
-     p%X_l = (/ - 1.0_wp, - infty, - infty /)
-     p%X_u = (/ 1.0_wp, infty, 2.0_wp /)
+     p%n = n ; p%m = m ; p%f = 1.0_rp_
+     p%G = (/ 0.0_rp_, 2.0_rp_, 0.0_rp_ /)
+     p%C_l = (/ 1.0_rp_, 2.0_rp_ /)
+     p%C_u = (/ 4.0_rp_, infty /)
+     p%X_l = (/ - 1.0_rp_, - infty, - infty /)
+     p%X_u = (/ 1.0_rp_, infty, 2.0_rp_ /)
 
      ALLOCATE( p%H%val( h_ne ), p%H%row( 0 ), p%H%col( h_ne ) )
      ALLOCATE( p%A%val( a_ne ), p%A%row( 0 ), p%A%col( a_ne ) )
      IF ( ALLOCATED( p%H%type ) ) DEALLOCATE( p%H%type )
      CALL SMT_put( p%H%type, 'SPARSE_BY_ROWS', smt_stat ) 
[…]
