+++ src/scaling/scaling.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 11:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 16:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*- G A L A H A D _ S C A L I N G   M O D U L E  *-*-*-*-*-*-*-*-*-
 
@@ -9,8 +11,8 @@
 !  Updated for GALAHAD 2.0 July 14th, 2006
 !  Name changed from oldscale to scaling GALAHAD 3.3 May 20th, 2021
 
-   MODULE GALAHAD_SCALING_double
-
+   MODULE GALAHAD_SCALING_precision
+            
 !     ------------------------------------------
 !     | Scale data for the quadratic program   |
 !     |                                        |
@@ -21,6 +23,8 @@
 !     | Nick Gould                             |
 !     | September 1999                         |
 !     ------------------------------------------
+
+      USE GALAHAD_PRECISION
 
       IMPLICIT NONE
 
@@ -34,18 +38,15 @@
       END INTERFACE SCALING_apply_factors
 
 !--------------------
-!   P r e c i s i o n
-!--------------------
 
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
       TYPE, PUBLIC :: SCALING_control_type
-        INTEGER :: print_level, out, out_error
+        INTEGER ( KIND = ip_ ) :: print_level, out, out_error
       END TYPE
 
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: two = 2.0_rp_
 
    CONTAINS
 
@@ -94,31 +95,35 @@
 
 !  Dummy arguments
 
-      INTEGER, INTENT( IN ) :: n, m
-      INTEGER, INTENT( OUT ) :: ifail
-      INTEGER, INTENT( IN ), DIMENSION( m + 1 ) :: A_ptr
-      INTEGER, INTENT( IN ), DIMENSION( A_ptr( m + 1 ) - 1 ) :: A_col
-      INTEGER, INTENT( IN ), DIMENSION( n + 1 ) :: H_ptr
-      INTEGER, INTENT( IN ), DIMENSION( H_ptr( n + 1 ) - 1 ) :: H_col
-      REAL ( KIND = wp ), INTENT( IN ), DIMENSION( A_ptr( m + 1 ) - 1 ) :: A_val
-      REAL ( KIND = wp ), INTENT( IN ), DIMENSION( H_ptr( n + 1 ) - 1 ) :: H_val
-      REAL ( KIND = wp ), INTENT( OUT ), DIMENSION( n ) :: SH
-      REAL ( KIND = wp ), INTENT( OUT ), DIMENSION( m ) :: SA
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: n, m
+      INTEGER ( KIND = ip_ ), INTENT( OUT ) :: ifail
+      INTEGER ( KIND = ip_ ), INTENT( IN ), DIMENSION( m + 1 ) :: A_ptr
+      INTEGER ( KIND = ip_ ), INTENT( IN ),                                    &
[…]
