+++ src/fdc/fdcs.f90
@@ -17,6 +17,8 @@
    A_ptr = (/ 1, 5, 9, 11 /)
    B = (/ 5.0_wp, 10.0_wp, 0.0_wp /)
    CALL FDC_initialize( data, control, inform )  ! Initialize control parameters
+   control%use_sls = .TRUE.
+   control%symmetric_linear_solver = 'sytr'
    CALL FDC_find_dependent( n, m, A_val, A_col, A_ptr, B, n_depen, DEPEN,      &
                             data, control, inform )  ! Check for dependencies
    IF ( inform%status == 0 ) THEN              ! Successful return

[…]
