### 2022-12-21 eaa1379
basic qp packages updated to new precision format
