+++ src/cqp/usecqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 27/01/2020 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 08:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D   U S E C Q P   M O D U L E  -*-*-*-*-*-*-*-*-
 
@@ -11,24 +13,25 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-    MODULE GALAHAD_USECQP_double
+    MODULE GALAHAD_USECQP_precision
 
 !    -------------------------------------------------------------
 !    | CUTEst/AMPL interface to CQP, an interior-point algorithm |
 !    | for convex quadratic & least-distance programming         |
 !    -------------------------------------------------------------
 
+      USE GALAHAD_PRECISION
 !$    USE omp_lib
-      USE CUTEst_interface_double
+      USE CUTEst_interface_precision
       USE GALAHAD_CLOCK
-      USE GALAHAD_RAND_double
-      USE GALAHAD_QPT_double
-      USE GALAHAD_SORT_double, ONLY: SORT_reorder_by_rows
-      USE GALAHAD_NORMS_double, ONLY: TWO_NORM
-      USE GALAHAD_CQP_double
-      USE GALAHAD_SLS_double
-      USE GALAHAD_PRESOLVE_double
-      USE GALAHAD_SPECFILE_double
+      USE GALAHAD_RAND_precision
+      USE GALAHAD_QPT_precision
+      USE GALAHAD_SORT_precision, ONLY: SORT_reorder_by_rows
+      USE GALAHAD_NORMS_precision, ONLY: TWO_NORM
+      USE GALAHAD_CQP_precision
+      USE GALAHAD_SLS_precision
+      USE GALAHAD_PRESOLVE_precision
+      USE GALAHAD_SPECFILE_precision
       USE GALAHAD_STRING, ONLY: STRING_upper_word
       USE GALAHAD_COPYRIGHT
       USE GALAHAD_SYMBOLS,                                                     &
@@ -37,7 +40,7 @@
           DEBUG                 => GALAHAD_DEBUG,                              &
           GENERAL               => GALAHAD_GENERAL,                            &
           ALL_ZEROS             => GALAHAD_ALL_ZEROS
-      USE GALAHAD_SCALE_double
+      USE GALAHAD_SCALE_precision
 
       IMPLICIT NONE
 
@@ -65,40 +68,39 @@
 
 !  Dummy argument
 
-      INTEGER, INTENT( IN ) :: input
+      INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
       LOGICAL, OPTIONAL, INTENT( IN ) :: close_input
 
 !  Parameters
 
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-      REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 19
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
[…]
