+++ src/presolve/presolve.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-27 AT 13:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-20 AT 08:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
@@ -271,21 +273,23 @@
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 !-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*
 
-   MODULE GALAHAD_PRESOLVE_double
+   MODULE GALAHAD_PRESOLVE_precision
 
 !-------------------------------------------------------------------------------
 !   U s e d   m o d u l e s   a n d   s y m b o l s
 !-------------------------------------------------------------------------------
 
-      USE GALAHAD_SMT_double      ! the matrix data type
+      USE GALAHAD_PRECISION
 
-      USE GALAHAD_QPT_double      ! the quadratic problem data type
+      USE GALAHAD_SMT_precision      ! the matrix data type
 
-      USE GALAHAD_SORT_double     ! sorting and permutation operations
+      USE GALAHAD_QPT_precision      ! the quadratic problem data type
 
-      USE GALAHAD_SPECFILE_double ! specfile manipulation
+      USE GALAHAD_SORT_precision     ! sorting and permutation operations
 
-      USE GALAHAD_SPACE_double    ! allocating, extending and deleting storage
+      USE GALAHAD_SPECFILE_precision ! specfile manipulation
+
+      USE GALAHAD_SPACE_precision   ! allocating, extending and deleting storage
 
 !     Matrix storage schemes
 
@@ -412,25 +416,22 @@
      END INTERFACE PRESOLVE_terminate
 
 !-------------------------------------------------------------------------------
-!   P r e c i s i o n
-!-------------------------------------------------------------------------------
 
-      INTEGER, PRIVATE, PARAMETER :: sp = KIND( 1.0 )
-      INTEGER, PRIVATE, PARAMETER :: dp = KIND( 1.0D+0 )
-      INTEGER, PRIVATE, PARAMETER :: wp = dp
+      INTEGER ( KIND = ip_ ), PRIVATE, PARAMETER :: sp = sp_
+      INTEGER ( KIND = ip_ ), PRIVATE, PARAMETER :: dp = dp_
 
 !-------------------------------------------------------------------------------
 !   C o n s t a n t s
 !-------------------------------------------------------------------------------
 
-      REAL ( KIND = wp ), PRIVATE, PARAMETER :: ZERO    = 0.0_wp
-      REAL ( KIND = wp ), PRIVATE, PARAMETER :: HALF    = 0.5_wp
-      REAL ( KIND = wp ), PRIVATE, PARAMETER :: ONE     = 1.0_wp
[…]
