+++ src/presolve/runpresolve_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.0 - 24/10/2017 AT 15:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-20 AT 13:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D   R U N P R E S O L V E _ S I F  *-*-*-*-*-*-*-*-
 
@@ -7,14 +9,15 @@
 !  March 14th 2003
 
    PROGRAM RUNPRESOLVE_SIF
-   USE GALAHAD_USEPRESOLVE_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USEPRESOLVE_precision
 
 !  Main program for the SIF/CUTEr interface to PRESOLVE, a preprocessing
 !  algorithm for quadratic programs
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
 
 !  Open the data input file
