+++ src/dqp/dqpt.F90
@@ -1,18 +1,31 @@
-! THIS VERSION: GALAHAD 2.5 - 01/08/2012 AT 08:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 08:00 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_DQP_EXAMPLE
-   USE GALAHAD_DQP_double                            ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_DQP_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )         ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infty = 10.0_wp ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: infty = 10.0_rp_ ** 20
    TYPE ( QPT_problem_type ) :: p
    TYPE ( DQP_data_type ) :: data
    TYPE ( DQP_control_type ) :: control
    TYPE ( DQP_inform_type ) :: info
-   INTEGER :: n, m, h_ne, a_ne, tests, smt_stat
-   INTEGER :: data_storage_type, i, status, scratch_out = 56
+   INTEGER ( KIND = ip_ ) :: n, m, h_ne, a_ne, tests, smt_stat
+   INTEGER ( KIND = ip_ ) :: data_storage_type, i, status, scratch_out = 56
    CHARACTER ( len = 1 ) :: st
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: C_stat, X_stat
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: C_stat, X_stat
+   CHARACTER ( LEN = 30 ) :: symmetric_linear_solver = REPEAT( ' ', 30 )
+   CHARACTER ( LEN = 30 ) :: definite_linear_solver = REPEAT( ' ', 30 )
+!  symmetric_linear_solver = 'ssids'
+!  symmetric_linear_solver = 'ma97'
+   symmetric_linear_solver = 'sils'
+   symmetric_linear_solver = 'sytr'
+   symmetric_linear_solver = 'ssids'
+!  definite_linear_solver = 'ssids'
+!  definite_linear_solver = 'ma97'
+   definite_linear_solver = 'sils'
+   definite_linear_solver = 'sytr'
+   definite_linear_solver = 'ssids'
 
    n = 3 ; m = 2 ; h_ne = 4 ; a_ne = 4
    ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
@@ -59,6 +72,8 @@
      CALL DQP_initialize( data, control, info )
      control%infinity = infty
      control%restore_problem = 1
+     control%SBLS_control%symmetric_linear_solver = symmetric_linear_solver
+     control%SBLS_control%definite_linear_solver = definite_linear_solver
 ! control%print_level = 1
 
 !control%print_level = 3
@@ -67,50 +82,50 @@
 !control%SBLS_control%preconditioner = 3
 
      p%new_problem_structure = .TRUE.
[…]
