### 2022-12-21 4073b17
basic qp packages updated to new precision format
