+++ src/qp/qp_spec.F90
@@ -1,38 +1,39 @@
-! THIS VERSION: GALAHAD 2.4 - 10/01/2011 AT 07:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-20 AT 15:50 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_QP_EXAMPLE
-   USE GALAHAD_QP_double         ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_QP_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: infinity = 10.0_rp_ ** 20
    TYPE ( QPT_problem_type ) :: p
    TYPE ( QP_data_type ) :: data
    TYPE ( QP_control_type ) :: control        
    TYPE ( QP_inform_type ) :: inform
-   INTEGER :: s
-   INTEGER, PARAMETER :: n = 3, m = 2, h_ne = 4, a_ne = 4 
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: C_stat, B_stat
+   INTEGER ( KIND = ip_ ) :: s
+   INTEGER ( KIND = ip_ ), PARAMETER :: n = 3, m = 2, h_ne = 4, a_ne = 4 
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: C_stat, B_stat
 ! start problem data
    ALLOCATE( p%G( n ), p%X_l( n ), p%X_u( n ) )
    ALLOCATE( p%C( m ), p%C_l( m ), p%C_u( m ) )
    ALLOCATE( p%X( n ), p%Y( m ), p%Z( n ) )
    ALLOCATE( B_stat( n ), C_stat( m ) )
    p%new_problem_structure = .TRUE.           ! new structure
-   p%n = n ; p%m = m ; p%f = 1.0_wp           ! dimensions & objective constant
-   p%G = (/ 0.0_wp, 2.0_wp, 0.0_wp /)         ! objective gradient
-   p%C_l = (/ 1.0_wp, 2.0_wp /)               ! constraint lower bound
-   p%C_u = (/ 2.0_wp, 2.0_wp /)               ! constraint upper bound
-   p%X_l = (/ - 1.0_wp, - infinity, - infinity /) ! variable lower bound
-   p%X_u = (/ 1.0_wp, infinity, 2.0_wp /)     ! variable upper bound
-   p%X = 0.0_wp ; p%Y = 0.0_wp ; p%Z = 0.0_wp ! start from zero
+   p%n = n ; p%m = m ; p%f = 1.0_rp_           ! dimensions & objective constant
+   p%G = (/ 0.0_rp_, 2.0_rp_, 0.0_rp_ /)         ! objective gradient
+   p%C_l = (/ 1.0_rp_, 2.0_rp_ /)               ! constraint lower bound
+   p%C_u = (/ 2.0_rp_, 2.0_rp_ /)               ! constraint upper bound
+   p%X_l = (/ - 1.0_rp_, - infinity, - infinity /) ! variable lower bound
+   p%X_u = (/ 1.0_rp_, infinity, 2.0_rp_ /)     ! variable upper bound
+   p%X = 0.0_rp_ ; p%Y = 0.0_rp_ ; p%Z = 0.0_rp_ ! start from zero
[…]

[…]
