### 2022-12-21 ca9a8ce
basic qp packages updated to new precision format
