+++ src/cqp/incqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 20/05/2021 AT 11:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 08:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D   R U N C Q P _ D A T A  *-*-*-*-*-*-*-*-*-*-
 
@@ -19,32 +21,33 @@
 !    | for convex quadratic programming                        |
 !    -----------------------------------------------------------
 
+   USE GALAHAD_PRECISION
 !$ USE omp_lib
    USE GALAHAD_CLOCK
-   USE GALAHAD_RAND_double
-   USE GALAHAD_QPT_double
-   USE GALAHAD_RPD_double
-   USE GALAHAD_SMT_double, only: SMT_put
-   USE GALAHAD_CQP_double
-   USE GALAHAD_SORT_double, only: SORT_reorder_by_rows
-   USE GALAHAD_NORMS_double, ONLY: TWO_NORM
-   USE GALAHAD_SLS_double
-   USE GALAHAD_PRESOLVE_double
-   USE GALAHAD_SPECFILE_double
+   USE GALAHAD_RAND_precision
+   USE GALAHAD_QPT_precision
+   USE GALAHAD_RPD_precision
+   USE GALAHAD_SMT_precision, only: SMT_put
+   USE GALAHAD_CQP_precision
+   USE GALAHAD_SORT_precision, only: SORT_reorder_by_rows
+   USE GALAHAD_NORMS_precision, ONLY: TWO_NORM
+   USE GALAHAD_SLS_precision
+   USE GALAHAD_PRESOLVE_precision
+   USE GALAHAD_SPECFILE_precision
    USE GALAHAD_STRING, ONLY: STRING_upper_word
    USE GALAHAD_COPYRIGHT
-   USE GALAHAD_SCALING_double
+   USE GALAHAD_SCALING_precision
    USE GALAHAD_SYMBOLS,                                                        &
        ACTIVE                => GALAHAD_ACTIVE,                                &
        TRACE                 => GALAHAD_TRACE,                                 &
        DEBUG                 => GALAHAD_DEBUG,                                 &
        GENERAL               => GALAHAD_GENERAL,                               &
        ALL_ZEROS             => GALAHAD_ALL_ZEROS
-   USE GALAHAD_SCALE_double
+   USE GALAHAD_SCALE_precision
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 5
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 5
 
 !  --------------------------------------------
 !
@@ -116,31 +119,30 @@
 
 !  Parameters
 
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-      REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 19
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: infinity = ten ** 19
 
 !  Scalars
 
[…]
