### 2022-12-21 9866edd
basic qp packages updated to new precision format
