+++ src/cqp/cqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-11 AT 15:25 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-20 AT 16:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D _ C Q P    M O D U L E  -*-*-*-*-*-*-*-*-
 
@@ -12,7 +14,7 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-    MODULE GALAHAD_CQP_double
+    MODULE GALAHAD_CQP_precision
 
 !     ------------------------------------------------
 !     |                                              |
@@ -35,29 +37,32 @@
 !     |                                              |
 !     ------------------------------------------------
 
+      USE GALAHAD_PRECISION
 !$    USE omp_lib
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
       USE GALAHAD_STRING, ONLY: STRING_pleural, STRING_verb_pleural,           &
                                        STRING_ies, STRING_are, STRING_ordinal
-      USE GALAHAD_SPACE_double
-      USE GALAHAD_SMT_double
-      USE GALAHAD_QPT_double
-      USE GALAHAD_SPECFILE_double
-      USE GALAHAD_QPP_double, CQP_dims_type => QPT_dimensions_type
-      USE GALAHAD_QPD_double, CQP_data_type => QPD_data_type,                  &
-                              CQP_AX => QPD_AX, CQP_HX => QPD_HX,              &
-                              CQP_abs_AX => QPD_abs_AX, CQP_abs_HX => QPD_abs_HX
-      USE GALAHAD_ROOTS_double
-      USE GALAHAD_LMS_double, ONLY: LMS_data_type, LMS_apply_lbfgs
-      USE GALAHAD_SORT_double, ONLY: SORT_inverse_permute
-      USE GALAHAD_FDC_double
-      USE GALAHAD_SBLS_double
-      USE GALAHAD_CRO_double
-      USE GALAHAD_FIT_double
-      USE GALAHAD_NORMS_double, ONLY: TWO_norm
-      USE GALAHAD_CHECKPOINT_double
-      USE GALAHAD_RPD_double, ONLY: RPD_inform_type, RPD_write_qp_problem_data
+      USE GALAHAD_SPACE_precision
+      USE GALAHAD_SMT_precision
+      USE GALAHAD_QPT_precision
+      USE GALAHAD_SPECFILE_precision
+      USE GALAHAD_QPP_precision, CQP_dims_type => QPT_dimensions_type
+      USE GALAHAD_QPD_precision, CQP_data_type => QPD_data_type,               &
+                                 CQP_AX => QPD_AX, CQP_HX => QPD_HX,           &
+                                 CQP_abs_AX => QPD_abs_AX,                     &
+                                 CQP_abs_HX => QPD_abs_HX
+      USE GALAHAD_ROOTS_precision
+      USE GALAHAD_LMS_precision, ONLY: LMS_data_type, LMS_apply_lbfgs
+      USE GALAHAD_SORT_precision, ONLY: SORT_inverse_permute
+      USE GALAHAD_FDC_precision
+      USE GALAHAD_SBLS_precision
+      USE GALAHAD_CRO_precision
+      USE GALAHAD_FIT_precision
+      USE GALAHAD_NORMS_precision, ONLY: TWO_norm
+      USE GALAHAD_CHECKPOINT_precision
+      USE GALAHAD_RPD_precision, ONLY: RPD_inform_type,                        &
+                                       RPD_write_qp_problem_data
 
[…]
