+++ src/cqp/runcqp_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.8 - 24/08/2016 AT 12:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 08:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D   R U N C Q P _ S I F  *-*-*-*-*-*-*-*-*-*-
 
@@ -19,11 +21,12 @@
 !    | quadratic & least-distance programming       |
 !    ------------------------------------------------
 
-   USE GALAHAD_USECQP_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USECQP_precision
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
 
 !  Open the data input file

[…]
