### 2022-12-22 540dfd5
auxiliary qp packages updated to new precision format
