+++ src/lpa/lpati.F90
@@ -1,41 +1,42 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-01-10 AT 11:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 08:50 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_LPA_interface_test
-   USE GALAHAD_LPA_double                       ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_LPA_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: infinity = 10.0_rp_ ** 20
    TYPE ( LPA_control_type ) :: control
    TYPE ( LPA_inform_type ) :: inform
    TYPE ( LPA_full_data_type ) :: data
-   INTEGER :: n, m, A_ne
-   INTEGER :: data_storage_type, status
-   REAL ( KIND = wp ) :: f
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: X, Z, X_l, X_u, G
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: Y, C, C_l, C_u
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: A_row, A_col, A_ptr
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: A_val, A_dense
-   INTEGER, ALLOCATABLE, DIMENSION( : ) :: C_stat, X_stat
+   INTEGER ( KIND = ip_ ) :: n, m, A_ne
+   INTEGER ( KIND = ip_ ) :: data_storage_type, status
+   REAL ( KIND = rp_ ) :: f
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: X, Z, X_l, X_u, G
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: Y, C, C_l, C_u
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: A_row, A_col, A_ptr
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: A_val, A_dense
+   INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: C_stat, X_stat
    CHARACTER ( len = 2 ) :: st
 
 ! set up problem data
 
    n = 3 ;  m = 2 ; A_ne = 4
-   f = 1.0_wp
+   f = 1.0_rp_
    ALLOCATE( X( n ), Z( n ), X_l( n ), X_u( n ), G( n ), X_stat( n ) )
    ALLOCATE( C( m ), Y( m ), C_l( m ), C_u( m ), C_stat( m ) )
-   G = (/ 0.0_wp, 2.0_wp, 0.0_wp /)         ! objective gradient
-   C_l = (/ 1.0_wp, 2.0_wp /)               ! constraint lower bound
-   C_u = (/ 2.0_wp, 2.0_wp /)               ! constraint upper bound
-   X_l = (/ - 1.0_wp, - infinity, - infinity /) ! variable lower bound
-   X_u = (/ 1.0_wp, infinity, 2.0_wp /)     ! variable upper bound
[…]
