### 2022-12-22 f0c377e
auxiliary qp packages updated to new precision format
