### 2022-12-22 69c59c7
auxiliary qp packages updated to new precision format
