### 2022-12-22 e1c6f97
auxiliary qp packages updated to new precision format
