+++ src/lqr/lqr.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 08/10/2021 AT 09:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-22 AT 10:50 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _ L Q R  double  M O D U L E  *-*-*-*-*-*-*-
 
@@ -11,7 +13,7 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_LQR_double
+   MODULE GALAHAD_LQR_precision
 
 !      ----------------------------------------------------------------------
 !      |                                                                    |
@@ -24,13 +26,13 @@
 !      |                                                                    |
 !      ----------------------------------------------------------------------
 
+      USE GALAHAD_PRECISION
       USE GALAHAD_SYMBOLS
-      USE GALAHAD_SPACE_double
-!     USE GALAHAD_ROOTS_double, ONLY: ROOTS_quadratic, ROOTS_quartic
-      USE GALAHAD_ROOTS_double
-      USE GALAHAD_SPECFILE_double
+      USE GALAHAD_SPACE_precision
+!     USE GALAHAD_ROOTS_precision, ONLY: ROOTS_quadratic, ROOTS_quartic
+      USE GALAHAD_ROOTS_precision
+      USE GALAHAD_SPECFILE_precision
       USE GALAHAD_LAPACK_interface, ONLY : LAEV2
-
 
       IMPLICIT NONE
 
@@ -39,25 +41,19 @@
                 LQR_terminate, LQR_full_initialize, LQR_full_terminate,        &
                 LQR_solve_2d, LQR_import, LQR_information
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-      REAL ( KIND = wp ), PARAMETER :: four = 4.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-      REAL ( KIND = wp ), PARAMETER :: epsmch = EPSILON( one )
-      REAL ( KIND = wp ), PARAMETER :: biginf = HUGE( one )
-      REAL ( KIND = wp ), PARAMETER :: boundary_tol = epsmch ** 0.75
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: half = 0.5_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: two = 2.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: four = 4.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: epsmch = EPSILON( one )
+      REAL ( KIND = rp_ ), PARAMETER :: biginf = HUGE( one )
+      REAL ( KIND = rp_ ), PARAMETER :: boundary_tol = epsmch ** 0.75
       LOGICAL :: roots_debug = .FALSE.
[…]
