+++ src/lpqp/lpqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 24/07/2019 AT 13:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-22 AT 09:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ L P Q P   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -12,12 +14,12 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_LPQP_double
+   MODULE GALAHAD_LPQP_precision
 
 !     ----------------------------------------------------
 !     |                                                  |
-!     | Convert an ordinary quadratic program into       |
-!     | an l_p quadratic program                         |
+!     |  Convert an ordinary quadratic program into      |
+!     |  an l_p quadratic program                        |
 !     |                                                  |
 !     |    minimize     1/2 x(T) H x + g(T) x + f        |
 !     |     + rho || max( 0, c_l - A x, A x - c_u ) ||_p |
@@ -25,13 +27,14 @@
 !     |                                                  |
 !     ----------------------------------------------------
 
+     USE GALAHAD_PRECISION
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_QPT_double
-     USE GALAHAD_SORT_double
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SMT_double, ONLY: SMT_put, SMT_get
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_QPT_precision
+     USE GALAHAD_SORT_precision
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SMT_precision, ONLY: SMT_put, SMT_get
 
      IMPLICIT NONE
 
@@ -39,32 +42,26 @@
      PUBLIC :: LPQP_initialize, LPQP_read_specfile, LPQP_formulate,            &
                LPQP_restore, LPQP_terminate, QPT_problem_type
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-     INTEGER, PARAMETER :: a_output_sparse_by_rows = 0
-     INTEGER, PARAMETER :: a_output_coordinate = 1
-     INTEGER, PARAMETER :: a_output_dense = 2
-     INTEGER, PARAMETER :: h_output_sparse_by_rows = 0
-     INTEGER, PARAMETER :: h_output_coordinate = 1
-     INTEGER, PARAMETER :: h_output_dense = 2
-     INTEGER, PARAMETER :: h_output_diagonal = 3
-     INTEGER, PARAMETER :: h_output_none = 4
-     INTEGER, PARAMETER :: h_output_lbfgs = 5
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
[…]

[…]
