### 2022-12-22 3f690ca
auxiliary qp packages updated to new precision format
