+++ src/ccqp/ccqpt.F90
@@ -1,30 +1,33 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-05-18 AT 08:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 11:55 GMT.
+#include "galahad_modules.h"
    PROGRAM GALAHAD_CCQP_EXAMPLE
-   USE GALAHAD_CCQP_double                            ! double precision version
+   USE GALAHAD_PRECISION
+   USE GALAHAD_CCQP_precision
    USE GALAHAD_SYMBOLS
-   USE GALAHAD_LMS_double
+   USE GALAHAD_LMS_precision
    IMPLICIT NONE
-   INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )         ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infty = 10.0_wp ** 20
+   REAL ( KIND = rp_ ), PARAMETER :: infty = 10.0_rp_ ** 20
    TYPE ( QPT_problem_type ) :: p
    TYPE ( CCQP_data_type ) :: data
    TYPE ( CCQP_control_type ) :: control
    TYPE ( CCQP_inform_type ) :: info
    TYPE ( LMS_control_type ) :: LMS_control
    TYPE ( LMS_inform_type ) :: LMS_inform
-   INTEGER :: n, m, h_ne, a_ne, tests, smt_stat
-   INTEGER :: data_storage_type, i, status, scratch_out = 56
-   REAL ( KIND = wp ) :: delta
+   INTEGER ( KIND = ip_ ) :: n, m, h_ne, a_ne, tests, smt_stat
+   INTEGER ( KIND = ip_ ) :: data_storage_type, i, status, scratch_out = 56
+   REAL ( KIND = rp_ ) :: delta
    CHARACTER ( len = 1 ) :: st
    CHARACTER ( LEN = 30 ) :: symmetric_linear_solver = REPEAT( ' ', 30 )
    CHARACTER ( LEN = 30 ) :: definite_linear_solver = REPEAT( ' ', 30 )
-   REAL ( KIND = wp ), ALLOCATABLE, DIMENSION( : ) :: S, Y
+   REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: S, Y
 !  symmetric_linear_solver = 'ssids'
 !  symmetric_linear_solver = 'ma97'
    symmetric_linear_solver = 'sils'
+   symmetric_linear_solver = 'sytr'
 !  definite_linear_solver = 'ssids'
 !  definite_linear_solver = 'ma97'
    definite_linear_solver = 'sils'
+   definite_linear_solver = 'sytr'
 
 !go to 111
    n = 3 ; m = 2 ; h_ne = 4 ; a_ne = 4
@@ -80,47 +83,47 @@
 !control%SBLS_control%preconditioner = 3
 
      p%new_problem_structure = .TRUE.
-     p%n = n ; p%m = m ; p%f = 1.0_wp
-     p%G = (/ 0.0_wp, 2.0_wp, 0.0_wp /)
-     p%C_l = (/ 1.0_wp, 2.0_wp /)
-     p%C_u = (/ 4.0_wp, infty /)
-     p%X_l = (/ - 1.0_wp, - infty, - infty /)
-     p%X_u = (/ 1.0_wp, infty, 2.0_wp /)
[…]
