+++ src/spral/matrix_util.F90
@@ -1,20 +1,25 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-12-23 AT 08:00 GMT.
+
+#include "spral_procedures.h"
+
 ! COPYRIGHT (c) 2010, 2013 Science and Technology Facilities Council
 ! Authors: Jonathan Hogg, John Reid, Jennifer Scott and Sue Thorne
 !
 ! Based on modified versions of hsl_mc34 and hsl_mc69
 !
-module spral_matrix_util
+module spral_matrix_util_precision
+   use spral_precision
    implicit none
 
    private
    ! Following are all types of matrix we may need to deal with
-   public :: SPRAL_MATRIX_UNSPECIFIED,                              &
-      SPRAL_MATRIX_REAL_RECT, SPRAL_MATRIX_CPLX_RECT,             &
-      SPRAL_MATRIX_REAL_UNSYM, SPRAL_MATRIX_CPLX_UNSYM,           &
-      SPRAL_MATRIX_REAL_SYM_PSDEF, SPRAL_MATRIX_CPLX_HERM_PSDEF,  &
-      SPRAL_MATRIX_REAL_SYM_INDEF, SPRAL_MATRIX_CPLX_HERM_INDEF,  &
-      SPRAL_MATRIX_CPLX_SYM,                                      &
-      SPRAL_MATRIX_REAL_SKEW, SPRAL_MATRIX_CPLX_SKEW
+   public :: SPRAL_MATRIX_UNSPECIFIED,                                         &
+             SPRAL_MATRIX_REAL_RECT, SPRAL_MATRIX_CPLX_RECT,                   &
+             SPRAL_MATRIX_REAL_UNSYM, SPRAL_MATRIX_CPLX_UNSYM,                 &
+             SPRAL_MATRIX_REAL_SYM_PSDEF, SPRAL_MATRIX_CPLX_HERM_PSDEF,        &
+             SPRAL_MATRIX_REAL_SYM_INDEF, SPRAL_MATRIX_CPLX_HERM_INDEF,        &
+             SPRAL_MATRIX_CPLX_SYM,                                            &
+             SPRAL_MATRIX_REAL_SKEW, SPRAL_MATRIX_CPLX_SKEW
    public ::                  &
       apply_conversion_map,   & ! Applies a map from previous conversion call
       cscl_verify,            & ! Verifies if matrix in CSC-lower form
@@ -23,54 +28,51 @@ module spral_matrix_util
       half_to_full,           & ! Expands a matrix from CSC-lower to CSC-full
       print_matrix              ! Pretty-print a CSC-lower matrix (summary)
 
-   integer, parameter :: wp = kind(0d0)
-   integer, parameter :: long = selected_int_kind(18)
-   real(wp), parameter :: zero = 0.0_wp
+   real(rp_), parameter :: zero = 0.0_rp_
 
    ! matrix types : real
-   integer, parameter :: SPRAL_MATRIX_UNSPECIFIED    =  0 ! undefined/unknown
-   integer, parameter :: SPRAL_MATRIX_REAL_RECT      =  1 ! real rectangular
-   integer, parameter :: SPRAL_MATRIX_REAL_UNSYM     =  2 ! real unsymmetric
-   integer, parameter :: SPRAL_MATRIX_REAL_SYM_PSDEF =  3 ! real symmetric
-      ! pos def
-   integer, parameter :: SPRAL_MATRIX_REAL_SYM_INDEF =  4 ! real symmetric indef
-   integer, parameter :: SPRAL_MATRIX_REAL_SKEW      =  6 ! real skew symmetric
[…]
