+++ src/ssids/ssids.F90
@@ -1,29 +1,34 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-12-23 AT 15:00 GMT.
+
+#include "spral_procedures.h"
+
 !> \file
 !> \copyright 2011-2016 The Science and Technology Facilities Council (STFC)
 !> \licence   BSD licence, see LICENCE file for details
 !> \author    Jonathan Hogg and Jennifer Scott
 !> \note      Originally based on HSL_MA97 v2.2.0
-module spral_ssids_double
+module spral_ssids_precision
 !$  use omp_lib
-  use, intrinsic :: iso_c_binding
-  use spral_hw_topology, only : guess_topology, numa_region
-  use spral_match_order, only : match_order_metis
-  use spral_matrix_util, only : SPRAL_MATRIX_REAL_SYM_INDEF, &
-                                SPRAL_MATRIX_REAL_SYM_PSDEF, &
-                                convert_coord_to_cscl, clean_cscl_oop, &
-                                apply_conversion_map
+  use spral_precision
+  use spral_hw_topology_precision, only : guess_topology, numa_region
+  use spral_match_order_precision, only : match_order_metis
+  use spral_matrix_util_precision, only : SPRAL_MATRIX_REAL_SYM_INDEF, &
+                                          SPRAL_MATRIX_REAL_SYM_PSDEF, &
+                                          convert_coord_to_cscl, &
+                                          clean_cscl_oop, &
+                                          apply_conversion_map
   use spral_metis_wrapper, only : metis_order
-  use spral_scaling, only : auction_scale_sym, equilib_scale_sym, &
-                            hungarian_scale_sym, &
-                            equilib_options, equilib_inform, &
-                            hungarian_options, hungarian_inform
-  use spral_ssids_anal, only : analyse_phase, check_order, expand_matrix, &
-                               expand_pattern
-  use spral_ssids_datatypes
-  use spral_ssids_akeep, only : ssids_akeep
-  use spral_ssids_fkeep, only : ssids_fkeep
-  use spral_ssids_inform, only : ssids_inform
-  use spral_rutherford_boeing, only : rb_write_options, rb_write
+  use spral_scaling_precision, only : auction_scale_sym, equilib_scale_sym, &
+                                      hungarian_scale_sym, &
+                                      equilib_options, equilib_inform, &
+                                      hungarian_options, hungarian_inform
+  use spral_ssids_anal_precision, only : analyse_phase, check_order, &
+                                         expand_matrix, expand_pattern
+  use spral_ssids_types_precision
+  use spral_ssids_akeep_precision, only : ssids_akeep
+  use spral_ssids_fkeep_precision, only : ssids_fkeep
+  use spral_ssids_inform_precision, only : ssids_inform
+  use spral_rutherford_boeing_precision, only : rb_write_options, rb_write
   implicit none
 
   private
@@ -44,43 +49,43 @@ module spral_ssids_double
   !> \brief Caches user OpenMP ICV values for later restoration
   type :: omp_settings
      logical :: nested, dynamic
-     integer :: max_active_levels
+     integer(ip_) :: max_active_levels
   end type omp_settings
 
   ! Make interfaces generic.
   interface ssids_analyse
-     module procedure analyse_double, analyse_double_ptr32
+     module procedure analyse_precision, analyse_precision_ptr32
   end interface ssids_analyse
 
   interface ssids_analyse_coord
-     module procedure ssids_analyse_coord_double
+     module procedure ssids_analyse_coord_precision
[…]

[…]
