+++ src/spral/blas_iface.F90
@@ -1,165 +1,0 @@
-! Define BLAS API in Fortran module
-module spral_blas_iface
-  implicit none
-
-  private
-  public :: daxpy, dcopy, ddot, dnrm2, dscal
-  public :: zaxpy, zcopy, zdotc, dznrm2, zscal
-  public :: dgemv, dtrsv
-  public :: dgemm, dsyrk, dtrsm
-  public :: zgemm, ztrsm
-
-  ! Level 1 BLAS
-  interface
-    subroutine daxpy( n, a, x, incx, y, incy )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx, incy
-      real(PRECISION), intent(in) :: a
-      real(PRECISION), intent(in   ), dimension(*) :: x
-      real(PRECISION), intent(inout), dimension(*) :: y
-    end subroutine daxpy
-    subroutine dcopy( n, x, incx, y, incy )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx, incy
-      real(PRECISION), intent(in ), dimension(*) :: x
-      real(PRECISION), intent(out), dimension(*) :: y
-    end subroutine dcopy
-    function ddot( n, x, incx, y, incy )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      real(PRECISION) :: ddot
-      integer, intent(in) :: n, incx, incy
-      real(PRECISION), intent(in), dimension(*) :: x
-      real(PRECISION), intent(in), dimension(*) :: y
-    end function ddot
-    function dnrm2( n, x, incx )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      real(PRECISION) :: dnrm2
-      integer, intent(in) :: n, incx
-      real(PRECISION), intent(in), dimension(*) :: x
-    end function dnrm2
-    subroutine dscal( n, a, x, incx )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx
-      real(PRECISION), intent(in) :: a
-      real(PRECISION), intent(in), dimension(*) :: x
-    end subroutine dscal
-    subroutine zcopy( n, x, incx, y, incy )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx, incy
-      complex(PRECISION), intent(in ), dimension(*) :: x
-      complex(PRECISION), intent(out), dimension(*) :: y
-    end subroutine zcopy
-    function dznrm2( n, x, incx )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      real(PRECISION) :: dznrm2
-      integer, intent(in) :: n, incx
-      complex(PRECISION), intent(in), dimension(*) :: x
-    end function dznrm2
-    function zdotc( n, x, incx, y, incy )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx, incy
-      complex(PRECISION), intent(in), dimension(*) :: x
-      complex(PRECISION), intent(in), dimension(*) :: y
-      complex(PRECISION) :: zdotc
-    end function zdotc
-    subroutine zscal( n, a, x, incx )
-      implicit none
-      integer, parameter :: PRECISION = kind(1.0D+0)
-      integer, intent(in) :: n, incx
-      complex(PRECISION), intent(in) :: a
-      complex(PRECISION), intent(in), dimension(*) :: x
-    end subroutine zscal
[…]

[…]
