+++ src/spral/metis5_wrapper.F90
@@ -1,3 +1,5 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-12-23 AT 09:45 GMT.
+
 ! COPYRIGHT (c) 2007-2013 Science & Technology Facilities Council
 ! Authors: Sue Thorne and Jonathan Hogg
 ! Origin: Heavily modified version of hsl_mc68
@@ -10,24 +12,25 @@
 #endif
 
 module spral_metis_wrapper
+   use spral_precision
    use, intrinsic :: iso_c_binding
    implicit none
 
    private
    public :: metis_order ! Calls metis on a symmetric matrix
 
-   integer, parameter :: long = C_INT64_T
+   integer(ip_), parameter :: long = C_INT64_T
 
 #if SPRAL_HAVE_METIS_H
 ! metis header is available, check for index types 
 #if SIZEOF_IDX_T == 8
-   integer, parameter :: metis_idx_t = c_int64_t
+   integer(ip_), parameter :: metis_idx_t = c_int64_t
 #else
-   integer, parameter :: metis_idx_t = c_int   
+   integer(ip_), parameter :: metis_idx_t = c_int   
 #endif
 #else
    ! metis header is not available, default to 32-bit index types 
-   integer, parameter :: metis_idx_t = c_int
+   integer(ip_), parameter :: metis_idx_t = c_int
 #endif
    
    ! We use the C interface to METIS via the following interoperable interfaces
@@ -50,8 +53,8 @@ module spral_metis_wrapper
 
    interface METIS_NodeND
       ! METIS_NodeND 32-bit integer interface
-      integer(c_int) function METIS_NodeND_32(nvtxs, xadj, adjncy, vwgt, options, &
-            perm, iperm) bind(C, name="METIS_NodeND")
+      integer(c_int) function METIS_NodeND_32(nvtxs, xadj, adjncy, vwgt,       &
+            options, perm, iperm) bind(C, name="METIS_NodeND")
          use iso_c_binding
          implicit none
          integer(c_int), intent(in) :: nvtxs
@@ -60,8 +63,8 @@ module spral_metis_wrapper
          integer(c_int), dimension(*), intent(out) :: perm, iperm
        end function METIS_NodeND_32
        ! METIS_NodeND 64-bit integer interface
-      integer(c_int) function METIS_NodeND_64(nvtxs, xadj, adjncy, vwgt, options, &
-            perm, iperm) bind(C, name="METIS_NodeND")
+      integer(c_int) function METIS_NodeND_64(nvtxs, xadj, adjncy, vwgt,       &
+            options, perm, iperm) bind(C, name="METIS_NodeND")
          use iso_c_binding
          implicit none
          integer(c_int64_t), intent(in) :: nvtxs
@@ -72,9 +75,9 @@ module spral_metis_wrapper
    end interface METIS_NodeND
    
    ! Following array size based on #define in metis.h
-   integer, parameter :: METIS_NOPTIONS = 40
+   integer(ip_), parameter :: METIS_NOPTIONS = 40
    ! Following are based on enum in metis.h, adjusted to Fortran indexing.
-   integer, parameter :: METIS_OPTION_CTYPE     =  3, &
+   integer(ip_), parameter :: METIS_OPTION_CTYPE     =  3, &
                          METIS_OPTION_RTYPE     =  5, &
                          METIS_OPTION_DBGLVL    =  6, &
                          METIS_OPTION_NITER     =  7, &
@@ -87,24 +90,26 @@ module spral_metis_wrapper
                          METIS_OPTION_UFACTOR   = 17, &
                          METIS_OPTION_NUMBERING = 18
    ! Following return codes based on enum in metis.h
[…]
