+++ src/spral/match_order.F90
@@ -1,3 +1,7 @@
+! THIS VERSION: GALAHAD 4.1 - 2022-12-23 AT 08:00 GMT.
+
+#include "spral_procedures.h"
+
 ! COPYRIGHT (c) 2012-3 Science and Technology Facilities Council
 ! Authors: Jonathan Hogg and Jennifer Scott
 ! Note: This code is a heavily modified version of HSL_MC80
@@ -8,31 +12,32 @@
 ! It optionally computes scaling factors.
 
 ! FIXME: At some stage replace call to mo_match() with call to
-! a higher level routine from spral_scaling instead (NB: have to cope with
-! fact we are currently expecting a full matrix, even if it means 2x more log
-! operations)
+! a higher level routine from spral_scaling_precision instead (NB: have to 
+! cope with fact we are currently expecting a full matrix, even if it means 
+! 2x more log operations)
 
-module spral_match_order
+module spral_match_order_precision
 
+  use spral_precision
   use spral_metis_wrapper, only : metis_order
-  use spral_scaling, only : hungarian_match
+  use spral_scaling_precision, only : hungarian_match
   implicit none
 
   private
   public :: match_order_metis ! Find a matching-based ordering using the
     ! Hungarian algorithm for matching and METIS for ordering.
 
-  integer, parameter :: wp = kind(0d0)
-  integer, parameter :: long = selected_int_kind(18)
+  integer(ip_), parameter :: wp = kind(0d0)
+  integer(ip_), parameter :: long = selected_int_kind(18)
 
   ! Error flags
-  integer, parameter :: SUCCESS               = 0
-  integer, parameter :: ERROR_ALLOCATION      = -1
-  integer, parameter :: ERROR_A_N_OOR         = -2
-  integer, parameter :: ERROR_UNKNOWN         = -99
+  integer(ip_), parameter :: SUCCESS               = 0
+  integer(ip_), parameter :: ERROR_ALLOCATION      = -1
+  integer(ip_), parameter :: ERROR_A_N_OOR         = -2
+  integer(ip_), parameter :: ERROR_UNKNOWN         = -99
 
   ! warning flags
-  integer, parameter :: WARNING_SINGULAR      = 1
+  integer(ip_), parameter :: WARNING_SINGULAR      = 1
 
   interface match_order_metis
      module procedure match_order_metis_ptr32, match_order_metis_ptr64
@@ -51,25 +56,26 @@ contains
   subroutine match_order_metis_ptr32(n, ptr, row, val, order, scale, &
        flag, stat)
     implicit none
-    integer, intent(in) :: n
-    integer, dimension(:), intent(in) :: ptr
-    integer, dimension(:), intent(in) :: row
-    real(wp), dimension(:), intent(in) :: val
-    integer, dimension(:), intent(out) :: order ! order(i)  holds the position
-      ! of variable i in the elimination order (pivot sequence). 
-    real(wp), dimension(n), intent(out) :: scale ! returns the mc64 symmetric
+    integer(ip_), intent(in) :: n
+    integer(ip_), dimension(:), intent(in) :: ptr
+    integer(ip_), dimension(:), intent(in) :: row
+    real(rp_), dimension(:), intent(in) :: val
+    integer(ip_), dimension(:), intent(out) :: order ! order(i)  holds the
+      ! position of variable i in the elimination order (pivot sequence). 
+    real(rp_), dimension(n), intent(out) :: scale ! returns the mc64 symmetric
       ! scaling
-    integer, intent(out) :: flag ! return value
-    integer, intent(out) :: stat ! stat value returned on failed allocation
+    integer(ip_), intent(out) :: flag ! return value
+    integer(ip_), intent(out) :: stat ! stat value returned on failed allocation
 
[…]
