### 2022-12-28 #4882 @nimgould
Thanks, Julian. As I and much of the rest of the world
are in hibernation until early January, don't ruin your
downtime with such trifles before then.

Happy new year to all

Nick
--
Nicholas I. M. Gould
Scientific Computing Department            email: ***@***.***
G39, R71 Rutherford Appleton Laboratory    phone: 0(44)1235 445801
Chilton, Oxon OX11 0QX England EU
www: http://www.numerical.rl.ac.uk/people/nimg

This email and any attachments are intended solely for the use of the named recipients. If you are not the intended recipient you must not use, disclose, copy or distribute this email or any of its attachments and should notify the sender immediately and delete this email from your system. UK Research and Innovation (UKRI) has taken every reasonable precaution to minimise risk of this email or any attachments containing viruses or malware but the recipient should carry out its own virus and malware checks before opening the attachments. UKRI does not accept any liability for any losses or damages which the recipient may sustain due to presence of any viruses.
