### 2022-12-29 6fedd8b
ampl, forthcoming and oblivion packages updated to new precision format
