+++ src/oblivion/funnel/funnel.1.F90
@@ -1,2249 +1,0 @@
-! THIS VERSION: GALAHAD 2.1 - 17/10/2007 AT 12:00 GMT.
-
-!-*-*-*-*-  G A L A H A D _ F U N N E L _ E Q U A L I T Y   M O D U L E -*-*-*-
-
-!  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
-!  Principal author: Nick Gould
-
-!  History -
-!   originally released GALAHAD Version 2.1. October 17th 2007
-
-!  For full documentation, see 
-!   http://galahad.rl.ac.uk/galahad-www/specs.html
-
-   MODULE GALAHAD_FUNNEL_equality_double
-
-!     ----------------------------------------------------------
-!    |                                                          |
-!    | FUNNEL, a filter trust-funnel method for                 |
-!    |  nonlinear optimization                                  |
-!    |                                                          |
-!    | Aim: to find a (local) minimizer of the nonlinear        |
-!    | programming problem                                      |
-!    |                                                          |
-!    |  minimize               f (x)                            |
-!    |  subject to          a_i^T x   = b_i      i in E_l       !
-!    |                       c_i (x)  =  0       i in E_g       |
-!    |                                                          |
-!     ----------------------------------------------------------
-
-     USE CUTEr_interface_double
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type
-     USE GALAHAD_LLS_double
-     USE GALAHAD_EQP_double
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_NORMS_double, ONLY: NORM => TWO_NORM
-
-     IMPLICIT NONE     
-
-     PRIVATE
-     PUBLIC :: FUNNEL_initialize, FUNNEL_read_specfile, FUNNEL_solve,          &
-               FUNNEL_terminate, NLPT_problem_type
-
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
-!  Set other parameters
-
-     INTEGER, PARAMETER :: wsout = 0
-!    INTEGER, PARAMETER :: wsout = 78
-
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: tenth = 0.1_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-     REAL ( KIND = wp ), PARAMETER :: tenm5 = 0.00001_wp
-     REAL ( KIND = wp ), PARAMETER :: point9 = 0.9_wp
-     REAL ( KIND = wp ), PARAMETER :: point1 = ten ** ( - 1 )
-     REAL ( KIND = wp ), PARAMETER :: point01 = ten ** ( - 2 )
-     REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 19
-     REAL ( KIND = wp ), PARAMETER :: epsmch = EPSILON( one )
-     REAL ( KIND = wp ), PARAMETER :: teneps = ten * epsmch
-     REAL ( KIND = wp ), PARAMETER :: mu_tiny = ten ** ( - 6 )
-     REAL ( KIND = wp ), PARAMETER :: y_tiny = ten ** ( - 6 )
-     REAL ( KIND = wp ), PARAMETER :: z_tiny = ten ** ( - 6 )
-
[…]
