+++ src/forthcoming/pdqp/pdqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 14/04/2021 AT 08:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-28 AT 14:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*- G A L A H A D _ P D Q P   M O D U L E -*-*-*-*-*-*-*-*-
 
@@ -13,7 +15,7 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_PDQP_double
+   MODULE GALAHAD_PDQP_precision
 
 !     -------------------------------------------
 !     | Solve the convex quadratic program      |
@@ -26,12 +28,13 @@
 !     | primal-dual active-set method           |
 !     -------------------------------------------
 
+      USE GALAHAD_PRECISION
       USE GALAHAD_SYMBOLS
-      USE GALAHAD_SPACE_double
-      USE GALAHAD_QPT_double
-      USE GALAHAD_SORT_double
-      USE GALAHAD_SPECFILE_double
-      USE GALAHAD_SBLS_double
+      USE GALAHAD_SPACE_precision
+      USE GALAHAD_QPT_precision
+      USE GALAHAD_SORT_precision
+      USE GALAHAD_SPECFILE_precision
+      USE GALAHAD_SBLS_precision
 
       IMPLICIT NONE
 
@@ -39,21 +42,15 @@
       PUBLIC :: PDQP_initialize, PDQP_read_specfile, PDQP_solve,               &
                 PDQP_terminate, QPT_problem_type
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-      INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-      REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-      REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-      REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-      REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-      REAL ( KIND = wp ), PARAMETER :: infinity = HUGE( one )
+      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: half = 0.5_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: ten = 10.0_rp_
+      REAL ( KIND = rp_ ), PARAMETER :: infinity = HUGE( one )
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -63,33 +60,33 @@
 
 !  unit for error messages
 
-        INTEGER :: error = 6
+        INTEGER ( KIND = ip_ ) :: error = 6
 
 !  unit for monitor output
 
-        INTEGER :: out = 6
+        INTEGER ( KIND = ip_ ) :: out = 6
 
 !  controls level of diagnostic output
 
-        INTEGER :: print_level = 0
+        INTEGER ( KIND = ip_ ) :: print_level = 0
 
 !   the maximum number of iterations permitted
 
-        INTEGER :: maxit = 100
+        INTEGER ( KIND = ip_ ) :: maxit = 100
 
[…]
