### 2022-12-29 6dbb2f6
ampl, forthcoming and oblivion packages updated to new precision format
