+++ src/oblivion/s2qp/s2qpt.F90
@@ -1,10 +1,13 @@
-! THIS VERSION: GALAHAD 3.3 - 30/05/2021 AT 08:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:20 GMT.
+#include "galahad_modules.h"
 PROGRAM S2QP_test !! far from comprehensive
-  USE GALAHAD_SMT_double
-  USE GALAHAD_NLPT_double
-  USE GALAHAD_S2QP_double
-  USE GALAHAD_SPACE_double
-  USE GALAHAD_SMT_double
+
+  USE GALAHAD_USERDATA_precision
+  USE GALAHAD_SMT_precision
+  USE GALAHAD_NLPT_precision
+  USE GALAHAD_S2QP_precision
+  USE GALAHAD_SPACE_precision
+  USE GALAHAD_SMT_precision
   USE GALAHAD_SYMBOLS
   IMPLICIT NONE
   !*************************************************!
@@ -21,26 +24,25 @@ PROGRAM S2QP_test !! far from comprehensive
   !                        -2 <=  x4 <= 2           !
   !                                                 !
   !*************************************************!
-  INTEGER, PARAMETER :: wp = KIND(  1.0D+0 )  ! Set precision.
-  REAL ( KIND = wp ), PARAMETER :: zero     = 0.0_wp
-  REAL ( KIND = wp ), PARAMETER :: one      = 1.0_wp
-  REAL ( KIND = wp ), PARAMETER :: two      = 2.0_wp
-  REAL ( KIND = wp ), PARAMETER :: three    = 3.0_wp
-  REAL ( KIND = wp ), PARAMETER :: four     = 4.0_wp
-  REAL ( KIND = wp ), PARAMETER :: ten      = 10.0_wp
-  REAL ( KIND = wp ), PARAMETER :: infinity = 1.0D+18
-  INTEGER, PARAMETER :: out   = 6
-  INTEGER, PARAMETER :: error = 6
-  INTEGER :: m, m_a, n, Ane, Jne, Hne, status, alloc_status
-  INTEGER :: num_recursive_calls, max_num_recursive_calls
+  REAL ( KIND = rp_ ), PARAMETER :: zero     = 0.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: one      = 1.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: two      = 2.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: three    = 3.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: four     = 4.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: ten      = 10.0_rp_
+  REAL ( KIND = rp_ ), PARAMETER :: infinity = 1.0D+18
+  INTEGER ( KIND = ip_ ), PARAMETER :: out   = 6
+  INTEGER ( KIND = ip_ ), PARAMETER :: error = 6
+  INTEGER ( KIND = ip_ ) :: m, m_a, n, Ane, Jne, Hne, status, alloc_status
+  INTEGER ( KIND = ip_ ) :: num_recursive_calls, max_num_recursive_calls
   LOGICAL :: is_specfile, print_debug
   CHARACTER ( LEN = 16 ) :: spec_name = "RUNS2QP.SPC"
-  INTEGER, PARAMETER :: spec_device = 60
+  INTEGER ( KIND = ip_ ), PARAMETER :: spec_device = 60
[…]
