+++ src/oblivion/trimsqp/trimsqp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 16:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D _ T R I M S Q P   M O D U L E  *-*-*-*-*-*-*-*
 
@@ -6,8 +8,7 @@
 !
 !  Author: Nick Gould and Daniel Robinson
 
-
-MODULE GALAHAD_TRIMSQP_double
+MODULE GALAHAD_TRIMSQP_precision
 
 !     ------------------------------------------------------------------
 !    |                                                                  |
@@ -25,82 +26,81 @@ MODULE GALAHAD_TRIMSQP_double
 !    |                                                                  |
 !     ------------------------------------------------------------------
 
-  USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type, NLPT_userdata_type, &
-                                 NLPT_write_problem
-  USE GALAHAD_LSQP_double
-  USE GALAHAD_QPC_double
-  USE GALAHAD_EQP_double
-  USE GALAHAD_QPT_double
-  USE GALAHAD_QPD_double, LSQP_data_type => QPD_data_type
-  USE GALAHAD_SMT_double
-  USE GALAHAD_SPECFILE_double
-  USE GALAHAD_SPACE_double
+  USE GALAHAD_USERDATA_precision
+  USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type, NLPT_write_problem
+  USE GALAHAD_LSQP_precision
+  USE GALAHAD_QPC_precision
+  USE GALAHAD_EQP_precision
+  USE GALAHAD_QPT_precision
+  USE GALAHAD_QPD_precision, LSQP_data_type => QPD_data_type
+  USE GALAHAD_SMT_precision
+  USE GALAHAD_SPECFILE_precision
+  USE GALAHAD_SPACE_precision
   USE GALAHAD_COPYRIGHT
   USE GAlAHAD_SYMBOLS
-  USE GALAHAD_NORMS_double
-  USE GALAHAD_CHECK_double
-  USE GALAHAD_SORT_double
-  USE GALAHAD_mop_double
+  USE GALAHAD_NORMS_precision
+  USE GALAHAD_CHECK_precision
+  USE GALAHAD_SORT_precision
+  USE GALAHAD_MOP_precision
 
   IMPLICIT NONE
 
   PRIVATE
-  PUBLIC :: TRIMSQP_initialize, TRIMSQP_read_specfile,   &
+  PUBLIC :: TRIMSQP_initialize, TRIMSQP_read_specfile,                         &
             TRIMSQP_solve, TRIMSQP_terminate
 
 !  Set precision
 
-  INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !  Set other parameters
 
-  REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-  REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-  REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-  REAL ( KIND = wp ), PARAMETER :: five = 5.0_wp
-  REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
[…]
