+++ src/oblivion/s2qp/uses2qp.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:20 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D   U S E _ S 2 Q P  -*-*-*-*-*-*-*-*-*-*-
 
@@ -6,22 +8,23 @@
 !  Copyright reserved
 !  Started: December 22th 2007
 
-   MODULE GALAHAD_USES2QP_double
+   MODULE GALAHAD_USES2QP_precision
 
 !  This is the driver program for running S2QP for a variety of computing
 !  systems. It opens and closes all the files, allocates arrays, reads and
 !  checks data, and calls the appropriate package.
 
+     USE GALAHAD_PRECISION
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_S2QP_double
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_S2QP_precision
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_COPYRIGHT
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type, NLPT_userdata_type
-     USE GALAHAD_SMT_double
-     USE GALAHAD_CUTEST_FUNCTIONS_double
-     USE GALAHAD_CHECK_double
-     USE CUTEST_interface_double
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type, NLPT_userdata_type
+     USE GALAHAD_SMT_precision
+     USE GALAHAD_CUTEST_FUNCTIONS_precision
+     USE GALAHAD_CHECK_precision
+     USE CUTEst_interface_precision
 
      IMPLICIT NONE
 
@@ -36,16 +39,12 @@
 
 !  Dummy argument
 
-     INTEGER, INTENT( IN ) :: input
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
-
-!  Set precision
-
-     INTEGER, PARAMETER :: wp    = KIND( 1.0D+0 )
-     REAL ( KIND = wp ) :: zero  = 0.0_wp
-     REAL ( KIND = wp ) :: one   = 1.0_wp
-     REAL ( KIND = wp ) :: two   = 2.0_wp
-     REAL ( KIND = wp ) :: three = 3.0_wp
+     REAL ( KIND = rp_ ) :: zero  = 0.0_rp_
+     REAL ( KIND = rp_ ) :: one   = 1.0_rp_
+     REAL ( KIND = rp_ ) :: two   = 2.0_rp_
+     REAL ( KIND = rp_ ) :: three = 3.0_rp_
 
 !-------------------------------
 !   D e r i v e d   T y p e s
@@ -68,26 +67,26 @@
 
 !  Problem input characteristics
 
-     INTEGER :: iores, i
+     INTEGER ( KIND = ip_ ) :: iores, i
      LOGICAL :: filexx, is_specfile
-     REAL( kind = wp ) :: dummy_real
-     REAL( kind = wp ), dimension(:), allocatable :: X_saved, Y_saved
+     REAL( kind = rp_ ) :: dummy_real
+     REAL( kind = rp_ ), dimension(:), allocatable :: X_saved, Y_saved
 
 !  Specfile characteristics
 
-     INTEGER, PARAMETER :: input_specfile = 34
-     INTEGER, PARAMETER :: lspec = 29
[…]
