+++ src/oblivion/pqp/runpqp_sif.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.1 - 22/03/2007 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:00 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D   R U N P Q P _ S I F  *-*-*-*-*-*-
 
@@ -7,14 +9,15 @@
 !  September 14th 2004
 
    PROGRAM RUNPQP_SIF
-   USE GALAHAD_USEPQP_double
+   USE GALAHAD_PRECISION
+   USE GALAHAD_USEPQP_precision
 
 !  Main program for the SIF/CUTEr interface to QPA, a working-set 
 !  algorithm for solving quadratic programs
 
 !  Problem input characteristics
 
-   INTEGER, PARAMETER :: input = 55
+   INTEGER ( KIND = ip_ ), PARAMETER :: input = 55
    CHARACTER ( LEN = 16 ) :: prbdat = 'OUTSDIF.d'
 
 !  Open the data input file
