+++ src/dum/mi26d.f
@@ -1,0 +1,3 @@
+! THIS VERSION: 2022-12-29 AT 10:00:00 GMT.
+
+!-*-*-*-*-  G A L A H A D  -  D U M M Y   M I 2 6   M O D U L E  -*-*-*-
