### 2022-12-29 f7fd480
ampl, forthcoming and oblivion packages updated to new precision format
