+++ src/oblivion/funnel/usefunnel.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 2.1 - 17/10/2007 AT 12:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 12:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-*-*-*-  G A L A H A D   U S E _ F U N N E L  -*-*-*-*-*-*-*-*-*-*-
 
@@ -6,18 +8,19 @@
 !  Copyright reserved
 !  Started: October 17th 2007
 
-   MODULE GALAHAD_USEFUNNEL_double
+   MODULE GALAHAD_USEFUNNEL_precision
 
 !  This is the driver program for running FUNNEL for a variety of computing
 !  systems. It opens and closes all the files, allocate arrays, reads and
 !  checks data, and calls the appropriate minimizers
 
-     USE GALAHAD_FUNNEL_double
-     USE GALAHAD_FUNNEL_equality_double
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_PRECISION
+     USE GALAHAD_FUNNEL_precision
+     USE GALAHAD_FUNNEL_equality_precision
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_COPYRIGHT
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_CUTEST_FUNCTIONS_double
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_CUTEST_FUNCTIONS_precision
      USE GALAHAD_SYMBOLS
      IMPLICIT NONE
 
@@ -32,11 +35,10 @@
 
 !  Dummy argument
 
-     INTEGER, INTENT( IN ) :: input
+     INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
 !  Set precision
 
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
 
 !-------------------------------
 !   D e r i v e d   T y p e s
@@ -56,24 +58,24 @@
 
 !  Problem input characteristics
 
-     INTEGER :: iores, i
+     INTEGER ( KIND = ip_ ) :: iores, i
      LOGICAL :: filexx, is_specfile
 
 !  Specfile characteristics
 
-     INTEGER, PARAMETER :: input_specfile = 34
-     INTEGER, PARAMETER :: lspec = 31
+     INTEGER ( KIND = ip_ ), PARAMETER :: input_specfile = 34
+     INTEGER ( KIND = ip_ ), PARAMETER :: lspec = 31
      CHARACTER ( LEN = 16 ) :: specname = 'RUNFUNNEL'
      TYPE ( SPECFILE_item_type ), DIMENSION( lspec ) :: spec
      CHARACTER ( LEN = 16 ) :: runspec = 'RUNFUNNEL.SPC'
 
 !  Default values for specfile-defined parameters
 
-     INTEGER :: dfiledevice = 26
-     INTEGER :: rfiledevice = 47
-     INTEGER :: sfiledevice = 62
-     INTEGER :: vfiledevice = 63
-     INTEGER :: wfiledevice = 59
+     INTEGER ( KIND = ip_ ) :: dfiledevice = 26
+     INTEGER ( KIND = ip_ ) :: rfiledevice = 47
+     INTEGER ( KIND = ip_ ) :: sfiledevice = 62
+     INTEGER ( KIND = ip_ ) :: vfiledevice = 63
+     INTEGER ( KIND = ip_ ) :: wfiledevice = 59
      LOGICAL :: write_problem_data    = .FALSE.
      LOGICAL :: write_solution        = .FALSE.
[…]
