+++ src/oblivion/fastr/fastr.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 27/01/2020 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 11:30 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D _ F A S T R   M O D U L E  *-*-*-*-*-*-*-*
 
@@ -12,7 +14,7 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_FASTR_double
+   MODULE GALAHAD_FASTR_precision
 
 !     ----------------------------------------------------------
 !    |                                                          |
@@ -31,60 +33,54 @@
 !    |                                                          |
 !     ----------------------------------------------------------
 
+     USE GALAHAD_USERDATA_precision
      USE GALAHAD_CLOCK
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_NLPT_double, ONLY: NLPT_problem_type, NLPT_userdata_type
-     USE GALAHAD_CQP_double
-     USE GALAHAD_EQP_double
-     USE GALAHAD_SPECFILE_double
-     USE GALAHAD_SPACE_double
+     USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type
+     USE GALAHAD_CQP_precision
+     USE GALAHAD_EQP_precision
+     USE GALAHAD_SPECFILE_precision
+     USE GALAHAD_SPACE_precision
      USE GALAHAD_STRING
-     USE GALAHAD_FILTER_double
-     USE GALAHAD_OPT_double
-     USE GALAHAD_NORMS_double, ONLY: TWO_norm, INFINITY_norm
+     USE GALAHAD_FILTER_precision
+     USE GALAHAD_OPT_precision
+     USE GALAHAD_NORMS_precision, ONLY: TWO_norm, INFINITY_norm
 !PLPLOT USE PLPLOT
 
      IMPLICIT NONE     
 
      PRIVATE
      PUBLIC :: FASTR_initialize, FASTR_read_specfile, FASTR_solve,             &
-               FASTR_terminate, NLPT_problem_type, NLPT_userdata_type,         &
+               FASTR_terminate, NLPT_problem_type, GALAHAD_userdata_type,      &
                SMT_type, SMT_put
-
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-     INTEGER, PARAMETER :: long = SELECTED_INT_KIND( 18 )
 
 !----------------------
 !   P a r a m e t e r s
 !----------------------
 
-     INTEGER, PARAMETER :: wsout = 0
-!    INTEGER, PARAMETER :: wsout = 78
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: tenth = 0.1_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-     REAL ( KIND = wp ), PARAMETER :: hundred = 100.0_wp
-     REAL ( KIND = wp ), PARAMETER :: point9 = 0.9_wp
[…]

[…]
