+++ src/tru/trus2.f90
@@ -1,4 +1,4 @@
-   PROGRAM GALAHAD_TRU2_EXAMPLE    !  GALAHAD 2.3 - 24/07/2008 AT 07:15 GMT
+   PROGRAM GALAHAD_TRU2_EXAMPLE    !  GALAHAD 4.1 - 2022-12-29 AT 11:15 GMT
    USE GALAHAD_TRU_double                       ! double precision version
    IMPLICIT NONE
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )    ! set precision
@@ -6,7 +6,7 @@
    TYPE ( TRU_control_type ) :: control
    TYPE ( TRU_inform_type ) :: inform
    TYPE ( TRU_data_type ) :: data
-   TYPE ( NLPT_userdata_type ) :: userdata
+   TYPE ( GALAHAD_userdata_type ) :: userdata
    EXTERNAL :: FUN, GRAD, HESSPROD
    INTEGER, PARAMETER :: n = 3, h_ne = 5
    REAL ( KIND = wp ), PARAMETER :: p = 4.0_wp
@@ -36,12 +36,12 @@
    END PROGRAM GALAHAD_TRU2_EXAMPLE
 
    SUBROUTINE FUN( status, X, userdata, f )     ! Objective function
-   USE GALAHAD_NLPT_double, ONLY: NLPT_userdata_type
+   USE GALAHAD_USERDATA_double, ONLY: GALAHAD_userdata_type
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
    INTEGER, INTENT( OUT ) :: status
    REAL ( KIND = wp ), INTENT( OUT ) :: f
    REAL ( KIND = wp ), DIMENSION( : ),INTENT( IN ) :: X
-   TYPE ( NLPT_userdata_type ), INTENT( INOUT ) :: userdata
+   TYPE ( GALAHAD_userdata_type ), INTENT( INOUT ) :: userdata
    f = ( X( 1 ) + X( 3 ) + userdata%real( 1 ) ) ** 2 +                         &
        ( X( 2 ) + X( 3 ) ) ** 2 + COS( X( 1 ) )
    status = 0
@@ -49,12 +49,12 @@
    END SUBROUTINE FUN
 
    SUBROUTINE GRAD( status, X, userdata, G )    ! gradient of the objective
-   USE GALAHAD_NLPT_double, ONLY: NLPT_userdata_type
+   USE GALAHAD_USERDATA_double, ONLY: GALAHAD_userdata_type
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
    INTEGER, INTENT( OUT ) :: status
    REAL ( KIND = wp ), DIMENSION( : ), INTENT( IN ) :: X
    REAL ( KIND = wp ), DIMENSION( : ), INTENT( OUT ) :: G
-   TYPE ( NLPT_userdata_type ), INTENT( INOUT ) :: userdata
+   TYPE ( GALAHAD_userdata_type ), INTENT( INOUT ) :: userdata
    G( 1 ) = 2.0_wp * ( X( 1 ) + X( 3 ) + userdata%real( 1 ) ) - SIN( X( 1 ) )
    G( 2 ) = 2.0_wp * ( X( 2 ) + X( 3 ) )
    G( 3 ) = 2.0_wp * ( X( 1 ) + X( 3 ) + userdata%real( 1 ) ) +                &
@@ -64,14 +64,14 @@
    END SUBROUTINE GRAD
 
[…]
