+++ src/bgo/bgot.F90
@@ -55,6 +55,8 @@
    CALL BGO_terminate( data, control, inform )  ! delete internal workspace
    DEALLOCATE( nlp%X, nlp%G, nlp%X_l, nlp%X_u )
    DEALLOCATE( nlp%H%row, nlp%H%col, nlp%H%val, nlp%H%type, userdata%real )
+   WRITE( 6, "( /, ' tests completed' )" )
+
    END PROGRAM GALAHAD_BGO_TEST
 
    SUBROUTINE FUN( status, X, userdata, f )     ! Objective function
