+++ src/oblivion/superb/superb.F90
@@ -1,4 +1,6 @@
-! THIS VERSION: GALAHAD 3.3 - 27/01/2020 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:40 GMT.
+
+#include "galahad_modules.h"
 
 !-*-*-*-*-*-*-  G A L A H A D _ S U P E R B   M O D U L E  *-*-*-*-*-*-*-*
 
@@ -12,7 +14,7 @@
 !  For full documentation, see 
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-   MODULE GALAHAD_SUPERB_double
+   MODULE GALAHAD_SUPERB_precision
 
 !     ----------------------------------------------------------
 !    |                                                          |
@@ -32,18 +34,19 @@
 !    |                                                          |
 !     ----------------------------------------------------------
 
-     USE CUTEst_interface_double
-     USE GALAHAD_NORMS_double
+     USE GALAHAD_PRECISION
+     USE CUTEst_interface_precision
+     USE GALAHAD_NORMS_precision
      USE GALAHAD_SYMBOLS
-     USE GALAHAD_SPACE_double
-     USE GALAHAD_SILS_double
-!    USE GALAHAD_QPT_double
-     USE GALAHAD_WCP_double
-     USE GALAHAD_SPECFILE_double
+     USE GALAHAD_SPACE_precision
+     USE GALAHAD_SILS_precision
+!    USE GALAHAD_QPT_precision
+     USE GALAHAD_WCP_precision
+     USE GALAHAD_SPECFILE_precision
      USE GALAHAD_STRING, ONLY: STRING_pleural
-     USE GALAHAD_SORT_double
-     USE GALAHAD_GLTR_double
-     USE GALAHAD_PTRANS_double
+     USE GALAHAD_SORT_precision
+     USE GALAHAD_GLTR_precision
+     USE GALAHAD_PTRANS_precision
 
      IMPLICIT NONE     
 
@@ -51,40 +54,36 @@
      PUBLIC :: SUPERB_initialize, SUPERB_read_specfile, SUPERB_solve,          &
                SUPERB_terminate
 
-!  Set precision
-
-     INTEGER, PARAMETER :: wp = KIND( 1.0D+0 )
-
 !  Set other parameters
 
-     REAL ( KIND = wp ), PARAMETER :: zero = 0.0_wp
-     REAL ( KIND = wp ), PARAMETER :: one = 1.0_wp
-     REAL ( KIND = wp ), PARAMETER :: half = 0.5_wp
-     REAL ( KIND = wp ), PARAMETER :: two = 2.0_wp
-     REAL ( KIND = wp ), PARAMETER :: three = 3.0_wp
-     REAL ( KIND = wp ), PARAMETER :: point9 = 0.9_wp
-     REAL ( KIND = wp ), PARAMETER :: point99 = 0.99_wp
-     REAL ( KIND = wp ), PARAMETER :: point1 = 0.1_wp
-     REAL ( KIND = wp ), PARAMETER :: point01 = 0.01_wp
-     REAL ( KIND = wp ), PARAMETER :: tenm5 = 0.00001_wp
-     REAL ( KIND = wp ), PARAMETER :: ten = 10.0_wp
-     REAL ( KIND = wp ), PARAMETER :: hundred = 100.0_wp
-     REAL ( KIND = wp ), PARAMETER :: infinity = ten ** 19
[…]
