+++ src/dum/hsl_mi32d.f90
@@ -1,0 +1,6 @@
+! THIS VERSION: 2022-12-29 AT 10:00:00 GMT.
+
+!-*-*-*-*-  G A L A H A D  -  D U M M Y   M I 3 2   M O D U L E  -*-*-*-
+
+   module hsl_mi32_double
+   end module hsl_mi32_double

[…]
