+++ src/wcp/wcpti.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 10:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_WCP_interface_test
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_WCP_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
