+++ src/lsp/lsp.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-22 AT 11:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -38,7 +38,7 @@
 !     |                                                            |
 !     --------------------------------------------------------------
 
-      USE GALAHAD_PRECISION
+      USE GALAHAD_KINDS
       USE GALAHAD_SYMBOLS
       USE GALAHAD_SMT_precision
       USE GALAHAD_QPT_precision
