+++ src/bqp/inbqp.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 11:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -22,7 +22,7 @@
 !    | bound-constrained convex quadratic programming  |
 !    ---------------------------------------------------
 
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_QPT_precision
    USE GALAHAD_RPD_precision
    USE GALAHAD_SMT_precision, only: SMT_put
