### 2022-12-30 48c3af3
change name of global precision files to galahad_kinds and spral_kinds updates to ampl make file
