+++ src/oblivion/superb/superb.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -34,8 +34,8 @@
 !    |                                                          |
 !     ----------------------------------------------------------
 
-     USE GALAHAD_PRECISION
-     USE CUTEst_interface_precision
+     USE GALAHAD_KINDS
+     USE CUTEST_INTERFACE_precision
      USE GALAHAD_NORMS_precision
      USE GALAHAD_SYMBOLS
      USE GALAHAD_SPACE_precision
