+++ src/qpa/useqpa.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-20 AT 14:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -21,8 +21,8 @@
 !    ! algorithm for quadratic programming          |
 !    ------------------------------------------------
 
-      USE GALAHAD_PRECISION
-      USE CUTEst_interface_precision
+      USE GALAHAD_KINDS
+      USE CUTEST_INTERFACE_precision
       USE GALAHAD_CLOCK
       USE GALAHAD_SYMBOLS
       USE GALAHAD_QPT_precision
