+++ src/dum/pastixf_interfaces.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-10-26 AT 16:25 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 ! G A L A H A D  -  D U M M Y   P A S T I X F _ I N T E R F A C E S  M O D U L E
 
@@ -20,7 +20,7 @@
 
    INTERFACE pastixInitParam
      SUBROUTINE pastixInitParam_f08( iparm, dparm )
-       USE GALAHAD_PRECISION, ONLY : rpc_
+       USE GALAHAD_KINDS, ONLY : rpc_
        USE pastixf_enums, ONLY : pastix_int_t
        IMPLICIT NONE
        INTEGER ( KIND = pastix_int_t ), INTENT( INOUT ), target :: iparm( : )
@@ -30,7 +30,7 @@
 
    INTERFACE pastixInit
      SUBROUTINE pastixInit_f08( pastix_data, pastix_comm, iparm, dparm )
-       USE GALAHAD_PRECISION, ONLY : rpc_
+       USE GALAHAD_KINDS, ONLY : rpc_
        USE spmf_enums, ONLY : MPI_Comm
        USE pastixf_enums, ONLY : pastix_data_t, pastix_int_t
        IMPLICIT NONE
@@ -43,7 +43,7 @@
 
    INTERFACE pastix_task_analyze
      SUBROUTINE pastix_task_analyze_f08( pastix_data, spm, info )
-       USE GALAHAD_PRECISION, ONLY : ipc_
+       USE GALAHAD_KINDS, ONLY : ipc_
        USE spmf_enums, ONLY : spmatrix_t
        USE pastixf_enums, ONLY : pastix_data_t
        IMPLICIT NONE
@@ -55,7 +55,7 @@
 
    INTERFACE pastix_task_numfact
      SUBROUTINE pastix_task_numfact_f08( pastix_data, spm, info )
-       USE GALAHAD_PRECISION, ONLY : ipc_
+       USE GALAHAD_KINDS, ONLY : ipc_
        USE spmf_enums, ONLY : spmatrix_t
        USE pastixf_enums, ONLY : pastix_data_t
        IMPLICIT NONE
@@ -67,7 +67,7 @@
 
    INTERFACE pastix_task_solve
      SUBROUTINE pastix_task_solve_f08( pastix_data, nrhs, B, ldb, info )
-       USE GALAHAD_PRECISION, ONLY : ipc_
+       USE GALAHAD_KINDS, ONLY : ipc_
        USE pastixf_enums, ONLY : pastix_data_t, pastix_int_t
        IMPLICIT NONE
        TYPE ( pastix_data_t ), INTENT( INOUT ), TARGET :: pastix_data
@@ -81,7 +81,7 @@
    INTERFACE pastix_task_refine
      SUBROUTINE pastix_task_refine_f08( pastix_data, n, nrhs, B, ldb, X, ldx,  &
                                         info )
-       USE GALAHAD_PRECISION, ONLY : ipc_
+       USE GALAHAD_KINDS, ONLY : ipc_
        USE pastixf_enums, ONLY : pastix_data_t, pastix_int_t
        IMPLICIT NONE
        TYPE ( pastix_data_t ), INTENT( INOUT ), TARGET   :: pastix_data
