+++ src/lancelot/cg.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -10,7 +10,7 @@
 
    MODULE LANCELOT_CG_precision
             
-      USE GALAHAD_PRECISION
+      USE GALAHAD_KINDS
 
       IMPLICIT NONE
 

[…]
