### 2022-12-30 df52ebb
change name of global precision files to galahad_kinds and spral_kinds updates to ampl make file
