+++ src/forthcoming/barc/runbarc_sif.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-28 AT 11:15 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -9,7 +9,7 @@
 !  February 7th 2008
 
    PROGRAM RUNBARC_SIF
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_USEBARC_precision
 
 !  Main program for the SIF interface to BARC, an adaptive cubic overestimation
