+++ src/clls/cllst.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 10:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_CLLS_EXAMPLE
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_CLLS_precision
    USE GALAHAD_SYMBOLS
    USE GALAHAD_LMS_precision
