+++ src/rqs/rqst.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 14:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_RQS_test_deck
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_RQS_precision
    USE GALAHAD_SYMBOLS
    IMPLICIT NONE
