+++ src/scale/scale_test.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-17 AT 09:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_SCALE_TEST
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_SCALE_precision
    USE GALAHAD_SMT_precision
    USE GALAHAD_SYMBOLS
