+++ src/rpd/rpd.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-22 AT 10:50 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -56,7 +56,7 @@
 !     |                                                  |
 !      --------------------------------------------------
 
-      USE GALAHAD_PRECISION
+      USE GALAHAD_KINDS
       USE GALAHAD_SYMBOLS
       USE GALAHAD_SMT_precision, ONLY: SMT_put
       USE GALAHAD_QPT_precision
