+++ src/filtrane/filtrane.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-19 AT 15:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -349,7 +349,7 @@
 
    MODULE GALAHAD_FILTRANE_precision
             
-      USE GALAHAD_PRECISION
+      USE GALAHAD_KINDS
 
 !-------------------------------------------------------------------------------
 !   U s e d   m o d u l e s   a n d   s y m b o l s
