+++ src/ccqp/ccqpt.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 11:55 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_CCQP_EXAMPLE
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_CCQP_precision
    USE GALAHAD_SYMBOLS
    USE GALAHAD_LMS_precision
