+++ src/rand/randt.F90
@@ -1,7 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-16 AT 10:30 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 #include "galahad_modules.h"
    PROGRAM GALAHAD_RAND_test
-   USE GALAHAD_PRECISION
+   USE GALAHAD_KINDS
    USE GALAHAD_RAND_precision
    IMPLICIT NONE
    TYPE (RAND_seed) seed
