+++ src/oblivion/sqp/usesqp.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-29 AT 15:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
@@ -10,7 +10,7 @@
 
    MODULE GALAHAD_USESQP_precision
 
-     USE GALAHAD_PRECISION
+     USE GALAHAD_KINDS
      USE GALAHAD_CUTEST_FUNCTIONS_precision
      USE GALAHAD_COPYRIGHT
      USE GALAHAD_SMT_precision
