### 2022-12-30 08c3d26
change name of global precision files to galahad_kinds and spral_kinds updates to ampl make file
