+++ src/external/wsmp/wsmpt.F90
@@ -1,10 +1,10 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-21 AT 10:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
 
 #include "galahad_modules.h"
 
 PROGRAM TEST_WSMP
 
-  USE GALAHAD_PRECISION
+  USE GALAHAD_KINDS
   USE GALAHAD_SYMBOLS
   IMPLICIT NONE
 
