+++ src/fdc/C/fdc_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-09-28 AT 16:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:15 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  F D C    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_FDC_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_FDC_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_FDC_double, ONLY:                                              &
+    USE GALAHAD_FDC_precision, ONLY:                                           &
         f_fdc_control_type        => FDC_control_type,                         &
         f_fdc_time_type           => FDC_time_type,                            &
         f_fdc_inform_type         => FDC_inform_type,                          &
@@ -24,7 +27,7 @@
         f_fdc_find_dependent_rows => FDC_find_dependent_rows,                  &
         f_fdc_terminate           => FDC_terminate
 
-    USE GALAHAD_SLS_double_ciface, ONLY:                                       &
+    USE GALAHAD_SLS_precision_ciface, ONLY:                                    &
         sls_inform_type,                                                       &
         sls_control_type,                                                      &
         copy_sls_inform_in   => copy_inform_in,                                &
@@ -32,7 +35,7 @@
         copy_sls_control_in  => copy_control_in,                               &
         copy_sls_control_out => copy_control_out
 
-    USE GALAHAD_ULS_double_ciface, ONLY:                                       &
+    USE GALAHAD_ULS_precision_ciface, ONLY:                                    &
         uls_inform_type,                                                       &
         uls_control_type,                                                      &
         copy_uls_inform_in   => copy_inform_in,                                &
@@ -42,27 +45,20 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: fdc_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: out
-      INTEGER ( KIND = C_INT ) :: print_level
[…]
