+++ src/uls/C/uls_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-20 AT 16:05 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 10:25 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  U L S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_ULS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_ULS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_ULS_double, ONLY:                                              &
+    USE GALAHAD_ULS_precision, ONLY:                                           &
         f_uls_control_type => ULS_control_type,                                &
         f_uls_inform_type => ULS_inform_type,                                  &
         f_uls_full_data_type => ULS_full_data_type,                            &
@@ -26,7 +29,7 @@
         f_uls_information => ULS_information,                                  &
         f_uls_terminate => ULS_terminate
 
-    USE GALAHAD_GLS_double_ciface, ONLY:                                       &
+    USE GALAHAD_GLS_precision_ciface, ONLY:                                    &
         gls_control,                                                           &
         gls_ainfo,                                                             &
         gls_finfo,                                                             &
@@ -38,7 +41,7 @@
         copy_gls_finfo_out => copy_finfo_out,                                  &
         copy_gls_sinfo_out => copy_sinfo_out
 
-    USE HSL_MA48_double_ciface, ONLY:                                          &
+    USE HSL_MA48_precision_ciface, ONLY:                                       &
         ma48_control,                                                          &
         ma48_ainfo,                                                            &
         ma48_finfo,                                                            &
@@ -49,61 +52,54 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: uls_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: warning
-      INTEGER ( KIND = C_INT ) :: out
[…]
