+++ src/convert/C/convert_ciface.F90
@@ -1,6 +1,9 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-02-25 AT 07:13 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 07:40 GMT.
 
-!-*-*-*-*-*-*-*-  G A L A H A D _  C O N V E R T    C   I N T E R F A C E  -*-*-*-*-*-
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
+
+!-*-*-*-*-  G A L A H A D _  C O N V E R T    C   I N T E R F A C E  -*-*-*-*-
 
 !  Copyright reserved, Gould/Orban/Toint, for GALAHAD productions
 !  Principal authors: Jaroslav Fowkes & Nick Gould
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_CONVERT_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_CONVERT_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_CONVERT_double, ONLY:                                          &
+    USE GALAHAD_CONVERT_precision, ONLY:                                       &
         f_convert_control_type   => CONVERT_control_type,                      &
         f_convert_time_type      => CONVERT_time_type,                         &
         f_convert_inform_type    => CONVERT_inform_type,                       &
@@ -24,16 +27,9 @@
 !       f_convert_import         => CONVERT_import,                            &
 !       f_convert_reset_control  => CONVERT_reset_control,                     &
 !       f_convert_information    => CONVERT_information,                       &
-!       f_convert_terminate => CONVERT_terminate
+!       f_convert_terminate      => CONVERT_terminate
 
     IMPLICIT NONE
-
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -41,9 +37,9 @@
 
     TYPE, BIND( C ) :: convert_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: out
-      INTEGER ( KIND = C_INT ) :: print_level
+      INTEGER ( KIND = ipc_ ) :: error
+      INTEGER ( KIND = ipc_ ) :: out
+      INTEGER ( KIND = ipc_ ) :: print_level
       LOGICAL ( KIND = C_BOOL ) :: transpose
       LOGICAL ( KIND = C_BOOL ) :: sum_duplicates
       LOGICAL ( KIND = C_BOOL ) :: order
@@ -53,14 +49,14 @@
     END TYPE convert_control_type
 
     TYPE, BIND( C ) :: convert_time_type
-      REAL ( KIND = wp ) :: total
-      REAL ( KIND = wp ) :: clock_total
+      REAL ( KIND = rp_ ) :: total
+      REAL ( KIND = rp_ ) :: clock_total
     END TYPE convert_time_type
 
     TYPE, BIND( C ) :: convert_inform_type
-      INTEGER ( KIND = C_INT ) :: status
[…]

[…]
