+++ src/ccqp/C/ccqp_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-20 AT 16:05 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 07:40 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  C C Q P    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_CCQP_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_CCQP_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_CCQP_double, ONLY:                                             &
+    USE GALAHAD_CCQP_precision, ONLY:                                          &
         f_ccqp_control_type   => CCQP_control_type,                            &
         f_ccqp_time_type      => CCQP_time_type,                               &
         f_ccqp_inform_type    => CCQP_inform_type,                             &
@@ -28,7 +31,7 @@
         f_ccqp_information    => CCQP_information,                             &
         f_ccqp_terminate      => CCQP_terminate
 
-    USE GALAHAD_SBLS_double_ciface, ONLY:                                      &
+    USE GALAHAD_SBLS_precision_ciface, ONLY:                                   &
         sbls_inform_type,                                                      &
         sbls_control_type,                                                     &
         copy_sbls_inform_in   => copy_inform_in,                               &
@@ -36,7 +39,7 @@
         copy_sbls_control_in  => copy_control_in,                              &
         copy_sbls_control_out => copy_control_out
 
-    USE GALAHAD_FDC_double_ciface, ONLY:                                       &
+    USE GALAHAD_FDC_precision_ciface, ONLY:                                    &
         fdc_inform_type,                                                       &
         fdc_control_type,                                                      &
         copy_fdc_inform_in   => copy_inform_in,                                &
@@ -44,7 +47,7 @@
         copy_fdc_control_in  => copy_control_in,                               &
         copy_fdc_control_out => copy_control_out
 
-    USE GALAHAD_FIT_double_ciface, ONLY:                                       &
+    USE GALAHAD_FIT_precision_ciface, ONLY:                                    &
         fit_inform_type,                                                       &
         fit_control_type,                                                      &
         copy_fit_inform_in   => copy_inform_in,                                &
@@ -52,7 +55,7 @@
[…]
