+++ src/dps/C/dps_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-09-29 AT 16:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 07:45 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  D P S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_DPS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_DPS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_DPS_double, ONLY:                                              &
+    USE GALAHAD_DPS_precision, ONLY:                                           &
         f_dps_control_type       => DPS_control_type,                          &
         f_dps_time_type          => DPS_time_type,                             &
         f_dps_inform_type        => DPS_inform_type,                           &
@@ -30,7 +33,7 @@
         f_dps_information        => DPS_information,                           &
         f_dps_terminate          => DPS_terminate
 
-    USE GALAHAD_SLS_double_ciface, ONLY:                                       &
+    USE GALAHAD_SLS_precision_ciface, ONLY:                                    &
         sls_inform_type,                                                       &
         sls_control_type,                                                      &
         copy_sls_inform_in   => copy_inform_in,                                &
@@ -40,30 +43,23 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: dps_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: out
-      INTEGER ( KIND = C_INT ) :: problem
-      INTEGER ( KIND = C_INT ) :: print_level
-      INTEGER ( KIND = C_INT ) :: new_h
-      INTEGER ( KIND = C_INT ) :: taylor_max_degree
-      REAL ( KIND = wp ) :: eigen_min
-      REAL ( KIND = wp ) :: lower
-      REAL ( KIND = wp ) :: upper
-      REAL ( KIND = wp ) :: stop_normal
-      REAL ( KIND = wp ) :: stop_absolute_normal
+      INTEGER ( KIND = ipc_ ) :: error
+      INTEGER ( KIND = ipc_ ) :: out
[…]

[…]
