+++ src/trb/C/trb_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-21 AT 13:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 10:05 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-*-  G A L A H A D _ T R B   C   I N T E R F A C E  -*-*-*-*-*-*-
 
@@ -13,10 +16,10 @@
 
 !  C interface module to GALAHAD_TRB types and interfaces
 
-  MODULE GALAHAD_TRB_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_TRB_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_TRB_double, only:                                              &
+    USE GALAHAD_TRB_precision, ONLY:                                           &
         f_trb_time_type                 => TRB_time_type,                      &
         f_trb_inform_type               => TRB_inform_type,                    &
         f_trb_control_type              => TRB_control_type,                   &
@@ -32,10 +35,10 @@
         f_trb_information               => TRB_information,                    &
         f_trb_terminate                 => TRB_terminate
 
-    USE GALAHAD_NLPT_double, ONLY:                                             &
-        f_nlpt_userdata_type => NLPT_userdata_type
+    USE GALAHAD_USERDATA_precision, ONLY:                                      &
+        f_galahad_userdata_type => GALAHAD_userdata_type
 
-    USE GALAHAD_TRS_double_ciface, ONLY:                                       &
+    USE GALAHAD_TRS_precision_ciface, ONLY:                                    &
         trs_inform_type,                                                       &
         trs_control_type,                                                      &
         copy_trs_inform_in   => copy_inform_in,                                &
@@ -43,7 +46,7 @@
         copy_trs_control_in  => copy_control_in,                               &
         copy_trs_control_out => copy_control_out
 
-    USE GALAHAD_GLTR_double_ciface, ONLY:                                      &
+    USE GALAHAD_GLTR_precision_ciface, ONLY:                                   &
         gltr_inform_type,                                                      &
         gltr_control_type,                                                     &
         copy_gltr_inform_in   => copy_inform_in,                               &
@@ -51,7 +54,7 @@
         copy_gltr_control_in  => copy_control_in,                              &
         copy_gltr_control_out => copy_control_out
 
-    USE GALAHAD_PSLS_double_ciface, ONLY:                                      &
+    USE GALAHAD_PSLS_precision_ciface, ONLY:                                   &
         psls_inform_type,                                                      &
[…]
