+++ src/sbls/C/sbls_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-20 AT 16:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:35 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  S B L S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_SBLS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_SBLS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_SBLS_double, ONLY:                                             &
+    USE GALAHAD_SBLS_precision, ONLY:                                          &
         f_sbls_control_type     => SBLS_control_type,                          &
         f_sbls_time_type        => SBLS_time_type,                             &
         f_sbls_inform_type      => SBLS_inform_type,                           &
@@ -28,7 +31,7 @@
         f_sbls_information      => SBLS_information,                           &
         f_sbls_terminate        => SBLS_terminate
 
-    USE GALAHAD_SLS_double_ciface, ONLY:                                       &
+    USE GALAHAD_SLS_precision_ciface, ONLY:                                    &
         sls_inform_type,                                                       &
         sls_control_type,                                                      &
         copy_sls_inform_in   => copy_inform_in,                                &
@@ -36,7 +39,7 @@
         copy_sls_control_in  => copy_control_in,                               &
         copy_sls_control_out => copy_control_out
 
-    USE GALAHAD_ULS_double_ciface, ONLY:                                       &
+    USE GALAHAD_ULS_precision_ciface, ONLY:                                    &
         uls_inform_type,                                                       &
         uls_control_type,                                                      &
         copy_uls_inform_in   => copy_inform_in,                                &
@@ -46,44 +49,37 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: sbls_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
[…]
