+++ src/trs/C/trs_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-01-06 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 10:05 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  T R S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_TRS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_TRS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_TRS_double, ONLY:                                              &
+    USE GALAHAD_TRS_precision, ONLY:                                           &
         f_trs_control_type => TRS_control_type,                                &
         f_trs_time_type => TRS_time_type,                                      &
         f_trs_history_type => TRS_history_type,                                &
@@ -30,7 +33,7 @@
         f_trs_information => TRS_information,                                  &
         f_trs_terminate => TRS_terminate
 
-    USE GALAHAD_SLS_double_ciface, ONLY:                                       &
+    USE GALAHAD_SLS_precision_ciface, ONLY:                                    &
         sls_inform_type,                                                       &
         sls_control_type,                                                      &
         copy_sls_inform_in => copy_inform_in,                                  &
@@ -38,7 +41,7 @@
         copy_sls_control_in => copy_control_in,                                &
         copy_sls_control_out => copy_control_out
 
-    USE GALAHAD_IR_double_ciface, ONLY:                                        &
+    USE GALAHAD_IR_precision_ciface, ONLY:                                     &
         ir_inform_type,                                                        &
         ir_control_type,                                                       &
         copy_ir_inform_in => copy_inform_in,                                   &
@@ -48,38 +51,31 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: trs_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
[…]

[…]
