+++ src/lsrt/C/lsrt_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.0 - 2022-01-06 AT 09:00 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:20 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  L S R T    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_LSRT_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_LSRT_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_LSRT_double, ONLY:                                             &
+    USE GALAHAD_LSRT_precision, ONLY:                                          &
         f_lsrt_control_type => LSRT_control_type,                              &
         f_lsrt_inform_type => LSRT_inform_type,                                &
         f_lsrt_full_data_type => LSRT_full_data_type,                          &
@@ -27,55 +30,48 @@
 
     IMPLICIT NONE
 
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
-
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
 !-------------------------------------------------
 
     TYPE, BIND( C ) :: lsrt_control_type
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: error
-      INTEGER ( KIND = C_INT ) :: out
-      INTEGER ( KIND = C_INT ) :: print_level
-      INTEGER ( KIND = C_INT ) :: start_print
-      INTEGER ( KIND = C_INT ) :: stop_print
-      INTEGER ( KIND = C_INT ) :: print_gap
-      INTEGER ( KIND = C_INT ) :: itmin
-      INTEGER ( KIND = C_INT ) :: itmax
-      INTEGER ( KIND = C_INT ) :: bitmax
-      INTEGER ( KIND = C_INT ) :: extra_vectors
-      INTEGER ( KIND = C_INT ) :: stopping_rule
-      INTEGER ( KIND = C_INT ) :: freq
-      REAL ( KIND = wp ) :: stop_relative
-      REAL ( KIND = wp ) :: stop_absolute
-      REAL ( KIND = wp ) :: fraction_opt
-      REAL ( KIND = wp ) :: time_limit
+      INTEGER ( KIND = ipc_ ) :: error
+      INTEGER ( KIND = ipc_ ) :: out
+      INTEGER ( KIND = ipc_ ) :: print_level
+      INTEGER ( KIND = ipc_ ) :: start_print
+      INTEGER ( KIND = ipc_ ) :: stop_print
+      INTEGER ( KIND = ipc_ ) :: print_gap
+      INTEGER ( KIND = ipc_ ) :: itmin
+      INTEGER ( KIND = ipc_ ) :: itmax
+      INTEGER ( KIND = ipc_ ) :: bitmax
[…]
