+++ src/sls/C/sls_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-11-20 AT 16:10 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 15:20 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  S L S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,10 +14,10 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_SLS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_SLS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_SLS_double, ONLY:                                              &
+    USE GALAHAD_SLS_precision, ONLY:                                           &
         f_sls_control_type => SLS_control_type,                                &
         f_sls_time_type => SLS_time_type,                                      &
         f_sls_inform_type => SLS_inform_type,                                  &
@@ -29,7 +32,7 @@
         f_sls_information => SLS_information,                                  &
         f_sls_terminate => SLS_terminate
 
-    USE GALAHAD_SILS_double_ciface, ONLY:                                      &
+    USE GALAHAD_SILS_precision_ciface, ONLY:                                   &
         sils_control,                                                          &
         sils_ainfo,                                                            &
         sils_finfo,                                                            &
@@ -41,7 +44,7 @@
         copy_sils_finfo_out => copy_finfo_out,                                 &
         copy_sils_sinfo_out => copy_sinfo_out
 
-    USE HSL_MA57_double_ciface, ONLY:                                          &
+    USE HSL_MA57_precision_ciface, ONLY:                                       &
         ma57_control,                                                          &
         ma57_ainfo,                                                            &
         ma57_finfo,                                                            &
@@ -50,37 +53,37 @@
         copy_ma57_finfo_out => copy_finfo_out,                                 &
         copy_ma57_sinfo_out => copy_sinfo_out
 
-    USE HSL_MA77_double_ciface, ONLY:                                          &
+    USE HSL_MA77_precision_ciface, ONLY:                                       &
         ma77_info,                                                             &
         ma77_control,                                                          &
         copy_ma77_info_out => copy_info_out,                                   &
         copy_ma77_control_in => copy_control_in
 
-    USE HSL_MA86_double_ciface, ONLY:                                          &
+    USE HSL_MA86_precision_ciface, ONLY:                                       &
         ma86_info,                                                             &
[…]
