+++ src/gls/C/gls_ciface.F90
@@ -1,4 +1,7 @@
-! THIS VERSION: GALAHAD 3.3 - 30/11/2021 AT 09:45 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:15 GMT.
+
+#include "galahad_modules.h"
+#include "galahad_cfunctions.h"
 
 !-*-*-*-*-*-*-*-  G A L A H A D _  G L S    C   I N T E R F A C E  -*-*-*-*-*-
 
@@ -11,28 +14,21 @@
 !  For full documentation, see
 !   http://galahad.rl.ac.uk/galahad-www/specs.html
 
-  MODULE GALAHAD_GLS_double_ciface
-    USE iso_c_binding
+  MODULE GALAHAD_GLS_precision_ciface
+    USE GALAHAD_KINDS
     USE GALAHAD_common_ciface
-    USE GALAHAD_GLS_double, ONLY: &
-        f_gls_control => GLS_control, &
-        f_gls_ainfo => GLS_ainfo, &
-        f_gls_finfo => GLS_finfo, &
-        f_gls_sinfo => GLS_sinfo, &
-        f_gls_full_data_type => GLS_full_data_type, &
-        f_gls_initialize => GLS_initialize, &
-        f_gls_reset_control => GLS_reset_control, &
-        f_gls_information => GLS_information, &
-        f_gls_finalize => gls_finalize
+    USE GALAHAD_GLS_precision, ONLY:                                           &
+        f_gls_control => GLS_control,                                          &
+        f_gls_ainfo => GLS_ainfo,                                              &
+        f_gls_finfo => GLS_finfo,                                              &
+        f_gls_sinfo => GLS_sinfo,                                              &
+        f_gls_full_data_type => GLS_full_data_type,                            &
+        f_gls_initialize => GLS_initialize,                                    &
+        f_gls_reset_control => GLS_reset_control,                              &
+        f_gls_information => GLS_information,                                  &
+        f_gls_finalize => GLS_finalize
 
     IMPLICIT NONE
-
-!--------------------
-!   P r e c i s i o n
-!--------------------
-
-    INTEGER, PARAMETER :: wp = C_DOUBLE ! double precision
-    INTEGER, PARAMETER :: sp = C_FLOAT  ! single precision
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -40,63 +36,63 @@
 
     TYPE, BIND( C ) :: gls_control
       LOGICAL ( KIND = C_BOOL ) :: f_indexing
-      INTEGER ( KIND = C_INT ) :: lp
-      INTEGER ( KIND = C_INT ) :: wp
-      INTEGER ( KIND = C_INT ) :: mp
-      INTEGER ( KIND = C_INT ) :: ldiag
-      INTEGER ( KIND = C_INT ) :: btf
-      INTEGER ( KIND = C_INT ) :: maxit
-      INTEGER ( KIND = C_INT ) :: factor_blocking
-      INTEGER ( KIND = C_INT ) :: solve_blas
-      INTEGER ( KIND = C_INT ) :: la
[…]
