+++ src/sls/sls.F90
@@ -3893,8 +3893,13 @@
          data%spm%dof = 1
          CALL spmUpdateComputedFields( data%spm )
          CALL spmAlloc( data%spm )
+#ifdef GALAHAD_SINGLE
+         CALL spmGetArray( data%spm, colptr = data%PTR,                        &
+                           rowptr = data%ROW, svalues = data%VAL )
+#else
          CALL spmGetArray( data%spm, colptr = data%PTR,                        &
                            rowptr = data%ROW, dvalues = data%VAL )
+#endif
 
 !  set the matrix
 

[…]
