+++ src/spral/config.h
@@ -16,7 +16,12 @@
 
 /* Define if you have hwloc library */
 /* #undef HAVE_HWLOC */
+/* #define HAVE_HWLOC 1 */
+#ifdef GALAHAD_NO_HWLOC
+#undef HAVE_HWLOC
+#else
 #define HAVE_HWLOC 1
+#endif
 
 /* Define to 1 if you have the <inttypes.h> header file. */
 #define HAVE_INTTYPES_H 1
@@ -34,8 +39,13 @@
 /* #undef HAVE_NVCC */
 
 /* Define to 1 if you have sched_getcpu(). */
+/* #define HAVE_SCHED_GETCPU 1 */
+/* #undef HAVE_SCHED_GETCPU */
+#ifdef GALAHAD_NO_SCHED_GETCPU
+#undef HAVE_SCHED_GETCPU
+#else
 #define HAVE_SCHED_GETCPU 1
-/*#undef HAVE_SCHED_GETCPU*/
+#endif
 
 /* Define to 1 if you have the <stdint.h> header file. */
 #define HAVE_STDINT_H 1
