+++ src/ssids/cpu_subtree.F90
@@ -48,26 +48,26 @@ module spral_ssids_cpu_subtree_precision
      type(C_PTR) function c_create_symbolic_subtree(n, sa, en, sptr, sparent, &
           rptr, rlist, nptr, nlist, ncontrib, contrib_idx, options) &
           bind(C, name="spral_ssids_cpu_create_symbolic_subtree")
-       use, intrinsic :: iso_c_binding
+       use :: spral_kinds
        import :: cpu_factor_options
        implicit none
-       integer(C_INT), value :: n
-       integer(C_INT), value :: sa
-       integer(C_INT), value :: en
-       integer(C_INT), dimension(*), intent(in) :: sptr
-       integer(C_INT), dimension(*), intent(in) :: sparent
-       integer(C_INT64_T), dimension(*), intent(in) :: rptr
-       integer(C_INT), dimension(*), intent(in) :: rlist
-       integer(C_INT64_T), dimension(*), intent(in) :: nptr
-       integer(C_INT64_T), dimension(*), intent(in) :: nlist
-       integer(C_INT), value :: ncontrib
-       integer(C_INT), dimension(*), intent(in) :: contrib_idx
+       integer(C_IP_), value :: n
+       integer(C_IP_), value :: sa
+       integer(C_IP_), value :: en
+       integer(C_IP_), dimension(*), intent(in) :: sptr
+       integer(C_IP_), dimension(*), intent(in) :: sparent
+       integer(CLONG_), dimension(*), intent(in) :: rptr
+       integer(C_IP_), dimension(*), intent(in) :: rlist
+       integer(CLONG_), dimension(*), intent(in) :: nptr
+       integer(CLONG_), dimension(*), intent(in) :: nlist
+       integer(C_IP_), value :: ncontrib
+       integer(C_IP_), dimension(*), intent(in) :: contrib_idx
        type(cpu_factor_options), intent(in) :: options
      end function c_create_symbolic_subtree
 
      subroutine c_destroy_symbolic_subtree(subtree) &
           bind(C, name="spral_ssids_cpu_destroy_symbolic_subtree")
-       use, intrinsic :: iso_c_binding
+       use :: spral_kinds
        implicit none
        type(C_PTR), value :: subtree
      end subroutine c_destroy_symbolic_subtree
@@ -75,12 +75,12 @@ module spral_ssids_cpu_subtree_precision
      type(C_PTR) function c_create_numeric_subtree(posdef, symbolic_subtree, &
           aval, scaling, child_contrib, options, stats) &
           bind(C, name="spral_ssids_cpu_create_num_subtree_dbl")
-       use, intrinsic :: iso_c_binding
+       use :: spral_kinds
        import :: cpu_factor_options, cpu_factor_stats
        implicit none
        logical(C_BOOL), value :: posdef
        type(C_PTR), value :: symbolic_subtree
-       real(C_DOUBLE), dimension(*), intent(in) :: aval
+       real(C_RP_), dimension(*), intent(in) :: aval
        type(C_PTR), value :: scaling
        type(C_PTR), dimension(*), intent(inout) :: child_contrib
        type(cpu_factor_options), intent(in) :: options
@@ -89,63 +89,63 @@ module spral_ssids_cpu_subtree_precision
 
      subroutine c_destroy_numeric_subtree(posdef, subtree) &
           bind(C, name="spral_ssids_cpu_destroy_num_subtree_dbl")
-       use, intrinsic :: iso_c_binding
+       use :: spral_kinds
        implicit none
        logical(C_BOOL), value :: posdef
[…]
