### 2023-01-03 d9ab902
ssids enhanced to allow single-precision factorization
