### 2023-01-03 a5163e5
ssids enhanced to allow single-precision factorization
