### 2023-01-03 aa6ce41
ssids enhanced to allow single-precision factorization
