+++ src/ssids/cpu_subtree.F90
@@ -71,7 +71,128 @@ module spral_ssids_cpu_subtree_precision
        implicit none
        type(C_PTR), value :: subtree
      end subroutine c_destroy_symbolic_subtree
+  end interface
 
+
+#ifdef SPRAL_SINGLE
+  interface
+     type(C_PTR) function c_create_numeric_subtree(posdef, symbolic_subtree, &
+          aval, scaling, child_contrib, options, stats) &
+          bind(C, name="spral_ssids_cpu_create_num_subtree_sgl")
+       use :: spral_kinds
+       import :: cpu_factor_options, cpu_factor_stats
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: symbolic_subtree
+       real(C_RP_), dimension(*), intent(in) :: aval
+       type(C_PTR), value :: scaling
+       type(C_PTR), dimension(*), intent(inout) :: child_contrib
+       type(cpu_factor_options), intent(in) :: options
+       type(cpu_factor_stats), intent(out) :: stats
+     end function c_create_numeric_subtree
+
+     subroutine c_destroy_numeric_subtree(posdef, subtree) &
+          bind(C, name="spral_ssids_cpu_destroy_num_subtree_sgl")
+       use :: spral_kinds
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: subtree
+     end subroutine c_destroy_numeric_subtree
+
+     integer(C_IP_) function c_subtree_solve_fwd(posdef, subtree, nrhs, x, &
+          ldx) &
+          bind(C, name="spral_ssids_cpu_subtree_solve_fwd_sgl")
+       use :: spral_kinds
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: subtree
+       integer(C_IP_), value :: nrhs
+       real(C_RP_), dimension(*), intent(inout) :: x
+       integer(C_IP_), value :: ldx
+     end function c_subtree_solve_fwd
+
+     integer(C_IP_) function c_subtree_solve_diag(posdef, subtree, nrhs, x, &
+          ldx) &
+          bind(C, name="spral_ssids_cpu_subtree_solve_diag_sgl")
+       use :: spral_kinds
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: subtree
+       integer(C_IP_), value :: nrhs
+       real(C_RP_), dimension(*), intent(inout) :: x
+       integer(C_IP_), value :: ldx
+     end function c_subtree_solve_diag
+
+     integer(C_IP_) function c_subtree_solve_diag_bwd(posdef, subtree, nrhs, &
+          x, ldx) &
+          bind(C, name="spral_ssids_cpu_subtree_solve_diag_bwd_sgl")
+       use :: spral_kinds
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: subtree
+       integer(C_IP_), value :: nrhs
+       real(C_RP_), dimension(*), intent(inout) :: x
+       integer(C_IP_), value :: ldx
+     end function c_subtree_solve_diag_bwd
+     
+     integer(C_IP_) function c_subtree_solve_bwd(posdef, subtree, nrhs, x, &
+          ldx) &
+          bind(C, name="spral_ssids_cpu_subtree_solve_bwd_sgl")
+       use :: spral_kinds
+       implicit none
+       logical(C_BOOL), value :: posdef
+       type(C_PTR), value :: subtree
+       integer(C_IP_), value :: nrhs
+       real(C_RP_), dimension(*), intent(inout) :: x
+       integer(C_IP_), value :: ldx
[…]
