### 2023-01-03 69cfc4e
ssids enhanced to allow single-precision factorization
