+++ src/ssids/contrib.h
@@ -13,19 +13,19 @@
 extern "C" {
 #endif
 
-#ifdef GALAHAD_SINGLE
+#ifdef SPRAL_SINGLE
 void spral_ssids_contrib_get_data_single(const void *const contrib, 
       int *const n, const float* *const val, int *const ldval, 
       const int* *const rlist, int *const ndelay, const int* *const delay_perm,
       const float* *const delay_val, int *const lddelay);
+void spral_ssids_contrib_free_sgl(void *const contrib);
 #else
 void spral_ssids_contrib_get_data_double(const void *const contrib, 
       int *const n, const double* *const val, int *const ldval, 
       const int* *const rlist, int *const ndelay, const int* *const delay_perm,
       const double* *const delay_val, int *const lddelay);
-#endif
-
 void spral_ssids_contrib_free_dbl(void *const contrib);
+#endif
 
 #ifdef __cplusplus
 }
