### 2023-01-03 653919e
ssids enhanced to allow single-precision factorization
