### 2023-01-03 a9de853
spral and ssids directories simplified and unified
