+++ src/ssids/cpu/subtree.F90
@@ -1,430 +1,0 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:55 GMT.
-
-#include "galahad_modules.h"
-#include "galahad_cfunctions.h"
-
-!> \file
-!> \copyright 2016 The Science and Technology Facilities Council (STFC)
-!> \licence   BSD licence, see LICENCE file for details
-!> \author    Jonathan Hogg
-module spral_ssids_cpu_subtree
-  use, intrinsic :: GALAHAD_KINDS
-  use spral_ssids_contrib, only : contrib_type
-  use spral_ssids_cpu_iface ! fixme only
-  use spral_ssids_datatypes
-  use spral_ssids_inform, only : ssids_inform
-  use spral_ssids_subtree, only : symbolic_subtree_base, numeric_subtree_base
-  implicit none
-
-  private
-  public :: cpu_symbolic_subtree, construct_cpu_symbolic_subtree
-  public :: cpu_numeric_subtree, cpu_free_contrib
-
-  type, extends(symbolic_subtree_base) :: cpu_symbolic_subtree
-     integer ( kind = ip_ ) :: n
-     type(C_PTR) :: csubtree
-   contains
-     procedure :: factor
-     procedure :: cleanup => symbolic_cleanup
-  end type cpu_symbolic_subtree
-
-  type, extends(numeric_subtree_base) :: cpu_numeric_subtree
-     logical(C_BOOL) :: posdef
-     type(cpu_symbolic_subtree), pointer :: symbolic
-     type(C_PTR) :: csubtree
-   contains
-     procedure :: get_contrib
-     procedure :: solve_fwd
-     procedure :: solve_diag
-     procedure :: solve_diag_bwd
-     procedure :: solve_bwd
-     procedure :: enquire_posdef
-     procedure :: enquire_indef
-     procedure :: alter
-     procedure :: cleanup => numeric_cleanup
-  end type cpu_numeric_subtree
-
-  interface
-     type(C_PTR) function c_create_symbolic_subtree(n, sa, en, sptr, sparent, &
-          rptr, rlist, nptr, nlist, ncontrib, contrib_idx, options) &
-          bind(C, name="spral_ssids_cpu_create_symbolic_subtree")
-       use, intrinsic :: GALAHAD_KINDS
-       import :: cpu_factor_options
-       implicit none
-       integer(ipc_), value :: n
-       integer(ipc_), value :: sa
-       integer(ipc_), value :: en
-       integer(ipc_), dimension(*), intent(in) :: sptr
-       integer(ipc_), dimension(*), intent(in) :: sparent
-       integer(C_INT64_T), dimension(*), intent(in) :: rptr
-       integer(ipc_), dimension(*), intent(in) :: rlist
-       integer(C_INT64_T), dimension(*), intent(in) :: nptr
-       integer(C_INT64_T), dimension(*), intent(in) :: nlist
-       integer(ipc_), value :: ncontrib
-       integer(ipc_), dimension(*), intent(in) :: contrib_idx
-       type(cpu_factor_options), intent(in) :: options
-     end function c_create_symbolic_subtree
-
-     subroutine c_destroy_symbolic_subtree(subtree) &
-          bind(C, name="spral_ssids_cpu_destroy_symbolic_subtree")
-       use, intrinsic :: GALAHAD_KINDS
-       implicit none
-       type(C_PTR), value :: subtree
-     end subroutine c_destroy_symbolic_subtree
-
-     type(C_PTR) function c_create_numeric_subtree(posdef, symbolic_subtree, &
-          aval, scaling, child_contrib, options, stats) &
[…]
