### 2023-01-03 84a7a67
spral and ssids directories simplified and unified
