### 2023-01-03 bc7ce47
spral and ssids directories simplified and unified
