### 2023-01-03 ff98d33
spral and ssids directories simplified and unified
