+++ src/ssids/config.h
@@ -1,1 +1,0 @@
-../spral/config.h
