+++ src/ssids/cpu/cpu_iface.F90
@@ -1,173 +1,0 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-31 AT 09:55 GMT.
-
-#include "galahad_modules.h"
-#include "galahad_cfunctions.h"
-
-!> \file
-!> \copyright 2016 The Science and Technology Facilities Council (STFC)
-!> \licence   BSD licence, see LICENCE file for details
-!> \author    Jonathan Hogg
-module spral_ssids_cpu_iface
-   use :: GALAHAD_KINDS
-   use spral_ssids_datatypes, only : ssids_options
-   use spral_ssids_inform, only : ssids_inform
-   implicit none
-
-   private
-   public :: cpu_factor_options, cpu_factor_stats
-   public :: cpu_copy_options_in, cpu_copy_stats_out
-
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-   !> @brief Interoperable subset of ssids_options
-   !> @details Interoperates with cpu_factor_options C++ type
-   !> @sa spral_ssids_datatypes::ssids_options
-   !> @sa spral::ssids::cpu::cpu_factor_options
-   type, bind(C) :: cpu_factor_options
-      integer(ipc_) :: print_level
-      logical(C_BOOL) :: action
-      real(rpc_) :: small
-      real(rpc_) :: u
-      real(rpc_) :: multiplier
-      integer(long_) :: small_subtree_threshold
-      integer(ipc_) :: cpu_block_size
-      integer(ipc_) :: pivot_method
-      integer(ipc_) :: failed_pivot_method
-   end type cpu_factor_options
-
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-   !> @brief Interoperable subset of ssids_inform
-   !> @details Interoperates with ThreadStats C++ type
-   !> @sa spral_ssids_inform::ssids_inform
-   !> @sa spral::ssids::cpu::ThreadStats
-   type, bind(C) :: cpu_factor_stats
-      integer(ipc_) :: flag
-      integer(ipc_) :: num_delay
-      integer(ipc_) :: num_neg
-      integer(ipc_) :: num_two
-      integer(ipc_) :: num_zero
-      integer(ipc_) :: maxfront
-      integer(ipc_) :: not_first_pass
-      integer(ipc_) :: not_second_pass
-   end type cpu_factor_stats
-
-contains
-
-!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!> @brief Copy subset of ssids_options to interoperable type
-subroutine cpu_copy_options_in(foptions, coptions)
-   type(ssids_options), intent(in) :: foptions
-   type(cpu_factor_options), intent(out) :: coptions
-
-   coptions%print_level    = foptions%print_level
-   coptions%action         = foptions%action
-   coptions%small          = foptions%small
-   coptions%u              = foptions%u
-   coptions%multiplier     = foptions%multiplier
-   coptions%small_subtree_threshold = foptions%small_subtree_threshold
-   coptions%cpu_block_size = foptions%cpu_block_size
-   coptions%pivot_method   = min(3, max(1, foptions%pivot_method))
-   coptions%failed_pivot_method = min(2, max(1, foptions%failed_pivot_method))
[…]
