### 2023-01-03 ac3aff2
spral and ssids directories simplified and unified
