### 2023-01-04 59f32d6
dummy interface to C ssids corrected
