### 2023-01-04 6623af2
3rd time lucky to fix that pesky dummy c interface to ssids
