### 2023-01-05 3621277
Fix some errors found with Meson
