### 2023-01-05 fcf5f2b
Fix some errors found with Meson
