### 2023-01-05 f267994
Fix some errors found with Meson
