### 2023-01-05 d05412d
ongoing work on sha codes
