+++ src/oblivion/trtn/usetrtn.F90
@@ -34,9 +34,6 @@
 
      INTEGER ( KIND = ip_ ), INTENT( IN ) :: input
 
-!  Set precision
-
-
 !-------------------------------
 !   D e r i v e d   T y p e s
 !-------------------------------
