+++ src/trb/usetrb.F90
@@ -20,6 +20,8 @@
      USE GALAHAD_TRB_precision
      USE GALAHAD_SPECFILE_precision
      USE GALAHAD_SPACE_precision
+     USE GALAHAD_USERDATA_precision
+     USE GALAHAD_NLPT_precision, ONLY: NLPT_problem_type
      USE GALAHAD_CUTEST_FUNCTIONS_precision
      IMPLICIT NONE
 
@@ -44,7 +46,7 @@
      TYPE ( TRB_inform_type ) :: inform
      TYPE ( TRB_data_type ) :: data
      TYPE ( NLPT_problem_type ) :: nlp
-     TYPE ( NLPT_userdata_type ) :: userdata
+     TYPE ( GALAHAD_userdata_type ) :: userdata
      TYPE ( CUTEST_FUNCTIONS_control_type ) :: cutest_control
      TYPE ( CUTEST_FUNCTIONS_inform_type ) :: cutest_inform
 
