+++ src/lancelot/bndsl.F90
@@ -18,9 +18,6 @@
      PRIVATE
      PUBLIC :: BAND_factor, BAND_solve
 
-!!  Set precision
-
-
 !  Set other parameters
 
      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
