+++ src/lancelot/lancelot.pointers.F90
@@ -57,9 +57,6 @@
                LANCELOT_control_type, LANCELOT_inform_type, LANCELOT_data_type,&
                LANCELOT_problem_pointer_type
 
-!  Set precision
-
-
 !  Set other parameters
 
      REAL ( KIND = rp_ ), PARAMETER :: zero = 0.0_rp_
