+++ src/external/mumps/mumpst.F90
@@ -1,4 +1,4 @@
-! THIS VERSION: GALAHAD 4.1 - 2022-12-30 AT 09:40 GMT.
+! THIS VERSION: GALAHAD 4.1 - 2023-01-06 AT 15:30 GMT.
 ! This file is a modified version of dsimpletest.F from MUMPS 5.5.1, 
 ! originally released on Tue Jul 12 13:17:24 UTC 2022
 
@@ -8,7 +8,6 @@
    USE GALAHAD_KINDS
    USE GALAHAD_MUMPS_TYPES_precision
    IMPLICIT NONE
-   INTEGER ( KIND = ip_ ), PARAMETER :: MPI_COMM_WORLD = 0
    INTEGER ( KIND = ip_ ), PARAMETER :: n = 5
    INTEGER ( KIND = ip_ ), PARAMETER :: ne = 7
    INTEGER ( KIND = ip_ ), PARAMETER :: neu = 11
