+++ src/lancelot/strutr.F90
@@ -102,9 +102,9 @@
        INTEGER ( KIND = ip_ ), INTENT( IN ) :: ielemn, nelvar, ninvar, ieltyp
        INTEGER ( KIND = ip_ ), INTENT( IN ) :: lw1, lw2
        LOGICAL, INTENT( IN ) :: transp
-       REAL ( KIND = KIND( 1.0D+0 ) ), INTENT( IN ), DIMENSION ( lw1 ) :: W1
+       REAL ( KIND = rp_ ), INTENT( IN ), DIMENSION ( lw1 ) :: W1
 !      REAL ( KIND = KIND( 1.0D+0 ) ), INTENT( OUT ), DIMENSION ( lw2 ) :: W2
-       REAL ( KIND = KIND( 1.0D+0 ) ), DIMENSION ( lw2 ) :: W2
+       REAL ( KIND = rp_ ), DIMENSION ( lw2 ) :: W2
        END SUBROUTINE RANGE
      END INTERFACE
 
