### 2023-01-09 ea40aff
correction to sls to allow proper single precision mumps
