### 2023-01-09 20739cb
correction to sls to allow proper single precision mumps
