### 2023-01-09 9aa9cfd
correction to sls to allow proper single precision mumps
