+++ src/sha/sha.F90
@@ -51,6 +51,7 @@
 !----------------------
 
      REAL ( KIND = rp_ ), PARAMETER :: one = 1.0_rp_
+     REAL ( KIND = rp_ ), PARAMETER :: eps_singular = EPSILON( one )
 
 !-------------------------------------------------
 !  D e r i v e d   t y p e   d e f i n i t i o n s
@@ -90,7 +91,7 @@
 !    3 : singular-value decomposition
 !    4 : singular-value decomposition with divide-and-conquer
 
-       INTEGER ( KIND = ip_ ) :: dense_linear_solver = 1
+       INTEGER ( KIND = ip_ ) :: dense_linear_solver = 3
 
 !  the maximum sparse degree if the combined version is used
 
@@ -151,16 +152,11 @@
      TYPE, PUBLIC :: SHA_solve_system_data_type
 
        INTEGER ( KIND = ip_ ) :: out = 0
-       INTEGER ( KIND = ip_ ) :: la_save1 = - 1
-       INTEGER ( KIND = ip_ ) :: la_save2 = - 1
-       INTEGER ( KIND = ip_ ) :: lb_save = - 1
        INTEGER ( KIND = ip_ ) :: lwork = - 1
        INTEGER ( KIND = ip_ ) :: liwork = - 1
        INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: IWORK
        REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: S, WORK
-       REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : ) :: B_save
        REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: U, VT
-       REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: A_save
      END TYPE SHA_solve_system_data_type
 
 !  - - - - - - - - - -
@@ -180,6 +176,9 @@
        INTEGER ( KIND = ip_ ) :: la1 = - 1
        INTEGER ( KIND = ip_ ) :: la2 = - 1
        INTEGER ( KIND = ip_ ) :: lb1 = - 1
+       INTEGER ( KIND = ip_ ) :: la_save1 = - 1
+       INTEGER ( KIND = ip_ ) :: la_save2 = - 1
+       INTEGER ( KIND = ip_ ) :: lb_save = - 1
        INTEGER ( KIND = ip_ ) :: ls = - 1
 
 !  SHA_analyse_called is true once SHA_analyse has been called
@@ -200,7 +199,9 @@
        INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: PU
        INTEGER ( KIND = ip_ ), ALLOCATABLE, DIMENSION( : ) :: PK
        REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: A
+       REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: A_save
        REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: B
+       REAL ( KIND = rp_ ), ALLOCATABLE, DIMENSION( : , : ) :: B_save
        TYPE ( SHA_solve_system_data_type ) :: solve_system_data
      END TYPE SHA_data_type
 
@@ -416,12 +417,12 @@
 
       SUBROUTINE SHA_analyse( n, nz, ROW, COL, data, control, inform )
 !
-!  ***************************************************************
-!  *                                                             *
-!  *   Compute a permutation of a symmetric matrix to try to     *
-!  *   minimize the number of super-diagonal entries in each row *
-!  *                                                             *
-!  ***************************************************************
[…]
