### 2023-01-10 0fdeae5
fixed small errors in sls, cro, slls and their tests
