### 2023-01-10 6e05b23
fixed small errors in sls, cro, slls and their tests
