### 2023-01-10 6366e9f
fixed small errors in sls, cro, slls and their tests
