+++ src/slls/sllss4.f90
@@ -12,7 +12,6 @@
    USE GALAHAD_RAND_double
    IMPLICIT NONE
    INTEGER, PARAMETER :: wp = KIND( 1.0D+0 ) ! set precision
-   REAL ( KIND = wp ), PARAMETER :: infinity = 10.0_wp ** 20
    TYPE ( QPT_problem_type ) :: p_x, p_y
    TYPE ( SLLS_data_type ) :: data_x, data_y
    TYPE ( SLLS_control_type ) :: control_x, control_y
@@ -95,13 +94,11 @@ write(6,"( ' MU', /, ( 5ES12.4 ) )" ) MU
 ! problem data complete. Initialze data and control parameters
 
    CALL SLLS_initialize( data_x, control_x, inform_x )
-   control_x%infinity = infinity
    control_x%print_level = 1
    control_x%exact_arc_search = .FALSE.
 !  control_x%convert_control%print_level = 3
 
    CALL SLLS_initialize( data_y, control_y, inform_y )
-   control_y%infinity = infinity
    control_y%print_level = 1
    control_y%exact_arc_search = .FALSE.
 !  control_y%convert_control%print_level = 3
