+++ src/sls/sls.F90
@@ -888,6 +888,70 @@
        END SUBROUTINE MA27ID
      END INTERFACE MA27I
 
+     INTERFACE MC61A
+       SUBROUTINE MC61A( job, n, lirn, IRN, ICPTR, PERM, liw, IW, W,           &
+                         ICNTL, CNTL, INFO, RINFO )
+       USE GALAHAD_KINDS
+       INTEGER ( KIND = ip_ ), INTENT( IN ) :: job, n, liw, lirn
+       INTEGER ( KIND = ip_ ), DIMENSION( lirn ), INTENT( INOUT ) :: IRN
+       INTEGER ( KIND = ip_ ), DIMENSION( n + 1 ), INTENT( INOUT ) :: ICPTR
+       INTEGER ( KIND = ip_ ), DIMENSION( n ), INTENT( INOUT ) :: PERM
+       INTEGER ( KIND = ip_ ), DIMENSION( liw ), INTENT( OUT ) :: IW
+       REAL ( KIND = sp_ ), DIMENSION( n ), INTENT( OUT ) :: W
+       INTEGER ( KIND = ip_ ), DIMENSION( 10 ), INTENT( IN ) :: ICNTL
+       REAL ( KIND = sp_ ), DIMENSION( 5 ), INTENT( IN ) :: CNTL
+       INTEGER ( KIND = ip_ ), DIMENSION( 10 ), INTENT( OUT ) :: INFO
+       REAL ( KIND = sp_ ), DIMENSION( 15 ), INTENT( OUT ) :: RINFO
+       END SUBROUTINE MC61A
+
+       SUBROUTINE MC61AD( job, n, lirn, IRN, ICPTR, PERM, liw, IW, W,          &
+                          ICNTL, CNTL, INFO, RINFO )
+       USE GALAHAD_KINDS
+       INTEGER ( KIND = ip_ ), INTENT( IN ) :: job, n, liw, lirn
+       INTEGER ( KIND = ip_ ), DIMENSION( lirn ), INTENT( INOUT ) :: IRN
+       INTEGER ( KIND = ip_ ), DIMENSION( n + 1 ), INTENT( INOUT ) :: ICPTR
+       INTEGER ( KIND = ip_ ), DIMENSION( n ), INTENT( INOUT ) :: PERM
+       INTEGER ( KIND = ip_ ), DIMENSION( liw ), INTENT( OUT ) :: IW
+       REAL ( KIND = dp_ ), DIMENSION( n ), INTENT( OUT ) :: W
+       INTEGER ( KIND = ip_ ), DIMENSION( 10 ), INTENT( IN ) :: ICNTL
+       REAL ( KIND = dp_ ), DIMENSION( 5 ), INTENT( IN ) :: CNTL
+       INTEGER ( KIND = ip_ ), DIMENSION( 10 ), INTENT( OUT ) :: INFO
+       REAL ( KIND = dp_ ), DIMENSION( 15 ), INTENT( OUT ) :: RINFO
+       END SUBROUTINE MC61AD
+     END INTERFACE MC61A
+
+     INTERFACE MC77A
+       SUBROUTINE MC77A( job, m, n, nnz, JCST, IRN, A, IW, liw, DW, ldw,       &
+                         ICNTL, CNTL, INFO, RINFO )
+       USE GALAHAD_KINDS
+       INTEGER ( KIND = ip_ ), INTENT( IN ) :: job, m, n, nnz, liw, ldw
+       INTEGER ( KIND = ip_ ), DIMENSION( n + 1 ), INTENT( IN ) :: JCST
[…]
