### 2023-01-10 ae07151
fixed small errors in sls, cro, slls and their tests
