### 2023-01-10 8da97ff
Remove infinity from slls tests
