### 2023-01-10 4c9ff96
Remove infinity from slls tests
