### 2023-01-10 7b8b2da
abort fdc gracefully if analysis fails
