### 2023-01-10 917f6c7
generally enable single precision c interfaces
