+++ include/galahad_dps.h
@@ -216,6 +216,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sls.h"
