+++ src/cqp/C/cqpt.c
@@ -3,6 +3,8 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_cqp.h"
 
 int main(void) {
@@ -19,21 +21,21 @@ int main(void) {
     int H_row[] = {0, 1, 2 };   // row indices, NB lower triangle
     int H_col[] = {0, 1, 2};    // column indices, NB lower triangle
     int H_ptr[] = {0, 1, 2, 3}; // row pointers
-    double H_val[] = {1.0, 1.0, 1.0 };   // values
-    double g[] = {0.0, 2.0, 0.0};   // linear term in the objective
-    double f = 1.0;  // constant term in the objective
+    real_wp_ H_val[] = {1.0, 1.0, 1.0 };   // values
+    real_wp_ g[] = {0.0, 2.0, 0.0};   // linear term in the objective
+    real_wp_ f = 1.0;  // constant term in the objective
     int A_ne = 4; // Jacobian elements
     int A_row[] = {0, 0, 1, 1}; // row indices
     int A_col[] = {0, 1, 1, 2}; // column indices
     int A_ptr[] = {0, 2, 4}; // row pointers
-    double A_val[] = {2.0, 1.0, 1.0, 1.0 }; // values
-    double c_l[] = {1.0, 2.0};   // constraint lower bound
-    double c_u[] = {2.0, 2.0};   // constraint upper bound
-    double x_l[] = {-1.0, - INFINITY, - INFINITY}; // variable lower bound
-    double x_u[] = {1.0, INFINITY, 2.0}; // variable upper bound
+    real_wp_ A_val[] = {2.0, 1.0, 1.0, 1.0 }; // values
+    real_wp_ c_l[] = {1.0, 2.0};   // constraint lower bound
+    real_wp_ c_u[] = {2.0, 2.0};   // constraint upper bound
+    real_wp_ x_l[] = {-1.0, - INFINITY, - INFINITY}; // variable lower bound
+    real_wp_ x_u[] = {1.0, INFINITY, 2.0}; // variable upper bound
 
     // Set output storage
-    double c[m]; // constraint values
+    real_wp_ c[m]; // constraint values
     int x_stat[n]; // variable status
     int c_stat[m]; // constraint status
     char st;
@@ -52,9 +54,9 @@ int main(void) {
         control.f_indexing = false; // C sparse matrix indexing
 
         // Start from 0
-        double x[] = {0.0,0.0,0.0};
-        double y[] = {0.0,0.0};
-        double z[] = {0.0,0.0,0.0};
+        real_wp_ x[] = {0.0,0.0,0.0};
+        real_wp_ y[] = {0.0,0.0};
+        real_wp_ z[] = {0.0,0.0,0.0};
 
         switch(d){
             case 1: // sparse co-ordinate storage
@@ -80,8 +82,8 @@ int main(void) {
                 st = 'D';
                 int H_dense_ne = 6; // number of elements of H
                 int A_dense_ne = 6; // number of elements of A
-                double H_dense[] = {1.0, 0.0, 1.0, 0.0, 0.0, 1.0};
-                double A_dense[] = {2.0, 1.0, 0.0, 0.0, 1.0, 1.0};
+                real_wp_ H_dense[] = {1.0, 0.0, 1.0, 0.0, 0.0, 1.0};
+                real_wp_ A_dense[] = {2.0, 1.0, 0.0, 0.0, 1.0, 1.0};
                 cqp_import( &control, &data, &status, n, m,
                             "dense", H_ne, NULL, NULL, NULL,
[…]
