+++ include/galahad_ccqp.h
@@ -356,6 +356,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_fdc.h"
