+++ include/galahad_trs.h
@@ -261,6 +261,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sls.h"
