+++ include/galahad_lpb.h
@@ -272,6 +272,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_fdc.h"
