### 2023-01-10 15a29ff
generally enable single precision c interfaces
