+++ include/galahad_qpb.h
@@ -388,6 +388,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_lsqp.h"
