+++ include/galahad_clls.h
@@ -345,6 +345,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_fdc.h"
