+++ src/sls/C/slst.c
@@ -5,9 +5,11 @@
 #include <math.h>
 #include <stdlib.h>
 #include <float.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_sls.h"
 
-int maxabsarray(double a[],int n, double *maxabs);
+int maxabsarray(double a[],int n, real_wp_ *maxabs);
 
 int main(void) {
 
@@ -23,17 +25,17 @@ int main(void) {
     int row[] = {0, 1, 1, 2, 2, 3, 4}; // row indices, NB lower triangle
     int col[] = {0, 0, 4, 1, 2, 2, 4}; // column indices
     int ptr[] = {0, 1, 3, 5, 6, 7}; // pointers to indices
-    double val[] = {2.0, 3.0, 6.0, 4.0,  1.0, 5.0, 1.0}; // values
-    double dense[] = {2.0, 3.0, 0.0, 0.0, 4.0, 1.0, 0.0, 
+    real_wp_ val[] = {2.0, 3.0, 6.0, 4.0,  1.0, 5.0, 1.0}; // values
+    real_wp_ dense[] = {2.0, 3.0, 0.0, 0.0, 4.0, 1.0, 0.0, 
                       0.0, 5.0, 0.0, 0.0, 6.0, 0.0, 0.0, 1.0};
-    double rhs[] = {8.0, 45.0, 31.0, 15.0, 17.0};
-    double sol[] = {1.0, 2.0, 3.0, 4.0, 5.0};
+    real_wp_ rhs[] = {8.0, 45.0, 31.0, 15.0, 17.0};
+    real_wp_ sol[] = {1.0, 2.0, 3.0, 4.0, 5.0};
     int i, status;
-    double x[n];
-    double error[n];
+    real_wp_ x[n];
+    real_wp_ error[n];
 
-    double norm_residual;
-    double good_x = pow( DBL_EPSILON, 0.3333 );
+    real_wp_ norm_residual;
+    real_wp_ good_x = pow( DBL_EPSILON, 0.3333 );
 
     printf(" C sparse matrix indexing\n\n");
 
@@ -132,10 +134,10 @@ int main(void) {
     }
 }
 
-int maxabsarray(double a[],int n, double *maxabs)
+int maxabsarray(double a[],int n, real_wp_ *maxabs)
  {
     int i;
-    double b,max;
+    real_wp_ b,max;
     max=abs(a[0]);
     for(i=1; i<n; i++)
     {
