+++ include/galahad_nls.h
@@ -406,6 +406,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_rqs.h"
