+++ include/galahad_slls.h
@@ -242,6 +242,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sbls.h"
