+++ include/galahad_sbls.h
@@ -324,6 +324,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sls.h"
