### 2023-01-10 fb74db9
generally enable single precision c interfaces
