+++ include/galahad_sha.h
@@ -54,6 +54,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 /**
  * control derived type as a C struct
