+++ src/lhs/C/lhst.c
@@ -3,6 +3,8 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_lhs.h"
 
 int main(void) {
