+++ src/ugo/C/ugos.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <math.h>
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_ugo.h"
 
 struct userdata_type {
