+++ src/lsrt/C/lsrtt.c
@@ -3,6 +3,8 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_lsrt.h"
 
 int main(void) {
@@ -17,11 +19,11 @@ int main(void) {
     int m = 2 * n;
 
     int status;
-    double power = 3.0;
-    double weight = 1.0;
-    double x[n];
-    double u[m];
-    double v[n];
+    real_wp_ power = 3.0;
+    real_wp_ weight = 1.0;
+    real_wp_ x[n];
+    real_wp_ u[m];
+    real_wp_ v[n];
 
     // Initialize lsrt
     lsrt_initialize( &data, &control, &status );
