+++ include/galahad_tru.h
@@ -215,6 +215,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_trs.h"
