+++ include/galahad_lpa.h
@@ -306,6 +306,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_rpd.h"
