+++ src/trb/C/trbt.c
@@ -3,31 +3,33 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_trb.h"
 
 // Custom userdata struct
 struct userdata_type {
-   double p;
+   real_wp_ p;
 };
 
 // Function prototypes
-int fun( int n, const double x[], double *f, const void * );
-int grad( int n, const double x[], double g[], const void * );
-int hess( int n, int ne, const double x[], double hval[], const void * );
-int hess_dense( int n, int ne, const double x[], double hval[], const void * );
-int hessprod( int n, const double x[], double u[], const double v[], 
+int fun( int n, const real_wp_ x[], real_wp_ *f, const void * );
+int grad( int n, const real_wp_ x[], real_wp_ g[], const void * );
+int hess( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hess_dense( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hessprod( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], 
               bool got_h, const void * );
-int shessprod( int n, const double x[], int nnz_v, const int index_nz_v[], 
-               const double v[], int *nnz_u, int index_nz_u[], double u[], 
+int shessprod( int n, const real_wp_ x[], int nnz_v, const int index_nz_v[], 
+               const real_wp_ v[], int *nnz_u, int index_nz_u[], real_wp_ u[], 
               bool got_h, const void * );
-int prec( int n, const double x[], double u[], const double v[], const void * );
-int fun_diag( int n, const double x[], double *f, const void * );
-int grad_diag( int n, const double x[], double g[], const void * );
-int hess_diag( int n, int ne, const double x[], double hval[], const void * );
-int hessprod_diag( int n, const double x[], double u[], const double v[], 
+int prec( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], const void * );
+int fun_diag( int n, const real_wp_ x[], real_wp_ *f, const void * );
+int grad_diag( int n, const real_wp_ x[], real_wp_ g[], const void * );
+int hess_diag( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hessprod_diag( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], 
                    bool got_h, const void * );
-int shessprod_diag( int n, const double x[], int nnz_v, const int index_nz_v[],
-                    const double v[], int *nnz_u, int index_nz_u[], double u[],
+int shessprod_diag( int n, const real_wp_ x[], int nnz_v, const int index_nz_v[],
+                    const real_wp_ v[], int *nnz_u, int index_nz_u[], real_wp_ u[],
                     bool got_h, const void * );
 
 int main(void) {
@@ -44,14 +46,14 @@ int main(void) {
     // Set problem data
     int n = 3; // dimension
     int ne = 5; // Hesssian elements
-    double x_l[] = {-10,-10,-10}; 
-    double x_u[] = {0.5,0.5,0.5};
+    real_wp_ x_l[] = {-10,-10,-10}; 
+    real_wp_ x_u[] = {0.5,0.5,0.5};
     int H_row[] = {0, 1, 2, 2, 2}; // Hessian H
     int H_col[] = {0, 1, 0, 1, 2}; // NB lower triangle
     int H_ptr[] = {0, 1, 2, 5};    // row pointers
 
[…]
