+++ include/galahad_ugo.h
@@ -111,6 +111,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 /*
  * control derived type as a C struct
