### 2023-01-10 6e5942e
generally enable single precision c interfaces
