+++ include/galahad_uls.h
@@ -174,6 +174,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_gls.h"
