+++ include/galahad_fdc.h
@@ -109,6 +109,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sls.h"

[…]
