### 2023-01-10 5e1738c
generally enable single precision c interfaces
