+++ src/arc/C/arct.c
@@ -3,25 +3,27 @@
 
 #include <stdio.h>
 #include <math.h>
+#include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 #include "galahad_arc.h"
 
 // Custom userdata struct
 struct userdata_type {
-   double p;
+   real_wp_ p;
 };
 
 // Function prototypes
-int fun( int n, const double x[], double *f, const void * );
-int grad( int n, const double x[], double g[], const void * );
-int hess( int n, int ne, const double x[], double hval[], const void * );
-int hess_dense( int n, int ne, const double x[], double hval[], const void * );
-int hessprod( int n, const double x[], double u[], const double v[], bool got_h,
+int fun( int n, const real_wp_ x[], real_wp_ *f, const void * );
+int grad( int n, const real_wp_ x[], real_wp_ g[], const void * );
+int hess( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hess_dense( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hessprod( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], bool got_h,
               const void * );
-int prec( int n, const double x[], double u[], const double v[], const void * );
-int fun_diag( int n, const double x[], double *f, const void * );
-int grad_diag( int n, const double x[], double g[], const void * );
-int hess_diag( int n, int ne, const double x[], double hval[], const void * );
-int hessprod_diag( int n, const double x[], double u[], const double v[], 
+int prec( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], const void * );
+int fun_diag( int n, const real_wp_ x[], real_wp_ *f, const void * );
+int grad_diag( int n, const real_wp_ x[], real_wp_ g[], const void * );
+int hess_diag( int n, int ne, const real_wp_ x[], real_wp_ hval[], const void * );
+int hessprod_diag( int n, const real_wp_ x[], real_wp_ u[], const real_wp_ v[], 
                    bool got_h, const void * );
 
 int main(void) {
@@ -43,7 +45,7 @@ int main(void) {
     int H_ptr[] = {0, 1, 2, 5};    // row pointers
 
     // Set storage
-    double g[n]; // gradient
+    real_wp_ g[n]; // gradient
     char st;
     int status;
 
@@ -61,7 +63,7 @@ int main(void) {
         //control.print_level = 1;
 
         // Start from 1.5
-        double x[] = {1.5,1.5,1.5}; 
+        real_wp_ x[] = {1.5,1.5,1.5}; 
 
         switch(d){
             case 1: // sparse co-ordinate storage
@@ -123,10 +125,10 @@ int main(void) {
 
     // reverse-communication input/output
     int eval_status;
-    double f = 0.0;
-    double u[n], v[n];
+    real_wp_ f = 0.0;
+    real_wp_ u[n], v[n];
     int index_nz_u[n], index_nz_v[n];
-    double H_val[ne], H_dense[n*(n+1)/2], H_diag[n];
+    real_wp_ H_val[ne], H_dense[n*(n+1)/2], H_diag[n];
  
     for( int d=1; d <= 5; d++){
 
@@ -138,7 +140,7 @@ int main(void) {
         //control.print_level = 1;
 
         // Start from 1.5
-        double x[] = {1.5,1.5,1.5}; 
+        real_wp_ x[] = {1.5,1.5,1.5}; 
 
         switch(d){
             case 1: // sparse co-ordinate storage
@@ -295,18 +297,18 @@ int main(void) {
 }
 
[…]
