+++ include/galahad_bgo.h
@@ -250,6 +250,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_trb.h"
