+++ include/galahad_bqp.h
@@ -213,6 +213,7 @@ extern "C" {
 
 // precision
 #include "galahad_precision.h"
+#include "galahad_cfunctions.h"
 
 // required packages
 #include "galahad_sbls.h"

[…]
