### 2023-03-08 #4882 @jajhall
I'm working through the checklist, and have now installed Galahad. Am I right in thinking that I should produce a report expanding on my views/experience of the checklist items?
