### 2023-03-08 #4882 @danielskatz
Hi @jajhall - just to step in quickly (and @melissawm should also comment if she wants), if something on the checklist is satisfied, you don't really need to say anything about it, though if you want to provide non-binding suggestions for improvements, that's fine.  If something is not satisfied, however, then yes, you do need to explain what you see the problem as, and perhaps how it might be solved so that you can check the item.  Best practice is typically to open an issue in the repo of the code being reviewed, and to mention `https://github.com/openjournals/joss-reviews/issues/4882` in it, so that a link appears here that shows the status of the issue. In some cases, you could also create a PR in the repo (and again mention this review issue) if it's easier to show the potential solution/fix that way.
