### 2023-04-06 #4882 @danielskatz
👋 @jajhall - Sorry, I don't want to rush you, but I would like to see what progress has been made in the month since the last comments above, and when you think you might be done with your review. JOSS normally tries to get reviews done in 2-6 weeks, and this one is taking quite a bit longer, though some of the reason is quite understandable.
