### 2023-04-17 ralna/GALAHAD#113 @jajhall
Failure? on ALLINITC.SIF and when using MPS file
When testing in relation to https://github.com/openjournals/joss-reviews/issues/4882, I tried

`sdgal binarybuilder.bb.gfo qpa ALLINITC.SIF`

and got the opaque error `ERROR return from PRESOLVE (exitc =-21)`. What does this mean?

I also tried to solve a simple LP using the command

`sdgal binarybuilder.bb.gfo lpa ml.SIF`

where `ml.SIF` is an MPS file [ml.SIF.txt](https://github.com/ralna/GALAHAD/files/11248862/ml.SIF.txt) with renamed extension. I got the message

> File successfully decoded
 Problem: AVGAS SIZE
 problem dimensions:  n = 8, m = 10, a_ne = 30
 Stopping with inform%status = -29
 Total CPU, clock times =    0.000,    0.000
 Problem: AVGAS SIZE
                     objective          < ------ time ----- >
 Method  iterations    value     status setup   solve   total
> ------  ----------   -------    ------ -----    ----   -----
> LPA       1       -1.0000E+00    -29    0.00    0.00    0.00

Although the file seems to have been read correctly, the optimal objective value is incorrect. `Status=-29` is opaque. I am unable to find documentation to explain what this status represents. What does this mean? Can LPA solve LPs defined by MPS files?
