### 2023-04-17 #4882 @jajhall
Installation is OK, but complex and opaque, even on a standard Linux machine. I had to abandon my first attempt (I think due to an error in an environment variable) and reinstall from scratch. I did get the software to solve some problems, although there were some failures that yielded opaque status codes. https://github.com/ralna/GALAHAD/issues/113

In the Documentation there is no "statement of need". This only appears in the paper, where it states that "we have now provided interfaces from Fortran to C for a significant subset of the Fortran packages." Surely the interfaces are from C to Fortran.
