### 2023-04-17 ralna/GALAHAD#113 @jajhall
Ah, I now see a raft of READMEs and pdf in https://github.com/ralna/GALAHAD/tree/v4.0.0/doc. I did wonder whether I was trying to apply  a QP solver to a nonlinear problem.

As for AVGAS - you get the right objective value, but I just get a `status = -29` error. I guess that I didn't install `lpb`
