### 2023-04-17 #4882 @jajhall
Installation is OK, but complex and opaque, even on a standard Linux machine. Thus I dispute that "GALAHAD is easy to install using its own make-based system". I had to abandon my first attempt (I think due to an error in an environment variable) and reinstall from scratch.

I did get the software to solve some problems, although there were some failures that yielded status codes that were opaque until I was referred to the relevant pdf file in https://github.com/ralna/GALAHAD/tree/v4.0.0/doc

### Documentation

In the Documentation there is no direct "statement of need". This only appears in the paper, which is available from within the documentation as https://github.com/ralna/GALAHAD/blob/v4.0.0/doc/paper_v4/paper.md.

I see no `Community guidelines:`. Even if there is no expectation of community contribution, this should be stated - with explanation. Given the use of GitHub, there is no need to document how to `Report issues or problems with the software` or `Seek support`

### Software paper

In the `Summary`, the "description of the high-level functionality and purpose of the software for a diverse, non-specialist audience" is very terse indeed, but understandable given the scope of the software.

In its `Statement of need`, the paper states that "we have now provided interfaces from Fortran to C for a significant subset of the Fortran packages." Surely the interfaces are from C to Fortran.

In the context of `State of the field`, another "Rival open-source solver" is mentioned, and one commercial solver, but there is no "comparison"

I'm surprised to see no (link) to performance comparisons such as http://plato.asu.edu/ftp/qpbench.html
