### 2023-04-17 #4882 @nimgould
Thank you. In response

1. I had intended that the README provides many of the details of what is required from the user to install things successfully, and to use it afterwards. Once the pieces (optional software, etc) are in place, the install script aims to guide the user to multi-compiler (and possibly multi-platform) installs from a single point. Complications are generally in choosing which subsets of the library to install, with which operating system(s), which compilers and with which other language interfaces.

2. Of course, we have no expectation that the software is bug free, or even that it will succeed in all cases (that is the nature of nonlinear problems). We do try to respond to any issues, and there is an active issues page on github page.  There is ongoing community collaboration, particularly  on moving beyond C to other supported languages such as Python and Julia that is as a result of the Python interfaces.

3. I had assumed that the "need" was obvious, and justified ever since the original GALAHAD paper in TOMS. I had intended that the v4 paper referred to would be the modern equivalent. But perhaps I am wrong. Certainly, the suggestion from JOSS was that this should not be as detailed as might be expected in a full journal paper.

4. The "interface to" seems to be used in both directions on the web. I will change this to "between" to keep both sides happy.

5. You are right, there is no global comparison with other software. Individual packages have been compared (at the time) in the articles that gave rise to them, but no attempt has been made to keep this up to date, the team is simply too small.

6. Specifically on http://plato.asu.edu/ftp/qpbench.html, there used to be (a decade ago) some comparisons, but the site maintainer stopped supporting fortran. However, I just looked again., and comparison with GALAHAD's CQP has reappeared; the code seems to be generally reliable, and reasonably efficient in comparison with the competition. Those examples for which CQP fails are well understood, and related to the default sparse linear-equation solver used.
Should we mention this,  it is just one of roughly 100 GALAHAD packages?

Thank you very much for the review and comments, they have been very useful, and some changes have been made to the planned version 5 as a consequence.
