### 2023-06-22 #4882 @melissawm
@nimgould could you fix the invalid DOI pointed above (for KNITRO)? A small note (not blocking, but might be worth checking) is that the first column on the table on page 2 could have a lower width, making the text fit better on the page. But it is legible as it is 🙂

Once you fix that and you are happy with the result, can you make a tagged release for GALAHAD and archive it on zenodo, reporting the version number and archive DOI as a comment here? Thanks!
