### 2023-06-23 #4882 @nimgould
@melissawm I have uploaded the most recent version (4.1 that includes bug fixes and enhancements) of the software to zenodo, and it appears that it has a DOI 10.5281/zenodo.8075232

I couldn't get the table columns to shrink, this seems to be a defect for md files and much commented on in google. Sorry!
