### 2023-07-25 #4882 @melissawm
@nimgould It looks like this is good to go, sorry for the delay.

I am particularly grateful for your work on this as a former academic and user of GALAHAD. Cheers!
