### 2023-07-25 #4882 @danielskatz
Congratulations to @nimgould (Nicholas Gould) and your co-author on your publication!!

JOSS wouldn't be successful without your voluntary efforts
